% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adae}
\alias{adae}
\title{Adverse Event Analysis Dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1191 rows and 101 columns.}
\source{
Derived from the \link{adsl} and \link{ae} datasets using \code{{admiral}} (\url{https://github.com/Roche-GSK/admiral/blob/master/inst/templates/ad_adae.R})
}
\usage{
adae
}
\description{
An example adverse event analysis dataset
}
\keyword{datasets}
