% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADaM-package.R
\docType{package}
\name{ADaM-package}
\alias{ADaM}
\alias{ADaM-package}
\title{ADaM: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
