% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adcm}
\alias{adcm}
\title{Concomitant Medication Analysis Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7510 rows and 90 columns.
}
\source{
Derived from the \link{adsl} and \link{cm} datasets using \code{{admiral}} (\url{https://github.com/Roche-GSK/admiral/blob/master/inst/templates/ad_adcm.R})
}
\usage{
adcm
}
\description{
An example concomitant medication analysis dataset
}
\keyword{datasets}
