% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_dtm.R
\name{+,iso_dtm,Period-method}
\alias{+,iso_dtm,Period-method}
\alias{+,Period,iso_dtm-method}
\alias{+,iso_dtm,Duration-method}
\alias{+,Duration,iso_dtm-method}
\title{Add Methods}
\usage{
\S4method{+}{iso_dtm,Period}(e1, e2)

\S4method{+}{Period,iso_dtm}(e1, e2)

\S4method{+}{iso_dtm,Duration}(e1, e2)

\S4method{+}{Duration,iso_dtm}(e1, e2)
}
\arguments{
\item{e1}{First object}

\item{e2}{Second object}
}
\description{
Add Methods
}
