% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adeg}
\alias{adeg}
\title{Electrocardiogram Analysis Dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 78614 rows and 105 columns.}
\source{
Derived from the \link{adsl} and \link{eg} datasets using \code{{admiral}} (\url{https://github.com/Roche-GSK/admiral/blob/master/inst/example_scripts/ad_adeg.R})
}
\usage{
adeg
}
\description{
An example electrocardiogram analysis dataset
}
\keyword{datasets}
