% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adex}
\alias{adex}
\title{Exposure Analysis Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6315 rows and 88 columns.
}
\source{
Derived from the \link{adsl} and \link{ex} datasets using \code{{admiral}} (\url{https://github.com/Roche-GSK/admiral/blob/master/inst/templates/ad_adex.R})
}
\usage{
adex
}
\description{
An example exposure analysis dataset
}
\keyword{datasets}
