% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dt_dtm_utils.R
\name{adjust_last_day_imputation}
\alias{adjust_last_day_imputation}
\title{Adjust Last Day Imputation}
\usage{
adjust_last_day_imputation(imputed_dtc, partial, is_datetime)
}
\arguments{
\item{imputed_dtc}{A character vector of imputed date/datetime strings.}

\item{partial}{A list of partial date/time components.}

\item{is_datetime}{A logical indicating whether it's a datetime adjustment.}
}
\value{
A character vector of adjusted date/datetime strings.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This is a helping function for \code{get_dt_dtm_range()}
Adjusts the imputed date/datetime to the last day of the month when necessary.
}
\details{
This function is used when the date imputation method is set to "last" and
the day component is missing. It adjusts the date to the last day of the month.
}
\keyword{internal}
