% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiral_adcm}
\alias{admiral_adcm}
\title{Concomitant Medication Analysis Dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7510 rows and 89 columns.}
\source{
Derived from the \code{adsl} and \code{cm} datasets using \code{{admiral}} (\url{https://github.com/pharmaverse/admiral/blob/main/inst/templates/ad_adcm.R})
}
\usage{
admiral_adcm
}
\description{
An example concomitant medication analysis dataset
}
\keyword{datasets}
