% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiral_adeg}
\alias{admiral_adeg}
\title{ECG Analysis Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 78614 rows and 104 columns.
}
\source{
Derived from the \code{adsl} and \code{eg} datasets using \code{{admiral}} (\url{https://github.com/pharmaverse/admiral/blob/main/inst/templates/ad_adeg.R})
}
\usage{
admiral_adeg
}
\description{
An example ECG analysis dataset
}
\keyword{datasets}
