% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiral_adex}
\alias{admiral_adex}
\title{Subject Level Analysis Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6315 rows and 87 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6315 rows and 87 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6315 rows and 87 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6315 rows and 87 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6315 rows and 87 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6315 rows and 87 columns.
}
\source{
Derived from the \code{dm} and \code{ds} datasets using \code{{admiral}} (\url{https://github.com/pharmaverse/admiral/blob/main/inst/templates/ad_adsl.R})

Derived from the \code{adsl} and \code{ae} datasets using \code{{admiral}} (\url{https://github.com/pharmaverse/admiral/blob/main/inst/templates/ad_adae.R})

Derived from the \code{adsl} and \code{cm} datasets using \code{{admiral}} (\url{https://github.com/pharmaverse/admiral/blob/main/inst/templates/ad_adcm.R})

Derived from the \code{adsl} and \code{ex} datasets using \code{{admiral}} (\url{https://github.com/pharmaverse/admiral/blob/main/inst/templates/ad_adex.R})

Derived from the \code{adsl} and \code{eg} datasets using \code{{admiral}} (\url{https://github.com/pharmaverse/admiral/blob/main/inst/templates/ad_adeg.R})

Derived from the \code{adsl} and \code{vs} datasets using \code{{admiral}} (\url{https://github.com/pharmaverse/admiral/blob/main/inst/templates/ad_advs.R})
}
\usage{
admiral_adex

admiral_adex

admiral_adex

admiral_adex

admiral_adex

admiral_adex
}
\description{
An example subject level analysis dataset

An example adverse event analysis dataset

An example concomitant medication analysis dataset

An example exposure analysis dataset

An example ECG analysis dataset

An example vital signs analysis dataset
}
\keyword{datasets}
