% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiral_adlb}
\alias{admiral_adlb}
\title{Lab Analysis Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 83652 rows and 111 columns.
}
\source{
Derived from the \code{adlb} template.
}
\usage{
admiral_adlb
}
\description{
An example of lab analysis dataset
}
\seealso{
Other datasets: 
\code{\link{admiral_adsl}},
\code{\link{ex_single}},
\code{\link{queries_mh}},
\code{\link{queries}}
}
\concept{datasets}
\keyword{datasets}
