% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiral_adpp}
\alias{admiral_adpp}
\title{Pharmacokinetics Parameters Analysis Dataset}
\format{
An object of class \code{data.frame} with 1344 rows and 64 columns.
}
\source{
Derived from the \code{adsl} and \code{pp} datasets using \code{{admiral}} (\url{https://github.com/Roche-GSK/admiral/blob/main/inst/templates/ad_adpp.R})
}
\usage{
admiral_adpp
}
\description{
An example vital signs analysis dataset
}
\keyword{datasets}
