% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiral_advs}
\alias{admiral_advs}
\title{Vital Signs Analysis Dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 57763 rows and 102 columns.}
\source{
Derived from the \code{adsl} and \code{vs} datasets using \code{{admiral}} (\url{https://github.com/pharmaverse/admiral/blob/main/inst/templates/ad_advs.R})
}
\usage{
admiral_advs
}
\description{
An example vital signs analysis dataset
}
\keyword{datasets}
