% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiral_options}
\alias{admiral_options}
\title{Admiral Options -- Subject Keys}
\format{
An object of class \code{list} of length 2.
}
\source{
Defined defaults from team.
}
\usage{
admiral_options
}
\description{
Modifiable function inputs, e.g. subject_keys
}
\seealso{
Other datasets: 
\code{\link{admiral_adsl}},
\code{\link{atoxgr_criteria_ctcv4}},
\code{\link{ex_single}},
\code{\link{queries_mh}},
\code{\link{queries}}
}
\concept{datasets}
\keyword{datasets}
