% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiral-package.R, R/admiraldev-package.R
\docType{package}
\name{admiraldev-package}
\alias{admiraldev-package}
\title{admiraldev: Utility Functions and Development Tools for the Admiral Package Family}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Utility functions to check data, variables and conditions for functions used in 'admiral' and 'admiral' extension packages. Additional utility helper functions to assist developers with maintaining documentation, testing and general upkeep of 'admiral' and 'admiral' extension packages.

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Utility functions to check data, variables and conditions for functions used in 'admiral' and 'admiral' extension packages. Additional utility helper functions to assist developers with maintaining documentation, testing and general upkeep of 'admiral' and 'admiral' extension packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiraldev/}
  \item \url{https://github.com/pharmaverse/admiraldev/}
  \item Report bugs at \url{https://github.com/pharmaverse/admiraldev/issues}
}


Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiraldev/}
  \item \url{https://github.com/pharmaverse/admiraldev/}
  \item Report bugs at \url{https://github.com/pharmaverse/admiraldev/issues}
}


Other internal: 
\code{\link{extract_duplicate_records}()},
\code{\link{format.basket_select}()},
\code{\link{signal_duplicate_records}()}
}
\author{
\strong{Maintainer}: Ben Straub \email{ben.x.straub@gsk.com}

Authors:
\itemize{
  \item Stefan Bundfuss
  \item Jeffrey Dickinson
  \item Ross Farrugia
  \item Pooja Kumari
  \item Edoardo Mancini
  \item Sadchla Mascary
  \item Zelos Zhu
}

Other contributors:
\itemize{
  \item Ania Golab [contributor]
  \item Samia Kabi [contributor]
  \item Syed Mubasheer [contributor]
  \item Thomas Neitmann [contributor]
  \item Ondrej Slama [contributor]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item GlaxoSmithKline LLC [copyright holder, funder]
}

}
\concept{internal}
\keyword{internal}
