% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adsl}
\alias{adsl}
\title{Subject Level Analysis Dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 306 rows and 35 columns.}
\source{
\url{https://github.com/Roche-GSK/admiral/blob/master/inst/example_scripts/ad_adsl.R}
}
\usage{
adsl
}
\description{
An example subject level analysis dataset
}
\keyword{datasets}
