% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{advs}
\alias{advs}
\title{Vital Signs Analysis Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 57763 rows and 103 columns.
}
\source{
Derived from the \link{adsl} and \link{vs} datasets using \code{{admiral}} (\url{https://github.com/Roche-GSK/admiral/blob/master/inst/templates/ad_advs.R})
}
\usage{
advs
}
\description{
An example vital signs analysis dataset
}
\keyword{datasets}
