% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{advs_param}
\alias{advs_param}
\title{ADVS Parameter Mapping}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19 rows and 4 columns.}
\usage{
advs_param
}
\description{
A dataset containing mappings from VS variables to ADVS \code{PARAM} values
}
\keyword{datasets}
