% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ae}
\alias{ae}
\title{Adverse Events Dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1191 rows and 35 columns.}
\source{
\url{https://bitbucket.cdisc.org/projects/CED/repos/sdtm-adam-pilot-project/browse/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/ae.xpt}
}
\usage{
ae
}
\description{
A SDTM AE dataset from the CDISC pilot project
}
\keyword{datasets}
