% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{are_records_same}
\alias{are_records_same}
\title{Checks the length of derived records and new values of are equal}
\usage{
are_records_same(x, y, x_arg, y_arg)
}
\arguments{
\item{x}{an R object}

\item{y}{an R object to compare the length with \code{x}.}

\item{x_arg}{Argument name of x.}

\item{y_arg}{Argument name of y.}
}
\value{
Logical value.
}
\description{
Checks the length of derived records and new values of are equal
}
\examples{
x <- list("x", "y")
y <- list("y", "z")
assert_that(are_records_same(x, y, "x", "y"))
}
