% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{arg_specified}
\alias{arg_specified}
\title{Argument Specified}
\usage{
arg_specified(arg)
}
\arguments{
\item{arg}{The argument to check}
}
\value{
``TRUE`` if the argument was specified, ``FALSE`` otherwise
}
\description{
Checks if an argument was specified
}
\examples{
assert_that(arg_specified(refdate), arg_specified(date))
}
\author{
Stefan Bundfuss
}
