% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{as_name}
\alias{as_name}
\title{Turn a Quosure into a String}
\usage{
as_name(x)
}
\arguments{
\item{x}{A \code{quosure}}
}
\value{
A \code{character} vector
}
\description{
Turn a Quosure into a String
}
\details{
This function is missing in earlier version of {rlang} which is why we re-
implement it here.
}
\seealso{
Developer Utility Functions: 
\code{\link{\%notin\%}()},
\code{\link{\%or\%}()},
\code{\link{arg_name}()},
\code{\link{contains_vars}()},
\code{\link{convert_dtm_to_dtc}()},
\code{\link{extract_vars}()},
\code{\link{filter_if}()},
\code{\link{valid_time_units}()},
\code{\link{vars2chr}()}
}
\concept{dev_utility}
\keyword{dev_utility}
