% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_character_scalar}
\alias{assert_character_scalar}
\title{Is an Argument a Character Scalar (String)?}
\usage{
assert_character_scalar(arg, values = NULL, optional = FALSE)
}
\arguments{
\item{arg}{A function argument to be checked}

\item{values}{A \code{character} vector of valid values for \code{arg}}

\item{optional}{Is the checked parameter optional? If set to \code{FALSE} and \code{arg}
is \code{NULL} then an error is thrown}
}
\value{
The function throws an error if \code{arg} is not a character vector or if \code{arg}
is a character vector but of length > 1 or if its value is not one of the \code{values}
specified. Otherwise, the input is returned invisibly.
}
\description{
Checks if an argument is a character scalar and (optionally) whether it matches
one of the provided \code{values}.
}
\examples{
example_fun <- function(msg_type) {
  assert_character_scalar(msg_type, values = c("warning", "error"))
}

example_fun("warning")

tryCatch(
  example_fun("message"),
  error = function(e) cat(e$message, "\n")
)

tryCatch(
  example_fun(TRUE),
  error = function(e) cat(e$message, "\n")
)
}
\author{
Thomas Neitmann
}
\keyword{assertion}
