% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_date_vector}
\alias{assert_date_vector}
\title{Is a Variable is a Date or Datetime Variable?}
\usage{
assert_date_vector(var, var_name = NULL)
}
\arguments{
\item{var}{The variable to check}

\item{var_name}{The name of the variable. If the argument is specified, the
specified name is displayed in the error message.}
}
\value{
The function throws an error if \code{var} is not a date or datetime variable
and returns the input invisibly otherwise.
}
\description{
Checks if a variable vector is a date or datetime variable
}
\examples{
library(tibble)
library(rlang)

example_fun <- function(var) {
  assert_date_vector(var)
}

my_dates_chr <- c("2018-08-23", "2022-01-30", "1993-07-14")
my_dates_fmt <- as.Date(my_dates_chr, tz = "UTC")

example_fun(
  var = my_dates_fmt
)

}
\author{
Sadchla Mascary
}
\keyword{assertion}
