% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dt_dtm_utils.R
\name{assert_dates_strict}
\alias{assert_dates_strict}
\title{Assert Strict Date Ranges}
\usage{
assert_dates_strict(min_dates_strict, max_dates_strict)
}
\arguments{
\item{min_dates_strict}{A list of minimum dates to check.

\describe{
\item{Default value}{none}
}}

\item{max_dates_strict}{A list of maximum dates to check.

\describe{
\item{Default value}{none}
}}
}
\value{
Invisibly returns \code{NULL} if the assertion passes.
}
\description{
Asserts that the minimum dates specified in \code{min_dates_strict} are not
greater than the maximum dates specified in \code{max_dates_strict}.
}
\details{
If any minimum date is greater than its corresponding maximum date, an error
is thrown with the invalid combinations of minimum and maximum dates.
}
\keyword{internal}
