% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_has_only_one_baseline_record}
\alias{assert_has_only_one_baseline_record}
\title{Are There Multiple Baseline Records?}
\usage{
assert_has_only_one_baseline_record(dataset, by)
}
\arguments{
\item{dataset}{A \code{data.frame}}

\item{by}{A \code{character} vector of variable names which uniquely identify a
set of records that should only contain a single baseline record}
}
\value{
The function throws an error if a subject has multiple baseline
records
}
\description{
Checks if a dataset contains multiple baseline records
}
\author{
Thomas Neitmann
}
