% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_is_symbol}
\alias{assert_is_symbol}
\title{Check whether an argument is a quosure with a symbol or not.}
\usage{
assert_is_symbol(x)
}
\arguments{
\item{x}{Test object}
}
\value{
TRUE or error.
}
\description{
Check whether an argument is a quosure with a symbol or not.
}
\examples{
assert_is_symbol(rlang::quo(x))
}
