% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_logical_scalar}
\alias{assert_logical_scalar}
\title{Is an Argument a Logical Scalar (Boolean)?}
\usage{
assert_logical_scalar(arg)
}
\arguments{
\item{arg}{A function argument to be checked}
}
\value{
The function throws an error if \code{arg} is not a \code{logical} vector or if \code{arg}
is a \code{logical} vector but of length > 1. Otherwise, the input is returned invisibly.
}
\description{
Checks if an argument is a logical scalar
}
\examples{
imputation_flag <- FALSE
add_one <- 1

assert_logical_scalar(imputation_flag)
tryCatch(
  assert_logical_scalar(add_one),
  error = function(e) cat(e$message)
)
}
\author{
Thomas Neitmann
}
\keyword{assertion}
