% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_computed.R
\name{assert_parameters_argument}
\alias{assert_parameters_argument}
\title{Asserts \code{parameters} Argument and Converts to List of Expressions}
\usage{
assert_parameters_argument(parameters, optional = TRUE)
}
\arguments{
\item{parameters}{The argument to check}

\item{optional}{Is the checked argument optional? If set to \code{FALSE} and
\code{parameters} is \code{NULL} then an error is thrown.}
}
\value{
The \code{parameters} argument (converted to a list of symbol, if it is a
character vector)
}
\description{
The function asserts that the argument is a character vector or a list of
expressions. If it is a character vector, it converts it to a list of
symbols.
}
\seealso{
Other Advanced Functions: 
\code{\link{assert_db_requirements}()},
\code{\link{assert_terms}()},
\code{\link{assert_valid_queries}()},
\code{\link{extend_source_datasets}()},
\code{\link{filter_date_sources}()},
\code{\link{format.basket_select}()},
\code{\link{get_hori_data}()},
\code{\link{list_tte_source_objects}()},
\code{\link{params}()},
\code{\link{validate_basket_select}()},
\code{\link{validate_query}()}
}
\concept{other_advanced}
\keyword{other_advanced}
