% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_symbol}
\alias{assert_symbol}
\title{Is an Argument a Symbol?}
\usage{
assert_symbol(arg, optional = FALSE)
}
\arguments{
\item{arg}{A function argument to be checked}

\item{optional}{Is the checked parameter optional? If set to \code{FALSE} and \code{arg}
is \code{NULL} then an error is thrown}
}
\value{
The function throws an error if \code{arg} is not a symbol and returns the input
invisibly otherwise. \code{arg} is expected to be a \code{quosure}.
}
\description{
Checks if an argument is a symbol
}
\examples{
data(dm)

example_fun <- function(dat, var) {
  var <- assert_symbol(rlang::enquo(var))
  dplyr::select(dat, !!var)
}

example_fun(dm, USUBJID)

tryCatch(
  example_fun(dm),
  error = function(e) cat(e$message, "\n")
)

tryCatch(
  example_fun(dm, "USUBJID"),
  error = function(e) cat(e$message, "\n")
)

tryCatch(
  example_fun(dm, toupper(PARAMCD)),
  error = function(e) cat(e$message, "\n")
)
}
\author{
Thomas Neitmann
}
\keyword{assertion}
