% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dt_dtm_utils.R
\name{assert_time_imputation}
\alias{assert_time_imputation}
\title{Assert \code{time_imputation}}
\usage{
assert_time_imputation(highest_imputation, time_imputation)
}
\arguments{
\item{highest_imputation}{Highest imputation level}

\item{time_imputation}{The value to impute time when missing}
}
\value{
asserted \code{time_imputation}

\code{invisible(NULL)}
}
\description{
Applies assertions on the \code{time_imputation} argument
}
\keyword{internal}
