% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_unit}
\alias{assert_unit}
\title{Asserts That a Parameter is Provided in the Expected Unit}
\usage{
assert_unit(dataset, param, unit_var, unit)
}
\arguments{
\item{dataset}{A \code{data.frame}}

\item{param}{Parameter code of the parameter to check}

\item{unit_var}{Variable providing the unit}

\item{unit}{Expected unit}
}
\value{
The function throws an error if the unit variable differs from the
unit for any observation of the parameter in the input dataset
}
\description{
Checks if a parameter (\code{PARAMCD}) in a dataset is provided in the expected
unit.
}
\examples{
data(advs)
assert_unit(advs, param = "WEIGHT", unit = "kg", unit_var = AVALU)
\dontrun{
assert_unit(advs, param = "WEIGHT", unit = "g", unit_var = AVALU)
}
}
\author{
Stefan Bundfuss
}
\keyword{assertion}
