% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_varval_list}
\alias{assert_varval_list}
\title{Is an Argument a Variable-Value List?}
\usage{
assert_varval_list(arg, optional = FALSE)
}
\arguments{
\item{arg}{The argument to checked}

\item{optional}{Is the checked parameter optional? If set to \code{FALSE} and \code{arg}
is \code{NULL} then an error is thrown.}
}
\description{
Checks if the argument is a list of quosures where the expressions are
variable-value pairs. The value can be a symbol, a string, or \code{NA}. More general
expression are not allowed.
}
\examples{
example_fun <- function(vars) {
  assert_varval_list(vars)
}
example_fun(vars(DTHDOM = "AE", DTHSEQ = AESEQ))

try(example_fun(vars("AE", DTSEQ = AESEQ)))
}
\author{
Stefan Bundfuss, Thomas Neitmann
}
\keyword{assertion}
