% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atoxgr_criteria_ctcv6}
\alias{atoxgr_criteria_ctcv6}
\title{Metadata Holding Grading Criteria for NCI-CTCAEv6 using SI unit where applicable}
\format{
An object of class \code{data.frame} with 44 rows and 13 columns.
}
\usage{
atoxgr_criteria_ctcv6
}
\description{
Metadata Holding Grading Criteria for NCI-CTCAEv6 using SI unit where applicable
}
\details{
This metadata has its origin in the ADLB Grading Spec json file \code{data-raw/adlb_grading/ncictcaev6.json}.
The variables \code{GRADE_NA_CODE}, \code{GRADE_4_CODE}, \code{GRADE_3_CODE}, \code{GRADE_2_CODE} and \code{GRADE_1_CODE} in
the json file are combined to create \code{GRADE_CRITERIA_CODE}, and then dropped from metadata.
The dataset contains the following columns:
\itemize{
\item \code{SOC}: variable to hold the SOC of the lab test criteria.
\item \code{TERM}: variable to hold the term describing the criteria applied to a particular lab test,
eg. 'Anemia' or 'INR Increased'. Note: the variable is case insensitive.
\item \verb{Grade 1}: Criteria defining lab value as Grade 1.
\item \verb{Grade 2}: Criteria defining lab value as Grade 2.
\item \verb{Grade 3}: Criteria defining lab value as Grade 3.
\item \verb{Grade 4}: Criteria defining lab value as Grade 4.
\item \verb{Grade 5}: Criteria defining lab value as Grade 5.
\item \code{Definition}: Holds the definition of the lab test abnormality.
\item \code{GRADE_CRITERIA_CODE}: variable to hold code that creates grade based on defined criteria.
\item \code{UNIT_CHECK}: variable to hold SI unit of particular lab test. Used to check against input data
if criteria is based on absolute values.
\item \code{VAR_CHECK}: List of variables required to implement lab grade criteria. Use to check against
input data.
\item \code{DIRECTION}: variable to hold the direction of the abnormality of a particular lab test
value. 'L' is for LOW values, 'H' is for HIGH values. Note: the variable is case insensitive.
\item \code{COMMENT}: Holds any information regarding rationale behind implementation of grading criteria.
}

Note: Variables \code{SOC}, \code{TERM}, \verb{Grade 1}, \verb{Grade 2},\verb{Grade 3},\verb{Grade 4},\verb{Grade 5}, \code{Definition}
are from the source document on NCI-CTC website defining the grading criteria.
\href{https://dctd.cancer.gov/research/ctep-trials/for-sites/adverse-events#ctep-ctcae}{\strong{Common Terminology Criteria for Adverse Events (CTCAE)v6.0}}
From these variables only 'TERM' is used in the \code{{admiral}} code, the rest are for information and
traceability only.
}
\seealso{
Other metadata:
\code{\link{atoxgr_criteria_ctcv4}},
\code{\link{atoxgr_criteria_ctcv4_uscv}},
\code{\link{atoxgr_criteria_ctcv5}},
\code{\link{atoxgr_criteria_ctcv5_uscv}},
\code{\link{atoxgr_criteria_ctcv6_uscv}},
\code{\link{atoxgr_criteria_daids}},
\code{\link{atoxgr_criteria_daids_uscv}},
\code{\link{country_code_lookup}},
\code{\link{dose_freq_lookup}}
}
\concept{metadata}
\keyword{metadata}
