% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_last_dose.R
\name{calc_ldosedtm}
\alias{calc_ldosedtm}
\title{Helper function to calculate last dose}
\usage{
calc_ldosedtm(exendtc, aestdtc, exdose)
}
\arguments{
\item{exendtc}{EX.EXENDTC}

\item{aestdtc}{ADAE.AESTDTC}

\item{exdose}{EX.EXDOSE}
}
\value{
date-time vector
}
\description{
Helper function to calculate last dose
}
