% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_user_fun.R
\name{call_user_fun}
\alias{call_user_fun}
\title{Calls a Function Provided by the User}
\usage{
call_user_fun(call)
}
\arguments{
\item{call}{Call to be executed

\describe{
\item{Default value}{none}
}}
}
\value{
The return value of the function call
}
\description{
r lifecycle::badge("deprecated")
Calls a function provided by the user and adds the function call to the error
message if the call fails.
}
\examples{
call_user_fun(compute_bmi(
  height = 172,
  weight = 60
))

try(call_user_fun(compute_bmi(
  height = 172,
  weight = "hallo"
)))
}
\seealso{
Other deprecated: 
\code{\link{date_source}()},
\code{\link{derive_param_extreme_record}()},
\code{\link{derive_var_dthcaus}()},
\code{\link{derive_var_extreme_dt}()},
\code{\link{derive_var_extreme_dtm}()},
\code{\link{dthcaus_source}()},
\code{\link{get_summary_records}()}
}
\concept{deprecated}
\keyword{deprecated}
