% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_utils.R
\name{chr2vars}
\alias{chr2vars}
\title{Turn a Character Vector into a List of Expressions}
\usage{
chr2vars(chr)
}
\arguments{
\item{chr}{A character vector}
}
\value{
A \code{list} of expressions as returned by \code{\link[=exprs]{exprs()}}
}
\description{
Turn a character vector into a list of expressions
}
\examples{
chr2vars(c("USUBJID", "AVAL"))
}
\seealso{
Utilities for working with quosures/list of expressions: 
\code{\link{negate_vars}()}
}
\concept{utils_quo}
\keyword{utils_quo}
