% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_aage.R
\name{compute_agegr}
\alias{compute_agegr}
\alias{compute_agegrn}
\title{Calculate age groups}
\usage{
compute_agegr(x, breaks, labels = NULL, leq = TRUE)

compute_agegrn(x, breaks, labels = NULL, leq = FALSE)
}
\arguments{
\item{x}{Numeric vector providing the age.}

\item{breaks}{Numeric vector providing the breaks for the age groups.}

\item{labels}{Character vector providing the labels of the age groups.
The length of \code{labels} needs to be greater by one than the length of \code{breaks}.}

\item{leq}{Should intervals be defined as "lower or equal" or "greater or equal"?
Defaults to the first.}
}
\value{
Factor (for \code{compute_agegr}) or numeric (for \code{compute_agegrn}) with defined age groups.
}
\description{
Calculates age groups by given breaks and returns either factor or numeric.
}
\section{Functions}{
\itemize{
\item \code{compute_agegrn}: Calculate age groups and return the number of the group.
}}

\examples{
compute_agegr(1:10, 5)
# provide arbitrary labels
compute_agegr(1:10, 5, labels = c("Less than 5 years old",
                                  "More or equal to 5 years old"))
# change the interval borders
compute_agegr(1:10, 5, leq = TRUE)
# add more breaks
compute_agegr(0:100, c(30, 60))
# compare factor with numeric output
compute_agegr(1:10, 5)
compute_agegrn(1:10, 5)
}
\author{
Ondrej Slama
}
