% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_advs_params.R
\name{compute_bsa}
\alias{compute_bsa}
\title{Derive BSA (Body Surface Area)}
\usage{
compute_bsa(height = height, weight = weight, method = "Mosteller")
}
\arguments{
\item{method}{Derivation method to use

The derivation method, e.g. Mosteller will use sqrt(height(cm) * weight(kg)) / 3600

Permitted Values: character value}

\item{height_code}{HEIGHT parameter code

The observations where \code{PARAMCD} equals the specified value are considered
as the HEIGHT assessments. It is expected that HEIGHT is measured in cm.

Permitted Values: character value}

\item{weight_code}{WEIGHT parameter code

The observations where \code{PARAMCD} equals the specified value are considered
as the WEIGHT assessments. It is expected that WEIGHT is measured in kg.

Permitted Values: character value}
}
\value{
The BSA (Body Surface Area) in m^2.
}
\description{
Derives BSA from HEIGHT and WEIGHT making use of the specified derivation formula
}
\details{
The BSA (Body Surface Area) is derived with the specified formula:
"Mosteller", "DuBois-DuBois", "Haycock", "Gehan-George", "Boyd",
""Fujimoto" or "Takahira".
}
\examples{
# derive BSA by the Mosteller method
compute_bsa(
  height = 170,
  weight = 75,
  method = "Mosteller")

# derive BSA by the DuBois & DuBois method
compute_bsa(
  height = 170,
  weight = 75,
  method = "DuBois-DuBois")

}
\author{
Eric Simms
}
\keyword{BSA}
\keyword{adam}
\keyword{computation}
