% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_crcl.R
\name{compute_crcl}
\alias{compute_crcl}
\title{Compute Creatinine Clearance (CRCL)}
\usage{
compute_crcl(creat, age, wt, sex)
}
\arguments{
\item{creat}{Creatinine (umol/L)

A numeric vector is expected.}

\item{age}{Age (years)

A numeric vector is expected.}

\item{wt}{Weight (kg)

A numeric vector is expected.}

\item{sex}{Gender

A character vector is expected.
Expected Values: 'M' 'F'}
}
\value{
A numeric vector of crcl values (mL/min)
}
\description{
Compute Creatinine Clearance (CRCL) as estimate of renal function
}
\details{
Calculates an estimate of Creatinine Clearance

\deqn{((140 - age) * weight (kg) * constant) / creatinine (umol/L)}

Constant = 1.04 for females, 1.23 for males
}
\examples{
compute_crcl(
  creat = 90, age = 53, wt = 85, sex = "M"
)

compute_crcl(
  creat = 70, age = 52, wt = 68, sex = "F"
)
}
\seealso{
BDS-Findings Functions that returns a vector: 
\code{\link{compute_bmi}()},
\code{\link{compute_bsa}()},
\code{\link{compute_framingham}()},
\code{\link{compute_map}()},
\code{\link{compute_qtc}()},
\code{\link{compute_qual_imputation_dec}()},
\code{\link{compute_qual_imputation}()},
\code{\link{compute_rr}()},
\code{\link{compute_scale}()}
}
\concept{com_bds_findings}
\keyword{com_bds_findings}
