% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_date_vars.R
\name{compute_imputed_dtc}
\alias{compute_imputed_dtc}
\title{Impute partial date/time portion of a --DTC variable}
\usage{
compute_imputed_dtc(dtc, impute = TRUE, day = 1, month = 1,
  year = "NONE", hour = 0, min = 0, sec = 0)
}
\arguments{
\item{dtc}{The --DTC date to impute

A character date is expected in a format like yyy-mm-yy}

\item{impute}{Whether to impute partial date or no

A logical value is expected.

Default: TRUE}

\item{day}{The value to impute the day when day is missing

Permitted Values: number from 1 to 31, text 'FIRST', 'MID', 'LAST'
'FIRST' must be used to impute to the first day of the month
'MID' must be used to impute to the 15th of the month
'LAST' must be used to impute to the last day of the month

Default: 1}

\item{month}{the value to impute the month when month is missing

Permitted Values: number from 1 to 12

Default: 1}

\item{year}{the value to impute the year when year is missing

Permitted Values: any 4-digit number or 'NONE' if a missing date shoudl not be imputed

Default: 'NONE'}
}
\value{
a character
when impute = TRUE: with a complete/time  : e.g. 2021-03-13T00:00:00
when impute = FALSE: with a complete date/time if the date is complete,
or a "" if the date is partial
}
\description{
Imputation partial date/time portion of a --DTC variable
based on user input
}
\examples{

date<-"2019-07-18T15:25"
compute_imputed_dtc(dtc="2019-07-18T15:25")#returns "2019-07-18T15:25:00"
compute_imputed_dtc(dtc="2019-07-18")#returns "2019-07-18T00:00:00"
compute_imputed_dtc(dtc="2019-07")#returns "2019-07-01T00:00:00"
compute_imputed_dtc(dtc="2019", day="LAST")#returns "2019-01-31T00:00:00"
compute_imputed_dtc(dtc="2019", day="LAST", month=12)#returns "2019-12-31T00:00:00"
compute_imputed_dtc(dtc="2019-07", day="LAST", hour=23, min=59, sec=59)#returns "2019-07-31T23:59:59"
ompute_imputed_dtc(dtc="2019-07", day="FIRST")#returns "2019-07-01T00:00:00"
compute_imputed_dtc(dtc="2019-07", day="MID")#returns "2019-07-15T00:00:00"
compute_imputed_dtc(dtc="2019")#returns "2019-01-01T00:00:00"
compute_imputed_dtc(dtc="2019---07")#returns "2019-01-01T00:00:00"

}
\seealso{
Other {general functions}: \code{\link{derive_merged_vars}},
  \code{\link{derive_vars_dtm}},
  \code{\link{derive_vars_dt}}, \code{\link{filter_first}}
}
\author{
Samia Kabi
}
\concept{{general functions}}
