% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_adeg_params.R
\name{compute_qtcb}
\alias{compute_qtcb}
\title{Compute Corrected QT Using Bazett's Formula}
\usage{
compute_qtcb(qt, rr)
}
\arguments{
\item{qt}{QT interval

A numeric vector is expected. It is expected that QT is measured in msec.}

\item{rr}{RR interval

A numeric vector is expected. It is expected that RR is measured in msec.}
}
\value{
QT interval in msec:
\deqn{\frac{QT}{\sqrt{\frac{RR}{1000}}}}{QT/\sqrt(RR/1000)}
}
\description{
Computes corrected QT using Bazett's formula.
}
\examples{
compute_qtcb(qt = 350, rr = 56.54)
}
\author{
Stefan Bundfuss
}
\keyword{adeg}
\keyword{computation}
