% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_adeg_params.R
\name{compute_qtcf}
\alias{compute_qtcf}
\title{Compute Corrected QT Using Fridericia's Formula}
\usage{
compute_qtcf(qt, rr)
}
\arguments{
\item{qt}{QT interval

A numeric vector is expected. It is expected that QT is measured in msec.}

\item{rr}{RR interval

A numeric vector is expected. It is expected that RR is measured in msec.}
}
\value{
QT interval in msec:
\deqn{\frac{QT}{\sqrt[3]{\frac{RR}{1000}}}}{QT/(RR/1000)^(1/3)}
}
\description{
Computes corrected QT using Fridericia's formula.
}
\examples{
compute_qtcf(qt = 350, rr = 56.54)
}
\author{
Stefan Bundfuss
}
\keyword{adeg}
\keyword{computation}
