% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_adeg_params.R
\name{compute_qtlc}
\alias{compute_qtlc}
\title{Compute Corrected QT Using Sagie's Formula}
\usage{
compute_qtlc(qt, rr)
}
\arguments{
\item{qt}{QT interval

A numeric vector is expected. It is expected that QT is measured in msec.}

\item{rr}{RR interval

A numeric vector is expected. It is expected that RR is measured in msec.}
}
\value{
QT interval in msec:
\deqn{1000\left(\frac{QT}{1000} + 0.154\left(1 - \frac{RR}{1000}\right)\right)}{
1000(QT/1000 + 0.154(1 - RR/1000))}
}
\description{
Computes corrected QT using Sagie's formula.
}
\examples{
compute_qtlc(qt = 350, rr = 56.54)
}
\author{
Stefan Bundfuss
}
\keyword{adeg}
\keyword{computation}
