% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_qualifier_imputation.R
\name{compute_qualifier_imputation}
\alias{compute_qualifier_imputation}
\title{Function to impute values when qualifier exists in character result}
\usage{
compute_qualifier_imputation(character_value, imputation_type = 1, factor = 0)
}
\arguments{
\item{character_value}{Character version of value to be imputed}

\item{imputation_type}{(default value=1)
Valid Values:
1: Strip <, >, = and convert to numeric.
2: imputation_type=1 and if the character value contains a < or >, the number of
of decimals associated with the character value is found and then a factor of
1/10^(number of decimals + 1) will be added/subtracted from the numeric value.
If no decimals exists, a factor of 1/10 will be added/subtracted from the value.}

\item{factor}{Numeric value (default=0), when using \code{imputation_type} = 1, this
value can be added or subtracted when the qualifier is removed.}
}
\value{
The imputed value
}
\description{
Derive an imputed value
}
\examples{
compute_qualifier_imputation("<40")
}
\author{
Alice Ehmann
}
