% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_qualifier_imputation.R
\name{compute_qualifier_imputation_decimal}
\alias{compute_qualifier_imputation_decimal}
\title{Function to compute factor for value imputation when character
value contains < or >.  The factor is calculated using the number of decimals.
If there are no decimals, the factor is 1, otherwise the factor = 1/10^decimal
place.  For example, the factor for 100 = 1, the factor for 5.4 = 1/10^1,
the factor for 5.44 = 1/10^2.  This results in no additional false precision
added to the value.  This is an intermediate function.}
\usage{
compute_qualifier_imputation_decimal(character_value_decimal)
}
\arguments{
\item{character_value_decimal}{Character value to determine decimal precision}
}
\value{
Decimal precision value to add or subtract
}
\description{
Derive an imputed value
}
\examples{
compute_qualifier_imputation_decimal("<40.1")
}
\author{
Alice Ehmann
}
