% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_bds_params.R
\name{compute_ratio}
\alias{compute_ratio}
\title{Compute ratio}
\usage{
compute_ratio(x, y)
}
\arguments{
\item{x}{Dividend

\emph{Permitted Values:} numeric vector}

\item{y}{Divisor

\emph{Permitted Values:} numeric vector}
}
\value{
The ratio computed as x / y.
}
\description{
Computes ratio as dividend (x) divided by divisor (y)
}
\details{
\code{x} and \code{y} must be of the same length.

If the divisor equals 0, the result will be NA.

Usually this computation function can not be used with \code{\%>\%}.
}
\examples{

# Example of Waist-to-Height Ratio

wstcir <- c(110, 108, 107)
hipcir <- c(125, 124, 123)

compute_ratio(wstcir, hipcir)
}
\seealso{
\code{\link[=derive_param_ratio]{derive_param_ratio()}}

BDS-Findings Functions that returns a vector: 
\code{\link{compute_bmi}()},
\code{\link{compute_bsa}()},
\code{\link{compute_egfr}()},
\code{\link{compute_framingham}()},
\code{\link{compute_map}()},
\code{\link{compute_qtc}()},
\code{\link{compute_qual_imputation}()},
\code{\link{compute_qual_imputation_dec}()},
\code{\link{compute_rr}()},
\code{\link{compute_scale}()}
}
\concept{com_bds_findings}
\keyword{com_bds_findings}
