% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_rr.R
\name{compute_rr}
\alias{compute_rr}
\title{Compute RR Interval From Heart Rate}
\usage{
compute_rr(hr)
}
\arguments{
\item{hr}{Heart rate

A numeric vector is expected. It is expected that heart rate is measured in
beats/min.

\describe{
\item{Default value}{none}
}}
}
\value{
RR interval in ms:
\deqn{\frac{60000}{HR}}{60000 / HR}
}
\description{
Computes RR interval from heart rate.
}
\details{
Usually this computation function can not be used with \verb{\%>\%}.
}
\examples{
compute_rr(hr = 70.14)
}
\seealso{
\code{\link[=derive_param_rr]{derive_param_rr()}}

BDS-Findings Functions that returns a vector:
\code{\link[=compute_bmi]{compute_bmi()}},
\code{\link[=compute_bsa]{compute_bsa()}},
\code{\link[=compute_egfr]{compute_egfr()}},
\code{\link[=compute_framingham]{compute_framingham()}},
\code{\link[=compute_map]{compute_map()}},
\code{\link[=compute_qtc]{compute_qtc()}},
\code{\link[=compute_qual_imputation]{compute_qual_imputation()}},
\code{\link[=compute_qual_imputation_dec]{compute_qual_imputation_dec()}},
\code{\link[=compute_scale]{compute_scale()}},
\code{\link[=transform_range]{transform_range()}}
}
\concept{com_bds_findings}
\keyword{com_bds_findings}
