% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{contains_vars}
\alias{contains_vars}
\title{Returns If Argument is a Vector of Symbols or Named Expressions}
\usage{
contains_vars(arg)
}
\arguments{
\item{arg}{Input vector}
}
\value{
\code{TRUE} if
\itemize{
\item The vector is a vector of quosures and
\item all elements are a symbol or a named expression like \code{CHG = AVAL - BASE}
}

\code{FALSE} otherwise
}
\description{
Returns If Argument is a Vector of Symbols or Named Expressions
}
\examples{

contains_vars(rlang::expr(AVAL))
contains_vars(dplyr::vars(AVAL))
contains_vars(dplyr::vars(1))
contains_vars(dplyr::vars(AVAL, ID = 1))
contains_vars(dplyr::vars(AVAL, CHG = AVAL - BASE))
}
\seealso{
Functions moved to admiraldev package: 
\code{\link{arg_name}()},
\code{\link{assert_db_requirements}()},
\code{\link{assert_terms}()},
\code{\link{assert_valid_queries}()},
\code{\link{backquote}()},
\code{\link{dquote}()},
\code{\link{enumerate}()},
\code{\link{extend_source_datasets}()},
\code{\link{extract_vars}()},
\code{\link{format.sdg_select}()},
\code{\link{format.smq_select}()},
\code{\link{get_constant_vars}()},
\code{\link{get_duplicates}()},
\code{\link{get_source_vars}()},
\code{\link{is_auto}()},
\code{\link{quo_c}()},
\code{\link{replace_values_by_names}()},
\code{\link{squote}()},
\code{\link{valid_time_units}()},
\code{\link{validate_query}()},
\code{\link{validate_sdg_select}()},
\code{\link{validate_smq_select}()},
\code{\link{what_is_it}()}
}
\concept{move_adm_dev}
\keyword{move_adm_dev}
