% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dt.R
\name{convert_dtc_to_dt}
\alias{convert_dtc_to_dt}
\title{Convert a Date Character Vector into a Date Object}
\usage{
convert_dtc_to_dt(
  dtc,
  highest_imputation = "n",
  date_imputation = "first",
  min_dates = NULL,
  min_dates_strict = NULL,
  max_dates = NULL,
  max_dates_strict = NULL,
  preserve = FALSE
)
}
\arguments{
\item{dtc}{The --DTC date to convert.

\describe{
\item{Permitted values}{a character date vector}
\item{Default value}{none}
}}

\item{highest_imputation}{Highest imputation level

The \code{highest_imputation} argument controls which components of the \code{--DTC}
value are imputed if they are missing. All components up to the specified
level are imputed.

If a component at a higher level than the highest imputation level is
missing, \code{NA} is returned. For example, for \code{highest_imputation = "D"}
\code{"2020"} results in \code{NA} because the month is missing.

If \code{"n"} (none, lowest level) is specified no imputation is performed, i.e.,
if any component is missing, \code{NA} is returned.

If \code{"Y"} (year, highest level) is specified, \code{date_imputation} must be \code{"first"} or \code{"last"}
and \code{min_dates} or \code{max_dates} must be specified respectively. Otherwise,
an error is thrown.

\describe{
\item{Permitted values}{\code{"Y"} (year, highest level), \code{"M"} (month), \code{"D"}
(day), \code{"n"} (none, lowest level)}
\item{Default value}{\code{"n"}}
}}

\item{date_imputation}{The value to impute the day/month when a datepart is
missing.

A character value is expected.
\itemize{
\item The\code{"first"} and \code{"last"} keywords allow imputation to the first/last
day/month. They can also be used to impute the year if used in conjunction
with the \code{min_dates} or \code{max_dates} arguments. Some examples of this are available
\href{https://pharmaverse.github.io/admiral/cran-release/articles/imputation.html#minimummaximum-dates}{here}.
\item When \code{highest_imputation} is \code{"M"} or \code{"D"}, the \code{"mid"} keyword can also be
specified to impute missing components to the middle of the possible range:
\itemize{
\item If both month and day are missing, they are imputed as \code{"06-30"}
(middle of the year).
\item If only day is missing, it is imputed as \code{"15"} (middle of the month).
}
\item \code{"<dd>"} can be specified only if \code{highest_imputation = "D"}. Missing days are
imputed by the specified day, e.g. \code{"10"} for the 10th day of the month.
The specified day should be valid for all months as otherwise an error might be
issued. For example, \code{date_imputation = "30"} results in an invalid date of
"2024-02-30" for the partial date "2024-02".
\item \code{"<mm>-<dd>"} can be specified only if \code{highest_imputation} is \code{"M"},  e.g. \code{"06-15"}
for the 15th of June.
}

\describe{
\item{Permitted values}{a key-word, i.e. \code{"first"}, \code{"mid"}, \code{"last"}, or \code{"<mm>-<dd>"} or \code{"<dd>"}}
\item{Default value}{\code{"first"}}
}}

\item{min_dates}{Minimum dates

A list of dates is expected. It is ensured that the imputed date is not
before any of the specified dates, e.g., that the imputed adverse event start
date is not before the first treatment date. Only dates which are in the
range of possible dates of the \code{dtc} value are considered. The possible dates
are defined by the missing parts of the \code{dtc} date (see example below). This
ensures that the non-missing parts of the \code{dtc} date are not changed.
A date or date-time object is expected.
For example

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(lubridate)
impute_dtc_dt(
  "2020-11",
  min_dates = list(
   ymd("2020-12-06"),
   ymd("2020-11-11")
  ),
  highest_imputation = "M"
)
#> [1] "2020-11-11"
}\if{html}{\out{</div>}}

returns \code{"2020-11-11"} because the possible dates for \code{"2020-11"} range from
\code{"2020-11-01"} to \code{"2020-11-30"}. Therefore \code{"2020-12-06"} is ignored.
Returning \code{"2020-12-06"} would have changed the month although it is not
missing (in the \code{dtc} date).

\describe{
\item{Permitted values}{a list of dates, e.g. \code{list(ymd_hms("2021-07-01T04:03:01"), ymd_hms("2022-05-12T13:57:23"))}}
\item{Default value}{\code{NULL}}
}}

\item{min_dates_strict}{Minimum dates (strict)

The argument works like the \code{min_dates} argument but it affects the behavior
of the \code{min_dates} and \code{max_dates} arguments. The range which is used to
determine which of the \code{min_dates} and \code{max_dates} are considered is
restricted by the dates specified for \code{min_dates_strict} (see example below
and the example in the "Avoid Imputed Dates Before a Particular Date" section
in \code{vignette("imputation")}).

This argument is useful if the \code{max_dates} argument is used and there are
strict restrictions on the imputed date like the event start date if event
end date is imputed.

For example

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(lubridate)
impute_dtc_dt(
  c("2020-11", "2020-11"),
  min_dates_strict = list(
   c(ymd("2020-11-24"), ymd("2020-11-11"))
  ),
  max_dates = list(
    c(ymd("2020-11-22"), ymd("2020-11-22"))
  ),
  highest_imputation = "M",
  date_imputation = "last"
)
#> [1] "2020-11-30" "2020-11-22"
}\if{html}{\out{</div>}}

returns \code{"2020-11-30"}. The possible dates for \code{"2020-11"} range from
\code{"2020-11-01"} to \code{"2020-11-30"}.

For the first element, the \code{min_dates_strict} argument restricts this range
to \code{"2020-11-24"} to \code{"2020-11-30"}. Therefore \code{"2020-11-22"} (from the
\code{max_dates} argument) is ignored as it is not within the restricted range.

For the second element, the \code{min_dates_strict} argument restricts this range
to \code{"2020-11-11"} to \code{"2020-11-30"}. In this case \code{"2020-11-22"} (from the
\code{max_dates} argument) is within the restricted range and is considered.

\describe{
\item{Permitted values}{a list of dates, e.g. \code{list(ymd_hms("2021-07-01T04:03:01"), ymd_hms("2022-05-12T13:57:23"))}}
\item{Default value}{\code{NULL}}
}}

\item{max_dates}{Maximum dates

A list of dates is expected. It is ensured that the imputed date is not after
any of the specified dates, e.g., that the imputed date is not after the data
cut off date. Only dates which are in the range of possible dates are
considered. A date or date-time object is expected.

\describe{
\item{Permitted values}{a list of dates, e.g. \code{list(ymd_hms("2021-07-01T04:03:01"), ymd_hms("2022-05-12T13:57:23"))}}
\item{Default value}{\code{NULL}}
}}

\item{max_dates_strict}{Maximum dates (strict)

The argument works like the \code{max_dates} argument but it affects the behavior
of the \code{min_dates} and \code{max_dates} arguments. The range which is used to
determine which of the \code{min_dates} and \code{max_dates} are considered is
restricted by the dates specified for \code{max_dates_strict} (see example below).

This argument is useful if the \code{min_dates} argument is used and there are
strict restrictions on the imputed date like date of death or the event end
date if event start date is imputed.

For example

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(lubridate)
impute_dtc_dt(
  c("2020-11", "2020-11"),
  max_dates_strict = list(
   c(ymd("2020-11-24"), ymd("2020-11-11"))
  ),
  min_dates = list(
    c(ymd("2020-11-22"), ymd("2020-11-22"))
  ),
  highest_imputation = "M",
  date_imputation = "first"
)
#> [1] "2020-11-22" "2020-11-01"
}\if{html}{\out{</div>}}

returns (\code{"2020-11-22"}, \code{"2020-11-01"}). The possible dates for \code{"2020-11"}
range from \code{"2020-11-01"} to \code{"2020-11-30"}.

For the first element, the \code{max_dates_strict} argument restricts this range
to \code{"2020-11-01"} to \code{"2020-11-24"}. The date \code{"2020-11-22"} (from the
\code{min_dates} argument) is within the restricted range and is therefore
considered.

For the second element, the \code{max_dates_strict} argument restricts this range
to \code{"2020-11-01"} to \code{"2020-11-11"}. In this case \code{"2020-11-22"} (from the
\code{max_dates} argument) is outside the restricted range and thus it is ignored.

\describe{
\item{Permitted values}{a list of dates, e.g. \code{list(ymd_hms("2021-07-01T04:03:01"), ymd_hms("2022-05-12T13:57:23"))}}
\item{Default value}{\code{NULL}}
}}

\item{preserve}{Preserve day if month is missing and day is present

For example \code{"2019---07"} would return \verb{"2019-06-07} if \code{preserve = TRUE}
(and \code{date_imputation = "MID"}).

\describe{
\item{Permitted values}{\code{TRUE}, \code{FALSE}}
\item{Default value}{\code{FALSE}}
}}
}
\value{
a date object
}
\description{
Convert a date character vector (usually \code{--DTC}) into a Date vector (usually \verb{*DT}).
}
\details{
This is a vector-oriented helper and is not usually called directly on a data
frame with \verb{\%>\%}.
}
\examples{
convert_dtc_to_dt("2019-07-18")
convert_dtc_to_dt("2019-07")
}
\seealso{
Date/Time Computation Functions that returns a vector:
\code{\link[=compute_age_years]{compute_age_years()}},
\code{\link[=compute_dtf]{compute_dtf()}},
\code{\link[=compute_duration]{compute_duration()}},
\code{\link[=compute_tmf]{compute_tmf()}},
\code{\link[=convert_date_to_dtm]{convert_date_to_dtm()}},
\code{\link[=convert_dtc_to_dtm]{convert_dtc_to_dtm()}},
\code{\link[=convert_xxtpt_to_hours]{convert_xxtpt_to_hours()}},
\code{\link[=impute_dtc_dt]{impute_dtc_dt()}},
\code{\link[=impute_dtc_dtm]{impute_dtc_dtm()}}
}
\concept{com_date_time}
\keyword{com_date_time}
