% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{convert_dtm_to_dtc}
\alias{convert_dtm_to_dtc}
\title{Helper Function to Convert Date (or Date-time) Objects to Characters of dtc Format
(-DTC type of variable)}
\usage{
convert_dtm_to_dtc(dtm)
}
\arguments{
\item{dtm}{date or date-time}
}
\value{
character
}
\description{
Helper Function to Convert Date (or Date-time) Objects to Characters of dtc Format
(-DTC type of variable)
}
\seealso{
Developer Utility Functions: 
\code{\link{\%notin\%}()},
\code{\link{arg_name}()},
\code{\link{as_name}()},
\code{\link{extract_vars}()},
\code{\link{replace_values_by_names}()}
}
\author{
Ondrej Slama
}
\concept{dev_utility}
\keyword{dev_utility}
