% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_derivation.R
\name{derivation_slice}
\alias{derivation_slice}
\title{Create a \code{derivation_slice} Object}
\usage{
derivation_slice(filter, args = NULL)
}
\arguments{
\item{filter}{An unquoted condition for defining the observations of the
slice

Comparing derived numeric variables to fixed values, e.g., \code{PCHG <= -90}, may give unexpected results due to floating point representation. For details and solutions see the "Floating Point Comparisons" section in \code{vignette("concepts_conventions")}.

\describe{
\item{Default value}{none}
}}

\item{args}{Arguments of the derivation to be used for the slice

A \code{params()} object is expected.

\describe{
\item{Default value}{\code{NULL}}
}}
}
\value{
An object of class \code{derivation_slice}
}
\description{
Create a \code{derivation_slice} object as input for \code{slice_derivation()}.
}
\seealso{
\code{\link[=slice_derivation]{slice_derivation()}}, \code{\link[=params]{params()}}

Higher Order Functions:
\code{\link[=call_derivation]{call_derivation()}},
\code{\link[=restrict_derivation]{restrict_derivation()}},
\code{\link[=slice_derivation]{slice_derivation()}}
}
\concept{high_order_function}
\keyword{high_order_function}
