% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_aage.R
\name{derive_aage}
\alias{derive_aage}
\title{Title}
\usage{
derive_aage(data)
}
\arguments{
\item{data}{An ADSL-like dataset}
}
\value{
The input datasets plus AAGE and AAGEU variables
}
\description{
Title
}
