% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_aval.R
\name{derive_aval}
\alias{derive_aval}
\title{Derive \code{AVALC}, \code{AVAL} and \code{AVALU}}
\usage{
derive_aval(dataset)
}
\arguments{
\item{dataset}{The input dataset}
}
\description{
Derive \code{AVALC}, \code{AVAL} and \code{AVALU}
}
\details{
The SDTM variables \code{--STRES}, \code{--STRESN} and \code{--STRESU} are mapped to the
ADaM variables \code{AVALC}, \code{AVAL} and \code{AVALU}, respectively.
}
\examples{
library(dplyr)
data(vs)
vs \%>\%
  derive_aval() \%>\%
  select(USUBJID, PARAMCD, VSSTRES, AVALC, VSSTRESN, AVAL, VSSTRESU, AVALU)
}
\author{
Thomas Neitmann
}
