% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_baseline.R
\name{derive_baseline}
\alias{derive_baseline}
\title{Derive Baseline}
\usage{
derive_baseline(dataset, by_vars, source_var, new_var)
}
\arguments{
\item{dataset}{The input dataset}

\item{by_vars}{Grouping variables uniquely identifying a set
of records for which to calculate \code{BASE}}

\item{source_var}{The column from which to extract the baseline value,
e.g. \code{AVAL}}

\item{new_var}{The name of the newly created baseline column, e.g.
\code{BASE}}
}
\value{
A new \code{data.frame} containing all records and variables of the input
dataset plus the \code{new_var} variable.
}
\description{
Derive a baseline variable, e.g. \code{BASE}, in a BDS dataset
}
\examples{
dataset <- tibble::tribble(
  ~STUDYID, ~USUBJID, ~PARAMCD,  ~AVALC,   ~AVISIT,    ~ABLFL,
  "TEST01", "PAT01",  "PARAM01", "LOW",    "Baseline", "Y",
  "TEST01", "PAT01",  "PARAM01", "LOW",    "Day 7",    "N",
  "TEST01", "PAT01",  "PARAM01", "MEDIUM", "Day 14",   "N",
  "TEST01", "PAT01",  "PARAM02", "HIGH",   "Baseline", "Y",
  "TEST01", "PAT01",  "PARAM02", "HIGH",   "Day 7",    "N",
  "TEST01", "PAT01",  "PARAM02", "MEDIUM", "Day 14",   "N",
  "TEST01", "PAT02",  "PARAM01", "MEDIUM", "Baseline", "Y",
  "TEST01", "PAT02",  "PARAM01", "LOW",    "Day 7",    "N",
  "TEST01", "PAT02",  "PARAM01", "MEDIUM", "Day 14",   "N",
)

derive_baseline(
  dataset,
  by_vars = vars(USUBJID, PARAMCD),
  source_var = AVALC,
  new_var = BASEC
)
}
\author{
Thomas Neitmann
}
\keyword{bds}
\keyword{derivation}
