% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_derived_param.R
\name{derive_derived_param}
\alias{derive_derived_param}
\title{Adds a Parameter Computed from the Analysis Value of Other Parameters}
\usage{
derive_derived_param(
  dataset,
  filter = NULL,
  parameters,
  by_vars,
  constant_parameters = NULL,
  constant_by_vars = NULL,
  analysis_value,
  set_values_to
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{by_vars} parameter, \code{PARAMCD}, and \code{AVAL}
are expected.

The variable specified by \code{by_vars} and \code{PARAMCD} must be a unique key of
the input dataset after restricting it by the filter condition (\code{filter}
parameter) and to the parameters specified by \code{parameters}.}

\item{filter}{Filter condition

The specified condition is applied to the input dataset before deriving the
new parameter, i.e., only observations fulfilling the condition are taken
into account.

\emph{Permitted Values:} a condition}

\item{parameters}{Required parameter codes

It is expected that all parameter codes (\code{PARAMCD}) which are required to
derive the new parameter are specified.

\emph{Permitted Values:} A character vector of \code{PARAMCD} values}

\item{by_vars}{Grouping variables

For each group defined by \code{by_vars} an observation is added to the output
dataset.

\emph{Permitted Values:} list of variables}

\item{analysis_value}{Definition of the analysis value

An expression defining the analysis value (\code{AVAL}) of the new parameter is
expected. The analysis values of the parameters specified by \code{parameters}
can be accessed using \verb{AVAL.<parameter code>}, e.g., \code{AVAL.SYSBP}.

\emph{Permitted Values:} An unquoted expression}

\item{set_values_to}{Variables to be set

The specified variables are set to the specified values for the new
observations. For example \code{vars(PARAMCD = "MAP")} defines the parameter
code for the new parameter.

\emph{Permitted Values:} List of variable-value pairs}
}
\value{
The input dataset with the new parameter added
}
\description{
Adds a parameter computed from the analysis value of other parameters. It is
expected that the analysis value of the new parameter is defined by an
expression using the analysis values of other parameters. For example mean
arterial pressure (MAP) can be derived from systolic (SYSBP) and diastolic
blood pressure (DIABP) with the formula \deqn{MAP = (SYSBP + 2 DIABP) / 3}
}
\details{
For each group (with respect to the variables specified for the
\code{by_vars} parameter) an observation is added to the output dataset if the
filtered input dataset contains exactly one observation for each parameter
code specified for \code{parameters}.

For the new observations \code{AVAL} is set to the value specified by
\code{analysis_value} and the variables specified for \code{set_values_to} are set to
the provided values. The values of the other variables of the input dataset
are retained if they are constant within each by group. Otherwise they are
set to \code{NA}.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
data("advs")
advs \%>\%
  select(USUBJID, PARAMCD, PARAM, AVAL, VSTESTCD, ANL01FL, DTYPE, AVISIT, AVISITN) \%>\%
  filter(
    USUBJID \%in\% c("01-701-1015", "01-701-1363") &
      PARAMCD \%in\% c("SYSBP", "DIABP") &
      AVISIT \%in\% c("Baseline", "Week 2")
  ) \%>\%
  derive_derived_param(
    advss,
    filter = ANL01FL == "Y" & DTYPE == "AVERAGE",
    parameters = c("SYSBP", "DIABP"),
    by_vars = vars(USUBJID, AVISIT),
    analysis_value = (AVAL.SYSBP + 2 * AVAL.DIABP) / 3,
    set_values_to = vars(
      PARAMCD = "MAP",
      PARAM = "Mean arterial pressure (mmHg)",
      DTYPE = NA
    )
  ) \%>\%
  filter(DTYPE == "AVERAGE" | PARAMCD == "MAP")
}
\author{
Stefan Bundfuss
}
\keyword{bds}
\keyword{derivation}
