% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_disposition_reason.R
\name{derive_disposition_reason}
\alias{derive_disposition_reason}
\title{Derive a disposition reason at a specific timepoint}
\usage{
derive_disposition_reason(dataset, dataset_ds, new_var, reason_var,
  new_var_spe = NULL, reason_var_spe = NULL,
  format_new_var = format_reason_default, filter_ds)
}
\arguments{
\item{dataset}{Input dataset.}

\item{dataset_ds}{Dataset containing the disposition information (e.g.: \code{ds}).

The variable specified in the \code{reason_var} parameter must be in \code{dataset_ds}.}

\item{new_var}{Name of the disposition reason variable.

a variable name is expected (e.g. \code{DCSREAS}).}

\item{reason_var}{The variable used to derive the disposition reason

A character vector is expected (e.g. \code{DSDECOD}).}

\item{new_var_spe}{Name of the disposition reason detail variable.

a variable name is expected (e.g. \code{DCSREASP}).}

\item{reason_var_spe}{The variable used to derive the disposition reason detail

A character vector is expected (e.g. \code{DSTERM}).}

\item{format_new_var}{The format used to derive the reason(s)

Default: format_reason_default defined as:
format_reason_default<-function(x, y=NULL){
case_when (
x == "COMPLETED" ~ x,
TRUE ~ NA_character
)
}
where x is the reason_var.}

\item{filter_ds}{Filter condition for the disposition data.

Filter used to select the relevant disposition data.

Permitted Values: logical expression.}
}
\value{
the input dataset with the disposition reason (\code{new_var}) added.
}
\description{
Derive a disposition reason from the the relevant records in the disposition domain.
}
\details{
This functions returns the main reason for discontinuation (e.g. \code{DCSREAS} or \code{DCTREAS}).
The reason for discontinuation is derived based on \code{reason_var} (e.g. \code{DSDECOD}) and
\code{format_new_var}.
If \code{new_var_spe} is not NULL, then the function will also returns the details associated
with the reason for discontinuation (e.g. \code{DCSREASP}).
The details associated with the reason for discontinuation are derived based on
\code{reason_var_spe} (e.g. \code{DSTERM}), \code{reason_var} and \code{format_new_var}. see
}
\examples{
library(dplyr)
data("dm")
data("ds")

# Derive DCSREAS using the default format
derive_disposition_reason(
  dataset = dm,
  dataset_ds = ds,
  new_var = DCSREAS,
  reason_var = DSDECOD,
  filter_ds = DSCAT == "DISPOSITION EVENT"
)
# Derive DCSREAS and DCSREASP using a study-specific format
format_dcsreas <- function(x, y = NULL) {
  out <- if (is.null(y)) x else y
  case_when(
    x \%!in\% c("COMPLETED", "SCREEN FAILURE") & !is.na(x) ~ out,
    TRUE ~ NA_character_
  )
}
derive_disposition_reason(
  dataset = dm,
  dataset_ds = ds,
  new_var = DCSREAS,
  reason_var = DSDECOD,
  new_var_spe = DCSREASP,
  reason_var_spe = DSTERM,
  format_new_var = format_dcsreas,
  filter_ds = DSCAT == "DISPOSITION EVENT"
)
}
\seealso{
\code{\link[=format_reason_default]{format_reason_default()}}
}
\author{
Samia Kabi
}
\keyword{adsl}
