% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_disposition_status.R
\name{derive_disposition_status}
\alias{derive_disposition_status}
\title{Derive a disposition status at a specific timepoint}
\usage{
derive_disposition_status(dataset, dataset_ds, new_var, status_var,
  format_new_var = format_eoxxstt_default, filter_ds)
}
\arguments{
\item{dataset}{Input dataset.}

\item{dataset_ds}{Dataset containing the disposition information (e.g.: ds).

The variable specified in the status_var parameter must be in dataset_ds.}

\item{new_var}{Name of the disposition date variable.

a variable name is expected (e.g. EOSSTT).}

\item{status_var}{The variable used to derive the disposition status.

A character vector is expected (e.g. DSDECOD).}

\item{format_new_var}{The format used to derive the status.

Default: format_eoxxstt_default defined as:
format_eoxxstt_default<-function(x){
case_when (
x %in% c("COMPLETED")~"COMPLETED",
!(x %in% c("COMPLETED")) & ! is.na(x)~"DISCONTINUED",
TRUE ~ "ONGOING"
)
}
where x is the status_var.}

\item{filter_ds}{Filter condition for the disposition data.

Filter used to select the relevant disposition data.

Permitted Values: logical expression.}
}
\value{
the input dataset with the disposition status (new_var) added.
}
\description{
Derive a disposition status from the the relevant records in the disposition domain.
}
\examples{
library(dplyr)
data("dm")
data("ds")

# Default derivation: EOSSTT =
#- COMPLETED when status_var = COMPLETED
#- DISCONTINUED when status_var != COMPLETED
#- ONGOING otherwise

derive_disposition_status(
  dataset = dm,
  dataset_ds = ds,
  new_var = EOSSTT,
  status_var = DSDECOD,
  filter_ds = DSCAT == "DISPOSITION EVENT"
)

# Specific derivation: EOSSTT =
#- COMPLETED when status_var = COMPLETED
#- DISCONTINUED DUE TO AE when status_var = ADVERSE EVENT
#- DISCONTINUED NOT DUE TO AE when status_var != ADVERSE EVENT nor COMPLETED nor missing
#- ONGOING otherwise

format_eoxxstt1 <- function(x) {
  case_when(
    x == "COMPLETED" ~ "COMPLETED",
    x == "ADVERSE EVENT" ~ "DISCONTINUED DUE TO AE",
    x \%!in\% c("ADVERSE EVENT", "COMPLETED") & !is.na(x) ~ "DISCONTINUED NOT DUE TO AE",
    TRUE ~ "ONGOING"
  )
}

derive_disposition_status(
  dataset = dm,
  dataset_ds = ds,
  new_var = EOSSTT,
  status_var = DSDECOD,
  format_new_var = format_eoxxstt1,
  filter_ds = DSCAT == "DISPOSITION EVENT"
)
}
\author{
Samia Kabi
}
\keyword{adsl}
