% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_dthcaus.R
\name{derive_dthcaus}
\alias{derive_dthcaus}
\title{Derive Death Cause and Domain.}
\usage{
derive_dthcaus(dataset, dataset_ae, dataset_ds, filter_ae = exprs(AEOUT
  == "FATAL"), filter_ds = exprs(DSDECOD == "DEATH",
  grepl("DEATH DUE TO", DSTERM)), dthcaus_ae = expr(AEDECOD),
  dthcaus_ds = expr(DSTERM))
}
\arguments{
\item{dataset}{Input dataset}

\item{dataset_ae}{Source AE dataset}

\item{dataset_ds}{Source DS dataset}

\item{filter_ae}{An expression to filter \code{dataset_ae}.}

\item{filter_ds}{An expression to filter \code{dataset_ds}.}

\item{dthcaus_ae}{Variable name (as symbol) or value (as character) to
be set as \code{DTHCAUS} from AE.}

\item{dthcaus_ds}{Variable name (as symbol) in DS or value (as character) to
be set as \code{DTHCAUS} from DS.}
}
\value{
The input dataset with \code{DTHCAT} added.
}
\description{
Derives death cause (\code{DTHCAUS}) and domain (\code{DTHDOM}).
}
\examples{
...

}
\author{
Shimeng Huang
}
\keyword{Roche}
