% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_duration.R
\name{derive_duration}
\alias{derive_duration}
\title{Derive Duration}
\usage{
derive_duration(
  dataset,
  new_var,
  new_var_unit = NULL,
  start_date,
  end_date,
  in_unit = "days",
  out_unit = "days",
  floor_in = TRUE,
  add_one = TRUE,
  trunc_out = FALSE
)
}
\description{
\code{derive_duration()} was renamed to \code{derive_vars_duration()} to create a
more consistent API.
}
\keyword{internal}
