% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_joined_vars.R
\name{derive_joined_vars}
\alias{derive_joined_vars}
\title{Derive variables that are based on data from a joined dataset}
\usage{
derive_joined_vars(dataset, dataset_join, by_vars = exprs(STUDYID,
  USUBJID), new_vars, order = NULL, pre_filter = NULL,
  post_filter = NULL)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{by_vars} parameter are expected.}

\item{dataset_join}{Dataset to join

The variables specified by the \code{by_vars} and the \code{pre_filter} are expected.}

\item{by_vars}{Grouping variables

Default: \code{exprs(STUDYID, USUBJID)}

Permitted Values: list of variables}

\item{new_vars}{Variables to derive

Expressions specifying the variables definitions. See Examples.}

\item{order}{Sort order

Within each by group the observations are ordered by the specified order.

Permitted Values: list of variables or functions of variables}

\item{pre_filter}{Filter condition before joining the join dataset

Only observations of the joining dataset which fulfill the specified condition(-s)
are used for joining.

Permitted Values: logical expression}

\item{post_filter}{Filter condition for resulting dataset

Only observations of the joined dataset which fulfill the specified condition(-s)
are returned.

Permitted Values: logical expression}
}
\value{
The input dataset with additional derived variables
}
\description{
Derives new variables which depend on either input dataset or joined dataset or both.
Additionally, pre-/post- filters can be supplied to filder the joined/resulting dataset.
}
\examples{
library(magrittr)

data(ae)
data(ex)
ae \%>\%
  derive_vars_dtm(new_vars_prefix = "AST", dtc = AESTDTC) \%>\%
  derive_joined_vars(
    dataset_join = ex,
    new_vars = exprs(LDOSEDTM = compute_dtm(dtc = EXENDTC,
                                            date_imputation = NULL,
                                            time_imputation = NULL)),
    pre_filter = exprs(is_valid_dose()),
    by_vars = exprs(USUBJID),
    order = exprs(EXENDTC),
    post_filter = exprs(ASTDTM <= c(dtc = EXENDTC))
  )

data(time_def)
ae \%>\%
  derive_vars_dtm(new_vars_prefix = "AST", dtc = AESTDTC) \%>\%
  derive_joined_vars(
    dataset_join = time_def,
    new_vars = exprs(APERIOD, APERIODC, TRTP, TRTA),
    by_vars = exprs(STUDYID, USUBJID),
    post_filter = exprs(STARTDTM <= ASTDTM && ASTDTM <= ENDDTM)
  )
}
\author{
Ondrej Slama
}
\keyword{adam}
\keyword{derivation}
