% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_last_dose_amt.R
\name{derive_last_dose_amt}
\alias{derive_last_dose_amt}
\title{Derive Last Dose Amount}
\usage{
derive_last_dose_amt(
  dataset,
  dataset_ex,
  filter_ex = NULL,
  by_vars = vars(STUDYID, USUBJID),
  dose_start,
  dose_end,
  analysis_date,
  dataset_seq_var,
  new_var,
  dose_var = EXDOSE,
  check_dates_only = FALSE,
  traceability_vars = NULL
)
}
\arguments{
\item{new_var}{The output variable.}

\item{dose_var}{The source dose amount variable. Defaults to \code{EXDOSE}.}
}
\value{
Input dataset with additional column \code{new_var}.
}
\description{
Derive Last Dose Amount
}
\details{
The last dose amount is derived as the dose amount where the \code{end_date} is lower to or
equal to the \code{analysis_date} per \code{by_vars} and \code{dataset_seq_var}.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
data(ae)
data(ex_single)

ae \%>\%
  head(100) \%>\%
  derive_last_dose_amt(
    head(ex_single, 100),
    filter_ex = (EXDOSE > 0 | (EXDOSE == 0 & grepl("PLACEBO", EXTRT))) &
      nchar(EXENDTC) >= 10,
    dose_start = EXSTDTC,
    dose_end = EXENDTC,
    analysis_date = AESTDTC,
    dataset_seq_var = AESEQ,
    new_var = LDOSE,
    dose_var = EXDOSE,
    check_dates_only = FALSE
  ) \%>\%
  select(STUDYID, USUBJID, AESEQ, AESTDTC, LDOSE)

# or with traceability variables
ae \%>\%
  head(100) \%>\%
  derive_last_dose_amt(
    head(ex_single, 100),
    filter_ex = (EXDOSE > 0 | (EXDOSE == 0 & grepl("PLACEBO", EXTRT))) &
      nchar(EXENDTC) >= 10,
    dose_start = EXSTDTC,
    dose_end = EXENDTC,
    analysis_date = AESTDTC,
    dataset_seq_var = AESEQ,
    new_var = LDOSE,
    dose_var = EXDOSE,
    check_dates_only = FALSE,
    traceability_vars = dplyr::vars(LDOSEDOM = "EX", LDOSESEQ = EXSEQ, LDOSEVAR = "EXDOSE")
  ) \%>\%
  select(STUDYID, USUBJID, AESEQ, AESTDTC, LDOSEDOM, LDOSESEQ, LDOSEVAR, LDOSE)
}
\seealso{
\code{\link[=get_last_dose]{get_last_dose()}}
}
\author{
Annie Yang
}
\keyword{adae}
\keyword{derivation}
