% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param.R
\name{derive_param}
\alias{derive_param}
\title{Derive PARAM and PARAMCD}
\usage{
derive_param(dataset, mapping, by_vars)
}
\arguments{
\item{dataset}{The input dataset}

\item{mapping}{A dataset containing the mapping of one or more SDTM variables
to \code{PARAM} and \code{PARAMCD}}

\item{by_vars}{The variables used to join \code{dataset} and \code{mapping}}
}
\description{
Derive PARAM and PARAMCD
}
\examples{
data(vs)
data(advs_param)

derive_param(vs, advs_param, by_vars = c("VSTESTCD", "VSPOS"))

## Disregard the `VSPOS` variable for mapping
derive_param(vs, advs_param, by_vars = c("VSTESTCD"))
}
\author{
Thomas Neitmann
}
