% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_first_event.R
\name{derive_param_first_event}
\alias{derive_param_first_event}
\title{Add a First Event Parameter}
\usage{
derive_param_first_event(
  dataset,
  dataset_adsl,
  source_param,
  condition,
  subject_keys = vars(STUDYID, USUBJID),
  set_values_to,
  check_type = "warning"
)
}
\arguments{
\item{dataset}{Input dataset

The variables \code{PARAMCD}, \code{ADT}, and those specified by the \code{subject_keys}
parameter are expected.}

\item{dataset_adsl}{ADSL input dataset

The variables specified for \code{subject_keys} are expected. For each
observation of the specified dataset a new observation is added to the
input dataset.}

\item{source_param}{Source parameter

Only the observations from the input dataset where \code{PARAMCD} equals the
specified value are considered for looking for an event.}

\item{condition}{Event condition

For subjects with at least one observation where the condition is fulfilled
\code{AVALC} is set to \code{"Y"}, \code{AVAL} to \code{1}, and \code{ADT} to the first date where
the condition is fulfilled.

For all other subjects \code{AVALC} is set to \code{"N"}, \code{AVAL} to \code{0}, and \code{ADT} to
\code{NA}.}

\item{subject_keys}{Variables to uniquely identify a subject

A list of symbols created using \code{vars()} is expected.}

\item{set_values_to}{Variables to set

A named list returned by \code{vars()} defining the variables to be set for the
new parameter, e.g. \code{vars(PARAMCD = "PD", PARAM = "Disease Progression")}
is expected. The values must be symbols, character strings, numeric values,
or \code{NA}.}

\item{check_type}{Check uniqueness?

If \code{"warning"} or \code{"error"} is specified, a message is issued if the
observations of the input dataset restricted to the source parameter
(\code{source_param}) are not unique with respect to the subject keys
(\code{subject_key} parameter) and \code{ADT}.

\emph{Default}: \code{"warning"}

\emph{Permitted Values}: \code{"none"}, \code{"warning"}, \code{"error"}}
}
\value{
The input dataset with a new parameter indicating if and when an
event occurred
}
\description{
Add a new parameter for the first event occurring within a parameter of the
input dataset. \code{AVALC} and \code{AVAL} indicate if an event occurred and \code{ADT} is
set to the date of the first event. For example, the function can derive a
parameter for the first disease progression.
}
\details{
\enumerate{
\item The input dataset is restricted to observations where \code{PARAMCD} equals
\code{source_param}.
\item For each subject (with respect to the variables specified for the
\code{subject_keys} parameter) the first observation (with respect \code{ADT}) where
the event condition (\code{condition} parameter) is fulfilled is selected.
\item For each observation in \code{dataset_adsl} a new observation is created. For
subjects with event \code{AVALC} is set to \code{"Y"}, \code{AVAL} to \code{1}, and \code{ADT} to
the first date where the event condition is fulfilled. For all other
subjects \code{AVALC} is set to \code{"N"}, \code{AVAL} to \code{0}, and \code{ADT} to \code{NA}.
\item The variables specified by the \code{set_values_to} parameter are added to
the new observations.
\item The new observations are added to input dataset.
}
}
\examples{
# Derive a new parameter for the first disease progression (PD)
adsl <- tibble::tribble(
  ~USUBJID,
  "1",
  "2",
  "3"
) \%>\%
  mutate(STUDYID = "XX1234")

adrs <- tibble::tribble(
  ~USUBJID, ~ADTC,        ~AVALC,
  "1",      "2020-01-02", "PR",
  "1",      "2020-02-01", "CR",
  "1",      "2020-03-01", "CR",
  "1",      "2020-04-01", "SD",
  "2",      "2021-06-15", "SD",
  "2",      "2021-07-16", "PD",
  "2",      "2021-09-14", "PD"
) \%>\%
  mutate(
    STUDYID = "XX1234",
    ADT = ymd(ADTC),
    PARAMCD = "OVR",
    PARAM = "Overall Response",
    ANL01FL = "Y"
  ) \%>\%
  select(-ADTC)

derive_param_first_event(
  adrs,
  dataset_adsl = adsl,
  source_param = "OVR",
  condition = AVALC == "PD",
  set_values_to = vars(
    PARAMCD = "PD",
    PARAM = "Disease Progression",
    ANL01FL = "Y"
  )
)
}
\author{
Stefan Bundfuss
}
\keyword{bds}
\keyword{derivation}
