% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_map.R
\name{derive_param_map}
\alias{derive_param_map}
\title{Adds a Parameter for Mean Arterial Pressure}
\usage{
derive_param_map(
  dataset,
  by_vars,
  set_values_to = exprs(PARAMCD = "MAP"),
  sysbp_code = "SYSBP",
  diabp_code = "DIABP",
  hr_code = NULL,
  get_unit_expr,
  filter = NULL
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{by_vars} argument are expected to be in the dataset.
\code{PARAMCD}, and \code{AVAL} are expected as well.

The variable specified by \code{by_vars} and \code{PARAMCD} must be a unique key of
the input dataset after restricting it by the filter condition (\code{filter}
parameter) and to the parameters specified by \code{sysbp_code}, \code{diabp_code}
and \code{hr_code}.

\describe{
\item{Default value}{none}
}}

\item{by_vars}{Grouping variables

For each group defined by \code{by_vars} an observation is added to the output
dataset. Only variables specified in \code{by_vars} will be populated
in the newly created records.

\describe{
\item{Permitted values}{list of variables created by \code{exprs()}, e.g., \code{exprs(USUBJID, VISIT)}}
\item{Default value}{none}
}}

\item{set_values_to}{Variables to be set

The specified variables are set to the specified values for the new
observations. For example \code{exprs(PARAMCD = "MAP")} defines the parameter code
for the new parameter.

\describe{
\item{Permitted values}{List of variable-value pairs}
\item{Default value}{\code{exprs(PARAMCD = "MAP")}}
}}

\item{sysbp_code}{Systolic blood pressure parameter code

The observations where \code{PARAMCD} equals the specified value are considered
as the systolic blood pressure assessments.

\describe{
\item{Permitted values}{character value}
\item{Default value}{\code{"SYSBP"}}
}}

\item{diabp_code}{Diastolic blood pressure parameter code

The observations where \code{PARAMCD} equals the specified value are considered
as the diastolic blood pressure assessments.

\describe{
\item{Permitted values}{character value}
\item{Default value}{\code{"DIABP"}}
}}

\item{hr_code}{Heart rate parameter code

The observations where \code{PARAMCD} equals the specified value are considered
as the heart rate assessments.

\describe{
\item{Permitted values}{character value}
\item{Default value}{\code{NULL}}
}}

\item{get_unit_expr}{An expression providing the unit of the parameter

The result is used to check the units of the input parameters.

\describe{
\item{Permitted values}{An expression which is evaluable in the input dataset
and results in a character value}
\item{Default value}{none}
}}

\item{filter}{Filter condition

The specified condition is applied to the input dataset before deriving the
new parameter, i.e., only observations fulfilling the condition are taken
into account.

Comparing derived numeric variables to fixed values, e.g., \code{PCHG <= -90}, may give unexpected results due to floating point representation. For details and solutions see the "Floating Point Comparisons" section in \code{vignette("concepts_conventions")}.

\describe{
\item{Permitted values}{an unquoted condition, e.g., \code{AVISIT == "BASELINE"}}
\item{Default value}{\code{NULL}}
}}
}
\value{
The input dataset with the new parameter added. Note, a variable will only
be populated in the new parameter rows if it is specified in \code{by_vars}.
}
\description{
Adds a record for mean arterial pressure (MAP) for each by group
(e.g., subject and visit) where the source parameters are available.

\strong{Note:} This is a wrapper function for the more generic \code{derive_param_computed()}.
}
\details{
The analysis value of the new parameter is derived as
\deqn{\frac{2DIABP + SYSBP}{3}}{(2DIABP + SYSBP) / 3}
if it is based on diastolic and systolic blood pressure and
\deqn{DIABP + 0.01 e^{4.14 - \frac{40.74}{HR}} (SYSBP - DIABP)}{
DIABP + 0.01 exp(4.14 - 40.74 / HR) (SYSBP - DIABP)}
if it is based on diastolic, systolic blood pressure, and heart rate.
}
\examples{
library(tibble)
library(dplyr, warn.conflicts = FALSE)

advs <- tibble::tribble(
  ~USUBJID, ~PARAMCD, ~PARAM, ~AVAL, ~VISIT,
  "01-701-1015", "PULSE", "Pulse (beats/min)", 59, "BASELINE",
  "01-701-1015", "PULSE", "Pulse (beats/min)", 61, "WEEK 2",
  "01-701-1015", "DIABP", "Diastolic Blood Pressure (mmHg)", 51, "BASELINE",
  "01-701-1015", "DIABP", "Diastolic Blood Pressure (mmHg)", 50, "WEEK 2",
  "01-701-1015", "SYSBP", "Systolic Blood Pressure (mmHg)", 121, "BASELINE",
  "01-701-1015", "SYSBP", "Systolic Blood Pressure (mmHg)", 121, "WEEK 2",
  "01-701-1028", "PULSE", "Pulse (beats/min)", 62, "BASELINE",
  "01-701-1028", "PULSE", "Pulse (beats/min)", 77, "WEEK 2",
  "01-701-1028", "DIABP", "Diastolic Blood Pressure (mmHg)", 79, "BASELINE",
  "01-701-1028", "DIABP", "Diastolic Blood Pressure (mmHg)", 80, "WEEK 2",
  "01-701-1028", "SYSBP", "Systolic Blood Pressure (mmHg)", 130, "BASELINE",
  "01-701-1028", "SYSBP", "Systolic Blood Pressure (mmHg)", 132, "WEEK 2"
)

# Derive MAP based on diastolic and systolic blood pressure
advs \%>\%
  derive_param_map(
    by_vars = exprs(USUBJID, VISIT),
    set_values_to = exprs(
      PARAMCD = "MAP",
      PARAM = "Mean Arterial Pressure (mmHg)"
    ),
    get_unit_expr = extract_unit(PARAM)
  ) \%>\%
  filter(PARAMCD != "PULSE")

# Derive MAP based on diastolic and systolic blood pressure and heart rate
derive_param_map(
  advs,
  by_vars = exprs(USUBJID, VISIT),
  hr_code = "PULSE",
  set_values_to = exprs(
    PARAMCD = "MAP",
    PARAM = "Mean Arterial Pressure (mmHg)"
  ),
  get_unit_expr = extract_unit(PARAM)
)
}
\seealso{
\code{\link[=compute_map]{compute_map()}}

BDS-Findings Functions for adding Parameters/Records:
\code{\link[=default_qtc_paramcd]{default_qtc_paramcd()}},
\code{\link[=derive_basetype_records]{derive_basetype_records()}},
\code{\link[=derive_expected_records]{derive_expected_records()}},
\code{\link[=derive_extreme_event]{derive_extreme_event()}},
\code{\link[=derive_extreme_records]{derive_extreme_records()}},
\code{\link[=derive_locf_records]{derive_locf_records()}},
\code{\link[=derive_param_bmi]{derive_param_bmi()}},
\code{\link[=derive_param_bsa]{derive_param_bsa()}},
\code{\link[=derive_param_computed]{derive_param_computed()}},
\code{\link[=derive_param_doseint]{derive_param_doseint()}},
\code{\link[=derive_param_exist_flag]{derive_param_exist_flag()}},
\code{\link[=derive_param_exposure]{derive_param_exposure()}},
\code{\link[=derive_param_framingham]{derive_param_framingham()}},
\code{\link[=derive_param_qtc]{derive_param_qtc()}},
\code{\link[=derive_param_rr]{derive_param_rr()}},
\code{\link[=derive_param_wbc_abs]{derive_param_wbc_abs()}},
\code{\link[=derive_summary_records]{derive_summary_records()}}
}
\concept{der_prm_bds_findings}
\keyword{der_prm_bds_findings}
