% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_merged_exist_flag.R
\name{derive_param_merged_exist_flag}
\alias{derive_param_merged_exist_flag}
\title{Add an Existence Flag Parameter}
\usage{
derive_param_merged_exist_flag(
  dataset = NULL,
  dataset_adsl,
  dataset_add,
  condition,
  true_value = "Y",
  false_value = NA_character_,
  missing_value = NA_character_,
  filter_add = NULL,
  aval_fun = yn_to_numeric,
  subject_keys = vars(STUDYID, USUBJID),
  set_values_to
)
}
\arguments{
\item{dataset}{Input dataset}

\item{dataset_adsl}{ADSL input dataset

The variables specified for \code{subject_keys} are expected. For each
observation of the specified dataset a new observation is added to the
input dataset.}

\item{dataset_add}{Additional dataset

The variables specified by the \code{subject_keys} parameter are expected.}

\item{condition}{Event condition

For subjects with at least one observation in \code{dataset_add} where the condition is fulfilled
\code{AVALC} is set to \code{true_value} and \code{AVAL} to \code{1}.

For subjects with observations in \code{dataset_add} but none of them is
fulfilling the condition \code{AVALC} is set to \code{false_value} and \code{AVAL} to \code{0}.

For all other subjects, i.e., those without an observation in \code{dataset_add}
\code{AVALC} is set to \code{missing_value}, \code{AVAL} to \code{0}, and \code{ADT} to \code{NA}.}

\item{true_value}{True value

For the new observations \code{AVALC} is set to the specified value for all
subjects with at least one observations in the additional dataset
(\code{dataset_add}) fulfilling the event condition (\code{condition}).

\emph{Default}: \code{"Y"}

\emph{Permitted Value}: A character scalar}

\item{false_value}{False value

For the new observations \code{AVALC} is set to the specified value for all
subjects with  observations in the additional dataset (\code{dataset_add})
but none of them is fulfilling the event condition (\code{condition}).

\emph{Default}: \code{NA_character_}

\emph{Permitted Value}: A character scalar}

\item{missing_value}{Values used for missing information

For the new observations \code{AVALC} is set to the specified value for all
subjects without observations in the additional dataset (\code{dataset_add}).

\emph{Default}: \code{NA_character_}

\emph{Permitted Value}: A character scalar}

\item{filter_add}{Filter for additional data

Only observations fulfilling the specified condition are taken into account
for flagging. If the parameter is not specified, all observations are
considered.

\emph{Permitted Values}: a condition}

\item{aval_fun}{Function to map \code{AVALC} values to \code{AVAL}

The (first) argument of the function must expect a character vector and the
function must return a numeric vector.

\emph{Default:} \code{yn_to_numeric} (see \code{yn_to_numeric()} for details)}

\item{subject_keys}{Variables to uniquely identify a subject

A list of symbols created using \code{vars()} is expected.}

\item{set_values_to}{Variables to set

A named list returned by \code{vars()} defining the variables to be set for the
new parameter, e.g. \code{vars(PARAMCD = "PD", PARAM = "Disease Progression")}
is expected. The values must be symbols, character strings, numeric values,
or \code{NA}.}
}
\value{
The input dataset with a new parameter indicating if and when an
event occurred
}
\description{
Add a new parameter indicating that a certain event exists in a dataset.
\code{AVALC} and \code{AVAL} indicate if an event occurred. For example, the function
can derive a parameter indicating if there is measureable disease at
baseline.
}
\details{
\enumerate{
\item The additional dataset (\code{dataset_add}) is restricted to the observations
matching the \code{filter_add} condition.
\item For each subject in \code{dataset_adsl} a new observation is created. The
\code{AVALC} variable is added and set to the true value (\code{true_value}) if for
the subject at least one observation exists in the (restricted) additional
dataset where the condition evaluates to \code{TRUE}. It is set to the false
value (\code{false_value}) if for the subject at least one observation exists
and for all observations the condition evaluates to \code{FALSE} or \code{NA}.
Otherwise, it is set to the missing value (\code{missing_value}), i.e., for
those subject not in \code{dataset_add}.
\item The \code{AVAL} variable is added and set to \code{aval_fun(AVALC)}.
\item The variables specified by the \code{set_values_to} parameter are added to
the new observations.
\item The new observations are added to input dataset.
}
}
\examples{
library(dplyr)
library(lubridate)

# Derive a new parameter for the first disease progression (PD)
adsl <- tibble::tribble(
  ~USUBJID,
  "1",
  "2",
  "3"
) \%>\%
  mutate(STUDYID = "XX1234")

tu <- tibble::tribble(
  ~USUBJID, ~VISIT,      ~TUSTRESC,
  "1",      "SCREENING", "TARGET",
  "1",      "WEEK 1",    "TARGET",
  "1",      "WEEK 5",    "TARGET",
  "1",      "WEEK 9",    "NON-TARGET",
  "2",      "SCREENING", "NON-TARGET",
  "2",      "SCREENING", "NON-TARGET"
) \%>\%
  mutate(
    STUDYID = "XX1234",
    TUTESTCD = "TUMIDENT"
  )

derive_param_merged_exist_flag(
  dataset_adsl = adsl,
  dataset_add = tu,
  filter_add = TUTESTCD == "TUMIDENT" & VISIT == "SCREENING",
  condition = TUSTRESC == "TARGET",
  false_value = "N",
  missing_value = "N",
  set_values_to = vars(
    PARAMCD = "MDIS",
    PARAM = "Measurable Disease at Baseline"
  )
)
}
\author{
Stefan Bundfuss
}
\keyword{bds}
\keyword{derivation}
