% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_adeg_params.R
\name{derive_param_qtlc}
\alias{derive_param_qtlc}
\title{Adds a parameter for corrected QT using Sagie's formula}
\usage{
derive_param_qtlc(
  dataset,
  filter = NULL,
  new_param = "QTLCR",
  qt_code = "QT",
  rr_code = "RR",
  by_vars,
  drop_values_from = vars(ends_with("RESU"))
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{by_vars} parameter are expected.}

\item{filter}{Filter condition

The specified condition is applied to the input dataset before deriving the
new parameter, i.e., only observations fulfilling the condition are taken
into account.

\emph{Permitted Values:} a condition}

\item{new_param}{Parameter code to add

For the new observations \code{PARAMCD} is set to the specified value.

Permitted Values: character value}

\item{qt_code}{QT parameter code

The observations where \code{PARAMCD} equals the specified value are considered
as the QT interval assessments.

Permitted Values: character value}

\item{rr_code}{RR parameter code

The observations where \code{PARAMCD} equals the specified value are considered
as the RR interval assessments.

Permitted Values: character value}

\item{by_vars}{Grouping variables

Permitted Values: list of variables}
}
\value{
The input dataset with the new flag variable added
}
\description{
Adds a parameter for corrected QT using Sagie's formula
}
\author{
Stefan Bundfuss
}
\keyword{adeg}
\keyword{derivation}
