% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_bds_params.R
\name{derive_param_ratio}
\alias{derive_param_ratio}
\title{Adds a Ratio Parameter Computed from the Analysis Value of Other Parameters}
\usage{
derive_param_ratio(
  dataset,
  by_vars,
  dividend_code,
  divisor_code,
  set_values_to,
  constant_dividend = FALSE,
  constant_divisor = FALSE,
  filter = NULL,
  constant_by_vars = NULL
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{by_vars} argument are expected to be in the dataset.
\code{PARAMCD}, and \code{AVAL} are expected as well.

The variable specified by \code{by_vars} and \code{PARAMCD} must be a unique key of
the input dataset after restricting it by the filter condition (\code{filter}
parameter) and to the parameters specified by \code{dividend_code} and \code{divisor_code}.}

\item{by_vars}{Grouping variables

For each group defined by \code{by_vars} an observation is added to the output
dataset. Only variables specified in \code{by_vars} will be populated
in the newly created records.

\emph{Permitted Values}: list of variables created by \code{exprs()}
e.g. \code{exprs(USUBJID, VISIT)}}

\item{dividend_code}{Dividend parameter code

The observations where \code{PARAMCD} equals the specified value are considered
as the dividend

\emph{Permitted Values:} character value}

\item{divisor_code}{Divisor parameter code

The observations where \code{PARAMCD} equals the specified value are considered
as the divisor

\emph{Permitted Values:} character value}

\item{set_values_to}{Variables to be set

The specified variables are set to the specified values for the new
observations. For example \code{exprs(PARAMCD = "MAP")} defines the parameter code
for the new parameter.

\emph{Permitted Values}: List of variable-value pairs}

\item{constant_dividend}{Is dividend parameter constant?

It is expected that the parameter code (PARAMCD) specified in \code{dividend_code}
which is required to derive the new parameter is measured only once. For example,
if Height-to-Weight Ratio should be derived and height is measured only once
while Weight is measured at each visit. Height could be specified
in the \code{dividend_code} argument and \code{constant_dividend} is to be set to \code{TRUE}.

\emph{Permitted Values:} logical scalar}

\item{constant_divisor}{Is divisor parameter constant?

It is expected that the parameter code (PARAMCD) specified in \code{dividend_code}
which is required to derive the new parameter is measured only once. For example,
if Waist-to-Height Ratio should be derived and height is measured only once
while Waist Circumference is measured at each visit. Height could be specified
in the \code{divisor_code} argument and \code{constant_divisor} is to be set to \code{TRUE}.

\emph{Permitted Values:} logical scalar}

\item{filter}{Filter condition

The specified condition is applied to the input dataset before deriving the
new parameter, i.e., only observations fulfilling the condition are taken
into account.

\emph{Permitted Values:} a condition}

\item{constant_by_vars}{By variables for when dividend and/or divisor is constant

When dividend and/or divisor is constant, the parameters (measured only once) are merged
to the other parameters using the specified variables.

If dividend and/or divisor is constant (e.g. only measured once at screening or baseline)
then use \code{constant_by_vars} to select the subject-level variable to merge on
(e.g. \code{USUBJID}). This will produce a generic Ratio parameter at all visits where dividend
and/or divisor is measured. Otherwise it will only be calculated at visits with both dividend
and divisor parameters collected.

\emph{Permitted Values}: list of variables created by \code{exprs()}
e.g. \code{exprs(USUBJID, VISIT)}}
}
\value{
The input dataset with the new parameter added. Note, a variable will only
be populated in the new parameter rows if it is specified in \code{by_vars}.
}
\description{
Adds a record for a generic Ratio parameter using two existing parameter
(dividend and divisor) each by group (e.g., subject and visit) where the source parameters
are available.

\strong{Note:} This is a wrapper function for the more generic \code{derive_param_computed()}
}
\details{
The analysis value of the new parameter is derived as
\deqn{RATIO = \frac{DIVIDENT}{DIVISOR}}
}
\examples{

library(tibble)

# Example 1: Derive Waist-to-Hip Ratio where both source parameters are measured multiple times

advs <- tribble(
  ~USUBJID, ~PARAMCD, ~PARAM, ~AVAL, ~AVALU, ~VISIT,
  "01-101-1001", "HIPCIR", "Hip Circumference (cm)", 125, "cm", "SCREENING",
  "01-101-1001", "HIPCIR", "Hip Circumference (cm)", 124, "cm", "WEEK 2",
  "01-101-1001", "HIPCIR", "Hip Circumference (cm)", 123, "cm", "WEEK 3",
  "01-101-1001", "WSTCIR", "Waist Circumference (cm)", 110, "cm", "SCREENING",
  "01-101-1001", "WSTCIR", "Waist Circumference (cm)", 108, "cm", "WEEK 2",
  "01-101-1001", "WSTCIR", "Waist Circumference (cm)", 107, "cm", "WEEK 3",
  "01-101-1002", "HIPCIR", "Hip Circumference (cm)", 135, "cm", "SCREENING",
  "01-101-1002", "HIPCIR", "Hip Circumference (cm)", 133, "cm", "WEEK 2",
  "01-101-1002", "HIPCIR", "Hip Circumference (cm)", 132, "cm", "WEEK 3",
  "01-101-1002", "WSTCIR", "Waist Circumference (cm)", 120, "cm", "SCREENING",
  "01-101-1002", "WSTCIR", "Waist Circumference (cm)", 118, "cm", "WEEK 2",
  "01-101-1002", "WSTCIR", "Waist Circumference (cm)", 117, "cm", "WEEK 3"
)

derive_param_ratio(
  advs,
  by_vars = exprs(USUBJID, VISIT),
  dividend_code = "WSTCIR",
  divisor_code = "HIPCIR",
  set_values_to = exprs(
    PARAMCD = "WAISTHIP",
    PARAM = "Waist-to-Hip Ratio"
  )
)

# Example 2: Derive Waist-to-Height Ratio where Height is measured only once

advs <- tribble(
  ~USUBJID, ~PARAMCD, ~PARAM, ~AVAL, ~AVALU, ~VISIT,
  "01-101-1001", "HEIGHT", "Height (cm)", 147, "cm", "SCREENING",
  "01-101-1001", "WSTCIR", "Waist Circumference (cm)", 110, "cm", "SCREENING",
  "01-101-1001", "WSTCIR", "Waist Circumference (cm)", 108, "cm", "WEEK 2",
  "01-101-1001", "WSTCIR", "Waist Circumference (cm)", 107, "cm", "WEEK 3",
  "01-101-1002", "HEIGHT", "Height (cm)", 163, "cm", "SCREENING",
  "01-101-1002", "WSTCIR", "Waist Circumference (cm)", 120, "cm", "SCREENING",
  "01-101-1002", "WSTCIR", "Waist Circumference (cm)", 118, "cm", "WEEK 2",
  "01-101-1002", "WSTCIR", "Waist Circumference (cm)", 117, "cm", "WEEK 3",
)

derive_param_ratio(
  advs,
  by_vars = exprs(USUBJID, VISIT),
  dividend_code = "WSTCIR",
  divisor_code = "HEIGHT",
  set_values_to = exprs(
    PARAMCD = "WAISTHGT",
    PARAM = "Waist-to-Height Ratio"
  ),
  constant_divisor = TRUE,
  constant_by_vars = exprs(USUBJID)
)
}
\seealso{
\code{\link[=compute_ratio]{compute_ratio()}}

BDS-Findings Functions for adding Parameters/Records: 
\code{\link{default_qtc_paramcd}()},
\code{\link{derive_expected_records}()},
\code{\link{derive_extreme_event}()},
\code{\link{derive_extreme_records}()},
\code{\link{derive_locf_records}()},
\code{\link{derive_param_bmi}()},
\code{\link{derive_param_bsa}()},
\code{\link{derive_param_computed}()},
\code{\link{derive_param_doseint}()},
\code{\link{derive_param_exist_flag}()},
\code{\link{derive_param_exposure}()},
\code{\link{derive_param_framingham}()},
\code{\link{derive_param_map}()},
\code{\link{derive_param_qtc}()},
\code{\link{derive_param_rr}()},
\code{\link{derive_param_wbc_abs}()},
\code{\link{derive_summary_records}()}
}
\concept{der_prm_bds_findings}
\keyword{der_prm_bds_findings}
