% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_query.R
\name{derive_query_vars}
\alias{derive_query_vars}
\title{Derive Query Variables}
\usage{
derive_query_vars(dataset, dataset_queries)
}
\description{
\code{derive_query_vars()} was renamed to \code{derive_vars_query()} to create a
more consistent API.
}
\keyword{internal}
