% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_reference_ranges.R
\name{derive_reference_ranges}
\alias{derive_reference_ranges}
\title{Derive Reference Ranges and Range Indicator}
\usage{
derive_reference_ranges(dataset, meta_ref_ranges, by_var = "PARAMCD")
}
\arguments{
\item{dataset}{The input dataset}

\item{meta_ref_ranges}{A dataset containing the references ranges. Required
variables are \code{by_var}, \code{ANRLO}, \code{ANRHI}. \code{A1LO} and \code{A1HI} are optional.}

\item{by_var}{A \code{character} vector of variables by which to join \code{dataset} and
\code{meta_ref_ranges}. By default, \code{"PARAMCD"}.}
}
\value{
The input dataset with additional columns \code{ANRLO}, \code{ANRHI}, \code{ANRIND}
and---if present in \code{meta_ref_ranges}---\code{A1LO} as well as \code{A1HI}
}
\description{
Derive Reference Ranges and Range Indicator
}
\details{
\code{ANRIND} is set to
\itemize{
\item \code{"NORMAL"} if \code{AVAL} is greater or equal \code{ANRLO} and less than
or equal \code{ANRHI}
\item \code{"LOW"} if \code{AVAL} is less than \code{ANRLO} and either \code{A1LO} is missing or \code{AVAL}
is greater than or equal \code{A1LO}
\item \code{"HIGH"} if \code{AVAL} is greater than \code{ANRHI} and either \code{A1HI} is missing or \code{AVAL}
is less than or equal \code{A1HI}
\item \code{"LOW LOW"} if \code{AVAL} is less than \code{A1LO}
\item \code{"HIGH HIGH"} if \code{AVAL} is greater than \code{A1HI}
}
}
\examples{
library(dplyr)
data(vs)
ref_ranges <- tibble::tribble(
  ~PARAMCD, ~ANRLO, ~ANRHI, ~A1LO, ~A1HI,
  "DIABP",  60,      80,    40,     90,
  "PULSE",  60,     100,    40,    110
)

vs \%>\%
  mutate(
    PARAMCD = VSTESTCD,
    AVAL = VSSTRESN
  ) \%>\%
  filter(PARAMCD \%in\% c("PULSE", "DIABP")) \%>\%
  derive_reference_ranges(ref_ranges) \%>\%
  select(USUBJID, PARAMCD, AVAL, ANRLO:ANRIND)
}
