% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_row_summary.R
\name{derive_row_summary}
\alias{derive_row_summary}
\title{Add new row within by groups using aggregation functions}
\usage{
derive_row_summary(dataset, by_vars, fns, set_values_to = list(),
  drop_values_from = NULL)
}
\arguments{
\item{dataset}{A data frame.}

\item{by_vars}{Variables to consider for generation of groupwise summary
rows. Providing the names of variables in \code{\link[=c]{c()}} will create a groupwise
summary and generate summary rows for the specified groups.}

\item{fns}{List of formulas specifying variable to use for aggregations.
This can include base functions like \code{mean}, \code{min}, \code{max}, \code{median}, \code{sd},
or \code{sum} or any other user-defined aggregation function. For example,
\code{fns = list(AVAL ~ mean, CHG ~ sum(., na.rm = TRUE)}. In the formula
representation, a \code{.} serves as the data to be summarized
(e.g., \code{sum(CHG, na.rm = TRUE)}).
\itemize{
\item LHS refer to the one or more variable to use for summarizing.
\item RHS refer to a \strong{single} summary function.
}}

\item{set_values_to}{A list of variable name-value pairs. Use this argument if
you need to change the values of any newly derived rows. For example,
\code{values_set = list(AVISITN = 9999, AVISIT= "Endpoint")} would change of
value of AVISITN to 9999 and AVISIT to Endpoint instead of retaining.}

\item{drop_values_from}{Providing the names of variables in \code{\link[dplyr:vars]{dplyr::vars()}}
will drop values and set as missing.}
}
\value{
A data frame with derived records appended to original dataset.
}
\description{
It is not uncommon to have an analysis need whereby one needs to derive an
analysis value (\code{AVAL}) from multiple rows. The ADaM basic dataset structure
variable \code{DTYPE} is available to indicate when a new derived row has been added to a
dataset.
}
\examples{
adeg <- tibble::tribble(
~USUBJID,~EGSEQ,~EGREPNUM,~PARAM,~VISIT,~AVISIT,~EGDTC,~AVAL,~TRTA,~SAFFL,
"XYZ-1001",1,1,"QTcF Interval (msec)","SCREENING","Baseline","2016-02-24T07:50:16",385,"","Y",
"XYZ-1001",2,2,"QTcF Interval (msec)","SCREENING","Baseline","2016-02-24T07:52:59",399,"","Y",
"XYZ-1001",3,3,"QTcF Interval (msec)","SCREENING","Baseline","2016-02-24T07:56:07",396,"","Y",
"XYZ-1001",4,1,"QTcF Interval (msec)","VISIT 2","Visit 2","2016-03-08T09:45:11",384,"Placebo","Y",
"XYZ-1001",5,2,"QTcF Interval (msec)","VISIT 2","Visit 2","2016-03-08T09:48:07",393,"Placebo","Y",
"XYZ-1001",6,3,"QTcF Interval (msec)","VISIT 2","Visit 2","2016-03-08T09:51:04",388,"Placebo","Y",
"XYZ-1001",7,1,"QTcF Interval (msec)","VISIT 3","Visit 3","2016-03-22T10:45:03",385,"Placebo","Y",
"XYZ-1001",8,2,"QTcF Interval (msec)","VISIT 3","Visit 3","2016-03-22T10:48:07",394,"Placebo","Y",
"XYZ-1001",9,3,"QTcF Interval (msec)","VISIT 3","Visit 3","2016-03-22T10:51:05",402,"Placebo","Y",
"XYZ-1002",1,1,"QTcF Interval (msec)","SCREENING","Baseline","2016-02-22T07:58:05",399,"","Y",
"XYZ-1002",2,2,"QTcF Interval (msec)","SCREENING","Baseline","2016-02-22T07:58:05",410,"","Y",
"XYZ-1002",3,3,"QTcF Interval (msec)","SCREENING","Baseline","2016-02-22T08:01:06",392,"","Y",
"XYZ-1002",4,1,"QTcF Interval (msec)","VISIT 2","Visit 2","2016-03-06T09:50:04",401,"Active 20mg","Y",
"XYZ-1002",5,2,"QTcF Interval (msec)","VISIT 2","Visit 2","2016-03-06T09:53:51",407,"Active 20mg","Y",
"XYZ-1002",6,3,"QTcF Interval (msec)","VISIT 2","Visit 2","2016-03-06T09:56:21",400,"Active 20mg","Y",
"XYZ-1002",7,1,"QTcF Interval (msec)","VISIT 3","Visit 3","2016-03-24T10:50:07",412,"Active 20mg","Y",
"XYZ-1002",8,2,"QTcF Interval (msec)","VISIT 3","Visit 3","2016-03-24T10:53:08",414,"Active 20mg","Y",
"XYZ-1002",9,3,"QTcF Interval (msec)","VISIT 3","Visit 3","2016-03-24T10:56:05",402,"Active 20mg","Y",
)

# summarize the average of the triplicate ECG interval values (AVAL)
derive_row_summary(adeg,
                   by_vars = vars(USUBJID, PARAM, AVISIT),
                   fns = list(AVAL ~ mean(., na.rm = TRUE)),
                   set_values_to = list(DTYPE = "AVERAGE"))

}
\concept{row summary}
