% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_adae_trtemfl.R
\name{derive_var_adae_trtemfl}
\alias{derive_var_adae_trtemfl}
\title{Derive Treatment Emergent Analysis Flag (TRTEMFL) for ADAE}
\usage{
derive_var_adae_trtemfl(
  dataset,
  dataset_faae = NULL,
  dataset_ex,
  sevtox = "AESEV",
  subject_keys = vars(STUDYID, USUBJID)
)
}
\arguments{
\item{dataset}{Input AE dataset}

\item{dataset_faae}{Input FAAE dataset (optional). The default is NULL}

\item{dataset_ex}{Input EX dataset}

\item{sevtox}{Name of variable in which severity or toxicity is stored. Permitted values are
"AESEV" or "AETOXGR". The default is "AESEV".}

\item{subject_keys}{The key variables to identify a subject provided as a list of variables.
The default is vars(STUDYID, USUBJID).}
}
\value{
The input AE dataset with the Treatment Emergent Analysis Flag 'TRTEMFL' added.
}
\description{
This function derives ADAE Treatment Emergent Analysis Flag (ADAE.TRTEMFL). \cr
The derivation is based on the default specifications for ADAE.TRTEMFL from the GDSR
(Version 2021-Oct.v1: Shenzou12). \cr
The following variables have to be available in the input data:
TRTSDTM, ASTDTM, AENDTM, either AESEV and AESEVIN or AETOXGR and AEITOXGR depending on parameter
setting, if FAAE is present: AEGRPID
}
\details{
The presence of 'TRTEMFL' is checked and a warning is issued if found. The existing variable
will be overwritten.
}
\examples{
\dontrun{
# Run function with default parameter setting and faae input
derive_var_adae_trtemfl(
  dataset = adae, 
  dataset_faae = faae, 
  dataset_ex = ex) 

# Run function with toxicity grade, custom subject key and no faae input
derive_var_adae_trtemfl(
  dataset = adae, 
  dataset_ex = ex,
  sevtox = "AETOXGR",
  subject_keys = vars(STUDYID, UNI_ID)
) 
}

}
\author{
Sandra Ott
}
\keyword{adae}
\keyword{adam}
\keyword{derivation}
\keyword{flag}
