% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_aage.R
\name{derive_var_agegr_fda}
\alias{derive_var_agegr_fda}
\alias{derive_var_agegr_ema}
\title{Derive Age Groups}
\usage{
derive_var_agegr_fda(dataset, age_var, age_unit = NULL, new_var)

derive_var_agegr_ema(dataset, age_var, age_unit = NULL, new_var)
}
\arguments{
\item{dataset}{Input dataset}

\item{age_var}{AGE variable}

\item{age_unit}{AGE unit variable}

\item{new_var}{New variable to create inside \code{dataset}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is \emph{deprecated}, please create a user defined function instead.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is \emph{deprecated}, please create a user defined function instead.
}
\seealso{
Other deprecated: 
\code{\link{derive_derived_param}()},
\code{\link{derive_param_first_event}()},
\code{\link{derive_var_confirmation_flag}()},
\code{\link{derive_var_disposition_status}()},
\code{\link{derive_var_worst_flag}()},
\code{\link{derive_vars_disposition_reason}()},
\code{\link{derive_vars_merged_dtm}()},
\code{\link{derive_vars_merged_dt}()},
\code{\link{filter_confirmation}()},
\code{\link{format_eoxxstt_default}()},
\code{\link{format_reason_default}()},
\code{\link{vars}()}

Other deprecated: 
\code{\link{derive_derived_param}()},
\code{\link{derive_param_first_event}()},
\code{\link{derive_var_confirmation_flag}()},
\code{\link{derive_var_disposition_status}()},
\code{\link{derive_var_worst_flag}()},
\code{\link{derive_vars_disposition_reason}()},
\code{\link{derive_vars_merged_dtm}()},
\code{\link{derive_vars_merged_dt}()},
\code{\link{filter_confirmation}()},
\code{\link{format_eoxxstt_default}()},
\code{\link{format_reason_default}()},
\code{\link{vars}()}

Other deprecated: 
\code{\link{derive_derived_param}()},
\code{\link{derive_param_first_event}()},
\code{\link{derive_var_confirmation_flag}()},
\code{\link{derive_var_disposition_status}()},
\code{\link{derive_var_worst_flag}()},
\code{\link{derive_vars_disposition_reason}()},
\code{\link{derive_vars_merged_dtm}()},
\code{\link{derive_vars_merged_dt}()},
\code{\link{filter_confirmation}()},
\code{\link{format_eoxxstt_default}()},
\code{\link{format_reason_default}()},
\code{\link{vars}()}
}
\concept{deprecated}
\keyword{deprecated}
