% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_analysis_ratio.R
\name{derive_var_analysis_ratio}
\alias{derive_var_analysis_ratio}
\title{Derive Ratio Variable}
\usage{
derive_var_analysis_ratio(
  dataset,
  numer_var,
  denom_var,
  ratio_var_suffix,
  override = FALSE,
  new_var
)
}
\arguments{
\item{dataset}{Input dataset}

\item{numer_var}{Variable containing values to be used in the numerator of
the ratio calculation.}

\item{denom_var}{Variable containing values to be used in the numerator of the of
the ratio calculation.}

\item{ratio_var_suffix}{User supplied suffix to be concatenated with \code{"R2"}}

\item{override}{A user wishing to remove the default prefix and supply
their own completely new variable can set this to TRUE.

Default is \code{"FALSE"}.}

\item{new_var}{Name of the new ratio variable being created.}
}
\value{
The input dataset with a ratio variable appended
}
\description{
Derives a ratio variable for a BDS dataset, e.g Ratio to Baseline \code{"AVAL / BASE"},
Ratio to Analysis Range y Lower Limit \code{"AVAL / AyLO"},  or Ratio to Analysis
Range y Upper Limit \code{"AVAL / AyHI"}.
}
\details{
Reference CDISC ADaM Implementation Guide Version 1.1
See section 3.3.4 Analysis Parameter Variables for BDS Datasets
}
\examples{
library(dplyr, warn.conflicts = FALSE)

data <-tibble::tribble(
 ~USUBJID, ~PARAMCD, ~SEQ,  ~AVAL, ~BASE, ~ANRLO, ~ANRHI,
 "P01",    "ALT",     1,     27,    27,    6,      34,
 "P01",    "ALT",     2,     41,    27,    6,      34,
 "P01",    "ALT",     3,     17,    27,    6,      34,
 "P02",    "ALB",     1,     38,    38,    33,     49,
 "P02",    "ALB",     2,     39,    38,    33,     49,
 "P02",    "ALB",     3,     37,    38,    33,     49
)

data \%>\%
 derive_var_analysis_ratio(numer_var = AVAL, denom_var = BASE, ratio_var_suffix = "BASE") \%>\%
 derive_var_analysis_ratio(numer_var = AVAL, denom_var = ANRLO, ratio_var_suffix = "ANR1LO") \%>\%
 derive_var_analysis_ratio(numer_var = AVAL, denom_var = ANRHI, ratio_var_suffix = "ANR1HI")

}
\author{
Ben Straub
}
\keyword{adam}
\keyword{adlb}
\keyword{bds}
\keyword{derivation}
