% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_basetype.R
\name{derive_var_basetype}
\alias{derive_var_basetype}
\title{Derive Basetype Variable}
\usage{
derive_var_basetype(dataset, basetypes)
}
\arguments{
\item{dataset}{Input dataset

The columns specified in the expressions inside \code{basetypes} are required.}

\item{basetypes}{A \emph{named} list of expressions created using the
\code{rlang::exprs()} function

The names corresponds to the values of the newly created \code{BASETYPE} variables
and the expressions are used to subset the input dataset.}
}
\value{
The input dataset with variable \code{BASETYPE} added
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is \emph{deprecated}, please use \code{derive_basetype_records()} instead.

Baseline Type \code{BASETYPE} is needed when there is more than one definition of
baseline for a given Analysis Parameter \code{PARAM} in the same dataset.  For a
given parameter, if Baseline Value \code{BASE} is populated, and there is more than
one definition of baseline, then \code{BASETYPE} must be non-null on all records of
any type for that parameter. Each value of \code{BASETYPE} refers to a definition of
baseline that characterizes the value of \code{BASE} on that row.  Please see
section 4.2.1.6 of the ADaM Implementation Guide, version 1.3 for further
background.

Adds the \code{BASETYPE} variable to a dataset and duplicates records based upon
the provided conditions.
}
\details{
For each element of \code{basetypes} the input dataset is subset based upon
the provided expression and the \code{BASETYPE} variable is set to the name of the
expression. Then, all subsets are stacked. Records which do not match any
condition are kept and \code{BASETYPE} is set to \code{NA}.
}
\seealso{
Other deprecated: 
\code{\link{derive_param_extreme_event}()},
\code{\link{derive_var_confirmation_flag}()},
\code{\link{derive_var_disposition_status}()},
\code{\link{derive_var_last_dose_amt}()},
\code{\link{derive_var_last_dose_date}()},
\code{\link{derive_var_last_dose_grp}()},
\code{\link{derive_var_merged_cat}()},
\code{\link{derive_var_merged_character}()},
\code{\link{derive_var_worst_flag}()},
\code{\link{derive_vars_disposition_reason}()},
\code{\link{derive_vars_last_dose}()},
\code{\link{format_eoxxstt_default}()},
\code{\link{format_reason_default}()}
}
\concept{deprecated}
\keyword{deprecated}
