% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_base.R
\name{derive_var_chg}
\alias{derive_var_chg}
\title{Derive BASE}
\usage{
derive_var_chg(bds_dataset)
}
\arguments{
\item{bds_dataset}{`data.frame`}
}
\value{
A new `data.frame` containing all records and variables of the input
dataset plus the `BASE` variable
}
\description{
Derive the `BASE` variable in a BDS dataset
}
\examples{
bds_dataset <- tibble::tribble(
  ~STUDYID, ~USUBJID, ~PARAMCD,  ~AVAL, ~ABL01FL,
  "TEST01", "PAT01",  "PARAM01", 10.12, "Y",
  "TEST01", "PAT01",  "PARAM01",  9.7,  "N",
  "TEST01", "PAT01",  "PARAM01", 15.01, "N",
  "TEST01", "PAT01",  "PARAM02",  8.35, "Y",
  "TEST01", "PAT01",  "PARAM02", NA,    "N",
  "TEST01", "PAT01",  "PARAM02",  8.35, "N",
)
derive_var_chg(bds_dataset)

}
\author{
Thomas Neitmann
}
