% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_dthcat.R
\name{derive_var_dthcat}
\alias{derive_var_dthcat}
\title{Derive Cause of Death Category}
\usage{
derive_var_dthcat(dataset)
}
\arguments{
\item{dataset}{Input dataset

The columns "DTHDOM" and "DTHCAUS" are expected.}
}
\value{
The input dataset with \code{DTHCAT} added
}
\description{
Derives cause of death category (\code{DTHCAT})
}
\details{
The cause of death category is derived based on DTHDOM and DTHCAUS
values.
}
\examples{
data <- tibble::tribble(
  ~DTHDOM, ~DTHCAUS,
  "ADVERSE EVENT", ""
)

derive_var_dthcat(data)
}
\author{
Shimeng Huang
}
\keyword{Roche}
