% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_extreme_date.R
\name{derive_var_extreme_dtm}
\alias{derive_var_extreme_dtm}
\title{Derive First or Last Datetime from Multiple Sources}
\usage{
derive_var_extreme_dtm(
  dataset,
  new_var,
  ...,
  source_datasets,
  mode,
  subject_keys = get_admiral_option("subject_keys")
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{subject_keys} argument are expected to be in the dataset.

\describe{
\item{Default value}{none}
}}

\item{new_var}{Name of variable to create

\describe{
\item{Default value}{none}
}}

\item{...}{Source(s) of dates. One or more \code{date_source()} objects are
expected.

\describe{
\item{Default value}{none}
}}

\item{source_datasets}{A named \code{list} containing datasets in which to search
for the first or last date

\describe{
\item{Default value}{none}
}}

\item{mode}{Selection mode (first or last)

If \code{"first"} is specified, the first date for each subject is selected. If
\code{"last"} is specified, the last date for each subject is selected.

\describe{
\item{Permitted values}{\code{"first"}, \code{"last"}}
\item{Default value}{none}
}}

\item{subject_keys}{Variables to uniquely identify a subject

A list of expressions where the expressions are symbols as returned by
\code{exprs()} is expected.

\describe{
\item{Default value}{\code{get_admiral_option("subject_keys")}}
}}
}
\value{
The input dataset with the new variable added.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

The \code{derive_var_extreme_dtm()}
function has been deprecated in favor of \code{derive_vars_extreme_event()}.

Add the first or last datetime from multiple sources to the dataset, e.g.,
the last known alive datetime (\code{LSTALVDTM}).
}
\details{
The following steps are performed to create the output dataset:
\enumerate{
\item For each source dataset the observations as specified by the \code{filter}
element are selected and observations where \code{date} is \code{NA} are removed.
Then for each patient the first or last observation (with respect to \code{date}
and \code{mode}) is selected.
\item The new variable is set to the variable or expression specified by the
\code{date} element. If this is a date variable (rather than datetime), then the
time is imputed as \code{"00:00:00"}.
\item The variables specified by the \code{set_values_to} element are added.
\item The selected observations of all source datasets are combined into a
single dataset.
\item For each patient the first or last observation (with respect to the new
variable and \code{mode}) from the single dataset is selected and the new
variable is merged to the input dataset.
}
}
\seealso{
\code{\link[=date_source]{date_source()}}, \code{\link[=derive_var_extreme_dt]{derive_var_extreme_dt()}},
\code{\link[=derive_vars_merged]{derive_vars_merged()}}

Other deprecated:
\code{\link[=date_source]{date_source()}},
\code{\link[=derive_param_extreme_record]{derive_param_extreme_record()}},
\code{\link[=derive_var_dthcaus]{derive_var_dthcaus()}},
\code{\link[=derive_var_extreme_dt]{derive_var_extreme_dt()}},
\code{\link[=derive_var_merged_summary]{derive_var_merged_summary()}},
\code{\link[=dthcaus_source]{dthcaus_source()}},
\code{\link[=get_summary_records]{get_summary_records()}}
}
\concept{deprecated}
\keyword{deprecated}
