% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_extreme_flag.R
\name{derive_var_extreme_flag}
\alias{derive_var_extreme_flag}
\title{Add a Variable Flagging the First or Last Observation Within Each By Group}
\usage{
derive_var_extreme_flag(
  dataset,
  by_vars,
  order,
  new_var,
  mode,
  check_type = "warning"
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{by_vars} parameter are expected.}

\item{by_vars}{Grouping variables

Permitted Values: list of variables}

\item{order}{Sort order

The first or last observation is determined with respect to the specified
order.

Permitted Values: list of variables or functions of variables}

\item{new_var}{Variable to add

The specified variable is added to the output dataset. It is set to \code{"Y"}
for the first or last observation (depending on the mode) of each by group.

Permitted Values: list of name-value pairs}

\item{mode}{Flag mode

Determines of the first or last observation is flagged.

Permitted Values: \code{"first"}, \code{"last"}}

\item{check_type}{Check uniqueness?

If \code{"warning"} or \code{"error"} is specified, the specified message is issued
if the observations of the input dataset are not unique with respect to the
by variables and the order.

Default: \code{"warning"}

Permitted Values: \code{"none"}, \code{"warning"}, \code{"error"}}
}
\value{
The input dataset with the new flag variable added
}
\description{
Add a variable flagging the first or last observation within each by group
}
\details{
For each group (with respect to the variables specified for the
\code{by_vars} parameter), \code{new_var} is set to "Y" for the first or last observation
(with respect to the order specified for the \code{order} parameter and the flag mode
specified for the \code{mode} parameter). Only observations included by the \code{filter} parameter
are considered for flagging.
Otherwise, \code{new_var} is set to \code{NA}. Thus, the direction of "worst" is considered fixed for
all parameters in the dataset depending on the \code{order} and the \code{mode}, i.e. for every
parameter the first or last record will be flagged across the whole dataset.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
vs <- tribble(
  ~STUDYID,  ~DOMAIN,  ~USUBJID, ~VSTESTCD, ~VSSTRESN,      ~VISIT, ~VSTPTNUM,
  "PILOT01",    "VS", "10-1187",   "DIABP",        70, "SCREENING",       815,
  "PILOT01",    "VS", "10-1187",   "DIABP",        78, "SCREENING",       816,
  "PILOT01",    "VS", "10-1187",   "DIABP",        80, "SCREENING",       817,
  "PILOT01",    "VS", "10-1187",   "DIABP",        68,  "BASELINE",       815,
  "PILOT01",    "VS", "10-1187",   "DIABP",        78,  "BASELINE",       816,
  "PILOT01",    "VS", "10-1187",   "DIABP",        74,  "BASELINE",       817,
  "PILOT01",    "VS", "10-1187",   "DIABP",        60,    "WEEK 4",       815,
  "PILOT01",    "VS", "10-1187",   "DIABP",        70,    "WEEK 4",       816,
  "PILOT01",    "VS", "10-1187",   "DIABP",        74,    "WEEK 4",       817,
  "PILOT01",    "VS", "10-1187",  "WEIGHT",      49.9, "SCREENING",        NA,
  "PILOT01",    "VS", "10-1187",  "WEIGHT",      49.9,  "BASELINE",        NA,
  "PILOT01",    "VS", "10-1187",  "WEIGHT",     49.44,    "WEEK 4",        NA,
  "PILOT01",    "VS", "10-1271",   "DIABP",        68, "SCREENING",       815,
  "PILOT01",    "VS", "10-1271",   "DIABP",        70, "SCREENING",       816,
  "PILOT01",    "VS", "10-1271",   "DIABP",        72, "SCREENING",       817,
  "PILOT01",    "VS", "10-1271",   "DIABP",        80,  "BASELINE",       815,
  "PILOT01",    "VS", "10-1271",   "DIABP",        86,  "BASELINE",       816,
  "PILOT01",    "VS", "10-1271",   "DIABP",        84,  "BASELINE",       817,
  "PILOT01",    "VS", "10-1271",   "DIABP",        80,    "WEEK 4",       815,
  "PILOT01",    "VS", "10-1271",   "DIABP",        88,    "WEEK 4",       816,
  "PILOT01",    "VS", "10-1271",   "DIABP",        86,    "WEEK 4",       817,
  "PILOT01",    "VS", "10-1271",  "WEIGHT",     47.17, "SCREENING",        NA,
  "PILOT01",    "VS", "10-1271",  "WEIGHT",     47.63,  "BASELINE",        NA,
  "PILOT01",    "VS", "10-1271",  "WEIGHT",     49.44,    "WEEK 4",        NA
)

# Flag last value for each patient, test, and visit, baseline observations are ignored
vs \%>\%
  restrict_derivation(
    derivation = derive_var_extreme_flag,
    args = params(
      by_vars = exprs(USUBJID, VSTESTCD, VISIT),
      order = exprs(VSTPTNUM),
      new_var = LASTFL,
      mode = "last"
    ),
    filter = VISIT != "BASELINE"
  ) \%>\%
  arrange(USUBJID, VSTESTCD, VISIT, VSTPTNUM) \%>\%
  select(USUBJID, VSTESTCD, VISIT, VSTPTNUM, VSSTRESN, LASTFL)

# Baseline (ABLFL) examples:

input <- tribble(
  ~STUDYID,    ~USUBJID,  ~PARAMCD,     ~AVISIT,          ~ADT, ~AVAL,    ~DTYPE,
  "PILOT01",  "10-1187", "PARAM01",  "BASELINE",  "2021-04-27",  15.0,        NA,
  "PILOT01",  "10-1187", "PARAM01",  "BASELINE",  "2021-04-25",  14.0,        NA,
  "PILOT01",  "10-1187", "PARAM01",  "BASELINE",  "2021-04-23",  15.0, "AVERAGE",
  "PILOT01",  "10-1187", "PARAM01",    "WEEK 1",  "2021-04-27",  10.0, "AVERAGE",
  "PILOT01",  "10-1187", "PARAM01",    "WEEK 2",  "2021-04-30",  12.0,        NA,
  "PILOT01",  "10-1271", "PARAM01", "SCREENING",  "2021-04-27",  15.0, "AVERAGE",
  "PILOT01",  "10-1271", "PARAM01",  "BASELINE",  "2021-04-25",  14.0, "AVERAGE",
  "PILOT01",  "10-1271", "PARAM01",  "BASELINE",  "2021-04-23",  15.0, "AVERAGE",
  "PILOT01",  "10-1271", "PARAM01",    "WEEK 1",  "2021-04-27",  10.0, "AVERAGE",
  "PILOT01",  "10-1271", "PARAM01",    "WEEK 2",  "2021-04-30",  12.0, "AVERAGE",
  "PILOT01",  "10-1271", "PARAM02", "SCREENING",  "2021-04-27",  15.0, "AVERAGE",
  "PILOT01",  "10-1271", "PARAM02", "SCREENING",  "2021-04-25",  14.0, "AVERAGE",
  "PILOT01",  "10-1271", "PARAM02", "SCREENING",  "2021-04-23",  15.0,        NA,
  "PILOT01",  "10-1271", "PARAM02",  "BASELINE",  "2021-04-27",  10.0, "AVERAGE",
  "PILOT01",  "10-1271", "PARAM02",    "WEEK 2",  "2021-04-30",  12.0,        NA,
  "PILOT01",  "10-1187", "PARAM02", "SCREENING",  "2021-04-27",  15.0,        NA,
  "PILOT01",  "10-1187", "PARAM02",  "BASELINE",  "2021-04-25",  14.0,        NA,
  "PILOT01",  "10-1187", "PARAM02",    "WEEK 1",  "2021-04-23",  15.0,        NA,
  "PILOT01",  "10-1187", "PARAM02",    "WEEK 1",  "2021-04-27",  10.0,        NA,
  "PILOT01",  "10-1187", "PARAM02",  "BASELINE",  "2021-04-30",  12.0,        NA
) \%>\%
  mutate(
    ADT = as.Date(ADT)
  )

# Last observation
restrict_derivation(
  input,
  derivation = derive_var_extreme_flag,
  args = params(
    by_vars = exprs(USUBJID, PARAMCD),
    order = exprs(ADT),
    new_var = ABLFL,
    mode = "last"
  ),
  filter = AVISIT == "BASELINE"
)

# Worst observation - Direction = High
restrict_derivation(
  input,
  derivation = derive_var_extreme_flag,
  args = params(
    by_vars = exprs(USUBJID, PARAMCD),
    order = exprs(AVAL, ADT),
    new_var = ABLFL,
    mode = "last"
  ),
  filter = AVISIT == "BASELINE"
)

# Worst observation - Direction = Lo
restrict_derivation(
  input,
  derivation = derive_var_extreme_flag,
  args = params(
    by_vars = exprs(USUBJID, PARAMCD),
    order = exprs(desc(AVAL), ADT),
    new_var = ABLFL,
    mode = "last"
  ),
  filter = AVISIT == "BASELINE"
)

# Average observation
restrict_derivation(
  input,
  derivation = derive_var_extreme_flag,
  args = params(
    by_vars = exprs(USUBJID, PARAMCD),
    order = exprs(ADT, desc(AVAL)),
    new_var = ABLFL,
    mode = "last"
  ),
  filter = AVISIT == "BASELINE" & DTYPE == "AVERAGE"
)

# OCCURDS Examples
ae <- tribble(
  ~STUDYID,  ~DOMAIN,  ~USUBJID, ~AESEQ,        ~AEDECOD,     ~AEBODSYS,     ~AESEV, ~AESTDY,
  "PILOT01",    "AE", "10-1187",      1,    "IRRITATION",     "GENERAL",     "MILD",      44,
  "PILOT01",    "AE", "10-1271",      3,  "FIBRILLATION",     "CARDIAC", "MODERATE",      56,
  "PILOT01",    "AE", "10-1271",      4,       "CARDIAC",     "CARDIAC", "MODERATE",      57,
  "PILOT01",    "AE", "10-1271",      1,      "DYSPNOEA", "RESPIRATORY", "MODERATE",      56,
  "PILOT01",    "AE", "10-1271",      5, "HYPONATRAEMIA",  "METABOLISM", "MODERATE",      57,
  "PILOT01",    "AE", "10-1271",      2,    "MYOCARDIAL",     "CARDIAC",   "SEVERE",      56
)

# Most severe AE first occurrence per patient
ae \%>\%
  mutate(
    TEMP_AESEVN =
      as.integer(factor(AESEV, levels = c("SEVERE", "MODERATE", "MILD")))
  ) \%>\%
  derive_var_extreme_flag(
    new_var = AOCCIFL,
    by_vars = exprs(USUBJID),
    order = exprs(TEMP_AESEVN, AESTDY, AESEQ),
    mode = "first"
  ) \%>\%
  arrange(USUBJID, AESTDY, AESEQ) \%>\%
  select(USUBJID, AEDECOD, AESEV, AESTDY, AESEQ, AOCCIFL)

# Most severe AE first occurrence per patient per body system
ae \%>\%
  mutate(
    TEMP_AESEVN =
      as.integer(factor(AESEV, levels = c("SEVERE", "MODERATE", "MILD")))
  ) \%>\%
  derive_var_extreme_flag(
    new_var = AOCCSIFL,
    by_vars = exprs(USUBJID, AEBODSYS),
    order = exprs(TEMP_AESEVN, AESTDY, AESEQ),
    mode = "first"
  ) \%>\%
  arrange(USUBJID, AESTDY, AESEQ) \%>\%
  select(USUBJID, AEBODSYS, AESEV, AESTDY, AOCCSIFL)
}
\seealso{
\code{\link[=derive_var_worst_flag]{derive_var_worst_flag()}}

General Derivation Functions for all ADaMs that returns variable appended to dataset:
\code{\link{derive_var_joined_exist_flag}()},
\code{\link{derive_var_last_dose_amt}()},
\code{\link{derive_var_last_dose_date}()},
\code{\link{derive_var_last_dose_grp}()},
\code{\link{derive_var_merged_exist_flag}()},
\code{\link{derive_var_merged_summary}()},
\code{\link{derive_var_obs_number}()},
\code{\link{derive_var_relative_flag}()},
\code{\link{derive_vars_joined}()},
\code{\link{derive_vars_last_dose}()},
\code{\link{derive_vars_merged_lookup}()},
\code{\link{derive_vars_merged}()},
\code{\link{derive_vars_transposed}()},
\code{\link{get_summary_records}()}
}
\concept{der_gen}
\keyword{der_gen}
