% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_last_dose_amt.R
\name{derive_var_last_dose_amt}
\alias{derive_var_last_dose_amt}
\title{Derive Last Dose Amount}
\usage{
derive_var_last_dose_amt(
  dataset,
  dataset_ex,
  filter_ex = NULL,
  by_vars = exprs(STUDYID, USUBJID),
  dose_id = exprs(),
  dose_date,
  analysis_date,
  single_dose_condition = (EXDOSFRQ == "ONCE"),
  new_var,
  dose_var = EXDOSE,
  traceability_vars = NULL
)
}
\arguments{
\item{new_var}{The new variable added to \code{dataset}.}

\item{dose_var}{The EX source dose amount variable. Defaults to \code{EXDOSE}.}
}
\value{
Input dataset with additional column \code{new_var}.
}
\description{
Add a variable for dose amount from the last dose to the input dataset.

\strong{Note:} This is a wrapper function for the function \code{derive_vars_last_dose()}.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is \emph{deprecated}, please use \code{derive_vars_joined()} instead.
}
\details{
The last dose amount is derived as the dose amount where the maximum \code{dose_date} is
lower to or equal to the \code{analysis_date} per \code{by_vars} for each observation in \code{dataset}.

If dose information is aggregated (i.e. is a dosing frequency other than \code{"ONCE"}
over a period defined by a start and end date) the function
\code{create_single_dose_dataset()} can be used to generate single doses from
aggregate dose information and satisfy \code{single_dose_condition}.
}
\seealso{
\code{\link[=create_single_dose_dataset]{create_single_dose_dataset()}}

Other deprecated: 
\code{\link{derive_param_extreme_event}()},
\code{\link{derive_var_basetype}()},
\code{\link{derive_var_confirmation_flag}()},
\code{\link{derive_var_last_dose_date}()},
\code{\link{derive_var_last_dose_grp}()},
\code{\link{derive_var_merged_cat}()},
\code{\link{derive_var_merged_character}()},
\code{\link{derive_vars_last_dose}()}
}
\concept{deprecated}
\keyword{deprecated}
