% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_last_dose_date.R
\name{derive_var_last_dose_date}
\alias{derive_var_last_dose_date}
\title{Derive Last Dose Date-Time}
\usage{
derive_var_last_dose_date(
  dataset,
  dataset_ex,
  filter_ex = NULL,
  by_vars = exprs(STUDYID, USUBJID),
  dose_id = exprs(),
  dose_date,
  analysis_date,
  single_dose_condition = (EXDOSFRQ == "ONCE"),
  new_var,
  output_datetime = TRUE,
  traceability_vars = NULL
)
}
\arguments{
\item{new_var}{The new date or datetime variable added to \code{dataset}.}

\item{output_datetime}{Display \code{new_var} as datetime or as date only. Defaults to \code{TRUE}.}
}
\value{
Input dataset with additional column \code{new_var}.
}
\description{
Add a variable for the dose date or datetime of the last dose to
the input dataset.

\strong{Note:} This is a wrapper function for the function \code{derive_vars_last_dose()}.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is \emph{deprecated}, please use \code{derive_vars_joined()} instead.
}
\details{
The last dose date is derived as the maximum dose date where the
\code{dose_date} is lower to or equal to the \code{analysis_date} per \code{by_vars} for
each observation in \code{dataset}. When \code{output_datetime} is \code{TRUE} and time is
missing, then the last dose date time is imputed to \code{00:00:00}. However, if
date is missing, then no imputation is done.

If dose information is aggregated (i.e. is a dosing frequency other than \code{"ONCE"}
over a period defined by a start and end date) the function
\code{create_single_dose_dataset()} can be used to generate single doses from
aggregate dose information and satisfy \code{single_dose_condition}.
}
\seealso{
\code{\link[=derive_vars_last_dose]{derive_vars_last_dose()}}, \code{\link[=create_single_dose_dataset]{create_single_dose_dataset()}}

Other deprecated: 
\code{\link{derive_param_extreme_event}()},
\code{\link{derive_var_basetype}()},
\code{\link{derive_var_confirmation_flag}()},
\code{\link{derive_var_last_dose_amt}()},
\code{\link{derive_var_last_dose_grp}()},
\code{\link{derive_var_merged_cat}()},
\code{\link{derive_var_merged_character}()},
\code{\link{derive_vars_last_dose}()}
}
\concept{deprecated}
\keyword{deprecated}
