% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_lstalvdt.R
\name{derive_var_lstalvdt}
\alias{derive_var_lstalvdt}
\title{Merge the first or last date from more than one dataset}
\usage{
derive_var_lstalvdt(dataset, ...)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{by_vars} parameter are expected.}

\item{sources}{List of sources

For each source the observations are selected as specified. The selected
observations of all sources are collected in one dataset and then for each
by group one observation (with respect to the \code{mode} parameter) is selected
and the date is merged to the input dataset.

Each element of the list must be a list with the following named elements.

\itemize{ \item \emph{dataset:} Dataset to add

The variables specified by the \code{by_vars} parameter, the \code{filter},
and the \code{order} element are expected.

\item \emph{filter:} Filter condition for dataset

Only observations of the dataset which fulfill the specified condition are
used for merging.

\emph{Permitted Values:} logical expression

\item \emph{var:} Variable to add

The specified variable is compared with the variables from the other
datasets. A date, a datetime, or a character variable containing ISO 8601
dates can be specified.

\emph{Permitted Values:} variable name (as symbol)

\item \emph{order:} Sort order

If the parameter is specified, the source dataset is ordered by the
specified order and only the first or last observation (depending on the
mode) in each by group is used for merging.

\emph{Permitted Values:} list of variables or functions of variables

\item \emph{mode:} mode

If the \code{order} element is specified, the mode determines if the first
or last observation of each by group is selected.

\emph{Permitted Values:} \code{"first"}, \code{"last"} }}
}
\value{
The input dataset with the additional variable
}
\description{
Merges the first or last date from more than one dataset. For each dataset
the observations to merge can be selected by a condition and/or by selecting
the first or last observation in each by group.
}
\details{
The following steps are performed to create the output dataset:

\enumerate{ \item For each source dataset the observations as specified by
the \code{filter} element are selected. If the \code{order} element is specified, for
each by group the first or last element (depending on the \code{mode} element)
with respect to the specified order is selected.

\item The new variable is set to the variable specified by the \code{var}
element. If the source variable is a datetime variable, only the datepart
is copied. If the source variable is a character variable, it is converted
to a date. If the date is imcomplete, it is set to NA.

\item The selected observations of all source datasets are combined into a
single dataset.

\item For each group (with respect to the variables specified for the
\code{by_vars} parameter) the first or last observation (with respect to the new
variable and the \code{mode} parameter) from the single dataset is selected and
the new variable is merged to the input dataset. }
}
\examples{
library(dplyr, warn.conflict = FALSE)
library(rlang, warn.conflict = FALSE)
library(stringr)
data("dm")
data("ae")
data("lb")
data("adsl")
ae_start <- list(dataset = ae,
                 var = expr(AESTDTC),
                 filter = expr(str_length(AESTDTC) >= 10))
ae_end <- list(dataset = ae,
               vars = expr(AEENDTC),
               filter = expr(str_length(AEENDTC) >= 10))
lb_date <- list(dataset = lb,
                var = expr(LBDTC),
                filter = expr(str_length(LBDTC) >= 10))

adsl_date <- list(dataset = adsl,
                  var = expr(TRTEDT))
derive_var_lstalvdt(dm,
                    sources = list(ae_start, ae_end, lb_date, adsl_date)) \%>\%
  select(USUBJID, LSTALVDT)

}
\author{
Stefan Bundfuss
}
\keyword{adam}
\keyword{derivation}
