% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_merged.R
\name{derive_var_merged_summary}
\alias{derive_var_merged_summary}
\title{Merge Summary Variables}
\usage{
derive_var_merged_summary(
  dataset,
  dataset_add,
  by_vars,
  new_vars = NULL,
  filter_add = NULL,
  missing_values = NULL
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{by_vars} argument are expected to be in the dataset.

\describe{
\item{Permitted values}{a dataset, i.e., a \code{data.frame} or tibble}
\item{Default value}{none}
}}

\item{dataset_add}{Additional dataset

The variables specified by the \code{by_vars} and the variables used on the left
hand sides of the \code{new_vars} arguments are expected.

\describe{
\item{Permitted values}{a dataset, i.e., a \code{data.frame} or tibble}
\item{Default value}{none}
}}

\item{by_vars}{Grouping variables

The expressions on the left hand sides of \code{new_vars} are evaluated by the
specified \emph{variables}. Then the resulting values are merged to the input
dataset (\code{dataset}) by the specified \emph{variables}.

\describe{
\item{Permitted values}{list of variables created by \code{exprs()}, e.g., \code{exprs(USUBJID, VISIT)}}
\item{Default value}{none}
}}

\item{new_vars}{New variables to add

The specified variables are added to the input dataset.

A named list of expressions is expected:
\itemize{
\item LHS refer to a variable.
\item RHS refers to the values to set to the variable. This can be a string, a
symbol, a numeric value, an expression or NA. If summary functions are
used, the values are summarized by the variables specified for \code{by_vars}.
Any expression on the RHS must result in a single value per by group.
}

For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  new_vars = exprs(
    DOSESUM = sum(AVAL),
    DOSEMEAN = mean(AVAL)
  )
}\if{html}{\out{</div>}}

\describe{
\item{Permitted values}{list of variables created by \code{exprs()}, e.g., \code{exprs(USUBJID, VISIT)}}
\item{Default value}{\code{NULL}}
}}

\item{filter_add}{Filter for additional dataset (\code{dataset_add})

Only observations fulfilling the specified condition are taken into account
for summarizing. If the argument is not specified, all observations are
considered.

Comparing derived numeric variables to fixed values, e.g., \code{PCHG <= -90}, may give unexpected results due to floating point representation. For details and solutions see the "Floating Point Comparisons" section in \code{vignette("concepts_conventions")}.

\describe{
\item{Permitted values}{an unquoted condition, e.g., \code{AVISIT == "BASELINE"}}
\item{Default value}{\code{NULL}}
}}

\item{missing_values}{Values for non-matching observations

For observations of the input dataset (\code{dataset}) which do not have a
matching observation in the additional dataset (\code{dataset_add}) the values
of the specified variables are set to the specified value. Only variables
specified for \code{new_vars} can be specified for \code{missing_values}.

\describe{
\item{Permitted values}{list of named expressions created by \code{exprs()}, e.g., \code{exprs(AVALC = VSSTRESC, AVAL = yn_to_numeric(AVALC))}}
\item{Default value}{\code{NULL}}
}}
}
\value{
The output dataset contains all observations and variables of the
input dataset and additionally the variables specified for \code{new_vars}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} The \code{derive_var_merged_summary()}
function has been deprecated in favor of \code{derive_vars_merged_summary()}.
}
\details{
\enumerate{
\item The records from the additional dataset (\code{dataset_add}) are restricted
to those matching the \code{filter_add} condition.
\item The new variables (\code{new_vars}) are created for each by group (\code{by_vars})
in the additional dataset (\code{dataset_add}) by calling \code{summarize()}. I.e.,
all observations of a by group are summarized to a single observation.
\item The new variables are merged to the input dataset. For observations
without a matching observation in the additional dataset the new variables
are set to \code{NA}. Observations in the additional dataset which have no
matching observation in the input dataset are ignored.
}
}
\seealso{
\code{\link[=derive_summary_records]{derive_summary_records()}}

Other deprecated:
\code{\link[=date_source]{date_source()}},
\code{\link[=derive_param_extreme_record]{derive_param_extreme_record()}},
\code{\link[=derive_var_dthcaus]{derive_var_dthcaus()}},
\code{\link[=derive_var_extreme_dt]{derive_var_extreme_dt()}},
\code{\link[=derive_var_extreme_dtm]{derive_var_extreme_dtm()}},
\code{\link[=dthcaus_source]{dthcaus_source()}},
\code{\link[=get_summary_records]{get_summary_records()}}
}
\concept{deprecated}
\keyword{deprecated}
