% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_trtemfl.R
\name{derive_var_trtemfl}
\alias{derive_var_trtemfl}
\title{Derive Treatment-emergent Flag}
\usage{
derive_var_trtemfl(
  dataset,
  new_var = TRTEMFL,
  start_date = ASTDTM,
  end_date = AENDTM,
  trt_start_date = TRTSDTM,
  trt_end_date = NULL,
  end_window = NULL,
  ignore_time_for_trt_end = TRUE,
  initial_intensity = NULL,
  intensity = NULL,
  group_var = NULL,
  subject_keys = get_admiral_option("subject_keys")
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by \code{start_date}, \code{end_date}, \code{trt_start_date},
\code{trt_end_date}, \code{initial_intensity}, and \code{intensity} are expected.}

\item{new_var}{New variable}

\item{start_date}{Event start date}

\item{end_date}{Event end date}

\item{trt_start_date}{Treatment start date}

\item{trt_end_date}{Treatment end date}

\item{end_window}{If the argument is specified (in 'days'), events starting more than the specified
number of days after end of treatment, are not flagged.}

\item{ignore_time_for_trt_end}{If the argument is set to \code{TRUE}, the time part is ignored for checking if
the event occurred more than \code{end_window} days after end of treatment.}

\item{initial_intensity}{Initial severity/intensity or toxicity

\code{initial_intensity} is ignored when \code{group_var} is specified.

If this argument is specified and \code{group_var} is \code{NULL}, events which start
before treatment start and end after treatment start (or are ongoing) and
worsened (i.e., the intensity is greater than the initial intensity), are
flagged.

The values of the specified variable must be comparable with the usual
comparison operators. I.e., if the intensity is greater than the initial
intensity \code{initial_intensity < intensity} must evaluate to \code{TRUE}.}

\item{intensity}{Severity/intensity or toxicity

If the argument is specified, events which start before treatment start and
end after treatment start (or are ongoing) and worsened (i.e., the
intensity is greater than the initial intensity), are flagged.

The values of the specified variable must be comparable with the usual
comparison operators. I.e., if the intensity is greater than the initial
intensity \code{initial_intensity < intensity} must evaluate to \code{TRUE}.}

\item{group_var}{Grouping variable

If the argument is specified, it assumes that AEs are recorded as one episode
of AE with multiple lines using a grouping variable.

Events starting during treatment or before treatment and worsening afterward
are flagged. Once an AE record in a group is flagged, all subsequent records
in the treatment window are flagged regardless of severity.}

\item{subject_keys}{Variables to uniquely identify a subject.

This argument is only used when \code{group_var} is specified.}
}
\value{
The input dataset with the variable specified by \code{new_var} added
}
\description{
Derive treatment emergent analysis flag (e.g., \code{TRTEMFL}).
}
\details{
For the derivation of the new variable the following cases are
considered in this order. The first case which applies, defines the value
of the variable.
\itemize{
\item \emph{not treated}: If \code{trt_start_date} is \code{NA}, it is set to \code{NA_character_}.
\item \emph{event before treatment}: If \code{end_date} is before \code{trt_start_date} (and
\code{end_date} is not \code{NA}), it is set to \code{NA_character_}.
\item \emph{no event date}: If \code{start_date} is \code{NA}, it is set to \code{"Y"} as in such
cases it is usually considered more conservative to assume the event was
treatment-emergent.
\item \emph{event started during treatment}:
\itemize{
\item if \code{end_window} is not specified:
if \code{start_date} is on or after \code{trt_start_date}, it is set to \code{"Y"},
\item if \code{end_window} is specified:
if \code{start_date} is on or after \code{trt_start_date} and \code{start_date} is on
or before \code{trt_end_date} + \code{end_window} days, it is set to \code{"Y"},
}
\item \emph{event started before treatment and (possibly) worsened on treatment}:
\itemize{
\item if \code{initial_intensity}, \code{intensity} is specified and \code{group_var} is not specified:
if \code{initial_intensity < intensity} and \code{start_date} is before \code{trt_start_date}
and \code{end_date} is on or after \code{trt_start_date} or \code{end_date} is \code{NA}, it
is set to \code{"Y"};
\item if \code{group_var} is specified:
if \code{intensity} at treatment start < \code{intensity} and \code{start_date} is after
\code{trt_start_date} and \code{end_date} is on or after \code{trt_start_date} or
\code{end_date} is \code{NA}, it is set to \code{"Y"};
}
\item Otherwise it is set to \code{NA_character_}.
}

The behavior of \code{derive_var_trtemfl()} is aligned with the proposed
treatment-emergent AE assignment in the following
\href{https://phuse.s3.eu-central-1.amazonaws.com/Deliverables/Safety+Analytics/WP-087+Recommended+Definition+of++Treatment-Emergent+Adverse+Events+in+Clinical+Trials+.pdf}{PHUSE White Paper}.
See the final example in the examples section below.
}
\seealso{
OCCDS Functions: 
\code{\link{derive_vars_atc}()},
\code{\link{derive_vars_query}()}
}
\concept{der_occds}
\keyword{der_occds}
