% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_trtsdtm.R
\name{derive_var_trtsdtm}
\alias{derive_var_trtsdtm}
\title{Derive datetime of First Exposure to Treatment (TRTSDTM)}
\usage{
derive_var_trtsdtm(dataset, dataset_ex, filter_ex = exprs(EXDOSE > 0 |
  (EXDOSE == 0 & str_detect(EXTRT, "PLACEBO"))))
}
\arguments{
\item{dataset}{Input dataset

  The variables specified by the `by_vars` parameter are expected.}

\item{dataset_ex}{`ex` dataset

  The variables `EXSTDTC`, `EXSEQ`, and those specified by the `filter_ex`
  parameter are expected.}

\item{filter_ex}{Fiter condition for the ex dataset

  Only observations of the ex dataset which fulfill the specified condition
  are considered for the treatment start date.

  Default: `exprs(EXDOSE > 0 | (EXDOSE == 0 & str_detect(EXTRT, 'PLACEBO'))`

  Permitted Values: logical expression}
}
\value{
The input dataset with `TRTSDTM` variable added
}
\description{
Derives datetime of First Exposure to Treatment (TRTSDTM)
}
\details{
For each group (with respect to the variables specified for the
  `by_vars` parameter) the first observation (with respect to the order
  specified for the `order` parameter) is included in the output dataset.
}
\examples{
data("ex")
data("dm")

derive_var_trtsdtm(dm)

}
\author{
Stefan Bundfuss
}
