% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_adae_astdtm.R
\name{derive_vars_adae_astdtm}
\alias{derive_vars_adae_astdtm}
\title{Derive ASTDTM and Flags for ADAE}
\usage{
derive_vars_adae_astdtm(dataset, drop_empty_flag = F)
}
\arguments{
\item{dataset}{Input AE dataset with derived TRTSDTM and AENDTM and source AESTDTC}

\item{drop_empty_flag}{Flag to indicate whether the date/time imputation flag(s)
should be dropped in case all values are missing. Possible Options are T(RUE) and
F(ALSE). The default is F}
}
\value{
The input AE dataset with the datetime 'ASTDTM' and the date/time imputation
flags 'ASTDTF' and 'ASTTMF' added.
}
\description{
This function derives ADAE Start Date/Time Variable (ADAE.ASTDTM) and imputation flags
(ADAE.ASTDTF and ADAE.ASTTMF). Admiral support funciton derive_vars_dtm is
used for initial date/time imputation. \cr
The derivation is based on the default specifications for ADAE.ASTDTM from the GDSR
(Version 2021-Oct.v1: Shenzou12). \cr
The following variables have to be available in the input data:
AESTDTC, TRTSDTM and AENDTM
}
\details{
The presence of 'ASTDTM', 'ASTDTF' and 'ASTTMF' is checked and a warning
is issued when any of them is found. The existing variables will be overwritten.
}
\examples{
\dontrun{
adae <- tibble::tribble(
  ~AESTDTC, ~TRTSDTM,  ~AENDTM,
  "2020-07-02T22:10:09", ymd_hms("2020-12-06T12:12:12"), ymd_hms("2020-07-08T12:12:12"),
  "2020-07-02T22:10", ymd_hms("2020-12-06T12:12:12"), ymd_hms("2020-07-08T12:12:12"),
  "2020-12", ymd_hms("2020-12-06T12:12:12"), ymd_hms("2020-12-06T12:12:12"),
  "2020", ymd_hms("2020-12-06T12:12:12"), ymd_hms("2020-12-06T12:12:12")
)

# Run function with default parameter setting
derive_vars_adae_astdtm(adae) 

# Run function and drop empty imputation flags
derive_vars_adae_astdtm(adae, drop_empty_flag = T) 
}

}
\author{
Sandra Ott
}
\keyword{adae}
\keyword{adam}
\keyword{derivation}
\keyword{timing}
