% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_atoxgr.R
\name{derive_vars_atoxgr}
\alias{derive_vars_atoxgr}
\title{Derive Lab High toxicity grade 0 - 4 and Low toxicity grades 0 - (-4)}
\usage{
derive_vars_atoxgr(
  dataset,
  new_var = ATOXGR,
  lotox_description_var = ATOXDSCL,
  lotox_var = ATOXGRL,
  hitox_description_var = ATOXDSCH,
  hitox_var = ATOXGRH
)
}
\arguments{
\item{dataset}{Input data set

The columns specified by \code{lotox_description_var}, \code{lotox_var},
\code{hitox_description_var} and \code{hitox_var} parameters are expected.}

\item{new_var}{Name of the character grade variable to create. Will contain
values "-4", "-3", "-2", "-1" for low values and "1", "2", "3", "4" for high values.
Contains 0 if value is gradable and does not satisfy any of the criteria for high or low
values. Is set to missing if information not available to give a grade.}

\item{lotox_description_var}{Variable containing the description of the grading
criteria eg. Anemia. Grade description for low values.}

\item{lotox_var}{Variable containing the grading for low values.}

\item{hitox_description_var}{Variable containing the description of the grading
criteria eg. Hemoglobin Increased. Grade description for high values.}

\item{hitox_var}{Variable containing the grading for high values.}
}
\value{
The input data set with the character grades (low and high) added
}
\description{
Derives character variable containing lab grade based on high and low
severity/toxicity grade(s).
}
\examples{
library(dplyr, warn.conflicts = FALSE)

data <- tibble::tribble(
  ~ATOXDSCL,          ~ATOXDSCH,        ~ATOXGRL,      ~ATOXGRH,
  "Hypoglycemia",     "Hyperglycemia",  NA_character_, "0",
  "Hypoglycemia",     "Hyperglycemia",  "0",           "1",
  "Hypoglycemia",     "Hyperglycemia",  "0",           "0",
  NA_character_,      "INR Increased",  NA_character_, "0",
  "Hypophosphatemia", NA_character_,    "1",           NA_character_,
)

derive_var_atoxgr(adlb)
}
\author{
Gordon Miller
}
\keyword{adam}
\keyword{adlb}
\keyword{bds}
\keyword{derivation}
