% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dtm_to_dt.R
\name{derive_vars_dtm_to_dt}
\alias{derive_vars_dtm_to_dt}
\title{Derive date variables from datetime variables}
\usage{
derive_vars_dtm_to_dt(dataset, source_vars)
}
\arguments{
\item{`dataset`}{Input dataset}

\item{`source_vars`}{A list of datetime variables from which dates are to be extracted}
}
\value{
A data frame containing all observations and variables of the input
dataset and adds the corresponding date variable of all specified datetime variables from
the source_vars option with the correct name.
}
\description{
This function creates a date as output from a datetime variable
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(lubridate, warn.conflicts = FALSE)

adcm <- tibble::tribble(
  ~STUDYID, ~USUBJID, ~TRTSDTM,              ~ASTDTM,               ~AENDTM,               ~ASTTMF,
  "TEST01", "PAT01",  "2012-02-25 23:00:00", "2012-02-28 19:00:00", "2012-02-25 23:00:00", NA,
  "TEST01", "PAT01",  NA,                    "2012-02-28 19:00:00", NA,                    NA,
  "TEST01", "PAT01",  "2017-02-25 23:00:00", "2013-02-25 19:00:00", "2014-02-25 19:00:00", NA,
  "TEST01", "PAT01",  "2017-02-25 16:00:00", "2017-02-25 14:00:00", "2017-03-25 23:00:00", "M",
  "TEST01", "PAT01",  "2017-02-25 16:00:00", "2017-02-25 14:00:00", "2017-04-29 14:00:00", NA
) \%>\%
  mutate(
    TRTSDTM = as_datetime(TRTSDTM),
    ASTDTM = as_datetime(ASTDTM),
    AENDTM = as_datetime(AENDTM)
  )

derive_vars_dtm_to_dt(adcm, vars(TRTSDTM, ASTDTM, AENDTM))
}
\seealso{

}
\author{
Teckla Akinyi
}
\keyword{ADaM}
\keyword{Date}
\keyword{Timing}
