% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_extreme_date.R
\name{derive_vars_extreme_dt}
\alias{derive_vars_extreme_dt}
\title{Derive Date of Any Extreme Dates}
\usage{
derive_vars_extreme_dt(
  dataset,
  subject_keys = vars(STUDYID, USUBJID),
  dataset_date,
  filter_dataset_date = NULL,
  order,
  new_var_prefix,
  dtc,
  date_imputation = NULL,
  flag_imputation = "AUTO",
  min_dates = NULL,
  max_dates = NULL,
  mode
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{order} and the \code{by_vars} parameter are
expected.}

\item{subject_keys}{Variables to uniquely identify a subject

The variables specified by the \code{subject_keys} parameter are expected.

Default: vars(STUDYID,USUBJID)}

\item{dataset_date}{dataset that contains the date values e.g. \code{ex}

The variables \code{--ENDTC}, \code{--SEQ}, and those specified by the \code{filter_dataset_date},
\code{dtc} and \code{subject_keys} parameters are expected.}

\item{filter_dataset_date}{Filter condition for the dataset with dates

Only observations of the dataset which fulfill the specified condition
are considered for the treatment start date.

Default: NULL

Permitted Values: logical expression

A list of quosures where the expressions are symbols as returned by
\code{vars()} is expected.}

\item{order}{Sort order

Within each by group the observations are ordered by the specified order.

Permitted Values: list of variables or functions of variables}

\item{new_var_prefix}{name of the extreme date derived
e.g.Treatment start and end date \code{TRTST}, \code{TRT01ED}
Phase start and end date \code{AP01ST}, \code{AP01ED}}

\item{dtc}{name of variable from which the extreme date is derived e.g. \code{EXENDTC}}

\item{date_imputation}{The value to impute the day/month when a datepart is
missing.

If \code{NULL}: no date imputation is performed and partial dates are returned as
missing.

Otherwise, a character value is expected, either as a
\itemize{
\item format with month and day specified as 'mm-dd': e.g. \code{"06-15"} for the 15th
of June,
\item or as a keyword: \code{"FIRST"}, \code{"MID"}, \code{"LAST"} to impute to the first/mid/last
day/month.
}

Default is \code{NULL}.}

\item{min_dates}{Minimum dates

A list of dates is expected. It is ensured that the imputed date is not
before any of the specified dates, e.g., that the imputed adverse event start
date is not before the first treatment date. Only dates which are in the
range of possible dates of the dtc value are considered. The possible dates
are defined by the missing parts of the dtc date (see example below). This
ensures that the non-missing parts of the dtc date are not changed. For
example\preformatted{impute_dtc(
  "2020-11",
  min_dates = list(ymd_hms("2020-12-06T12:12:12"),
                   ymd_hms("2020-11-11T11:11:11")),
  date_imputation = "first"
)
}

returns \code{"2020-11-11T11:11:11"} because the possible dates for \code{"2020-11"}
range from \code{"2020-11-01T00:00:00"} to \code{"2020-11-30T23:59:59"}. Therefore
\code{"2020-12-06T12:12:12"} is ignored. Returning \code{"2020-12-06T12:12:12"} would
have changed the month although it is not missing (in the dtc date).}

\item{max_dates}{Maximum dates

A list of dates is expected. It is ensured that the imputed date is not after
any of the specified dates, e.g., that the imputed date is not after the data
cut off date. Only dates which are in the range of possible dates are
considered.}

\item{mode}{Selection mode (first or last)

If \code{"first"} is specified, the first observation of each by group is
included in the output dataset. If \code{"last"} is specified, the last
observation of each by group is included in the output dataset.

Permitted Values:  \code{"first"}, \code{"last"}}
}
\value{
The input dataset with extreme datetime variables (and date imputation flags) added
e.g. Treatment start datetime and associated flags \code{TRTSTDT}, \code{TRTSTDTF}
Phase start datetime and associated flags \code{APO1STDTM}, \code{APO1STDTF}
}
\description{
Derives date of any extreme dates
e.g. Treatment, Phase or Period start or end dates
}
\details{
For each group (with respect to the variables specified for the
\code{by_vars} parameter) the first observation (with respect to the order
specified for the \code{order} parameter) is included in the output dataset.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
data("ex")
data("dm")

dm \%>\%
 derive_vars_extreme_date(subject_keys=vars(STUDYID,USUBJID),
                          dataset_date = ex,
                          filter_dataset_date = EXDOSE>0,
                          dtc = EXSTDTC,
                          order = vars(EXSTDTC,EXSEQ),
                          new_var_prefix=TRTEDTM,
                          date_imputation = "first",
                          time_imputation = "first",
                          flag_imputation = "AUTO",
                          min_dates = NULL,
                          max_dates = NULL,
                          mode="LAST") \%>\%
  select(USUBJID, TRTEDTM)
}
\author{
Teckla Akinyi
}
\keyword{adsl}
\keyword{derivation}
\keyword{timing}
