% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_merged.R
\name{derive_vars_merged_dtm}
\alias{derive_vars_merged_dtm}
\title{Merge a (imputed) Datetime Variable}
\usage{
derive_vars_merged_dtm(
  dataset,
  dataset_add,
  by_vars,
  order,
  new_vars_prefix,
  filter_add,
  mode,
  dtc,
  date_imputation = NULL,
  time_imputation = "00:00:00",
  flag_imputation = "auto",
  min_dates = NULL,
  max_dates = NULL
)
}
\description{
Merge a (imputed) Datetime Variable
}
\examples{
library(admiral.test)
data("dm")
data("ex")
derive_vars_merged_dtm(
  select(dm, STUDYID, USUBJID),
  dataset_add = ex,
  by_vars = vars(STUDYID, USUBJID),
  new_vars_prefix = "TRTS",
  date = EXSTDTC,
  date_imputation = "first",
  time_imputation = "first",
  order = vars(TRTSDTM),
  mode = "first"
)

derive_vars_merged_dtm(
  select(dm, STUDYID, USUBJID),
  dataset_add = ex,
  by_vars = vars(STUDYID, USUBJID),
  new_vars_prefix = "TRTE",
  date = EXENDTC,
  date_imputation = "last",
  time_imputation = "last",
  order = vars(TRTEDTM),
  mode = "last"
)
}
