% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_suppqual.R
\name{derive_vars_suppqual}
\alias{derive_vars_suppqual}
\title{Join Supplementary Qualifier Variables into the Parent SDTM Domain}
\usage{
derive_vars_suppqual(dataset, dataset_suppqual, domain = NULL)
}
\arguments{
\item{dataset}{A SDTM domain data set.}

\item{dataset_suppqual}{A Supplemental Qualifier (SUPPQUAL) data set.}

\item{domain}{Two letter domain value. Used when supplemental data set is
common across multiple SDTM domain.}
}
\value{
A data frame with SUPPQUAL variables appended to parent data set.
}
\description{
\ifelse{html}{\figure{lifecycle-deprecated.svg}{options: alt='Deprecated lifecycle'}}{\strong{Deprecated}}

\emph{Deprecated}, please use \code{metatools::combine_supp()} instead.

The SDTM does not allow any new variables beside ones assigned to each SDTM
domain. So, Supplemental Qualifier is introduced to supplement each SDTM
domain to contain non standard variables. \code{dataset_suppqual} can be either
a single SUPPQUAL dataset or separate supplementary data sets (SUPP) such
as SUPPDM, SUPPAE, and SUPPEX. When a \code{dataset_suppqual} is a single SUPPQUAL
dataset, specify two character\code{domain} value.

\code{derive_vars_suppqual()} expects \code{USUBJID}, \code{RDOMAIN}, \code{IDVAR}, \code{IDVARVAL},
\code{QNAM}, \code{QLABEL}, and \code{QVAL} variables to exist in \code{dataset_suppqual}.
}
\author{
Vignesh Thanikachalam
}
