% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_trt_start.R
\name{derive_vars_trt_start}
\alias{derive_vars_trt_start}
\title{Derive Datetime of Any Treatment Start Date}
\usage{
derive_vars_trt_start(
  dataset,
  dataset_ex,
  filter_ex = NULL,
  subject_keys = vars(STUDYID, USUBJID),
  new_var,
  dtc,
  date_imputation,
  time_imputation,
  flag_imputation = "AUTO",
  min_dates = NULL,
  max_dates = NULL,
  order,
  mode
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{subject_keys} parameter are expected.

Default: vars(STUDYID,USUBJID)}

\item{dataset_ex}{\code{ex} dataset

The variables \code{EXSTDTC}, \code{EXSEQ}, and those specified by the \code{filter_ex}
\code{dtc} and \code{subject_keys} parameters are expected.}

\item{filter_ex}{Filter condition for the ex dataset

Only observations of the ex dataset which fulfill the specified condition
are considered for the treatment start date.

Default: NULL

Permitted Values: logical expression}

\item{subject_keys}{Variables to uniquely identify a subject

A list of quosures where the expressions are symbols as returned by
\code{vars()} is expected.}

\item{new_var}{name of treatment start date e.g. \code{TRTSDTM}}

\item{dtc}{name of variable from which start date is derived e.g. \code{EXSTDTC}}

\item{date_imputation}{The value to impute the day/month when a datepart is
missing.

If \code{NULL}: no date imputation is performed and partial dates are returned as
missing.

Otherwise, a character value is expected, either as a
\itemize{
\item format with month and day specified as 'mm-dd': e.g. \code{"06-15"} for the 15th
of June,
\item or as a keyword: \code{"FIRST"}, \code{"MID"}, \code{"LAST"} to impute to the first/mid/last
day/month.
}

Default is \code{NULL}.}

\item{time_imputation}{The value to impute the time when a timepart is
missing.

A character value is expected, either as a
\itemize{
\item format with hour, min and sec specified as 'hh:mm:ss': e.g. \code{"00:00:00"}
for the start of the day,
\item or as a keyword: \code{"FIRST"},\code{"LAST"} to impute to the start/end of a day.
}

Default is \code{"00:00:00"}.}

\item{min_dates}{Minimum dates

A list of dates is expected. It is ensured that the imputed date is not
before any of the specified dates, e.g., that the imputed adverse event start
date is not before the first treatment date. Only dates which are in the
range of possible dates of the dtc value are considered. The possible dates
are defined by the missing parts of the dtc date (see example below). This
ensures that the non-missing parts of the dtc date are not changed. For
example\preformatted{impute_dtc(
  "2020-11",
  min_dates = list(ymd_hms("2020-12-06T12:12:12"),
                   ymd_hms("2020-11-11T11:11:11")),
  date_imputation = "first"
)
}

returns \code{"2020-11-11T11:11:11"} because the possible dates for \code{"2020-11"}
range from \code{"2020-11-01T00:00:00"} to \code{"2020-11-30T23:59:59"}. Therefore
\code{"2020-12-06T12:12:12"} is ignored. Returning \code{"2020-12-06T12:12:12"} would
have changed the month although it is not missing (in the dtc date).}

\item{max_dates}{Maximum dates

A list of dates is expected. It is ensured that the imputed date is not after
any of the specified dates, e.g., that the imputed date is not after the data
cut off date. Only dates which are in the range of possible dates are
considered.}
}
\value{
The input dataset with start date variable added e.g. \code{TRTSDTM}
}
\description{
Derives datetime of first exposure to treatment e.g. (\code{TRTSDTM})
}
\details{
For each group (with respect to the variables specified for the
\code{by_vars} parameter) the first observation (with respect to the order
specified for the \code{order} parameter) is included in the output dataset.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
data("ex")
data("dm")

dm \%>\%
 derive_vars_trt_start(dataset_ex = ex,
                       filter_ex = EXDOSE>0 ,
                       subject_keys=vars(STUDYID,USUBJID),
                       new_var=TRTSDTM,
                       dtc=EXSTDTC,
                       date_imputation = "first",
                       time_imputation="first",
                       flag_imputation = "AUTO",
                       min_dates = NULL,
                       max_dates = NULL,
                       order = vars(EXSTDTC,EXSEQ),
                       mode="first") \%>\%
 select(USUBJID, TRTSDTM)
}
\author{
Stefan Bundfuss, Teckla Akinyi
}
\keyword{adsl}
\keyword{derivation}
\keyword{timing}
