% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{arg_name}
\alias{arg_name}
\title{Extract Argument Name from an Expression}
\usage{
arg_name(expr)
}
\arguments{
\item{expr}{An expression created inside a function using \code{substitute()}}
}
\description{
Extract Argument Name from an Expression
}
\author{
Thomas Neitmann, Ondrej Slama
}
\keyword{dev_utility}
