% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote.R
\name{backquote}
\alias{backquote}
\title{Wrap a String in Backquotes}
\usage{
backquote(x)
}
\arguments{
\item{x}{A \code{character} vector}
}
\description{
Wrap a String in Backquotes
}
\seealso{
Helpers for working with Quosures and Quoting: 
\code{\link{dquote}()},
\code{\link{enumerate}()},
\code{\link{quo_c}()},
\code{\link{squote}()}
}
\author{
Thomas Neitmann
}
\concept{quo}
\keyword{quo}
