% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{get_source_vars}
\alias{get_source_vars}
\title{Get Source Variables from a List of Quosures}
\usage{
get_source_vars(quosures)
}
\arguments{
\item{quosures}{A list of quosures}
}
\value{
A list of quosures
}
\description{
Get Source Variables from a List of Quosures
}
\examples{
admiral:::get_source_vars(vars(USUBJID, AVISIT = VISIT, SRCDOM = "EX"))
}
\author{
Stefan Bundfuss
}
\keyword{dev_utility}
