% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Negated Value Matching}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{The values to be matched}

\item{table}{The values to be matched against}
}
\description{
Returns a \code{logical} vector indicating if there is \emph{no} match of the
left operand in the right operand.
}
\examples{
`\%notin\%` <- admiral:::`\%notin\%`
"a" \%notin\% c("b", "v", "k")
}
\author{
Thomas Neitmann
}
\keyword{dev_utility}
