% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote.R
\name{squote}
\alias{squote}
\title{Wrap a String in Single Quotes}
\usage{
squote(x)
}
\arguments{
\item{x}{A \code{character} vector}
}
\description{
Wrap a String in Single Quotes
}
\seealso{
Helpers for working with Quosures and Quoting: 
\code{\link{backquote}()},
\code{\link{dquote}()},
\code{\link{enumerate}()},
\code{\link{quo_c}()}
}
\author{
Thomas Neitmann
}
\concept{quo}
\keyword{quo}
