% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what.R
\name{what_is_it}
\alias{what_is_it}
\title{What Kind of Object is This?}
\usage{
what_is_it(x)
}
\arguments{
\item{x}{Any R object}
}
\description{
Returns a string describing what kind of object the input is.
}
\author{
Thomas Neitmann
}
\concept{what}
\keyword{what}
