% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_single_dose_dataset.R
\docType{data}
\name{dose_freq_lookup}
\alias{dose_freq_lookup}
\title{Pre-Defined Dose Frequencies}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 33 rows and 5 columns.
}
\usage{
dose_freq_lookup
}
\description{
These pre-defined dose frequencies are sourced from
\href{https://evs.nci.nih.gov/ftp1/CDISC/SDTM/SDTM\%20Terminology.pdf}{CDISC}. The
number of rows to generate using \code{create_single_dose_dataset()} arguments
\code{start_date} and \code{end_date} is derived from \code{DFFactor} and \code{DFUnit} with
appropriate functions from \code{lubridate}.
}
\details{
To see the entire table in the console, run \code{print(dose_freq_lookup)}.
}
\seealso{
\code{\link[=create_single_dose_dataset]{create_single_dose_dataset()}}
}
\keyword{dose_freq}
