% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{dquote}
\alias{dquote}
\title{Wrap a String in Double Quotes}
\usage{
dquote(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
If the input is \code{NULL}, the text \code{"NULL"} is returned. Otherwise, the
input in double quotes is returned.
}
\description{
Wrap a string in double quotes, e.g., for displaying character values in
messages.
}
\examples{
dquote("foo")
dquote(NULL)
}
\seealso{
Functions moved to admiraldev package: 
\code{\link{arg_name}()},
\code{\link{assert_db_requirements}()},
\code{\link{assert_terms}()},
\code{\link{assert_valid_queries}()},
\code{\link{backquote}()},
\code{\link{contains_vars}()},
\code{\link{enumerate}()},
\code{\link{extend_source_datasets}()},
\code{\link{extract_vars}()},
\code{\link{format.sdg_select}()},
\code{\link{format.smq_select}()},
\code{\link{get_constant_vars}()},
\code{\link{get_duplicates}()},
\code{\link{get_source_vars}()},
\code{\link{is_auto}()},
\code{\link{quo_c}()},
\code{\link{replace_values_by_names}()},
\code{\link{squote}()},
\code{\link{valid_time_units}()},
\code{\link{validate_query}()},
\code{\link{validate_sdg_select}()},
\code{\link{validate_smq_select}()},
\code{\link{what_is_it}()}
}
\author{
Stefan Bundfuss
}
\concept{move_adm_dev}
\keyword{move_adm_dev}
