% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_fcn.R
\name{dummy_function}
\alias{dummy_function}
\title{Dummy Function Example}
\usage{
dummy_function(a, b)
}
\arguments{
\item{a}{A numeric vector.

\describe{
\item{Default value}{none}
}}

\item{b}{A numeric vector.

\describe{
\item{Default value}{none}
}}
}
\value{
A numeric vector containing the element-wise sum of \code{a} and \code{b}.
}
\description{
This dummy function demonstrates how to document an R function using roxygen2.
It takes two numeric inputs and returns their element-wise sum. Make it pretty!!
Making it more pretty!! TEST!! More TESTs!! ??
}
\examples{
# Basic addition with scalars
dummy_function(1, 2)

# Element-wise addition with vectors
dummy_function(c(1, 2, 3), c(4, 5, 6))

}
