% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eg}
\alias{eg}
\title{Electrocardiogram Dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26669 rows and 24 columns.}
\usage{
eg
}
\description{
An example of standard SDTM EG dataset to be used in deriving ADEG dataset
}
\keyword{datasets}
