% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{enumerate}
\alias{enumerate}
\title{Enumerate Multiple Strings}
\usage{
enumerate(x, quote_fun = backquote, conjunction = "and")
}
\arguments{
\item{x}{A \code{character} vector}

\item{quote_fun}{Quoting function, defaults to \code{backquote}.}

\item{conjunction}{Character to be used in the message, defaults to "and".}
}
\description{
Enumerate Multiple Strings
}
\examples{
enumerate(c("STUDYID", "USUBJID", "PARAMCD"))
enumerate(letters[1:6], quote_fun = squote)
enumerate(
  c("date", "time", "both"),
  quote_fun = squote,
  conjunction = "or"
)
}
\author{
Thomas Neitmann
}
\keyword{move_adm_dev}
