% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex}
\alias{ex}
\title{Exposure Dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 591 rows and 17 columns.}
\source{
\url{https://bitbucket.cdisc.org/projects/CED/repos/sdtm-adam-pilot-project/browse/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/ex.xpt}
}
\usage{
ex
}
\description{
A SDTM EX dataset from the CDISC pilot project
}
\keyword{datasets}
