% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_single}
\alias{ex_single}
\title{Single Dose Exposure Dataset (Removed)}
\format{
An object of class \code{data.frame} with 0 rows and 0 columns.
}
\usage{
ex_single
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

The \code{ex_single} dataset has been removed from \code{{admiral}} to reduce package
size. Derive it on the fly using \code{\link[=create_single_dose_dataset]{create_single_dose_dataset()}}.
See \code{vignette("occds", package = "admiral")} for a full example.
}
\seealso{
Other deprecated: 
\code{\link{call_user_fun}()},
\code{\link{date_source}()},
\code{\link{derive_param_extreme_record}()},
\code{\link{derive_var_dthcaus}()},
\code{\link{derive_var_extreme_dt}()},
\code{\link{derive_var_extreme_dtm}()},
\code{\link{derive_var_merged_summary}()},
\code{\link{dthcaus_source}()},
\code{\link{get_summary_records}()}
}
\concept{deprecated}
\keyword{deprecated}
