% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_qs}
\alias{example_qs}
\title{Example \code{QS} Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 161 rows and 11 columns.
}
\source{
Created by (\url{https://github.com/pharmaverse/admiral/blob/main/inst/example_scripts/example_qs.R})
}
\usage{
example_qs
}
\description{
An example \code{QS} dataset based on the examples from the CDISC ADaM Supplements
\href{https://www.cdisc.org/standards/foundational/qrs/generalized-anxiety-disorder-7-item-version-2-0}{Generalized Anxiety Disorder 7-Item Version 2 (GAD-7)}
and \href{https://www.cdisc.org/standards/foundational/qrs/geriatric-depression-scale-short-form-0}{Geriatric Depression Scale Short Form (GDS-SF)}.
}
\seealso{
Other datasets: 
\code{\link{admiral_adlb}},
\code{\link{admiral_adsl}},
\code{\link{ex_single}},
\code{\link{queries}},
\code{\link{queries_mh}}
}
\concept{datasets}
\keyword{datasets}
