% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_dfs_equal.R
\name{expect_equal_tbl}
\alias{expect_equal_tbl}
\title{Expectation: Are Table Lists are Equal?}
\usage{
expect_equal_tbl(object, expected, ..., info = NULL)
}
\arguments{
\item{base}{Input dataset}

\item{compare}{Comparison dataset}
}
\value{
An error if \code{base} and \code{compare} do not match or \code{NULL} invisibly if they do
}
\description{
Uses \code{\link[diffdf:diffdf]{diffdf::diffdf()}} to compares 2 datasets for any differences
}
\seealso{
Helps with Testing: 
\code{\link{expect_dfs_equal}()}
}
\author{
Ben Straub
}
\concept{test_helper}
\keyword{test_helper}
