% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_cat.R
\name{extend_condition}
\alias{extend_condition}
\title{Extend a condition string by adding a new condition based on a variable and its value}
\usage{
extend_condition(cond, var, is)
}
\arguments{
\item{cond}{A character string representing an existing condition.}

\item{var}{A character string representing the name of the variable to check.}

\item{is}{A character string representing the value the variable should be equal to.}
}
\value{
A character string representing the extended condition.
}
\description{
This internal helper function extends a condition string by appending a new condition
that checks if a variable equals a specific value.
}
\examples{
# Extend an existing condition to include a check for 'AGE == "30"'
admiral:::extend_condition("SEX == 'M'", "AGE", "30")
}
\keyword{internal}
