% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_vars}
\alias{extract_vars}
\title{Extract All Symbols from a List of Quosures}
\usage{
extract_vars(quosures)
}
\arguments{
\item{quosures}{A list of quosures, e.g. created using \code{vars()}}
}
\description{
Extract All Symbols from a List of Quosures
}
\examples{
admiral:::extract_vars(vars(STUDYID, USUBJID, desc(ADTM)))
}
\author{
Thomas Neitmann
}
\keyword{dev_utility}
