% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_first.R
\name{filter_first}
\alias{filter_first}
\title{Filter out the first observation for each group}
\usage{
filter_first(dataset, order, by_vars)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{order} and the \code{by_vars} parameter are
expected.}

\item{order}{Sort order

Within each by group the observations are ordered by the specified order.

Permitted Values: list of variables or functions of variables}

\item{by_vars}{Grouping variables}
}
\value{
A dataset containing the first observation of each by group
}
\description{
Filters out the first observation for each group.
}
\details{
For each group (with respect to the variables specified for the
\code{by_vars} parameter) the first observation (with respect to the order
specified for the \code{order} parameter) is included in the output dataset.

Permitted Values: list of variables
}
\examples{
data("ex")

# selecting first dose for each patient
filter_first(ex,
             order = exprs(EXSEQ),
             by_vars = exprs(USUBJID))

# selecting highest dose for each patient
filter_first(ex,
             order = exprs(desc(EXDOSE)),
             by_vars = exprs(USUBJID))

}
\seealso{
Other {general functions}: \code{\link{derive_merged_vars}}
}
\author{
Stefan Bundfuss
}
\concept{{general functions}}
