% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query_data.R
\name{format.smq_select}
\alias{format.smq_select}
\title{Returns a Character Representation of a \code{smq_select()} Object}
\usage{
\method{format}{smq_select}(x, ...)
}
\arguments{
\item{x}{A \code{smq_select()} object}

\item{...}{Not used}
}
\value{
A character representation of the \code{smq_select()} object
}
\description{
The function returns a character representation of a \code{smq_select()} object.
It can be used for error messages for example.
}
\examples{

format(smq_select(id = 42, scope = "NARROW"))
}
\seealso{
\code{\link[=smq_select]{smq_select()}}

Source Specifications: 
\code{\link{assert_db_requirements}()},
\code{\link{assert_terms}()},
\code{\link{assert_valid_queries}()},
\code{\link{censor_source}()},
\code{\link{date_source}()},
\code{\link{death_event}},
\code{\link{derive_var_dthcaus}()},
\code{\link{event_source}()},
\code{\link{extend_source_datasets}()},
\code{\link{filter_date_sources}()},
\code{\link{format.sdg_select}()},
\code{\link{list_tte_source_objects}()},
\code{\link{params}()},
\code{\link{query}()},
\code{\link{sdg_select}()},
\code{\link{smq_select}()},
\code{\link{tte_source}()},
\code{\link{validate_query}()},
\code{\link{validate_sdg_select}()},
\code{\link{validate_smq_select}()}
}
\author{
Stefan Bundfuss
}
\concept{source_specifications}
\keyword{source_specifications}
