% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dt_dtm_utils.R
\name{format_imputed_dtc}
\alias{format_imputed_dtc}
\title{Format Imputed Date/Datetime}
\usage{
format_imputed_dtc(imputed, is_datetime)
}
\arguments{
\item{imputed}{A list of imputed date/time components.}

\item{is_datetime}{A logical indicating whether it's a datetime format.}
}
\value{
A character vector of formatted date/datetime strings.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This is a helping function for \code{get_dt_dtm_range()}
Formats imputed date/datetime components into a string representation.
}
\details{
The function formats the imputed components into "YYYY-MM-DD" for dates
and "YYYY-MM-DDTHH:MM:SS" for datetimes. It replaces any string containing
'x' with NA.
}
\keyword{internal}
