% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_disposition_reason.R
\name{format_reason_default}
\alias{format_reason_default}
\title{Default Format for the Disposition Reason}
\usage{
format_reason_default(reason, reason_spe = NULL)
}
\arguments{
\item{reason}{the disposition variable used for the mapping (e.g. \code{DSDECOD}).}

\item{reason_spe}{the disposition variable used for the mapping of the details
if required (e.g. \code{DSTERM}).}
}
\value{
A \code{character} vector
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is \emph{deprecated}. This function is a default for \code{derive_vars_disposition_reason()}
for the \code{format_new_vars} argument. Please use \code{derive_vars_merged()} and
specify the \code{filter_add} argument to derive the respective variables.

Define a function to map the disposition reason, to be used as a parameter in
\code{derive_vars_disposition_reason()}.
}
\details{
\code{format_reason_default(DSDECOD)} returns \code{DSDECOD} when \code{DSDECOD} is not \code{'COMPLETED'} nor \code{NA}.
\cr\code{format_reason_default(DSDECOD, DSTERM)} returns \code{DSTERM} when \code{DSDECOD} is
equal to \code{'OTHER'}.
\cr Usually this function can not be used with \verb{\%>\%}.
}
\seealso{
\code{\link[=derive_vars_disposition_reason]{derive_vars_disposition_reason()}}

Other deprecated: 
\code{\link{derive_derived_param}()},
\code{\link{derive_param_first_event}()},
\code{\link{derive_var_agegr_fda}()},
\code{\link{derive_var_confirmation_flag}()},
\code{\link{derive_var_disposition_status}()},
\code{\link{derive_var_worst_flag}()},
\code{\link{derive_vars_disposition_reason}()},
\code{\link{derive_vars_merged_dtm}()},
\code{\link{derive_vars_merged_dt}()},
\code{\link{filter_confirmation}()},
\code{\link{format_eoxxstt_default}()},
\code{\link{vars}()}
}
\concept{deprecated}
\keyword{deprecated}
