% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat_friendly_type.R
\name{friendly_type_of}
\alias{friendly_type_of}
\title{Return English-friendly messaging for object-types}
\usage{
friendly_type_of(x, value = TRUE, length = FALSE)
}
\arguments{
\item{x}{Any R object.}

\item{value}{Whether to describe the value of \code{x}.}

\item{length}{Whether to mention the length of vectors and lists.}
}
\value{
A string describing the type. Starts with an indefinite
article, e.g. "an integer vector".
}
\description{
Return English-friendly messaging for object-types
}
\details{
This helper function aids us in forming user-friendly messages that gets
called through \code{what_is_it()}, which is often used in the assertion functions
to identify what object-type the user passed through an argument instead of
an expected-type.
}
\seealso{
Developer Utility Functions: 
\code{\link{\%notin\%}()},
\code{\link{\%or\%}()},
\code{\link{arg_name}()},
\code{\link{contains_vars}()},
\code{\link{convert_dtm_to_dtc}()},
\code{\link{extract_vars}()},
\code{\link{filter_if}()},
\code{\link{valid_time_units}()},
\code{\link{vars2chr}()}
}
\concept{dev_utility}
\keyword{dev_utility}
