% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiral_options.R
\name{get_admiral_options}
\alias{get_admiral_options}
\title{Call a package-standard input}
\usage{
get_admiral_options(input)
}
\arguments{
\item{input}{The function input necessary.}
}
\value{
Call on an admiral option to be called for function inputs: e.g \code{subject_keys}.
}
\description{
Call a package-standard input that can be modified for advanced users.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(admiral.test)
get_admiral_options(subject_keys)
}
\seealso{
\code{\link[=vars]{vars()}}

Other XXX: 
\code{\link{set_admiral_options}()}
}
\author{
Zelos Zhu
}
\concept{XXX}
\keyword{XXX}
