% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{get_constant_vars}
\alias{get_constant_vars}
\title{Get Constant Variables}
\usage{
get_constant_vars(dataset, by_vars, ignore_vars = NULL)
}
\arguments{
\item{dataset}{A data frame.}

\item{by_vars}{By variables
The groups defined by the by variables are considered separately. I.e., if
a variable is constant within each by group, it is returned.}

\item{ignore_vars}{Variables to ignore
The specified variables are not considered, i.e., they are not returned
even if they are constant (unless they are included in the by variables).

\emph{Permitted Values:} A list of variable names or selector function calls
like \code{starts_with("EX")}}
}
\value{
Variable vector.
}
\description{
Get Constant Variables
}
\examples{
library(admiral.test)
data(admiral_vs)

get_constant_vars(admiral_vs, by_vars = vars(USUBJID, VSTESTCD))

get_constant_vars(
  admiral_vs,
  by_vars = vars(USUBJID, VSTESTCD),
  ignore_vars = vars(DOMAIN, tidyselect::starts_with("VS"))
)
}
\seealso{
Functions moved to admiraldev package: 
\code{\link{arg_name}()},
\code{\link{assert_db_requirements}()},
\code{\link{assert_terms}()},
\code{\link{assert_valid_queries}()},
\code{\link{backquote}()},
\code{\link{contains_vars}()},
\code{\link{dquote}()},
\code{\link{enumerate}()},
\code{\link{extend_source_datasets}()},
\code{\link{extract_vars}()},
\code{\link{format.sdg_select}()},
\code{\link{format.smq_select}()},
\code{\link{get_duplicates}()},
\code{\link{get_source_vars}()},
\code{\link{is_auto}()},
\code{\link{quo_c}()},
\code{\link{replace_values_by_names}()},
\code{\link{squote}()},
\code{\link{valid_time_units}()},
\code{\link{validate_query}()},
\code{\link{validate_sdg_select}()},
\code{\link{validate_smq_select}()},
\code{\link{what_is_it}()}
}
\concept{move_adm_dev}
\keyword{move_adm_dev}
