% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_dataset}
\alias{get_dataset}
\title{Retrieve a Dataset from the \code{.datasets} environment}
\usage{
get_dataset(name)
}
\arguments{
\item{name}{The name of the dataset to retrieve}
}
\description{
Retrieve a Dataset from the \code{.datasets} environment
}
\author{
Thomas Neitmann
}
