% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dt_dtm_utils.R
\name{get_highest_imputation_level}
\alias{get_highest_imputation_level}
\title{Get Highest Imputation Level}
\usage{
get_highest_imputation_level(is_datetime, highest_imputation)
}
\arguments{
\item{is_datetime}{A logical indicating whether the imputation is for a datetime.

\describe{
\item{Default value}{none}
}}

\item{highest_imputation}{A character indicating the highest imputation level.

\describe{
\item{Default value}{none}
}}
}
\value{
An integer representing the highest imputation level.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This is a helping function for \code{get_dt_dtm_range()}
Determines the highest imputation level based on whether it's a date or datetime.
}
\keyword{internal}
