% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dt_dtm_utils.R
\name{get_imputation_target_date}
\alias{get_imputation_target_date}
\title{Get Date Imputation Targets}
\usage{
get_imputation_target_date(date_imputation, month)
}
\arguments{
\item{date_imputation}{The value to impute the day/month when a datepart is
missing.

A character value is expected, as a:
\itemize{
\item format with month and day specified as \code{"mm-dd"}: e.g. \code{"06-15"} for the 15th
of June,
\item format with day specified as \code{"dd"}: e.g. \code{"15"} for the 15th day of a month
\item or as a keyword: \code{"first"}, \code{"mid"}, \code{"last"} to impute to the first/mid/last
day/month.
}

\describe{
\item{Default value}{none}
}}

\item{month}{Month component of the partial date

\describe{
\item{Default value}{none}
}}
}
\value{
A list of character vectors. The elements of the list are named
"year", "month", "day".
}
\description{
Get Date Imputation Targets
}
\details{
\itemize{
\item For \code{date_imputation = "first"}, \code{"0000"}, \code{"01"}, \code{"01"} are returned.
\item For \code{date_imputation = "mid"}, \code{"xxxx"}, \code{"06"}, \code{"30"} if \code{month} is \code{NA}.
otherwise \code{"15"} returned.
\item For \code{date_imputation = "last"}, \code{"9999"}, \code{"12"}, \code{"28"} are returned.
\item For \code{date_imputation = "<mm>-<dd>"}, \code{"xxxx"}, \code{"<mm>"}, \code{"<dd>"} are returned.
\item For \code{date_imputation = "<dd>"}, \code{"xxxx"}, \code{"xx"}, \code{"<dd>"} are returned.
}

\code{"xxxx"} indicates that the component is undefined. If an undefined
component occurs in the imputed \code{--DTC} value, the imputed \code{--DTC} value is set to
\code{NA_character_} in the imputation functions.
}
\examples{
# Get imputation target for "first"
admiral:::get_imputation_target_date("first", month = NA)

# Get imputation target for "mid" with specified month
admiral:::get_imputation_target_date("mid", month = "03")

# Get imputation target for "mid" with NA month
admiral:::get_imputation_target_date("mid", month = NA)

# Get imputation target for "last"
admiral:::get_imputation_target_date("last", month = NA)

# Get imputation target for custom date imputation "06-15"
admiral:::get_imputation_target_date("06-15", month = NA)

# Get imputation target for custom date imputation "11"
admiral:::get_imputation_target_date("11", month = NA)

}
\seealso{
\code{\link[=impute_dtc_dtm]{impute_dtc_dtm()}}, \code{\link[=impute_dtc_dt]{impute_dtc_dt()}}

Utilities used for date imputation:
\code{\link[=dt_level]{dt_level()}},
\code{\link[=dtm_level]{dtm_level()}},
\code{\link[=get_imputation_target_time]{get_imputation_target_time()}},
\code{\link[=get_partialdatetime]{get_partialdatetime()}},
\code{\link[=restrict_imputed_dtc_dt]{restrict_imputed_dtc_dt()}},
\code{\link[=restrict_imputed_dtc_dtm]{restrict_imputed_dtc_dtm()}}
}
\concept{utils_impute}
\keyword{internal}
