% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dt_dtm_utils.R
\name{get_partialdate}
\alias{get_partialdate}
\title{Extract Partial Date Components}
\usage{
get_partialdate(dtc)
}
\arguments{
\item{dtc}{A character vector of date strings to be parsed.

\describe{
\item{Default value}{none}
}}
}
\value{
A list with three elements: "year", "month", and "day", each containing
a character vector of the respective date components.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function extracts year, month, and day components from a vector of
date strings.
}
\details{
The function uses regular expressions to extract date components. It handles
partial dates and propagates NA values for missing higher-order components.
}
\keyword{internal}
