% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dt_dtm_utils.R
\name{get_partialdatetime}
\alias{get_partialdatetime}
\title{Parse DTC variable and Determine Components}
\usage{
get_partialdatetime(dtc)
}
\arguments{
\item{dtc}{The \code{'--DTC'} date to parse

A character date is expected in a format like \code{yyyy-mm-dd} or
\code{yyyy-mm-ddThh:mm:ss}. Trailing components can be omitted and \code{-} is a
valid value for any component.

\describe{
\item{Default value}{none}
}}
}
\value{
A list of character vectors. The elements of the list are named
"year", "month", "day", "hour", "minute", and "second". Missing components
are set to \code{NA_character_}.
}
\description{
Parse DTC variable and Determine Components
}
\details{
The function can be replaced by the parttime parser once it is
available.
}
\examples{
# Get partial datetime components for a complete datetime string
dtc_complete <- admiral:::get_partialdatetime("2020-03-15T12:34:56")
print(dtc_complete)

# Get partial datetime components for a partial datetime string
dtc_partial <- admiral:::get_partialdatetime("2020-03-15T12:34")
print(dtc_partial)

# Get partial datetime components for a date-only string
dtc_date_only <- admiral:::get_partialdatetime("2020-03-15")
print(dtc_date_only)

# Get partial datetime components for an incomplete year string
dtc_year_partial <- admiral:::get_partialdatetime("2020")
print(dtc_year_partial)

}
\seealso{
\code{\link[=impute_dtc_dtm]{impute_dtc_dtm()}}, \code{\link[=impute_dtc_dt]{impute_dtc_dt()}}

Utilities used for date imputation: 
\code{\link{dt_level}()},
\code{\link{dtm_level}()},
\code{\link{get_imputation_target_date}()},
\code{\link{get_imputation_target_time}()},
\code{\link{restrict_imputed_dtc_dt}()},
\code{\link{restrict_imputed_dtc_dtm}()}
}
\concept{utils_impute}
\keyword{internal}
