% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary_records.R
\name{get_summary_records}
\alias{get_summary_records}
\title{Create Summary Records}
\usage{
get_summary_records(dataset, by_vars, filter = NULL, set_values_to = NULL)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{by_vars} argument are expected to be in the dataset.

\describe{
\item{Default value}{none}
}}

\item{by_vars}{Grouping variables

Variables to consider for generation of groupwise summary records.

\describe{
\item{Default value}{none}
}}

\item{filter}{Filter condition as logical expression to apply during
summary calculation. By default, filtering expressions are computed within
\code{by_vars} as this will help when an aggregating, lagging, or ranking
function is involved.

For example,
\itemize{
\item \code{filter_rows = (AVAL > mean(AVAL, na.rm = TRUE))} will filter all AVAL
values greater than mean of AVAL with in \code{by_vars}.
\item \code{filter_rows = (dplyr::n() > 2)} will filter n count of \code{by_vars} greater
than 2.
}

Comparing derived numeric variables to fixed values, e.g., \code{PCHG <= -90}, may give unexpected results due to floating point representation. For details and solutions see the "Floating Point Comparisons" section in \code{vignette("concepts_conventions")}.

\describe{
\item{Default value}{\code{NULL}}
}}

\item{set_values_to}{Variables to be set

The specified variables are set to the specified values for the new
observations.

Set a list of variables to some specified value for the new records
\itemize{
\item LHS refer to a variable.
\item RHS refers to the values to set to the variable. This can be a string, a
symbol, a numeric value, an expression or NA. If summary functions are
used, the values are summarized by the variables specified for \code{by_vars}.
}

For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  set_values_to = exprs(
    AVAL = sum(AVAL),
    PARAMCD = "TDOSE",
    PARCAT1 = "OVERALL"
  )
}\if{html}{\out{</div>}}

\describe{
\item{Default value}{\code{NULL}}
}}
}
\value{
A data frame of derived records.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} The \code{get_summary_records()} has been
deprecated in favor of \code{derive_summary_records()} (call it with the \code{dataset_add}
argument and without the \code{dataset} argument).

It is not uncommon to have an analysis need whereby one needs to derive an
analysis value (\code{AVAL}) from multiple records. The ADaM basic dataset
structure variable \code{DTYPE} is available to indicate when a new derived
records has been added to a dataset.
}
\details{
This function only creates derived observations and does not append them
to the original dataset observations. If you would like to this instead,
see the \code{derive_summary_records()} function.
}
\seealso{
\code{\link[=derive_summary_records]{derive_summary_records()}}, \code{\link[=derive_vars_merged_summary]{derive_vars_merged_summary()}}

Other deprecated:
\code{\link[=date_source]{date_source()}},
\code{\link[=derive_param_extreme_record]{derive_param_extreme_record()}},
\code{\link[=derive_var_dthcaus]{derive_var_dthcaus()}},
\code{\link[=derive_var_extreme_dt]{derive_var_extreme_dt()}},
\code{\link[=derive_var_extreme_dtm]{derive_var_extreme_dtm()}},
\code{\link[=derive_var_merged_summary]{derive_var_merged_summary()}},
\code{\link[=dthcaus_source]{dthcaus_source()}}
}
\concept{deprecated}
\keyword{deprecated}
