% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Negated Value Matching}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{The values to be matched}

\item{table}{The values to be matched against}
}
\value{
A \code{logical} vector
}
\description{
Returns a \code{logical} vector indicating if there is \emph{no} match of the
left operand in the right operand.
}
\seealso{
Developer Utility Functions: 
\code{\link{\%or\%}()},
\code{\link{arg_name}()},
\code{\link{as_name}()},
\code{\link{convert_dtm_to_dtc}()},
\code{\link{extract_vars}()},
\code{\link{filter_if}()},
\code{\link{negate_vars}()},
\code{\link{replace_values_by_names}()},
\code{\link{valid_time_units}()},
\code{\link{vars2chr}()}
}
\author{
Thomas Neitmann
}
\concept{dev_utility}
\keyword{dev_utility}
