% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{\%or\%}
\alias{\%or\%}
\title{Or Guy - Update}
\usage{
lhs \%or\% rhs
}
\arguments{
\item{lhs}{something}

\item{rhs}{something}
}
\value{
something
}
\description{
Or Guy - Update
}
\seealso{
Developer Utility Functions: 
\code{\link{\%notin\%}()},
\code{\link{arg_name}()},
\code{\link{as_name}()},
\code{\link{convert_dtm_to_dtc}()},
\code{\link{extract_vars}()},
\code{\link{filter_if}()},
\code{\link{negate_vars}()},
\code{\link{replace_values_by_names}()},
\code{\link{valid_time_units}()},
\code{\link{vars2chr}()}
}
\concept{dev_utility}
\keyword{dev_utility}
