% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{hello_admiral}
\alias{hello_admiral}
\title{Derive Extension Example #'
Says hello admiral}
\usage{
hello_admiral(hw = TRUE)
}
\arguments{
\item{hw}{TRUE or FALSE}
}
\value{
Happy Message
}
\description{
Derive Extension Example #'
Says hello admiral
}
\details{
Something Something
}
\examples{
hello_admiral(hw = FALSE)

}
\author{
Open Source
}
\concept{der_adxx}
\keyword{der_adxx}
