% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{if_non_len0}
\alias{if_non_len0}
\title{Return y if x has length > 0}
\usage{
if_non_len0(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\description{
Return y if x has length > 0
}
\examples{
if_non_len0(character(0), "whatever")
if_non_len0(character(1), "updated if length > 0")
}
