% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_date_vars.R
\name{impute_dtc}
\alias{impute_dtc}
\title{Impute partial date/time portion of a --DTC variable}
\usage{
impute_dtc(dtc, date_imputation = NULL, time_imputation = "00:00:00")
}
\arguments{
\item{dtc}{The --DTC date to impute

A character date is expected in a format like yyyy-mm-dd or yyyy-mm-ddThh:mm:ss.
If the year part is not recorded (missing date), no imputation is performed.}

\item{date_imputation}{The value to impute the day/month when a datepart is missing.

If NULL: no date imputation is performed and partial dates are returned as missing.

Otherwise, a character value is expected, either as a
\itemize{
\item format with day and month specified as 'dd-mm': e.g. '15-06' for the 15th of June,
\item or as a keyword: 'FIRST', 'MID', 'LAST' to impute to the first/mid/last day/month.
}

Default is NULL.}

\item{time_imputation}{The value to impute the time when a timepart is missing.

A character value is expected, either as a
\itemize{
\item format with hour, min and sec specified as 'hh:mm:ss': e.g. '00:00:00' for
the start of the day,
\item or as a keyword: 'FIRST','LAST' to impute to the start/end of a day.
}

Default is '00:00:00'.}
}
\value{
a character vector
}
\description{
Imputation partial date/time portion of a --DTC variable.
based on user input.
}
\examples{
dates <- c(
  "2019-07-18T15:25:40",
  "2019-07-18T15:25",
  "2019-07-18T15",
  "2019-07-18",
  "2019-02",
  "2019",
  "2019",
  "2019---07",
  ""
)

# No date imputation (date_imputation defaulted to NULL)
# Missing time part imputed with 00:00:00 portion by default
impute_dtc(dtc = dates)

# No date imputation (date_imputation defaulted to NULL)
# Missing time part imputed with 23:59:59 portion
impute_dtc(
  dtc = dates,
  time_imputation = "23:59:59"
)

# Same as above
impute_dtc(
  dtc = dates,
  time_imputation = "LAST"
)

# Impute to first day/month if date is partial
# Missing time part imputed with 00:00:00 portion by default
impute_dtc(
  dtc = dates,
  date_imputation = "01-01"
)
# same as above
impute_dtc(
  dtc = dates,
  date_imputation = "FIRST"
)

# Impute to last day/month if date is partial
# Missing time part imputed with 23:59:59 portion
impute_dtc(
  dtc = dates,
  date_imputation = "LAST",
  time_imputation = "LAST"
)

# Impute to mid day/month if date is partial
# Missing time part imputed with 00:00:00 portion by default
impute_dtc(
  dtc = dates,
  date_imputation = "MID"
)
}
\author{
Samia Kabi
}
\keyword{computation}
\keyword{timing}
