% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dt_dtm_utils.R
\name{impute_values}
\alias{impute_values}
\title{Impute Missing Values}
\usage{
impute_values(partial, target, components)
}
\arguments{
\item{partial}{A list of partial date/time components.

\describe{
\item{Default value}{none}
}}

\item{target}{A list of target values for imputation.

\describe{
\item{Default value}{none}
}}

\item{components}{A character vector of component names.

\describe{
\item{Default value}{none}
}}
}
\value{
A list of imputed date/time components.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This is a helping function for \code{get_dt_dtm_range()}
Imputes missing values in partial date/time components using target values.
}
\examples{
# Impute missing values for date components
partial_date <- list(year = "2020", month = NA_character_, day = NA_character_)
target_date <- list(year = "2020", month = "01", day = "01")
components_date <- c("year", "month", "day")
imputed_date <- admiral:::impute_values(partial_date, target_date, components_date)
print(imputed_date)

# Impute missing values for datetime components
partial_datetime <- list(
  year = "2020", month = NA_character_, day = NA_character_,
  hour = "12", minute = NA_character_, second = NA_character_
)
target_datetime <- list(
  year = "2020", month = "01", day = "01",
  hour = "12", minute = "00", second = "00"
)
components_datetime <- c("year", "month", "day", "hour", "minute", "second")
imputed_datetime <- admiral:::impute_values(
  partial_datetime, target_datetime,
  components_datetime
)
print(imputed_datetime)

# Impute missing values when some components are already present
partial_mixed <- list(year = "2020", month = "06", day = NA_character_)
target_mixed <- list(year = "2020", month = "01", day = "01")
components_mixed <- c("year", "month", "day")
imputed_mixed <- admiral:::impute_values(partial_mixed, target_mixed, components_mixed)
print(imputed_mixed)

}
\keyword{internal}
