% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize}
\alias{initialize}
\title{Derive Predecessors Based Upon Specification}
\usage{
initialize(dataset_name, metadata, source_datasets)
}
\arguments{
\item{dataset_name}{The name of the ADaM dataset to create, e.g. \code{"ADVS"}}

\item{metadata}{The result of \code{\link[=read_dap_m3]{read_dap_m3()}}}

\item{source_datasets}{A \code{list} of source datasets}
}
\value{
A dataset containing all predecessor variables as listed in \code{metadata}
}
\description{
Derive all predecessor variables of an ADaM dataset based upon metadata
}
\examples{
\dontrun{
data("adsl")
data("vs")
meta <- read_dap_m3("DAP_M3.xlsx")
initialize("ADVS", meta, list(adsl, vs))
}
}
\author{
Thomas Neitmann
}
