% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_character}
\alias{is_character}
\title{Is the object a character?}
\usage{
is_character(arg)
}
\arguments{
\item{arg}{The argument to check}
}
\value{
\code{TRUE} if the argument is a character, \code{FALSE} otherwise
}
\description{
Checks if a character vector was specified
}
\examples{
date <- "2020-02-03"
assertthat::assert_that(is_character(date))
}
\author{
Samia Kabi
}
