% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_date}
\alias{is_date}
\title{Is Date/Date-time?}
\usage{
is_date(arg)
}
\arguments{
\item{arg}{The argument to check}
}
\value{
\code{TRUE} if the argument is a date or date-time, \code{FALSE} otherwise
}
\description{
Checks if a date or date-time vector was specified
}
\examples{
assert_that(is_date(refdate), is_date(date))
}
\author{
Stefan Bundfuss
}
