% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_named}
\alias{is_named}
\title{Is a named argument}
\usage{
is_named(x)
}
\arguments{
\item{x}{Any R object}
}
\value{
\code{TRUE} if the argument is named, \code{FALSE} otherwise
}
\description{
Is a named argument
}
\seealso{
Identifies type of Object with return of TRUE/FALSE: 
\code{\link{is_auto}()},
\code{\link{is_order_vars}()},
\code{\link{is_valid_dtc}()}
}
\concept{is}
\keyword{is}
