% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{is_order_vars}
\alias{is_order_vars}
\title{Is order vars?}
\usage{
is_order_vars(arg)
}
\arguments{
\item{arg}{A<n R object}
}
\value{
A TRUE if input was created with \code{vars()} or calls involving \code{desc()}
otherwise throws an error
}
\description{
Check if inputs are created using \code{vars()} or calls involving \code{desc()}
}
\seealso{
Developer Utility Functions: 
\code{\link{\%notin\%}()},
\code{\link{\%or\%}()},
\code{\link{arg_name}()},
\code{\link{as_name}()},
\code{\link{convert_dtm_to_dtc}()},
\code{\link{extract_vars}()},
\code{\link{filter_if}()},
\code{\link{negate_vars}()},
\code{\link{valid_time_units}()},
\code{\link{vars2chr}()}
}
\concept{dev_utility}
\keyword{dev_utility}
