% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_valid_date_entry}
\alias{is_valid_date_entry}
\title{Check validity of the date imputation input}
\usage{
is_valid_date_entry(arg)
}
\arguments{
\item{arg}{The argument to check}
}
\value{
\code{TRUE} if the argument is a valid date_imputation input, \code{FALSE} otherwise
}
\description{
Date_imputation format should be specified as "dd-mm" (e.g. "01-01")
or as a keyword: "FISRT", "MID", "LAST"
}
\examples{
assertthat::assert_that(is_valid_date_entry("01-02"))
assertthat::assert_that(is_valid_date_entry("FIRST"))
}
\author{
Samia Kabi
}
