% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_valid_day}
\alias{is_valid_day}
\title{Check Validity of the Day Portion in the Date Input}
\usage{
is_valid_day(arg)
}
\arguments{
\item{arg}{The argument to check}
}
\value{
\code{TRUE} if the argument is a day input, \code{FALSE} otherwise
}
\description{
Days are expected to range from 1 to 31
}
\examples{
assertthat::assert_that(is_valid_day(20))
}
\author{
Samia Kabi
}
\keyword{check}
