% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_valid_hour}
\alias{is_valid_hour}
\title{Check validity of the hour portion in the time input}
\usage{
is_valid_hour(arg)
}
\arguments{
\item{arg}{The argument to check}
}
\value{
\code{TRUE} if the argument is a valid hour input, \code{FALSE} otherwise
}
\description{
Hours are expected to range from 0 to 23
}
\examples{
assertthat::assert_that(is_valid_hour(20))
}
\author{
Samia Kabi
}
