% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_valid_month}
\alias{is_valid_month}
\title{Check validity of the month portion in the date input}
\usage{
is_valid_month(arg)
}
\arguments{
\item{arg}{The argument to check}
}
\value{
\code{TRUE} if the argument is a month input, \code{FALSE} otherwise
}
\description{
Days are expected to range from 1 to 12
}
\examples{
assertthat::assert_that(is_valid_month(20))
}
\author{
Samia Kabi
}
