% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_valid_sec_min}
\alias{is_valid_sec_min}
\title{Check Validity of the Minute/Second Portion of the Time Input}
\usage{
is_valid_sec_min(arg)
}
\arguments{
\item{arg}{The argument to check}
}
\value{
\code{TRUE} if the argument is a valid min/sec input, \code{FALSE} otherwise
}
\description{
Minutes and seconds are expected to range from 0 to 59
}
\examples{
assertthat::assert_that(is_valid_sec_min(59))
}
\author{
Samia Kabi
}
\keyword{check}
