% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_valid_time_entry}
\alias{is_valid_time_entry}
\title{Check Validity of the Time Imputation Input}
\usage{
is_valid_time_entry(arg)
}
\arguments{
\item{arg}{The argument to check}
}
\value{
\code{TRUE} if the argument is a valid time_imputation input, \code{FALSE} otherwise
}
\description{
Time_imputation format should be specified as "hh:mm:ss" (e.g. "00:00:00")
or as a keyword: "FIRST", "LAST"
}
\examples{
assertthat::assert_that(admiral:::is_valid_time_entry("23:59:59"))
assertthat::assert_that(admiral:::is_valid_time_entry("FIRST"))
}
\author{
Samia Kabi
}
\keyword{check}
