% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_value}
\alias{is_value}
\title{Is Variable-value List?}
\usage{
is_value(arg)
}
\arguments{
\item{arg}{The argument to check}
}
\value{
\code{TRUE} if the argument is a symbol, a string, or NA, \code{FALSE} otherwise
}
\description{
Checks if the argument is a symbol, a string, or NA.
}
\author{
Stefan Bundfuss
}
\keyword{check}
