% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_varval_list}
\alias{is_varval_list}
\title{Is Variable-Value List?}
\usage{
is_varval_list(arg)
}
\arguments{
\item{arg}{The argument to check}
}
\value{
\code{TRUE} if the argument is a variable-value list, \code{FALSE} otherwise
}
\description{
Checks if the argument is a list of quosures where the expressions are
variable-value pairs. The value can be a symbol, a string, or NA. More general
expression are not allowed.
}
\examples{
assertthat::assert_that(is_varval_list(vars(DTHDOM = "AE", DTHSEQ = AESEQ)))
}
\author{
Stefan Bundfuss
}
\keyword{check}
