% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_helpers.R
\name{list_all_templates}
\alias{list_all_templates}
\title{List All Available ADaM Templates}
\usage{
list_all_templates()
}
\value{
A \code{character} vector of all available templates
}
\description{
List All Available ADaM Templates
}
\examples{
list_all_templates()
}
\seealso{
Helper Utilities: 
\code{\link{extract_unit}()},
\code{\link{negate_vars}()},
\code{\link{use_ad_template}()},
\code{\link{vars2chr}()}
}
\author{
Shimeng Huang, Thomas Neitmann
}
\concept{utils_help}
\keyword{utils_help}
