% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_lstalvdt.R
\name{lstalvdt_source}
\alias{lstalvdt_source}
\title{Create an \code{lstalvdt_source} object}
\usage{
lstalvdt_source(dataset, filter = NULL, date_var,
  traceability_vars = NULL)
}
\arguments{
\item{dataset}{A data.frame containing a source dataset.}

\item{filter}{An unquoted condition for filtering \code{dataset}.}

\item{date_var}{An unquoted symbol to be used for sorting \code{dataset}.}

\item{traceabilty_vars}{A named list returned by \code{vars()} defining the traceability variables, e.g.
\code{vars(LALVDOM = "AE", LALVSEQ = AESEQ, LALVVAR = "AESTDTC")}.}
}
\value{
An object of class "lstalvdt_source".
}
\description{
Create an \code{lstalvdt_source} object
}
\author{
Stefan Bundfuss
}
