% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dt_dtm_utils.R
\name{mask_target_values}
\alias{mask_target_values}
\title{Mask Target Values}
\usage{
mask_target_values(target, components, highest_imputation_level, is_datetime)
}
\arguments{
\item{target}{A list of target values for imputation.

\describe{
\item{Default value}{none}
}}

\item{components}{A character vector of component names.

\describe{
\item{Default value}{none}
}}

\item{highest_imputation_level}{The highest allowed imputation level.

\describe{
\item{Default value}{none}
}}

\item{is_datetime}{A logical indicating whether it's a datetime imputation.

\describe{
\item{Default value}{none}
}}
}
\value{
A list of masked target values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Masks target values based on the highest allowed imputation level.
}
\keyword{internal}
