% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{my_first_fcn.R}
\alias{my_first_fcn.R}
\alias{hello_admiral}
\title{Hello Admiral Function}
\usage{
hello_admiral(hw = TRUE)
}
\arguments{
\item{hw}{TRUE or FALSE}
}
\value{
Happy Message
}
\description{
Hello Admiral Function
}
\details{
First function to generate welcome message
}
\examples{
hello_admiral(hw = FALSE)
}
\seealso{
ADSL Functions that returns variable appended to dataset: 
\code{\link{derive_var_age_years}()},
\code{\link{derive_vars_aage}()},
\code{\link{derive_vars_extreme_event}()},
\code{\link{derive_vars_period}()}
}
\concept{der_adsl}
\keyword{der_adsl}
