% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{my_first_fcn}
\alias{my_first_fcn}
\title{Derive onboarding dummy function}
\usage{
my_first_fcn(first_time = TRUE)
}
\arguments{
\item{first_time}{TRUE or FALSE}
}
\value{
a Welcome message
}
\description{
Derive onboarding dummy function
}
\details{
outputs "Welcome to the admiral family!".
}
\examples{
my_first_fcn(first_time = FALSE)
}
\author{
Fanny Gautier
}
