% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_function.R
\name{my_first_function}
\alias{my_first_function}
\title{My First Function}
\usage{
my_first_function()
}
\value{
character: "Welcome to the admiral family!"
}
\description{
My First Function
}
\examples{
my_first_function()
}
