% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_utils.R
\name{negate_vars}
\alias{negate_vars}
\title{Negate List of Variables}
\usage{
negate_vars(vars = NULL)
}
\arguments{
\item{vars}{List of variables created by \code{exprs()}

\describe{
\item{Default value}{\code{NULL}}
}}
}
\value{
A list of expressions
}
\description{
The function adds a minus sign as prefix to each variable.
}
\details{
This is useful if a list of variables should be removed from a dataset,
e.g., \code{select(!!!negate_vars(by_vars))} removes all by variables.
}
\examples{
negate_vars(exprs(USUBJID, STUDYID))
}
\seealso{
Utilities for working with quosures/list of expressions: 
\code{\link{chr2vars}()}
}
\concept{utils_quo}
\keyword{utils_quo}
