% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_nfrlt.R
\name{nfrlt_special}
\alias{nfrlt_special}
\title{Create a \code{nfrlt_special} Object}
\usage{
nfrlt_special(condition, value)
}
\arguments{
\item{condition}{Condition

The condition is evaluated at the input dataset of \code{derive_var_nfrlt()}.
For all records where it evaluates to \code{TRUE}, the new variable is set to
\code{value}.

\describe{
\item{Permitted values}{an unquoted condition, e.g., \code{AVISIT == "BASELINE"}}
\item{Default value}{none}
}}

\item{value}{Special value

The value is assigned as provided, i.e., it is not converted to the
\code{out_unit} of \code{derive_var_nfrlt()}.

\describe{
\item{Permitted values}{A numeric scalar, e.g., \code{99998}}
\item{Default value}{none}
}}
}
\value{
An object of class \code{nfrlt_special}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \code{nfrlt_special} object is used to define special values as input for
the \code{special_values} argument of \code{derive_var_nfrlt()}, e.g., to set the
nominal relative time to \code{99998} for unscheduled visits.
}
\examples{
nfrlt_special(condition = VISIT == "UNSCHEDULED", value = 99998)
}
\seealso{
\code{\link[=derive_var_nfrlt]{derive_var_nfrlt()}}

Source Objects:
\code{\link[=basket_select]{basket_select()}},
\code{\link[=censor_source]{censor_source()}},
\code{\link{death_event}},
\code{\link[=event]{event()}},
\code{\link[=event_joined]{event_joined()}},
\code{\link[=event_source]{event_source()}},
\code{\link[=flag_event]{flag_event()}},
\code{\link[=query]{query()}},
\code{\link[=tte_source]{tte_source()}}
}
\concept{source_specifications}
\keyword{source_specifications}
