% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiral_df.R
\name{occds_seq_var}
\alias{occds_seq_var}
\title{Find the Record Key of an Occurrence Dataset}
\usage{
occds_seq_var(dataset)
}
\arguments{
\item{dataset}{A data frame

\describe{
\item{Default value}{none}
}}
}
\value{
The name of the sequence variable which keys \code{dataset}, or a
zero-length vector (with a warning) when none can be determined.
}
\description{
Find the Record Key of an Occurrence Dataset
}
\details{
\code{ASEQ} is preferred, falling back to a populated SDTM domain sequence. If
no sequence is left -- or several domain sequences are, making the record
key ambiguous -- a warning is issued and nothing is returned.
}
\seealso{
\code{\link[=infer_admiral_keys]{infer_admiral_keys()}}, which uses this for \code{OCCDS} datasets

Other internal:
\code{\link{admiral-package}},
\code{\link[=extract_duplicate_records]{extract_duplicate_records()}},
\code{\link[=format.basket_select]{format.basket_select()}},
\code{\link[=get_admiral_df_type]{get_admiral_df_type()}},
\code{\link[=infer_admiral_keys]{infer_admiral_keys()}},
\code{\link[=is_adsl_structure]{is_adsl_structure()}},
\code{\link[=minimal_unique_key]{minimal_unique_key()}},
\code{\link[=signal_duplicate_records]{signal_duplicate_records()}}
}
\concept{internal}
\keyword{internal}
