% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dt_dtm_utils.R
\name{parse_partial_date_time}
\alias{parse_partial_date_time}
\title{Parse Partial Date or Datetime}
\usage{
parse_partial_date_time(dtc, is_datetime)
}
\arguments{
\item{dtc}{A character vector of date or datetime strings to be parsed.

\describe{
\item{Default value}{none}
}}

\item{is_datetime}{A logical value indicating whether the input strings include
time information.

\describe{
\item{Default value}{none}
}}
}
\value{
A list of character vectors, each representing a component of the date
or datetime.
\itemize{
\item For dates, the components are "year", "month", and "day".
\item For datetimes, the components also include "hour", "minute", and "second".
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This is a helping function for \code{get_dt_dtm_range()}
This function parses a vector of date or datetime strings into their component
parts.
}
\details{
The function uses different parsing methods depending on whether the input is
a date or a datetime:
\itemize{
\item For dates, it calls \code{get_partialdate()}.
\item For datetimes, it calls \code{get_partialdatetime()}.
}
}
\examples{
# Parse partial datetime components
dtc_datetime <- "2020-03-15T12:34"
parsed_datetime <- admiral:::parse_partial_date_time(dtc_datetime, is_datetime = TRUE)
print(parsed_datetime)

# Parse partial date components
dtc_date <- "2020-03"
parsed_date <- admiral:::parse_partial_date_time(dtc_date, is_datetime = FALSE)
print(parsed_date)

# Parse partial datetime with missing components
dtc_partial_datetime <- "2020-03T12"
parsed_partial_datetime <- admiral:::parse_partial_date_time(dtc_partial_datetime,
  is_datetime = TRUE
)
print(parsed_partial_datetime)

# Parse partial date with incomplete year
dtc_year <- "2020"
parsed_year <- admiral:::parse_partial_date_time(dtc_year, is_datetime = FALSE)
print(parsed_year)

}
\keyword{internal}
