% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs}{A value or the magrittr placeholder.

\describe{
\item{Default value}{\code{NULL}}
}}

\item{rhs}{A function call using the magrittr semantics.

\describe{
\item{Default value}{\code{NULL}}
}}
}
\description{
See \code{magrittr::\link[magrittr:\%>\%]{\%>\%}} for more details.
}
\keyword{reexport}
