% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_helpers.R
\name{print.adam_templates}
\alias{print.adam_templates}
\title{Print \code{adam_templates} Objects}
\usage{
\method{print}{adam_templates}(x, ...)
}
\arguments{
\item{x}{A \code{adam_templates} object

\describe{
\item{Default value}{none}
}}

\item{...}{Not used

\describe{
\item{Default value}{none}
}}
}
\value{
No return value, called for side effects
}
\description{
Print \code{adam_templates} Objects
}
\examples{
templates <- list_all_templates()
print(templates)
}
\seealso{
\code{\link[=list_all_templates]{list_all_templates()}}

Utilities for printing:
\code{\link[=check_admiral_df]{check_admiral_df()}},
\code{\link[=print.admiral_df_check]{print.admiral_df_check()}},
\code{\link[=print.duplicates]{print.duplicates()}},
\code{\link[=print.source]{print.source()}},
\code{\link[=print.summary_admiral_df]{print.summary_admiral_df()}},
\code{\link[=print_named_list]{print_named_list()}},
\code{\link[=summary.admiral_df]{summary.admiral_df()}}
}
\concept{utils_print}
\keyword{internal}
