% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_derivation.R
\name{print.derivation_slice}
\alias{print.derivation_slice}
\title{Print \code{derivation_slice} Objects}
\usage{
\method{print}{derivation_slice}(x, ...)
}
\arguments{
\item{x}{A \code{derivation_slice} object}

\item{...}{Not used}
}
\value{
No return value, called for side effects
}
\description{
Print \code{derivation_slice} Objects
}
\examples{
print(death_event)
}
\seealso{
\code{\link[=derivation_slice]{derivation_slice()}}

Higher Order Functions: 
\code{\link{call_derivation}()},
\code{\link{derivation_slice}()},
\code{\link{restrict_derivation}()},
\code{\link{slice_derivation}()}
}
\concept{high_order_function}
\keyword{high_order_function}
