% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{print.duplicates}
\alias{print.duplicates}
\title{Print \code{duplicates} Objects}
\usage{
\method{print}{duplicates}(x, ...)
}
\arguments{
\item{x}{A \code{duplicates} object

\describe{
\item{Default value}{none}
}}

\item{...}{Not used

\describe{
\item{Default value}{none}
}}
}
\value{
No return value, called for side effects
}
\description{
Print \code{duplicates} Objects
}
\seealso{
Utilities for printing:
\code{\link[=check_admiral_df]{check_admiral_df()}},
\code{\link[=print.adam_templates]{print.adam_templates()}},
\code{\link[=print.admiral_df_check]{print.admiral_df_check()}},
\code{\link[=print.source]{print.source()}},
\code{\link[=print.summary_admiral_df]{print.summary_admiral_df()}},
\code{\link[=print_named_list]{print_named_list()}},
\code{\link[=summary.admiral_df]{summary.admiral_df()}}
}
\concept{utils_print}
\keyword{internal}
