% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print}
\alias{print}
\title{Print Values}
\usage{
print(x, ...)
}
\arguments{
\item{x}{An object to print}

\item{...}{Further arguments passed to or from other methods}
}
\description{
\code{print()} prints its argument and returns it invisibly (via \code{\link[=invisible]{invisible()}}).
It is a generic function which means that new printing methods can be easily
added for new classes.
}
