% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_utils.R
\name{print.source}
\alias{print.source}
\title{Print \code{source} Objects}
\usage{
\method{print}{source}(x, ...)
}
\arguments{
\item{x}{An \code{source} object

\describe{
\item{Default value}{none}
}}

\item{...}{If \verb{indent = <numeric value>} is specified the output is indented
by the specified number of characters.

\describe{
\item{Default value}{none}
}}
}
\value{
No return value, called for side effects
}
\description{
Print \code{source} Objects
}
\examples{
print(death_event)
}
\seealso{
Utilities for printing:
\code{\link[=check_admiral_df]{check_admiral_df()}},
\code{\link[=print.adam_templates]{print.adam_templates()}},
\code{\link[=print.admiral_df_check]{print.admiral_df_check()}},
\code{\link[=print.duplicates]{print.duplicates()}},
\code{\link[=print.summary_admiral_df]{print.summary_admiral_df()}},
\code{\link[=print_named_list]{print_named_list()}},
\code{\link[=summary.admiral_df]{summary.admiral_df()}}
}
\concept{utils_print}
\keyword{internal}
