% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_tte.R
\name{print.tte_source}
\alias{print.tte_source}
\title{Print \code{tte_source} Objects}
\usage{
\method{print}{tte_source}(x, ...)
}
\arguments{
\item{x}{A \code{tte_source} object}

\item{...}{Not used}
}
\value{
No return value, called for side effects
}
\description{
Print \code{tte_source} Objects
}
\examples{
print(death_event)
}
\seealso{
\code{\link[=tte_source]{tte_source()}}, \code{\link[=censor_source]{censor_source()}}, \code{\link[=event_source]{event_source()}}
}
\author{
Thomas Neitmann
}
