% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1839_dummy_my_first_fcn.R
\name{professorp 1839 Dummy Function}
\alias{professorp 1839 Dummy Function}
\alias{my_first_fcn}
\title{professorp 1839 Dummy Function}
\usage{
my_first_fcn(x)
}
\arguments{
\item{x}{any input for the function}
}
\value{
"Welcome to the admiral family!"
}
\description{
This function prints "Welcome to the admiral family!"
}
\details{
This function will print the description message regardless of the input or an absence of input
for the 1839 dummy example.
}
\examples{
if(!require('devtools')){
install.packages('devtools')
}

devtools::dev_package_deps(dependencies = T)

my_first_fcn()
}
\concept{deprecated,internal}
\keyword{deprecated,}
\keyword{internal}
