% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{queries}
\alias{queries}
\title{Queries Dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14 rows and 6 columns.}
\usage{
queries
}
\description{
An example of standard query dataset to be used in deriving variables in ADAE and ADCM
}
\keyword{datasets}
