% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{queries_mh}
\alias{queries_mh}
\title{Queries MH Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 14 rows and 8 columns.
}
\source{
An example of standard query MH dataset to be used in deriving variables in ADMH
}
\usage{
queries_mh
}
\description{
Queries MH Dataset
}
\seealso{
Other datasets: 
\code{\link{admiral_adsl}},
\code{\link{ex_single}},
\code{\link{queries}}
}
\concept{datasets}
\keyword{datasets}
