% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{quo_c}
\alias{quo_c}
\title{Concatenate One or More Quosure(s)}
\usage{
quo_c(...)
}
\arguments{
\item{...}{One or more objects of class \code{quosure} or \code{quosures}}
}
\value{
An object of class \code{quosures}
}
\description{
Concatenate One or More Quosure(s)
}
\examples{
quo_c(rlang::quo(USUBJID))
quo_c(rlang::quo(STUDYID), rlang::quo(USUBJID))
quo_c(vars(USUBJID, ADTM))
quo_c(rlang::quo(BASETYPE), vars(USUBJID, PARAM), rlang::quo(ADTM))
}
\seealso{
Functions moved to admiraldev package: 
\code{\link{arg_name}()},
\code{\link{assert_db_requirements}()},
\code{\link{assert_terms}()},
\code{\link{assert_valid_queries}()},
\code{\link{backquote}()},
\code{\link{contains_vars}()},
\code{\link{dquote}()},
\code{\link{enumerate}()},
\code{\link{extend_source_datasets}()},
\code{\link{extract_vars}()},
\code{\link{format.sdg_select}()},
\code{\link{format.smq_select}()},
\code{\link{get_constant_vars}()},
\code{\link{get_duplicates}()},
\code{\link{get_source_vars}()},
\code{\link{is_auto}()},
\code{\link{replace_values_by_names}()},
\code{\link{squote}()},
\code{\link{valid_time_units}()},
\code{\link{validate_query}()},
\code{\link{validate_sdg_select}()},
\code{\link{validate_smq_select}()},
\code{\link{what_is_it}()}
}
\author{
Thomas Neitmann
}
\concept{move_adm_dev}
\keyword{move_adm_dev}
