% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet_rdx.R
\name{rdx_roclet}
\alias{rdx_roclet}
\title{Roclet Extending the Standard rd Roclet}
\usage{
rdx_roclet()
}
\description{
This roclet extends the standard \code{rd} roclet by allowing
\itemize{
\item to add permitted values and default values to the \verb{@param} tag and
\item to add a caption and a description to examples.
}
}
\details{
The following tags are supported:
\itemize{
\item \verb{@permitted}: Permitted values for the argument. Permitted value description
which are used for several arguments/functions can be stored in
\code{inst/roxygen/rdx_meta.R}. For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  rdx_permitted_values = list(
    mode = "`\\"first\\"`, `\\"last\\"`",
    msg_type = "`\\"none\\"`, `\\"message\\"`, `\\"warning\\"`, `\\"error\\"`"
  )
)
}\if{html}{\out{</div>}}

The reference to the permitted values is done by specifying the name of
the list element in square brackets, e.g., \verb{@permitted [mode]}.
\item \verb{@default}: Default value for the argument. By default the default value
from the function formals is displayed. This can be overwritten by using
the \verb{@default} tag.
\item \verb{@examplesx}: This tag can be used to mark the beginning of the examples
section but doesn't affect the output, i.e., it can be omitted.
\item \verb{@caption}: Caption for the example. The caption is displayed as a subsection
in the examples section. The caption can be followed by an arbitrary number
of \verb{@info} and \verb{@code} tags.
\item \verb{@info}: Description of the example.
\item \verb{@code}: Code of the example.
}

To use the roclet call \code{roxygen2::roxygenise(roclets = "admiral::rdx_roclet")} or add to the \code{DESCRIPTION} file:

\if{html}{\out{<div class="sourceCode">}}\preformatted{Roxygen: list(markdown = TRUE, roclets = c("collate", "namespace", "admiral::rdx_roclet"))
}\if{html}{\out{</div>}}
}
\keyword{internal}
