% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{read_dap_m3}
\alias{read_dap_m3}
\title{Read a DAP M3}
\usage{
read_dap_m3(file)
}
\arguments{
\item{file}{\code{character}. Path to a DAP M3 excel file}
}
\value{
A named \code{list} where each elements corresponds to one sheet in the original
excel workbook
}
\description{
Read all sheets of a DAP M3 into R
}
\examples{
\dontrun{
read_dap_m3("DAP_M3.xlsx")
}
}
\author{
Thomas Neitmann
}
