% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{vars}
\alias{desc}
\alias{\%>\%}
\alias{exprs}
\alias{expr}
\title{Function Re-exports}
\details{
Sometimes functions from other libraries become critically important while using admiral
and thus should be included as an export. This applies especially to functions which are
frequently called within \code{admiral} function arguments. The goal of these exports is such that
admiral comes ready "out of the box", similar to how one might think the pipe operator, \verb{\%>\%},
comes from \code{dplyr} but is actually native to \code{magrittr}.
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{desc}}, \code{\link[dplyr]{vars}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{rlang}{\code{\link[rlang:defusing-advanced]{exprs}}}
}}

