% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiral_options.R
\name{set_admiral_options}
\alias{set_admiral_options}
\title{Set a package-standard input}
\usage{
set_admiral_options(subject_keys, future_input)
}
\arguments{
\item{subject_keys}{Subjects used for several \code{derive_()} functions, defaults to
vars(STUDYID, USUBJID)}

\item{future_input}{Possible future input to figure out how to scale this}
}
\value{
Modify an admiral option which will affect downstream function inputs: e.g \code{subject_keys}.
}
\description{
Set a package-standard input that can be modified for advanced users.
}
\examples{
set_admiral_options(subject_keys = vars(STUDYID, USUBJID2))
}
\seealso{
\code{\link[=vars]{vars()}}

Other XXX: 
\code{\link{get_admiral_options}()}
}
\author{
Zelos Zhu
}
\concept{XXX}
\keyword{XXX}
