% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{set_dataset}
\alias{set_dataset}
\title{Set a Dataset in the \code{.datasets} environment}
\usage{
set_dataset(dataset, name)
}
\arguments{
\item{dataset}{A \code{data.frame}}

\item{name}{A name for \code{dataset}}
}
\value{
No return value, called for side effects
}
\description{
Set a Dataset in the \code{.datasets} environment
}
\details{
The object passed to the \code{dataset} argument will be assigned to \code{name} in
the \code{.datasets} environment. It can be retrieved later on using \code{\link[=get_dataset]{get_dataset()}}
}
\seealso{
Other datasets: 
\code{\link{get_dataset}()}
}
\author{
Thomas Neitmann
}
\concept{datasets}
\keyword{datasets}
