% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query_data.R
\name{smq_select}
\alias{smq_select}
\title{Create an \code{smq_select} object}
\usage{
smq_select(name = NULL, id = NULL, scope = NULL)
}
\arguments{
\item{name}{Name of the query used to select the definition of the query from
the company database.}

\item{id}{Identifier of the query used to select the definition of the query
from the company database.}

\item{scope}{Scope of the query used to select the definition of the query
from the company database.

\emph{Permitted Values}: \code{"BROAD"}, \code{"NARROW"}}
}
\value{
An object of class \code{smq_select}.
}
\description{
Create an \code{smq_select} object
}
\details{
Exactly one of \code{name} or \code{id} must be specified.
}
\seealso{
\code{\link[=create_query_data]{create_query_data()}}, \code{\link[=query]{query()}}

Source Specifications: 
\code{\link{censor_source}()},
\code{\link{date_source}()},
\code{\link{death_event}},
\code{\link{event_source}()},
\code{\link{list_tte_source_objects}()},
\code{\link{params}()},
\code{\link{query}()},
\code{\link{sdg_select}()}
}
\author{
Stefan Bundfuss
}
\concept{source_specifications}
\keyword{source_specifications}
