% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{suppress_warning}
\alias{suppress_warning}
\title{Suppress Warnings}
\usage{
suppress_warning(expr, regexpr)
}
\arguments{
\item{expr}{Expression to be executed}

\item{regexpr}{Regular expression matching warnings to suppress}
}
\value{
return value of the expression
}
\description{
Suppress certain warnings issued by an expression.
}
\details{
All warnings which are issued by the expression and match the
regular expression are suppressed.
}
\examples{
  \dontrun{
  suppress_warning(
    left_join(dataset,
              all_data,
              by = c("USUBJID")),
    regexpr = "^Column `USUBJID` has different attributes on LHS and RHS of join$")
  )
  }
}
\author{
Stefan Bundfuss
}
\keyword{warning}
