% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time_def}
\alias{time_def}
\title{Time definitions dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 460 rows and 8 columns.}
\source{
Created manually based on \link{dm}
}
\usage{
time_def
}
\description{
An example time definitions dataset
}
\keyword{datasets}
