% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_helpers.R
\name{use_ad_template}
\alias{use_ad_template}
\title{Open an ADaM Template Script}
\usage{
use_ad_template(
  adam_name = "adsl",
  save_path = paste0("./", adam_name, ".R"),
  open = interactive()
)
}
\arguments{
\item{adam_name}{An ADaM dataset name.}

\item{save_path}{Path to save the script.}

\item{open}{Whether to open the script right away.}
}
\description{
Open an ADaM Template Script
}
\examples{
if (interactive()) {
  use_ad_template("adsl")
}
}
\author{
Shimeng Huang
}
\keyword{user_utility}
