% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{not_in}
\alias{not_in}
\alias{\%!in\%}
\title{Negated Value Matching}
\usage{
x \%!in\% table
}
\arguments{
\item{x}{The values to be matched}

\item{table}{The values to be matched against}
}
\description{
Returns a \code{logical} vector indicating if there is \emph{no} match of the
left operand in the right operand.
}
\examples{
"a" \%!in\% c("b", "v", "k")
}
