% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{valid_time_units}
\alias{valid_time_units}
\title{Returns Valid Time Units}
\usage{
valid_time_units()
}
\value{
A vector of valid time units
}
\description{
Returns valid time units as expected for arguments expecting a time unit
}
\examples{

valid_time_units()
}
\keyword{dev_utility}
