% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query_data.R
\name{validate_basket_select}
\alias{validate_basket_select}
\title{Validate an object is indeed a \code{basket_select} object}
\usage{
validate_basket_select(obj)
}
\arguments{
\item{obj}{An object to be validated.

\describe{
\item{Default value}{none}
}}
}
\value{
The original object.
}
\description{
Validate an object is indeed a \code{basket_select} object
}
\seealso{
\code{\link[=basket_select]{basket_select()}}
}
\keyword{internal}
