% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_dthcaus.R
\name{validate_dthcaus_source}
\alias{validate_dthcaus_source}
\title{Validate an object is indeed a \code{dthcaus_source} object}
\usage{
validate_dthcaus_source(x)
}
\arguments{
\item{x}{An object to be validated.}
}
\value{
The original object.
}
\description{
Validate an object is indeed a \code{dthcaus_source} object
}
\author{
Shimeng Huang
}
