% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_lstalvdt.R
\name{validate_lstalvdt_source}
\alias{validate_lstalvdt_source}
\title{Validate an object is indeed a \code{lstalvdt_source} object}
\usage{
validate_lstalvdt_source(obj)
}
\arguments{
\item{obj}{An object to be validated.}
}
\value{
The original object.
}
\description{
Validate an object is indeed a \code{lstalvdt_source} object
}
\author{
Stefan Bundfuss
}
