% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query_data.R
\name{validate_smq_select}
\alias{validate_smq_select}
\title{Validate an object is indeed a \code{smq_select} object}
\usage{
validate_smq_select(obj)
}
\arguments{
\item{obj}{An object to be validated.}
}
\value{
The original object.
}
\description{
Validate an object is indeed a \code{smq_select} object
}
\seealso{
\code{\link[=smq_select]{smq_select()}}
}
\author{
Stefan Bundfuss
}
\keyword{move_adm_dev}
