% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_utils.R
\name{vars2chr}
\alias{vars2chr}
\title{Turn a List of Quosures into a Character Vector}
\usage{
vars2chr(quosures)
}
\arguments{
\item{quosures}{A \code{list} of \code{quosures} created using \code{\link[=vars]{vars()}}}
}
\value{
A character vector
}
\description{
Turn a List of Quosures into a Character Vector
}
\examples{
library(dplyr)

vars2chr(vars(USUBJID, AVAL))
}
\seealso{
Other utils_quo: 
\code{\link{chr2vars}()},
\code{\link{negate_vars}()}
}
\author{
Thomas Neitmann
}
\concept{utils_quo}
\keyword{utils_quo}
