% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{warn_if_incomplete_dtc}
\alias{warn_if_incomplete_dtc}
\title{Warn if incomplete dtc}
\usage{
warn_if_incomplete_dtc(dtc, n)
}
\arguments{
\item{dtc}{A \code{character} vector of date-times in ISO 8601 format}

\item{n}{A non-negative integer}
}
\value{
A warning if \code{dtc} contains any partial dates
}
\description{
Warn if incomplete dtc
}
\seealso{
Function that provide users with custom warnings
\code{\link{suppress_warning}()},
\code{\link{warn_if_inconsistent_list}()},
\code{\link{warn_if_invalid_dtc}()},
\code{\link{warn_if_vars_exist}()}
}
\concept{warnings}
\keyword{warnings}
