% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{warn_if_invalid_dtc}
\alias{warn_if_invalid_dtc}
\title{Warn If a vector contains unknown datetime format}
\usage{
warn_if_invalid_dtc(dtc, is_valid = is_valid_dtc(dtc))
}
\arguments{
\item{dtc}{a character vector containing the dates}

\item{is_valid}{a logical vector indicating whether elements in \code{dtc} are valid}
}
\description{
Warn if the vector contains unknown datetime format such as
"2003-12-15T-:15:18", "2003-12-15T13:-:19","--12-15","-----T07:15"
}
\examples{

## No warning as `dtc` is a valid date format
warn_if_invalid_dtc(dtc = "2021-04-06")

## Issues a warning
warn_if_invalid_dtc(dtc = "2021-04-06T-:30:30")
}
\author{
Samia Kabi
}
\keyword{warning}
