% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{welcome_fun}
\alias{welcome_fun}
\title{Dummy Function for Onboarding}
\usage{
welcome_fun(anything = NULL)
}
\arguments{
\item{anything}{Element in any format the user wants.}
}
\value{
A "Welcome to the admiral family!" string.
}
\description{
Returns "Welcome to the admiral family!"
}
\details{
This is a function that returns "Welcome to the admiral family!"
independent of its inputs.
}
\examples{
welcome_fun(anything = "a")
welcome_fun(anything = 2)
welcome_fun(anything = NULL)
welcome_fun(anything = data.frame(1:10, 2:11))
}
\seealso{
General Derivation Functions for all ADaMs that returns variable appended to dataset:
\code{\link{derive_var_extreme_flag}()},
\code{\link{derive_var_joined_exist_flag}()},
\code{\link{derive_var_merged_exist_flag}()},
\code{\link{derive_var_merged_summary}()},
\code{\link{derive_var_obs_number}()},
\code{\link{derive_var_relative_flag}()},
\code{\link{derive_vars_joined}()},
\code{\link{derive_vars_merged_lookup}()},
\code{\link{derive_vars_merged}()},
\code{\link{derive_vars_transposed}()},
\code{\link{get_summary_records}()}
}
\concept{der_gen}
\keyword{der_gen}
