% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{wlcm_admiral}
\alias{wlcm_admiral}
\title{Drive Extension Example}
\usage{
wlcm_admiral(hw = TRUE)
}
\arguments{
\item{hw}{TRUE or FALSE}
}
\value{
Welcome Message
}
\description{
Says Welcome to admiral
}
\details{
Created function to output "Welciome to admiral family"
}
\examples{
wlcm_admiral(hw = FALSE)
}
\author{
Open Source
}
\concept{der_adxx}
\keyword{der_adxx}
