% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_exist_flag.R
\name{yn_to_numeric}
\alias{yn_to_numeric}
\title{Map \code{"Y"} and \code{"N"} to Numeric Values}
\usage{
yn_to_numeric(arg)
}
\arguments{
\item{arg}{Character vector}
}
\value{
\code{1} if \code{arg} equals \code{"Y"}, \code{0} if \code{arg} equals \code{"N"}, \code{NA_real_} otherwise
}
\description{
Map \code{"Y"} and \code{"N"} to numeric values.
}
\examples{

yn_to_numeric(c("Y", "N", NA_character_))
}
\author{
Stefan Bundfuss
}
\keyword{user_utility}
