% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_dttm.R
\docType{methods}
\name{+,iso_dttm,Period-method}
\alias{+,iso_dttm,Period-method}
\alias{+,Period,iso_dttm-method}
\alias{+,iso_dttm,Duration-method}
\alias{+,Duration,iso_dttm-method}
\title{Add Methods}
\usage{
\S4method{+}{iso_dttm,Period}(e1, e2)

\S4method{+}{Period,iso_dttm}(e1, e2)

\S4method{+}{iso_dttm,Duration}(e1, e2)

\S4method{+}{Duration,iso_dttm}(e1, e2)
}
\arguments{
\item{e1}{First object}

\item{e2}{Second object}
}
\description{
Add Methods
}
