% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiral-package.R
\docType{package}
\name{admiral-package}
\alias{admiral}
\alias{admiral-package}
\title{admiral: ADaM in R Asset Library}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A toolbox for programming Clinical Data Interchange Standards Consortium (CDISC) compliant Analysis Data Model (ADaM) datasets in R. ADaM datasets are a mandatory part of any New Drug or Biologics License Application submitted to the United States Food and Drug Administration (FDA). Analysis derivations are implemented in accordance with the "Analysis Data Model Implementation Guide" (CDISC Analysis Data Model Team, 2021, \url{https://www.cdisc.org/standards/foundational/adam}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiral/}
  \item \url{https://github.com/pharmaverse/admiral}
  \item Report bugs at \url{https://github.com/pharmaverse/admiral/issues}
}


Other internal:
\code{\link[=as_admiral_df]{as_admiral_df()}},
\code{\link[=extract_duplicate_records]{extract_duplicate_records()}},
\code{\link[=format.basket_select]{format.basket_select()}},
\code{\link[=signal_duplicate_records]{signal_duplicate_records()}}
}
\author{
\strong{Maintainer}: Edoardo Mancini \email{edoardo.mancini@roche.com} (\href{https://orcid.org/0009-0006-4899-8641}{ORCID})

Authors:
\itemize{
  \item Edoardo Mancini \email{edoardo.mancini@roche.com} (\href{https://orcid.org/0009-0006-4899-8641}{ORCID})
  \item Stefan Bundfuss (\href{https://orcid.org/0009-0005-0027-1198}{ORCID})
  \item Arianna Cascone (\href{https://orcid.org/0000-0001-5948-2831}{ORCID})
  \item Kristin Dahnert
  \item Jeffrey Dickinson
  \item Ross Farrugia
  \item Fanny Gautier (\href{https://orcid.org/0009-0004-3581-0131}{ORCID})
  \item Liam Hobby \email{liam.f.hobby@gsk.com}
  \item Gordon Miller
  \item Lina Patil
  \item Ben Straub
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item GlaxoSmithKline LLC [copyright holder, funder]
}

}
\concept{internal}
\keyword{internal}
