% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiral-package.R
\docType{package}
\name{admiral-package}
\alias{admiral}
\alias{admiral-package}
\title{admiral: ADaM in R Asset Library}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A toolbox for programming Clinical Data Interchange Standards Consortium (CDISC) compliant Analysis Data Model (ADaM) datasets in R. ADaM datasets are a mandatory part of any New Drug or Biologics License Application submitted to the United States Food and Drug Administration (FDA). Analysis derivations are implemented in accordance with the "Analysis Data Model Implementation Guide" (CDISC Analysis Data Model Team, 2021, \url{https://www.cdisc.org/standards/foundational/adam/adamig-v1-3-release-package}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiral/}
  \item \url{https://github.com/pharmaverse/admiral}
  \item Report bugs at \url{https://github.com/pharmaverse/admiral/issues}
}

}
\author{
\strong{Maintainer}: Ben Straub \email{ben.x.straub@gsk.com}

Authors:
\itemize{
  \item Stefan Bundfuss
  \item Jeffrey Dickinson
  \item Ross Farrugia
  \item Adam Forys
  \item Daphne Grasselly
  \item Dinakar Kulkarni
  \item Edoardo Mancini
  \item Sadchla Mascary
  \item Gordon Miller
  \item Sophie Shapcott
  \item Eric Simms
  \item Stefan Thoma
  \item Kangjie Zhang
  \item Zelos Zhu
}

Other contributors:
\itemize{
  \item G Gayatri [contributor]
  \item Thomas Neitmann [contributor]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item GlaxoSmithKline LLC [copyright holder, funder]
}

}
\concept{internal}
\keyword{internal}
