% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiral-package.R
\docType{package}
\name{admiral-package}
\alias{admiral}
\alias{admiral-package}
\title{admiral: ADaM in R Asset Library}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A toolbox for programming Clinical Data Standards Interchange Consortium (CDISC) compliant Analysis Data Model (ADaM) datasets in R. ADaM datasets are a mandatory part of any New Drug or Biologics License Application submitted to the United States Food and Drug Administration (FDA). Analysis derivations are implemented in accordance with the "Analysis Data Model Implementation Guide" (CDISC Analysis Data Model Team, 2021, \url{https://www.cdisc.org/standards/foundational/adam/adamig-v1-3-release-package}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiral/index.html}
  \item \url{https://github.com/pharmaverse/admiral/}
}


Other internal: 
\code{\link{print.tte_source}()}
}
\author{
\strong{Maintainer}: Thomas Neitmann \email{thomas.neitmann@roche.com}

Authors:
\itemize{
  \item Stefan Bundfuss
  \item Ben Straub
  \item Samia Kabi
  \item Gordon Miller
  \item Teckla Akinyi
  \item Andrew Smith
  \item Konstantina Koukourikou
  \item Ross Farrugia
  \item Eric Simms
  \item Annie Yang
  \item Robin Koeger
  \item Sophie Shapcott
  \item Ojesh Upadhyay
  \item Jack McGavigan
  \item Kamila Duniec
  \item Gayatri G
  \item Alana Harris
  \item Mahdi About
  \item Pooja Kumari
  \item Claudia Carlucci
  \item Daniil Stefonishin
}

Other contributors:
\itemize{
  \item Michael Thorpe [contributor]
  \item Pavan Kumar [contributor]
  \item Hamza Rahal [contributor]
  \item Alice Ehmann [contributor]
  \item Tom Ratford [contributor]
  \item Vignesh Thanikachalam [contributor]
  \item Ondrej Slama [contributor]
  \item Shimeng Huang [contributor]
  \item James Kim [contributor]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item GlaxoSmithKline LLC [copyright holder, funder]
}

}
\concept{internal}
\keyword{internal}
