% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiral-package.R
\docType{package}
\name{admiral-package}
\alias{admiral}
\alias{admiral-package}
\title{admiral: ADaM in R Asset Library}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A toolbox for programming ADaM datasets in R.
}
\author{
\strong{Maintainer}: Thomas Neitmann \email{thomas.neitmann@roche.com}

Authors:
\itemize{
  \item Alice Ehmann \email{alice.x.ehmann@gsk.com}
  \item Stefan Bundfuss \email{stefan.bundfuss@roche.com}
  \item Samia Kabi \email{samia.kabi@roche.com}
  \item Vignesh Thanikachalam
  \item Ondrej Slama \email{ondrej.slama@roche.com}
  \item Shimeng Huang
  \item Robin Koeger
  \item Gordon Miller
}

}
\keyword{internal}
