% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiral-package.R
\docType{package}
\name{admiral-package}
\alias{admiral}
\alias{admiral-package}
\title{admiral: ADaM in R Asset Library}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A toolbox for programming Clinical Data Interchange Standards Consortium (CDISC) compliant Analysis Data Model (ADaM) datasets in R. ADaM datasets are a mandatory part of any New Drug or Biologics License Application submitted to the United States Food and Drug Administration (FDA). Analysis derivations are implemented in accordance with the "Analysis Data Model Implementation Guide" (CDISC Analysis Data Model Team, 2021, \url{https://www.cdisc.org/standards/foundational/adam/adamig-v1-3-release-package}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiral/}
  \item \url{https://github.com/pharmaverse/admiral}
}

}
\author{
\strong{Maintainer}: Thomas Neitmann \email{thomas.neitmann@roche.com}

Authors:
\itemize{
  \item Stefan Bundfuss
  \item Ben Straub
  \item Samia Kabi
  \item Gordon Miller
  \item Teckla Akinyi
  \item Andrew Smith
  \item Konstantina Koukourikou
  \item Ross Farrugia
  \item Eric Simms
  \item Annie Yang
  \item Robin Koeger
  \item Sophie Shapcott
  \item Ojesh Upadhyay
  \item Jack McGavigan
  \item Kamila Duniec
  \item Gayatri G
  \item Alana Harris
  \item Mahdi About
  \item Pooja Kumari
  \item Claudia Carlucci
  \item Daniil Stefonishin
  \item Sadchla Mascary
  \item Zelos Zhu
  \item Jeffrey Dickinson
  \item Ania Golab
}

Other contributors:
\itemize{
  \item Michael Thorpe [contributor]
  \item Declan Hodges [contributor]
  \item Jaxon Abercrombie [contributor]
  \item Nick Ramirez [contributor]
  \item Pavan Kumar [contributor]
  \item Hamza Rahal [contributor]
  \item Yohann Omnes [contributor]
  \item Alice Ehmann [contributor]
  \item Tom Ratford [contributor]
  \item Vignesh Thanikachalam [contributor]
  \item Ondrej Slama [contributor]
  \item Shimeng Huang [contributor]
  \item James Kim [contributor]
  \item Shan Lee [contributor]
  \item Bill Denney [contributor]
  \item Syed Mubasheer [contributor]
  \item Wenyi Liu [contributor]
  \item Dinakar Kulkarni [contributor]
  \item Franciszek Walkowiak [contributor]
  \item Tamara Senior [contributor]
  \item Jordanna Morrish [contributor]
  \item Anthony Howard [contributor]
  \item Barbara O'Reilly [contributor]
  \item John Kirkpatrick [contributor]
  \item James Black [contributor]
  \item Leena Khatri [contributor]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item GlaxoSmithKline LLC [copyright holder, funder]
}

}
\concept{internal}
\keyword{internal}
