% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiral_adlb}
\alias{admiral_adlb}
\title{Lab Analysis Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3779 rows and 111 columns.
}
\source{
Derived from the \code{adlb} template, then further filtered
due to dataset size by the following USUBJIDs:
01-701-1015, 01-701-1023, 01-701-1028, 01-701-1033,
01-701-1034, 01-701-1047, 01-701-1097, 01-705-1186,
01-705-1292, 01-705-1310, 01-708-1286
}
\usage{
admiral_adlb
}
\description{
An example of lab analysis dataset
}
\seealso{
Other datasets: 
\code{\link{admiral_adsl}},
\code{\link{ex_single}},
\code{\link{example_qs}},
\code{\link{queries}},
\code{\link{queries_mh}}
}
\concept{datasets}
\keyword{datasets}
