% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiral_adlb}
\alias{admiral_adlb}
\title{Laboratory Analysis Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 83580 rows and 100 columns.
}
\source{
Derived from the \code{adsl} and \code{lb} datasets using \code{{admiral}} (\url{https://github.com/pharmaverse/admiral/blob/main/inst/templates/ad_adlb.R})
}
\usage{
admiral_adlb
}
\description{
An example laboratory analysis dataset
}
\keyword{datasets}
