% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiraldev-package.R
\docType{package}
\name{admiraldev-package}
\alias{admiraldev}
\alias{admiraldev-package}
\title{admiraldev: ADaM in R Asset Library - Dev Tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A toolbox for developers working on (pharmaverse admiral packages, \url{https://github.com/pharmaverse}). Package contains functions that are used internally to admiral and admiral extension packages to check data, variables and conditions. Package also contains utilitiy functions to help developer with documentaion, testing and genearl upkeep of admiral and admiral extension packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiraldev}
  \item \url{https://github.com/pharmaverse/admiraldev}
}

}
\author{
\strong{Maintainer}: Ben Straub \email{ben.x.straub@gsk.com}

Authors:
\itemize{
  \item Stefan Bundfuss
  \item Thomas Neitmann
  \item Samia Kabi
  \item Ross Farrugia
  \item Annie Yang
  \item Robin Koeger
  \item Pooja Kumari
  \item Syed Mubasheer
}

Other contributors:
\itemize{
  \item Ondrej Slama [contributor]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item GlaxoSmithKline LLC [copyright holder, funder]
}

}
\keyword{internal}
