% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiraldev-package.R
\docType{package}
\name{admiraldev-package}
\alias{admiraldev}
\alias{admiraldev-package}
\title{admiraldev: ADaM in R Asset Library - Dev Tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A toolbox for programming Clinical Data Standards Interchange Consortium (CDISC) compliant Analysis Data Model (ADaM) datasets in R. ADaM datasets are a mandatory part of any New Drug or Biologics License Application submitted to the United States Food and Drug Administration (FDA). Analysis derivations are implemented in accordance with the "Analysis Data Model Implementation Guide" (CDISC Analysis Data Model Team, 2021, \url{https://www.cdisc.org/standards/foundational/adam/adamig-v1-3-release-package}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiraltemplate}
  \item \url{https://github.com/pharmaverse/admiraltemplate}
}

}
\author{
\strong{Maintainer}: Open Source \email{admiralext@pharmaverse.com}

}
\keyword{internal}
