% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiraldev-package.R
\docType{package}
\name{admiraldev-package}
\alias{admiraldev}
\alias{admiraldev-package}
\title{admiraldev: Development Tools for the admiral package family}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A toolbox for developers working on admiral packages. Package contains functions that are used internally to admiral and admiral extension packages to check data, variables and conditions. Package also contains utility functions to help developer with documentation, testing and general upkeep of admiral and admiral extension packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiraldev/main/}
  \item \url{https://github.com/pharmaverse/admiraldev/}
}

}
\author{
\strong{Maintainer}: Ben Straub \email{ben.x.straub@gsk.com}

Authors:
\itemize{
  \item Stefan Bundfuss
  \item Thomas Neitmann
  \item Samia Kabi
  \item Pooja Kumari
  \item Syed Mubasheer
}

Other contributors:
\itemize{
  \item Ondrej Slama [contributor]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item GlaxoSmithKline LLC [copyright holder, funder]
}

}
\keyword{internal}
