% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiraldev-package.R
\docType{package}
\name{admiraldev-package}
\alias{admiraldev}
\alias{admiraldev-package}
\title{admiraldev: Utility Functions and Development Tools for the Admiral Package Family}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Utility functions to check data, variables and conditions for functions used in 'admiral' and 'admiral' extension packages. Additional utility helper functions to assist developers with maintaining documentation, testing and general upkeep of 'admiral' and 'admiral' extension packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiraldev/main/}
  \item \url{https://github.com/pharmaverse/admiraldev/}
  \item Report bugs at \url{https://github.com/pharmaverse/admiraldev/issues}
}

}
\author{
\strong{Maintainer}: Ben Straub \email{ben.x.straub@gsk.com}

Authors:
\itemize{
  \item Stefan Bundfuss
  \item Thomas Neitmann
  \item Samia Kabi
  \item Pooja Kumari
  \item Syed Mubasheer
  \item Ross Farrugia
  \item Sadchla Mascary
  \item Zelos Zhu
  \item Jeffrey Dickinson
  \item Ania Golab
}

Other contributors:
\itemize{
  \item Ondrej Slama [contributor]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item GlaxoSmithKline LLC [copyright holder, funder]
}

}
\keyword{internal}
