% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adsl}
\alias{adsl}
\title{Subject Level Analysis Dataset}
<<<<<<< HEAD
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 306 rows and 49 columns.}
=======
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 306 rows and 48 columns.
}
>>>>>>> devel
\source{
Derived from the \code{dm} and \code{ds} datasets using \code{{admiral}} (\url{https://github.com/pharmaverse/admiral/blob/main/inst/templates/ad_adsl.R})
}
\usage{
adsl
}
\description{
An example subject level analysis dataset
}
\keyword{datasets}
