% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{arg_name}
\alias{arg_name}
\title{Extract Argument Name from an Expression}
\usage{
arg_name(expr)
}
\arguments{
\item{expr}{An expression created inside a function using \code{substitute()}}
}
\description{
Extract Argument Name from an Expression
}
\examples{
test_fun <- function(something) {
  arg_name(substitute(something))
}

inner_function <- function(x) x
test_fun2 <- function(something) {
  arg_name(substitute(inner_function(something)))
}
}
\seealso{
Functions moved to admiraldev package: 
\code{\link{assert_db_requirements}()},
\code{\link{assert_terms}()},
\code{\link{assert_valid_queries}()},
\code{\link{backquote}()},
\code{\link{contains_vars}()},
\code{\link{dquote}()},
\code{\link{enumerate}()},
\code{\link{extend_source_datasets}()},
\code{\link{extract_vars}()},
\code{\link{format.sdg_select}()},
\code{\link{format.smq_select}()},
\code{\link{get_constant_vars}()},
\code{\link{get_duplicates}()},
\code{\link{get_source_vars}()},
\code{\link{is_auto}()},
\code{\link{quo_c}()},
\code{\link{replace_values_by_names}()},
\code{\link{squote}()},
\code{\link{valid_time_units}()},
\code{\link{validate_query}()},
\code{\link{validate_sdg_select}()},
\code{\link{validate_smq_select}()},
\code{\link{what_is_it}()}
}
\author{
Thomas Neitmann, Ondrej Slama
}
\concept{move_adm_dev}
\keyword{move_adm_dev}
