% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_character_vector}
\alias{assert_character_vector}
\title{Is an Argument a Character Vector?}
\usage{
assert_character_vector(arg, optional = FALSE)
}
\arguments{
\item{arg}{A function argument to be checked}

\item{optional}{Is the checked parameter optional? If set to \code{FALSE} and \code{arg}
is \code{NULL} then an error is thrown}
}
\value{
The function throws an error if \code{arg} is not a character vector. Otherwise,
the input is returned invisibly.
}
\description{
Checks if an argument is a character vector
}
\examples{
assert_character_vector(letters)

numbers <- 1:10
tryCatch(
  assert_character_vector(numbers),
  error = function(e) cat(e$message)
)
}
\author{
Thomas Neitmann
}
\keyword{assertion}
