% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_data_frame}
\alias{assert_data_frame}
\title{Is an Argument a Data Frame?}
\usage{
assert_data_frame(arg, required_vars = NULL, optional = FALSE)
}
\arguments{
\item{arg}{A function argument to be checked}

\item{required_vars}{A list of variables created using \code{vars()}}

\item{optional}{Is the checked parameter optional? If set to \code{FALSE} and \code{arg}
is \code{NULL} then an error is thrown}
}
\value{
The function throws an error if \code{arg} is not a data frame or if \code{arg}
is a data frame but misses any variable specified in \code{required_vars}. Otherwise,
the input is returned invisibly.
}
\description{
Checks if an argument is a data frame and (optionally) whether is contains
a set of required variables
}
\examples{
data(dm)

example_fun <- function(dataset) {
  assert_data_frame(dataset, required_vars = vars(STUDYID, USUBJID))
}

example_fun(dm)

try(example_fun(dplyr::select(dm, -STUDYID)))

try(example_fun("Not a dataset"))
}
\author{
Thomas Neitmann
}
\keyword{assertion}
