% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dt.R
\name{assert_date_imputation}
\alias{assert_date_imputation}
\title{assert date imputation
Applies assertions on the \code{date_imputation} argument to reduce
cyclomatic complexity}
\usage{
assert_date_imputation(highest_imputation, date_imputation)
}
\arguments{
\item{highest_imputation}{Highest imputation level}

\item{date_imputation}{The value to impute the day/month when a datepart is
missing.}
}
\value{
\code{invisible(NULL)}
}
\description{
assert date imputation
Applies assertions on the \code{date_imputation} argument to reduce
cyclomatic complexity
}
\keyword{internal}
