% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dt_dtm_utils.R
\name{assert_date_imputation}
\alias{assert_date_imputation}
\title{Assert \code{date_imputation}}
\usage{
assert_date_imputation(date_imputation, highest_imputation)
}
\arguments{
\item{date_imputation}{The value to impute the day/month when a datepart is
missing.

\describe{
\item{Default value}{none}
}}

\item{highest_imputation}{Highest imputation level

\describe{
\item{Default value}{none}
}}
}
\value{
asserted \code{date_imputation}
}
\description{
Applies assertions on the \code{date_imputation} argument to reduce
cyclomatic complexity
}
\details{
Asserts that date_imputation is a scalar.
Asserts that the values in \code{date_imputation} are permitted.
The permitted values in \code{date_imputation} vary by \code{highest_imputation}
}
\keyword{internal}
