% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_date_vector}
\alias{assert_date_vector}
\title{Is a Variable is a Date or Datetime Variable?}
\usage{
assert_date_vector(arg, var_name = NULL, optional = FALSE)
}
\arguments{
\item{arg}{The function argument to be checked}

\item{var_name}{a character vector name of the \code{arg}.
\code{var_name} argument is optional.}

\item{optional}{Is the checked parameter optional? If set to \code{FALSE}
and \code{arg} is \code{NULL} then an error is thrown.}
}
\value{
The function returns an error if \code{arg} is not a date or datetime variable
but otherwise returns an invisible output.
}
\description{
Checks if a variable vector is a date or datetime variable
}
\examples{
example_fun <- function(arg) {
  assert_date_vector(arg)
}

example_fun(
  as.Date("2022-01-30", tz = "UTC")
)
try(example_fun(c("2018-08-23", "2022-01-30", "1993-07-14")))
}
\author{
Sadchla Mascary
}
\keyword{assertion}
