% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query_data.R
\name{assert_db_requirements}
\alias{assert_db_requirements}
\title{Check required parameters for SMQ/SDG}
\usage{
assert_db_requirements(version, fun, queries, i, type)
}
\arguments{
\item{version}{Version provided by user}

\item{fun}{Function provided by user}

\item{queries}{Queries provide by user}

\item{i}{Index of query being checked}

\item{type}{Type of query

Should be \verb{"SMQ}" or \code{"SDG"}.}
}
\value{
An error is issued if \code{version} or \code{fun} is null.
}
\description{
If SMQs or SDGs are requested, the version and a function to access the
database must be provided. The function checks these requirements.
}
\author{
Stefan Bundfuss
}
