% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query_data.R
\name{assert_db_requirements}
\alias{assert_db_requirements}
\title{Check required parameters for SMQ/SDG}
\usage{
assert_db_requirements(
  version,
  version_arg_name,
  fun,
  fun_arg_name,
  queries,
  i,
  type
)
}
\arguments{
\item{version}{Version provided by user}

\item{version_arg_name}{Name of the argument providing the version}

\item{fun}{Function provided by user}

\item{fun_arg_name}{Name of the argument providing the function}

\item{queries}{Queries provide by user}

\item{i}{Index of query being checked}

\item{type}{Type of query

Should be \verb{"SMQ}" or \code{"SDG"}.}
}
\value{
An error is issued if \code{version} or \code{fun} is null.
}
\description{
If SMQs or SDGs are requested, the version and a function to access the
database must be provided. The function checks these requirements.
}
\seealso{
Functions moved to admiraldev package: 
\code{\link{arg_name}()},
\code{\link{assert_terms}()},
\code{\link{assert_valid_queries}()},
\code{\link{backquote}()},
\code{\link{contains_vars}()},
\code{\link{dquote}()},
\code{\link{enumerate}()},
\code{\link{extend_source_datasets}()},
\code{\link{extract_vars}()},
\code{\link{format.sdg_select}()},
\code{\link{format.smq_select}()},
\code{\link{get_constant_vars}()},
\code{\link{get_duplicates}()},
\code{\link{get_source_vars}()},
\code{\link{is_auto}()},
\code{\link{quo_c}()},
\code{\link{replace_values_by_names}()},
\code{\link{squote}()},
\code{\link{valid_time_units}()},
\code{\link{validate_query}()},
\code{\link{validate_sdg_select}()},
\code{\link{validate_smq_select}()},
\code{\link{what_is_it}()}
}
\author{
Stefan Bundfuss
}
\concept{move_adm_dev}
\keyword{move_adm_dev}
