% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dt.R
\name{assert_dt_dtm_inputs}
\alias{assert_dt_dtm_inputs}
\title{Check input assertions for date and imputation parameters
Used in \code{derive_vars_dt()} and \code{derive_vars_dtm()}.}
\usage{
assert_dt_dtm_inputs(
  new_vars_prefix,
  max_dates,
  min_dates,
  flag_imputation,
  flag_imputation_values,
  highest_imputation,
  highest_imputation_values,
  date_imputation = NULL
)
}
\arguments{
\item{new_vars_prefix}{Prefix for new variables.}

\item{max_dates}{Maximum dates of events (can be \code{NULL}).}

\item{min_dates}{Minimum dates of events (can be \code{NULL}).}

\item{flag_imputation}{The value to impute.}

\item{flag_imputation_values}{Allowed values for the \code{flag_imputation} argument.}

\item{highest_imputation}{Highest imputation level.}

\item{highest_imputation_values}{Allowed values for the \code{highest_imputation} argument.}

\item{date_imputation}{The value to impute the day/month
when a date part is missing (can be \code{NULL}).}
}
\value{
\code{invisible(NULL)}
}
\description{
Applies assertions on the \code{new_vars_prefix}, \code{max_dates}, \code{min_dates},
\code{flag_imputation}, \code{highest_imputation}, and \code{date_imputation} arguments
to reduce cyclomatic complexity.
}
\keyword{internal}
