% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_function_param}
\alias{assert_function_param}
\title{Assert Argument is a Parameter of a Function}
\usage{
assert_function_param(arg, params)
}
\arguments{
\item{arg}{The name of a function passed as a string}

\item{params}{A character vector of function parameters}
}
\value{
The function throws an error if any elements of \code{params} is not an argument of
the function given by \code{arg}
}
\description{
Assert Argument is a Parameter of a Function
}
\examples{
hello <- function(name) {
  print(sprintf("Hello \%s", name))
}

assert_function_param("hello", "name")

try(assert_function_param("hello", "surname"))
}
\seealso{
Checks for valid input and returns warning or errors messages: 
\code{\link{assert_atomic_vector}()},
\code{\link{assert_character_scalar}()},
\code{\link{assert_character_vector}()},
\code{\link{assert_data_frame}()},
\code{\link{assert_date_vector}()},
\code{\link{assert_expr_list}()},
\code{\link{assert_expr}()},
\code{\link{assert_filter_cond}()},
\code{\link{assert_function}()},
\code{\link{assert_has_variables}()},
\code{\link{assert_integer_scalar}()},
\code{\link{assert_list_element}()},
\code{\link{assert_list_of}()},
\code{\link{assert_logical_scalar}()},
\code{\link{assert_named_exprs}()},
\code{\link{assert_named}()},
\code{\link{assert_numeric_vector}()},
\code{\link{assert_one_to_one}()},
\code{\link{assert_param_does_not_exist}()},
\code{\link{assert_s3_class}()},
\code{\link{assert_same_type}()},
\code{\link{assert_symbol}()},
\code{\link{assert_unit}()},
\code{\link{assert_vars}()},
\code{\link{assert_varval_list}()}
}
\concept{assertion}
\keyword{assertion}
