% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_function_param}
\alias{assert_function_param}
\title{Assert Argument is a Parameter of a Function}
\usage{
assert_function_param(arg, params)
}
\arguments{
\item{arg}{The name of a function passed as a string}

\item{params}{A character vector of function parameters}
}
\value{
The function throws an error if any elements of \code{params} is not an argument of
the function given by \code{arg}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is \emph{deprecated}, please use \code{assert_function()} instead.
}
\seealso{
Other deprecated: 
\code{\link{assert_has_variables}()},
\code{\link{assert_named_exprs}()}
}
\concept{deprecated}
\keyword{deprecated}
