% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_has_unique_records}
\alias{assert_has_unique_records}
\title{Are records unique?}
\usage{
assert_has_unique_records(dataset, by_vars = NULL, order = NULL,
  message = NULL, message_type = "error")
}
\arguments{
\item{dataset}{The input dataset to check}

\item{by_vars}{List of by variables}

\item{order}{Order of observation
If the parameter is specified, it is checked if the observations are unique
with respect to the by variables and the order. If the check fails, the
order values are written as variables in the output.}

\item{message}{Error message
The message to be displayed if the check fails.}

\item{message_type}{Message type
If \code{'error'} is specified, an error is issued if the check fails. Otherwise
an warning is issued.}
}
\value{
\code{TRUE} if the argument is a date or date-time, \code{FALSE} otherwise
}
\description{
Checks if the reords of a dateset are unique with respect to the specified
list of by variables and order.
}
\examples{
data(ex)
assert_has_unique_records(ex,
                          by_vars = exprs(USUBJID) ,
                          order = exprs(desc(EXENDTC)))
}
\author{
Stefan Bundfuss
}
