% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_has_variables}
\alias{assert_has_variables}
\title{Does a Dataset Contain All Required Variables?}
\usage{
assert_has_variables(dataset, required_vars)
}
\arguments{
\item{dataset}{A \code{data.frame}}

\item{required_vars}{A \code{character} vector of variable names}
}
\value{
The function throws an error if any of the required variables are
missing in the input dataset. Otherwise, the dataset is returned invisibly.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is \emph{deprecated}, please use \code{assert_data_frame()} instead.
}
\details{
Checks if a dataset contains all required variables
}
\seealso{
Other deprecated: 
\code{\link{assert_function_param}()},
\code{\link{assert_named_exprs}()}
}
\concept{deprecated}
\keyword{deprecated}
