% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_logical_scalar}
\alias{assert_logical_scalar}
\title{Is an Argument a Logical Scalar (Boolean)?}
\usage{
assert_logical_scalar(arg)
}
\arguments{
\item{arg}{A function argument to be checked}
}
\value{
The function throws an error if \code{arg} is neither \code{TRUE} or \code{FALSE}. Otherwise,
the input is returned invisibly.
}
\description{
Checks if an argument is a logical scalar
}
\examples{
example_fun <- function(flag) {
  assert_logical_scalar(flag)
}

example_fun(FALSE)

tryCatch(
  example_fun(NA),
  error = function(e) cat(e$message, "\n")
)

tryCatch(
  example_fun(c(TRUE, FALSE, FALSE)),
  error = function(e) cat(e$message, "\n")
)

tryCatch(
  example_fun(1:10),
  error = function(e) cat(e$message, "\n")
)
}
\author{
Thomas Neitmann
}
\keyword{assertion}
