% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_named_exprs}
\alias{assert_named_exprs}
\title{Assert Argument is a Named List of Expressions}
\usage{
assert_named_exprs(arg, optional = FALSE)
}
\arguments{
\item{arg}{A function argument to be checked}

\item{optional}{Is the checked argument optional? If set to \code{FALSE} and \code{arg}
is \code{NULL} then an error is thrown}
}
\value{
The function throws an error if \code{arg} is not a named \code{list} of expression or
returns the input invisibly otherwise
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is \emph{deprecated}, please use \code{assert_expr_list()} instead.
}
\seealso{
Other deprecated: 
\code{\link{assert_function}()},
\code{\link{assert_has_variables}()},
\code{\link{assert_order_vars}()}
}
\concept{deprecated}
\keyword{deprecated}
