% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_order_vars}
\alias{assert_order_vars}
\title{Is an Argument a List of Order Variables?}
\usage{
assert_order_vars(arg, optional = FALSE)
}
\arguments{
\item{arg}{A function argument to be checked}

\item{optional}{Is the checked argument optional? If set to \code{FALSE} and \code{arg}
is \code{NULL} then an error is thrown}
}
\value{
The function throws an error if \code{arg} is not a list of variables or
expressions created using \code{exprs()} and returns the input invisibly
otherwise.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is \emph{deprecated}, please use \code{assert_expr_list()} instead.

Checks if an argument is a valid list of order variables/expressions created
using \code{exprs()}
}
\seealso{
Other deprecated: 
\code{\link{quo_c}()},
\code{\link{quo_not_missing}()},
\code{\link{replace_symbol_in_quo}()}
}
\concept{deprecated}
\keyword{deprecated}
