% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_query_vars.R
\name{assert_valid_queries}
\alias{assert_valid_queries}
\title{Verify if a dataset has the required format as queries dataset.}
\usage{
assert_valid_queries(queries, queries_name)
}
\arguments{
\item{queries}{A data.frame.}

\item{queries_name}{Name of the queries dataset, a string.}
}
\value{
The function throws an error if any of the requirements not met.
}
\description{
Verify if a dataset has the required format as queries dataset.
}
\details{
Check if the dataset has the following columns
\itemize{
\item \code{VAR_PREFIX}, e.g., SMQ01, CQ12
\item \code{QUERY_NAME}, non NULL, must be unique per each \code{VAR_PREFIX}
\item \code{QUERY_ID}, could be NULL, must be unique per each \code{VAR_PREFIX}
\item \code{QUERY_SCOPE}, 'BROAD', 'NARROW', or NULL
\item \code{QUERY_SCOPE_NUM}, 1, 2, or NA
\item \code{TERM_LEVEL}, e.g., AEDECOD, AELLT, ...
\item \code{TERM_NAME}, non NULL
}
}
\examples{
data("queries")
assert_valid_queries(queries, "queries")
}
\author{
Shimeng Huang, Ondrej Slama
}
