% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_query_vars.R
\name{assert_valid_queries}
\alias{assert_valid_queries}
\title{Verify if a dataset has the required format as queries dataset.}
\usage{
assert_valid_queries(queries)
}
\arguments{
\item{A}{data.frame.}
}
\value{
The function throws an error if any of the requirements not met.
}
\description{
Verify if a dataset has the required format as queries dataset.
}
\details{
Check if the dataset has the following columns
\itemize{
\item VAR_PREFIX, e.g., SMQ01, CQ12
\item QUERY_NAME, non NULL
\item QUERY_ID, could be NULL
\item QUERY_SCOPE, ‘BROAD’, ‘NARROW’, or NULL
\item TERM_LEVEL, e.g., AEDECOD, AELLT, ...
\item TERM_NAME, non NULL
This will be check by \code{\link[=validate_queries]{validate_queries()}}.
}
}
\author{
Shimeng Huang
}
