% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assert_vars}
\alias{assert_vars}
\title{Is an Argument a valid list of variables created using \code{vars()}?}
\usage{
assert_vars(arg, optional = FALSE)
}
\arguments{
\item{arg}{A function argument to be checked}

\item{optional}{Is the checked parameter optional? If set to \code{FALSE} and \code{arg}
is \code{NULL} then an error is thrown}
}
\value{
The function throws an error if \code{arg} is not a list of variables created using \code{vars()}
and returns the input invisibly otherwise.
}
\description{
Checks if an argument is a valid list of variables created using \code{vars()}
}
\examples{
list_ok <- vars(USUBJID,PARAMCD)
assert_vars(list_ok)

list_invalid<-  exprs(USUBJID,PARAMCD)
tryCatch(
  assert_vars(list_invalid1),
  error = function(e) cat(e$message)
)
list_invalid2<-  vars("USUBJID","PARAMCD", "VISIT")
tryCatch(
  assert_vars(list_invalid2),
  error = function(e) cat(e$message)
)
}
\author{
Samia Kabi
}
\keyword{assertion}
