% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoxgr_sources.R
\docType{data}
\name{atoxgr_criteria_ctcv4}
\alias{atoxgr_criteria_ctcv4}
\title{Create Metadata Holding Grading Criteria for NCI-CTCAEv4}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 40 rows and 13 columns.
}
\usage{
atoxgr_criteria_ctcv4
}
\value{
Dataset with grading criteria
}
\description{
Reads in excel file and creates metadata data set
}
\details{
Reads in excel file containing grading criteria, the spreadsheet has the following columns:
\itemize{
\item \code{SOC}: variable to hold the SOC of the lab test criteria.
\item \code{TERM}: variable to hold the term describing the criteria applied to a particular lab test,
eg. 'Anemia' or 'INR Increased'. Note: the variable is case insensitive.
\item \verb{Grade 1}: Criteria defining lab value as Grade 1.
\item \verb{Grade 2}: Criteria defining lab value as Grade 2.
\item \verb{Grade 3}: Criteria defining lab value as Grade 3.
\item \verb{Grade 4}: Criteria defining lab value as Grade 4.
\item \verb{Grade 5}: Criteria defining lab value as Grade 5.
\item \code{Definition}: Holds the definition of the lab test abnormality.
\item \code{GRADE_CRITERIA_CODE}: variable to hold code that creates grade based on defined criteria.
\item \code{SI_UNIT_CHECK}: variable to hold unit of particular lab test. Used to check against input data
if criteria is based on absolute values.
\item \code{VAR_CHECK}: List of variables required to implement lab grade criteria. Use to check against
input data.
\item \code{DIRECTION}: variable to hold the direction of the abnormality of a particular lab test
value. 'L' is for LOW values, 'H' is for HIGH values. Note: the variable is case insensitive.
\item \code{COMMENT}: Holds any information regarding rationale behind implementation of grading criteria.
}

Note: Variables \code{SOC}, \code{TERM}, \verb{Grade 1}, \verb{Grade 2},\verb{Grade 3},\verb{Grade 4},\verb{Grade 5}, \code{Definition}
are from the source document on NCI-CTC website defining the grading criteria.
From these variables only 'TERM' is used in the {admiral} code, the rest are for information and
tracability only.
}
\seealso{
BDS-Findings Functions that returns variable appended to dataset: 
\code{\link{derive_extreme_records}()},
\code{\link{derive_param_rr}()},
\code{\link{derive_summary_records}()},
\code{\link{derive_var_analysis_ratio}()},
\code{\link{derive_var_anrind}()},
\code{\link{derive_var_atoxgr_dir}()},
\code{\link{derive_var_atoxgr}()},
\code{\link{derive_var_basetype}()},
\code{\link{derive_var_base}()},
\code{\link{derive_var_chg}()},
\code{\link{derive_var_ontrtfl}()},
\code{\link{derive_var_pchg}()},
\code{\link{derive_var_shift}()}
}
\author{
Gordon Miller
}
\concept{der_bds_findings}
\keyword{der_bds_findings}
