% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query_data.R
\name{basket_select}
\alias{basket_select}
\title{Create a \code{basket_select} object}
\usage{
basket_select(name = NULL, id = NULL, scope = NULL, type, ...)
}
\arguments{
\item{name}{Name of the query used to select the definition of the query from
the company database.

\describe{
\item{Default value}{\code{NULL}}
}}

\item{id}{Identifier of the query used to select the definition of the query
from the company database.

\describe{
\item{Default value}{\code{NULL}}
}}

\item{scope}{Scope of the query used to select the definition of the query
from the company database.

\describe{
\item{Permitted values}{\code{"BROAD"}, \code{"NARROW"}, \code{NA_character_}}
\item{Default value}{\code{NULL}}
}}

\item{type}{The type argument expects a character scalar. It is passed to the
company specific get_terms() function such that the function can determine
which sort of basket is requested

\describe{
\item{Default value}{none}
}}

\item{...}{Any number of \emph{named} function arguments. Can be used to pass in company
specific conditions or flags that will then be used in user-defined function that is
passed into argument \code{get_terms_fun} for function \code{create_query_data()}.

\describe{
\item{Default value}{none}
}}
}
\value{
An object of class \code{basket_select}.
}
\description{
Create a \code{basket_select} object
}
\details{
Exactly one of \code{name} or \code{id} must be specified.
}
\seealso{
\code{\link[=create_query_data]{create_query_data()}}, \code{\link[=query]{query()}}

Source Objects:
\code{\link{censor_source}()},
\code{\link{death_event}},
\code{\link{event}()},
\code{\link{event_joined}()},
\code{\link{event_source}()},
\code{\link{flag_event}()},
\code{\link{query}()},
\code{\link{records_source}()},
\code{\link{tte_source}()}
}
\concept{source_specifications}
\keyword{source_specifications}
