% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_user_fun.R
\name{call_user_fun}
\alias{call_user_fun}
\title{Calls a Function Provided by the User}
\usage{
call_user_fun(call)
}
\arguments{
\item{call}{Call to be executed}
}
\value{
The return value of the function call
}
\description{
Calls a function provided by the user and adds the function call to the error
message if the call fails.
}
\examples{
call_user_fun(compute_bmi(
  height = 172,
  weight = 60
))

try(call_user_fun(compute_bmi(
  height = 172,
  weight = "hallo"
)))
}
\seealso{
Helper Utilities: 
\code{\link{convert_blanks_to_na}()},
\code{\link{extract_unit}()},
\code{\link{list_all_templates}()},
\code{\link{negate_vars}()},
\code{\link{signal_duplicate_records}()},
\code{\link{use_ad_template}()},
\code{\link{vars2chr}()}
}
\author{
Stefan Bundfuss
}
\concept{utils_help}
\keyword{utils_help}
