% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_utils.R
\name{chr2vars}
\alias{chr2vars}
\title{Turn a Character Vector into a List of Quosures}
\usage{
chr2vars(chr)
}
\arguments{
\item{chr}{A character vector}
}
\value{
A \code{list} of \code{quosures} as returned by \code{\link[=vars]{vars()}}
}
\description{
Turn a character vector into a list of quosures
}
\examples{
chr2vars(c("USUBJID", "AVAL"))
}
\seealso{
Other utils_quo: 
\code{\link{negate_vars}()}
}
\author{
Stefan Bundfuss
}
\concept{utils_quo}
\keyword{utils_quo}
