% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_date_vars.R
\name{compute_dtf}
\alias{compute_dtf}
\title{Derive the date imputation flag}
\usage{
compute_dtf(dtc, dt)
}
\arguments{
\item{dtc}{The date character vector ('--DTC').

A character date is expected in a format like yyyy-mm-ddThh:mm:ss (partial or complete).}

\item{dt}{The  Date vector to compare.

A date object is expected.}
}
\value{
the date imputation flag ('--DTF') (character value of 'D', 'M' , 'Y' or NA )
}
\description{
Derive the date imputation flag ('--DTF') comparing a date character vector
('--DTC') with a Date vector ('--DT').
}
\examples{

compute_dtf(dtc = "2019-07", dt = as.Date("2019-07-18"))
compute_dtf(dtc = "2019", dt = as.Date("2019-07-18"))
}
\author{
Samia Kabi
}
\keyword{computation}
\keyword{timing}
