% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_date_vars.R
\name{compute_dtf}
\alias{compute_dtf}
\title{Derive --DTF}
\usage{
compute_dtf(dtc, dt)
}
\arguments{
\item{dtc}{The --DTC used as input for --DT

A character date is expected in a format like yyyy-mm-ddThh:mm:ss (partial or complete)}

\item{dt}{The --DT resulting from the imputation of --DTC

A date object is expected}
}
\value{
the date imputation flag --DTF (character value of 'D', 'M' , 'Y' or missing )
}
\description{
Derive --DTF based on --DTC and --DT
}
\examples{
compute_dtf(dtc = "2019-07", dt = as.Date("2019-07-18"))
compute_dtf(dtc = "2019", dt = as.Date("2019-07-18"))
}
\seealso{
Other {general functions}: \code{\link{compute_tmf}},
  \code{\link{convert_dtc_to_dtm}},
  \code{\link{convert_dtc_to_dt}}, \code{\link{impute_dtc}}
}
\author{
Samia Kabi
}
\concept{{general functions}}
