% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_framingham.R
\name{compute_framingham}
\alias{compute_framingham}
\title{Compute Framingham Heart Study Cardiovascular Disease 10-Year Risk Score}
\usage{
compute_framingham(sysbp, chol, cholhdl, age, sex, smokefl, diabetfl, trthypfl)
}
\arguments{
\item{sysbp}{Systolic blood pressure

A numeric vector is expected.}

\item{chol}{Total serum cholesterol (mg/dL)

A numeric vector is expected.}

\item{cholhdl}{HDL serum cholesterol (mg/dL)

A numeric vector is expected.}

\item{age}{Age (years)

A numeric vector is expected.}

\item{sex}{Gender

A character vector is expected.}

\item{smokefl}{Smoking Status

A character vector is expected.}

\item{diabetfl}{Diabetic Status

A character vector is expected.}

\item{trthypfl}{Treated for hypertension status

A character vector is expected.}
}
\value{
A numeric vector of Framingham values
}
\description{
Computes Framingham Heart Study Cardiovascular Disease 10-Year Risk Score
(FCVD101) based on systolic blood pressure, total serum cholesterol (mg/dL),
HDL serum cholesterol (mg/dL), sex, smoking status, diabetic status,
and treated for hypertension flag.
}
\details{
The predicted probability of having cardiovascular disease (CVD)
within 10-years according to Framingham formula \link{D'Agostino, 2008} is:

For women:
Age Factor = 2.32888;
Total Chol Factor = 1.20904;
HDL Chol Factor = -0.70833;
SysBPFactor=2.76157 if not on hypertension medications, 2.82263 if on
hypertension medications, Avg Risk = 26.1931 and Risk Period;
Smoker=.52873 if smoker, 0 otherwise;
Diabetes Present=.69154 is present, 0 otherwise;
Avg Risk = 26.1931;
RiskPeriodFactor = 0.95012;

For men:
Age Factor = 3.06117;
Total Chol Factor = 1.12370;
HDL Chol Factor = -0.93263;
SysBPFactor=1.93303 if not on hypertension medications, 2.99881 if on
hypertensino medications,;
Smoker=.65451 if smoker, 0 otherwise;
Diabetes Present=.57367 is present, 0 otherwise,;
Avg Risk = 23.9802;
Risk Period Factor = 0.88936;

Equation:
\deqn{RiskFactors = (log(Age) * AgeFactor) + (log(TotalChol) * TotalCholFactor)
+ (log(CholHDL) * CholHDLFactor) + (log(SysBP) * SysBPFactor) + Smoker
+ Diabetes Present - AvgRisk}

\deqn{Risk = 100 * (1 - RiskPeriodFactor ^ exp(RiskFactors))}
}
\examples{
compute_framingham(sysbp = 133, chol = 216.16, cholhdl = 54.91, age=53,
                   sex = "M", smokefl = "N", diabetfl = "N", trthypfl = "N")

compute_framingham(sysbp = 161, chol = 186.39, cholhdl = 64.19, age=52,
                   sex = "F", smokefl = "Y", diabetfl = "N", trthypfl = "Y")

}
\author{
Alice Ehmann
}
\keyword{computation}
