% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_last_dose.R
\name{compute_ldose_idx}
\alias{compute_ldose_idx}
\title{Helper function to get the index of last dose date}
\usage{
compute_ldose_idx(exendtc, aestdtc)
}
\arguments{
\item{exendtc}{dose end date}

\item{aestdtc}{analysis date}
}
\value{
index. The last dose date is then \code{exendtc[return_value]}
}
\description{
Helper function to get the index of last dose date
}
