% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_last_dose.R
\name{compute_ldosedtm}
\alias{compute_ldosedtm}
\title{Helper function to calculate last dose}
\usage{
compute_ldosedtm(exendtc, aestdtc)
}
\arguments{
\item{exendtc}{dose end date}

\item{aestdtc}{analysis date}
}
\value{
date-time vector
}
\description{
Helper function to calculate last dose
}
