% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_adeg_params.R
\name{compute_qtc}
\alias{compute_qtc}
\title{Compute Corrected QT}
\usage{
compute_qtc(qt, rr, method)
}
\arguments{
\item{qt}{QT interval

A numeric vector is expected. It is expected that QT is measured in msec.}

\item{rr}{RR interval

A numeric vector is expected. It is expected that RR is measured in msec.}

\item{method}{Method used to QT correction

Permitted Values: \code{"Bazett"}, \code{"Fridericia"}, \code{"Sagie"}}
}
\value{
QT interval in msec
}
\description{
Computes corrected QT using Bazett's, Fridericia's or Sagie's formula.
}
\details{
Depending on the chosen \code{method} one of the following formulae is used.

\emph{Bazett}: \deqn{\frac{QT}{\sqrt{\frac{RR}{1000}}}}{QT/\sqrt(RR/1000)}

\emph{Fridericia}: \deqn{\frac{QT}{\sqrt[3]{\frac{RR}{1000}}}}{QT/(RR/1000)^(1/3)}

\emph{Sagie}: \deqn{1000\left(\frac{QT}{1000} + 0.154\left(1 - \frac{RR}{1000}\right)\right)}{
1000(QT/1000 + 0.154(1 - RR/1000))}

Usually this computation function can not be used with \verb{\%>\%}.
}
\examples{
compute_qtc(qt = 350, rr = 56.54, method = "Bazett")

compute_qtc(qt = 350, rr = 56.54, method = "Fridericia")

compute_qtc(qt = 350, rr = 56.54, method = "Sagie")
}
\author{
Stefan Bundfuss
}
\keyword{adeg}
\keyword{computation}
