% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_adeg_params.R
\name{compute_rr}
\alias{compute_rr}
\title{Compute RR Interval From Heart Rate}
\usage{
compute_rr(hr)
}
\arguments{
\item{hr}{Heart rate

A numeric vector is expected. It is expected that heart rate is measured in
beats/min.}
}
\value{
RR interval in msec:
\deqn{\frac{60000}{HR}}{60000 / HR}
}
\description{
Computes RR interval from heart rate.
}
\details{
Usually this computation function can not be used with \verb{\%>\%}.
}
\examples{
compute_rr(hr = 70.14)
}
\author{
Stefan Bundfuss
}
\keyword{adeg}
\keyword{computation}
