% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_date_vars.R
\name{compute_tmf}
\alias{compute_tmf}
\title{Derive --TMF}
\usage{
compute_tmf(dtc, dtm)
}
\arguments{
\item{dtc}{The --DTC used as input for --DT

A character date is expected in a format like yyyy-mm-ddThh:mm:ss (partial or complete)}

\item{dtm}{The --DTM resulting from the imputation of --DTC

A datetime object is expected}
}
\value{
the time imputation flag --DTF (character value of 'H', 'M' , 'S' or missing)
}
\description{
Derive --TMF based on --DTC and --DT
}
\examples{

compute_tmf(dtc = "2019-07-18T15:25", dtm = as.POSIXct("2019-07-18T15:25:00"))
compute_tmf(dtc = "2019-07-18T15", dtm = as.POSIXct("2019-07-18T15:25:00"))
compute_tmf(dtc = "2019-07-18", dtm = as.POSIXct("2019-07-18"))
}
\seealso{
Other {general functions}: \code{\link{compute_dtf}},
  \code{\link{convert_dtc_to_dtm}},
  \code{\link{convert_dtc_to_dt}}, \code{\link{impute_dtc}}
}
\author{
Samia Kabi
}
\concept{{general functions}}
\keyword{computation}
\keyword{timing}
