% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{contains_vars}
\alias{contains_vars}
\title{Returns If Argument is a Vector of Symbols or Named Expressions}
\usage{
contains_vars(arg)
}
\arguments{
\item{arg}{Input vector}
}
\value{
\code{TRUE} if
\itemize{
\item The vector is a vector of quosures and
\item all elements are a symbol or a named expression like \code{CHG = AVAL - BASE}
}

\code{FALSE} otherwise
}
\description{
Returns If Argument is a Vector of Symbols or Named Expressions
}
\examples{

contains_vars(rlang::expr(AVAL))
contains_vars(dplyr::vars(AVAL))
contains_vars(dplyr::vars(1))
contains_vars(dplyr::vars(AVAL, ID = 1))
contains_vars(dplyr::vars(AVAL, CHG = AVAL - BASE))
}
\keyword{bye_fcn_amd_dev}
