% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dt_dtm_utils.R
\name{convert_date_to_dtm}
\alias{convert_date_to_dtm}
\title{Convert a Date into a Datetime Object}
\usage{
convert_date_to_dtm(
  dt,
  highest_imputation = "h",
  date_imputation = "first",
  time_imputation = "first",
  min_dates = NULL,
  max_dates = NULL,
  preserve = FALSE
)
}
\arguments{
\item{dt}{The date to convert.

A date or character date is expected in a format like \code{yyyy-mm-ddThh:mm:ss}.

\describe{
\item{Default value}{none}
}}

\item{highest_imputation}{Highest imputation level

The \code{highest_imputation} argument controls which components of the \code{--DTC}
value are imputed if they are missing. All components up to the specified
level are imputed.

If a component at a higher level than the highest imputation level is
missing, \code{NA} is returned. For example, for \code{highest_imputation = "D"}
\code{"2020"} results in \code{NA} because the month is missing.

If \code{"n"} is specified, no imputation is performed, i.e., if any component is
missing, \code{NA} is returned.

If \code{"Y"} is specified, \code{date_imputation} should be \code{"first"} or \code{"last"}
and \code{min_dates} or \code{max_dates} should be specified respectively. Otherwise,
\code{NA} is returned if the year component is missing.

\describe{
\item{Permitted values}{\code{"Y"} (year, highest level), \code{"M"} (month), \code{"D"}
(day), \code{"h"} (hour), \code{"m"} (minute), \code{"s"} (second), \code{"n"} (none, lowest
level)}
\item{Default value}{\code{"h"}}
}}

\item{date_imputation}{The value to impute the day/month when a datepart is
missing.

A character value is expected.
\itemize{
\item The\code{"first"} and \code{"last"} keywords allow imputation to the first/last
day/month. They can also be used to impute the year if used in conjunction
with the \code{min_dates} or \code{max_dates} arguments. Some examples of this are available
\href{https://pharmaverse.github.io/admiral/cran-release/articles/imputation.html#minimummaximum-dates}{here}.
\item When \code{highest_imputation} is \code{"M"} or \code{"D"}, the \code{"mid"} keyword can also be
specified to impute missing components to the middle of the possible range:
\itemize{
\item If both month and day are missing, they are imputed as \code{"06-30"}
(middle of the year).
\item If only day is missing, it is imputed as \code{"15"} (middle of the month).
}
\item \code{"<dd>"} can be specified only if \code{highest_imputation = "D"}. Missing days are
imputed by the specified day, e.g. \code{"10"} for the 10th day of the month.
The specified day should be valid for all months as otherwise an error might be
issued. For example, \code{date_imputation = "30"} results in an invalid date of
"2024-02-30" for the partial date "2024-02".
\item \code{"<mm>-<dd>"} can be specified only if \code{highest_imputation} is \code{"M"},  e.g. \code{"06-15"}
for the 15th of June.
}

\describe{
\item{Permitted values}{a key-word, i.e. \code{"first"}, \code{"mid"}, \code{"last"}, or \code{"<mm>-<dd>"} or \code{"<dd>"}}
\item{Default value}{\code{"first"}}
}}

\item{time_imputation}{The value to impute the time when a timepart is
missing.

A character value is expected, either as a
\itemize{
\item format with hour, min and sec specified as \code{"hh:mm:ss"}: e.g. \code{"00:00:00"}
for the start of the day,
\item or as a keyword: \code{"first"},\code{"last"} to impute to the start/end of a day.
}

The argument is ignored if \code{highest_imputation = "n"}.

\describe{
\item{Permitted values}{\code{"first"}, \code{"last"}, or user-defined}
\item{Default value}{\code{"first"}}
}}

\item{min_dates}{Minimum dates

A list of dates is expected. It is ensured that the imputed date is not
before any of the specified dates, e.g., that the imputed adverse event start
date is not before the first treatment date. Only dates which are in the
range of possible dates of the \code{dtc} value are considered. The possible dates
are defined by the missing parts of the \code{dtc} date (see example below). This
ensures that the non-missing parts of the \code{dtc} date are not changed.
A date or date-time object is expected.
For example

\if{html}{\out{<div class="sourceCode r">}}\preformatted{impute_dtc_dtm(
  "2020-11",
  min_dates = list(
   ymd_hm("2020-12-06T12:12"),
   ymd_hm("2020-11-11T11:11")
  ),
  highest_imputation = "M"
)
}\if{html}{\out{</div>}}

returns \code{"2020-11-11T11:11:11"} because the possible dates for \code{"2020-11"}
range from \code{"2020-11-01T00:00:00"} to \code{"2020-11-30T23:59:59"}. Therefore
\code{"2020-12-06T12:12:12"} is ignored. Returning \code{"2020-12-06T12:12:12"} would
have changed the month although it is not missing (in the \code{dtc} date).

For date variables (not datetime) in the list the time is imputed to
\code{"00:00:00"}. Specifying date variables makes sense only if the date is
imputed. If only time is imputed, date variables do not affect the result.

\describe{
\item{Permitted values}{a list of dates, e.g. \code{list(ymd_hms("2021-07-01T04:03:01"), ymd_hms("2022-05-12T13:57:23"))}}
\item{Default value}{\code{NULL}}
}}

\item{max_dates}{Maximum dates

A list of dates is expected. It is ensured that the imputed date is not after
any of the specified dates, e.g., that the imputed date is not after the data
cut off date. Only dates which are in the range of possible dates are
considered. A date or date-time object is expected.

For date variables (not datetime) in the list the time is imputed to
\code{"23:59:59"}. Specifying date variables makes sense only if the date is
imputed. If only time is imputed, date variables do not affect the result.

\describe{
\item{Permitted values}{a list of dates, e.g. \code{list(ymd_hms("2021-07-01T04:03:01"), ymd_hms("2022-05-12T13:57:23"))}}
\item{Default value}{\code{NULL}}
}}

\item{preserve}{Preserve lower level date/time part when higher order part
is missing, e.g. preserve day if month is missing or
preserve minute when hour is missing.

For example \code{"2019---07"} would return \verb{"2019-06-07} if \code{preserve = TRUE}
(and \code{date_imputation = "mid"}).

\describe{
\item{Permitted values}{\code{TRUE}, \code{FALSE}}
\item{Default value}{\code{FALSE}}
}}
}
\value{
A datetime object
}
\description{
Convert a date (datetime, date, or date character) into a Date
vector (usually \verb{*DTM}).

\strong{Note:} This is a wrapper function for the function \code{convert_dtc_to_dtm()}.
}
\details{
This is a vector-oriented helper and is not usually called directly on a data
frame with \verb{\%>\%}.
}
\examples{
convert_date_to_dtm("2019-07-18T15:25:00")
convert_date_to_dtm(Sys.time())
convert_date_to_dtm(as.Date("2019-07-18"), time_imputation = "23:59:59")
convert_date_to_dtm("2019-07-18", time_imputation = "23:59:59")
convert_date_to_dtm("2019-07-18")
}
\seealso{
Date/Time Computation Functions that returns a vector:
\code{\link[=compute_age_years]{compute_age_years()}},
\code{\link[=compute_dtf]{compute_dtf()}},
\code{\link[=compute_duration]{compute_duration()}},
\code{\link[=compute_tmf]{compute_tmf()}},
\code{\link[=convert_dtc_to_dt]{convert_dtc_to_dt()}},
\code{\link[=convert_dtc_to_dtm]{convert_dtc_to_dtm()}},
\code{\link[=convert_xxtpt_to_hours]{convert_xxtpt_to_hours()}},
\code{\link[=impute_dtc_dt]{impute_dtc_dt()}},
\code{\link[=impute_dtc_dtm]{impute_dtc_dtm()}}
}
\concept{com_date_time}
\keyword{com_date_time}
