% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_date_vars.R
\name{convert_dtc_to_dtm}
\alias{convert_dtc_to_dtm}
\title{Convert a Date Character Vector into a Datetime Object}
\usage{
convert_dtc_to_dtm(dtc)
}
\arguments{
\item{dtc}{The --DTC date to convert.

A character date is expected in a format like yyyy-mm-ddThh:mm:ss.
A partial datetime will issue a warning.
Note: you can use impute_dtc function to build a complete datetime.}
}
\value{
a datetime  object
}
\description{
Convert a date character vector (usually '--DTC') into a Date vector (usually '--DTM').
}
\examples{
convert_dtc_to_dtm("2019-07-18T15:25:00")
convert_dtc_to_dtm("2019-07-18T00:00:00") # note Time = 00:00:00 is not printed
convert_dtc_to_dtm("2019-07-18")
}
\author{
Samia Kabi
}
\keyword{computation}
\keyword{timing}
