% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_dtm_to_dtc}
\alias{convert_dtm_to_dtc}
\title{Helper function to convert date (or date-time) objects to characters of dtc format
(-DTC type of variable)}
\usage{
convert_dtm_to_dtc(dtm)
}
\arguments{
\item{dtm}{date or date-time}
}
\value{
character
}
\description{
Helper function to convert date (or date-time) objects to characters of dtc format
(-DTC type of variable)
}
\examples{
admiral:::convert_dtm_to_dtc(as.POSIXct(Sys.time()))
admiral:::convert_dtm_to_dtc(as.Date(Sys.time()))
}
\author{
Ondrej Slama
}
\keyword{dev_utility}
