% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{convert_dtm_to_dtc}
\alias{convert_dtm_to_dtc}
\title{Helper Function to Convert Date (or Date-time) Objects to Characters of dtc Format
(-DTC type of variable)}
\usage{
convert_dtm_to_dtc(dtm)
}
\arguments{
\item{dtm}{date or date-time}
}
\value{
character
}
\description{
Helper Function to Convert Date (or Date-time) Objects to Characters of dtc Format
(-DTC type of variable)
}
\examples{
convert_dtm_to_dtc(as.POSIXct(Sys.time()))
convert_dtm_to_dtc(as.Date(Sys.time()))
}
\seealso{
Date/Time Computation Functions that returns a vector: 
\code{\link{compute_dtf}()},
\code{\link{compute_duration}()},
\code{\link{compute_tmf}()},
\code{\link{convert_date_to_dtm}()},
\code{\link{convert_dtc_to_dtm}()},
\code{\link{convert_dtc_to_dt}()},
\code{\link{impute_dtc_dtm}()}
}
\author{
Ondrej Slama
}
\concept{com_date_time}
\keyword{com_date_time}
