% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{convert_dtm_to_dtc}
\alias{convert_dtm_to_dtc}
\title{Helper Function to Convert Date (or Date-time) Objects to Characters of dtc Format
(-DTC type of variable)}
\usage{
convert_dtm_to_dtc(dtm)
}
\arguments{
\item{dtm}{date or date-time}
}
\value{
character
}
\description{
Helper Function to Convert Date (or Date-time) Objects to Characters of dtc Format
(-DTC type of variable)
}
\examples{
convert_dtm_to_dtc(as.POSIXct(Sys.time()))
convert_dtm_to_dtc(as.Date(Sys.time()))
}
\author{
Ondrej Slama
}
\keyword{dev_utility}
