% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_single_dose_dataset.R
\name{create_single_dose_dataset}
\alias{create_single_dose_dataset}
\title{Create dataset of single doses}
\usage{
create_single_dose_dataset(
  dataset,
  by_vars,
  dose_freq = EXDOSFRQ,
  start_date = ASTDT,
  end_date = AENDT,
  lookup_table = dose_freq_lookup,
  lookup_column = CDISCSubmissionValue
)
}
\arguments{
\item{dataset}{Input dataset

The columns specified by \code{by_vars}, \code{dose_freq}, \code{start_date} and the \code{end_date}
parameters are expected.}

\item{by_vars}{Variables to group by (created by \code{dplyr::vars}).

These variables are used to identify unique rows for each aggregate dose.

Example: vars(STUDYID, USUBJID, EXTRT)}

\item{dose_freq}{The dose frequency

The aggregate dosing frequency used for multiple doses in a row.

Default: EXDOSFRQ

Permitted Values: defined by lookup table.}

\item{start_date}{The start date

A date or date-time object is expected.

Default: \code{ASTDT}}

\item{end_date}{The end date

A date or date-time object is expected.

Default: \code{AENDT}}

\item{lookup_table}{The dose frequency value lookup table

The table used to look up \code{dose_freq} values and determine the appropriate
multiplier to be used for row generation. If a lookup table other than the
default is used, it must have columns \code{DoseWindow}, \code{DoseCount}, and
\code{DayConversionFactor}

Default: \code{dose_freq_lookup}

Permitted Values for \code{DoseWindow}: "DAY", "WEEK", "MONTH", "YEAR"}

\item{lookup_column}{The dose frequency value column in the lookup table

The column of \code{lookup_table}.

Default: \code{CDISCSubmissionValue} (column of \code{dose_freq_lookup})}
}
\value{
The input dataset with a single dose per row.
}
\description{
Derives dataset of single dose from aggregate dose information
}
\details{
Each aggregate dose row is split into multiple rows which each represent
a single dose.The number of completed dose periods between \code{start_date} and
\code{end_date} is calculated with \code{compute_duration} and multiplied by the count
of doses per period. \code{start_date} and \code{end_date} are then adjusted by an
appropriate increment using \code{DayConversionFactor}.
}
\examples{
data <- tibble::tribble(
  ~USUBJID, ~EXDOSFRQ, ~ASTDT, ~AENDT,
  "P01", "Q2D", lubridate::ymd("2021-01-01"), lubridate::ymd("2021-01-07"),
  "P01", "Q3D", lubridate::ymd("2021-01-08"), lubridate::ymd("2021-01-15"),
  "P01", "EVERY 2 WEEKS", lubridate::ymd("2021-01-15"), lubridate::ymd("2021-01-29")
)

create_single_dose_dataset(data, by_vars = vars(USUBJID))
}
\author{
Michael Thorpe, Andrew Smith
}
