% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_single_dose_dataset.R
\name{create_single_dose_dataset}
\alias{create_single_dose_dataset}
\title{Create dataset of single doses}
\usage{
create_single_dose_dataset(
  dataset,
  by_vars,
  dose_freq = EXDOSFRQ,
  start_date = ASTDT,
  end_date = AENDT
)
}
\arguments{
\item{dataset}{Input dataset

The columns specified by \code{by_vars}, \code{dose_freq}, \code{start_date} and the \code{end_date}
parameters are expected.}

\item{by_vars}{Variables to group by (created by \code{dplyr::vars}).

These variables are used to identify unique rows for each aggregate dose.

Example: vars(STUDYID, USUBJID, EXTRT)}

\item{dose_freq}{The dose frequency

The aggregate dosing frequency used for multiple doses in a row.

Default: EXDOSFRQ

Permitted Values: 'QxD', 'QxW', 'EVERY x DAYS', 'EVERY x WEEKS'
where x is a positive integer}

\item{start_date}{The start date

A date or date-time object is expected.

Default: \code{ASTDT}}

\item{end_date}{The end date

A date or date-time object is expected.

Default: \code{AENDT}}
}
\value{
The input dataset with a single dose per row.
}
\description{
Derives dataset of single dose from aggregate dose information
}
\details{
Each aggregate dose row is split into multiple rows which each represent
a single dose.
}
\examples{
data <- tibble::tribble(
  ~USUBJID, ~EXDOSFRQ, ~ASTDT, ~AENDT,
  "P01", "Q2D", lubridate::ymd("2021-01-01"), lubridate::ymd("2021-01-07"),
  "P01", "Q3D", lubridate::ymd("2021-01-08"), lubridate::ymd("2021-01-15"),
  "P01", "EVERY 2 WEEKS", lubridate::ymd("2021-01-15"), lubridate::ymd("2021-01-29")
)

create_single_dose_dataset(data, by_vars = vars(USUBJID))
}
\author{
Michael Thorpe
}
