% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_vignette.R
\name{dataset_vignette}
\alias{dataset_vignette}
\title{Output a Dataset in a Vignette in the admiral Format}
\usage{
dataset_vignette(dataset, display_vars = NULL, filter = NULL)
}
\arguments{
\item{dataset}{Dataset to output in the vignette}

\item{display_vars}{Variables selected to demonstrate the outcome of the derivation

Permitted Values: list of variables

Default is NULL

If \code{display_vars} is not NULL, only the selected variables are visible in the vignette while the
other variables are hidden. They can be made visible by clicking the\verb{Choose the columns to display} button.}

\item{filter}{Filter condition

The specified condition is applied to the dataset before it is displayed.

Permitted Values: a condition}
}
\value{
A HTML table
}
\description{
Output a dataset in a vignette with the pre-specified admiral format.
}
\examples{
library(admiral)
library(DT)
library(dplyr)
library(admiral.test)
data("admiral_dm")

dataset_vignette(admiral_dm)
dataset_vignette(admiral_dm,
  display_vars = vars(USUBJID, RFSTDTC, DTHDTC),
  filter = ARMCD == "Pbo"
)
}
\seealso{
Functions moved to admiraldev package: 
\code{\link{arg_name}()},
\code{\link{assert_character_scalar}()},
\code{\link{assert_character_vector}()},
\code{\link{assert_data_frame}()},
\code{\link{assert_db_requirements}()},
\code{\link{assert_filter_cond}()},
\code{\link{assert_function}()},
\code{\link{assert_has_variables}()},
\code{\link{assert_integer_scalar}()},
\code{\link{assert_list_element}()},
\code{\link{assert_list_of}()},
\code{\link{assert_logical_scalar}()},
\code{\link{assert_numeric_vector}()},
\code{\link{assert_one_to_one}()},
\code{\link{assert_order_vars}()},
\code{\link{assert_param_does_not_exist}()},
\code{\link{assert_s3_class}()},
\code{\link{assert_symbol}()},
\code{\link{assert_terms}()},
\code{\link{assert_unit}()},
\code{\link{assert_valid_queries}()},
\code{\link{assert_vars}()},
\code{\link{assert_varval_list}()},
\code{\link{backquote}()},
\code{\link{contains_vars}()},
\code{\link{dquote}()},
\code{\link{enumerate}()},
\code{\link{expect_dfs_equal}()},
\code{\link{extend_source_datasets}()},
\code{\link{extract_vars}()},
\code{\link{filter_date_sources}()},
\code{\link{format.sdg_select}()},
\code{\link{format.smq_select}()},
\code{\link{get_constant_vars}()},
\code{\link{get_duplicates}()},
\code{\link{get_source_vars}()},
\code{\link{is_auto}()},
\code{\link{quo_c}()},
\code{\link{replace_values_by_names}()},
\code{\link{squote}()},
\code{\link{suppress_warning}()},
\code{\link{valid_time_units}()},
\code{\link{validate_query}()},
\code{\link{validate_sdg_select}()},
\code{\link{validate_smq_select}()},
\code{\link{warn_if_inconsistent_list}()},
\code{\link{warn_if_invalid_dtc}()},
\code{\link{warn_if_vars_exist}()},
\code{\link{what_is_it}()}
}
\concept{move_adm_dev}
\keyword{move_adm_dev}
