% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_adeg_params.R
\name{default_qtc_paramcd}
\alias{default_qtc_paramcd}
\title{Get Default Parameter Code for Corrected QT}
\usage{
default_qtc_paramcd(method)
}
\arguments{
\item{method}{Method used to QT correction

Permitted Values: \code{"Bazett"}, \code{"Fridericia"}, \code{"Sagie"}}
}
\value{
\code{"QTCBR"} if \code{method} is \code{"Bazett"}, \code{"QTCFR"} if it's \code{"Fridericia"} or
\code{"QTLCR"} if it's \code{"Sagie"}. An error otherwise.
}
\description{
Get Default Parameter Code for Corrected QT
}
\examples{
default_qtc_paramcd("Sagie")
}
\seealso{
BDS-Findings Functions that returns variable appended to dataset: 
\code{\link{derive_derived_param}()},
\code{\link{derive_extreme_records}()},
\code{\link{derive_param_bmi}()},
\code{\link{derive_param_bsa}()},
\code{\link{derive_param_doseint}()},
\code{\link{derive_param_exist_flag}()},
\code{\link{derive_param_exposure}()},
\code{\link{derive_param_first_event}()},
\code{\link{derive_param_map}()},
\code{\link{derive_param_qtc}()},
\code{\link{derive_param_rr}()},
\code{\link{derive_param_wbc_abs}()},
\code{\link{derive_summary_records}()},
\code{\link{derive_var_analysis_ratio}()},
\code{\link{derive_var_anrind}()},
\code{\link{derive_var_basetype}()},
\code{\link{derive_var_base}()},
\code{\link{derive_var_chg}()},
\code{\link{derive_var_obs_number}()},
\code{\link{derive_var_shift}()}
}
\author{
Thomas Neitmann
}
\concept{der_bds_findings}
\keyword{der_bds_findings}
