% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_adeg_params.R
\name{default_qtc_paramcd}
\alias{default_qtc_paramcd}
\title{Get Default Parameter Code for Corrected QT}
\usage{
default_qtc_paramcd(method)
}
\arguments{
\item{method}{Method used to QT correction

Permitted Values: \code{"Bazett"}, \code{"Fridericia"}, \code{"Sagie"}}
}
\description{
Get Default Parameter Code for Corrected QT
}
\examples{
default_qtc_paramcd("Sagie")
}
\author{
Thomas Neitmann
}
