% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_derivation.R
\name{derivation_slice}
\alias{derivation_slice}
\title{Create a \code{derivation_slice} Object}
\usage{
derivation_slice(filter, args)
}
\arguments{
\item{filter}{An unquoted condition for defining the observations of the
slice}

\item{args}{Arguments of the derivation to be used for the slice

A \code{params()} object is expected.}
}
\value{
An object of class \code{derivation_slice}

An object of class \code{slice}.
}
\description{
Create a \code{derivation_slice} object as input for \code{slice_derivation()}.
}
\seealso{
\code{\link[=slice_derivation]{slice_derivation()}}, \code{\link[=params]{params()}}
}
\author{
Stefan Bundfuss
}
\keyword{source_specifications}
