% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_aage.R
\name{derive_aage}
\alias{derive_aage}
\title{Derive Analysis Age}
\usage{
derive_aage(dataset, start_date = BRTHDT, end_date = RANDDT, unit = "years")
}
\description{
\code{derive_aage()} was renamed to \code{derive_vars_aage()} to create a
more consistent API.
}
\keyword{internal}
