% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_aage.R
\name{derive_agegr_fda}
\alias{derive_agegr_fda}
\alias{derive_agegr_ema}
\title{Derive Age Groups}
\usage{
derive_agegr_fda(dataset, age_var, age_unit = NULL, new_var)

derive_agegr_ema(dataset, age_var, age_unit = NULL, new_var)
}
\arguments{
\item{dataset}{Input dataset.}

\item{age_var}{AGE variable.}

\item{age_unit}{AGE unit variable.

The AGE unit variable is used to convert AGE to 'years' so that grouping can occur.
This is only used when the age_var variable does not have a corresponding unit in the dataset.

Default: NULL

Permitted Values: 'years', 'months', 'weeks', 'days', 'hours', 'minutes', 'seconds'}

\item{new_var}{New variable to be created.}
}
\value{
\code{dataset} with new column \code{new_var} of class factor.
}
\description{
\verb{r lifecycle::badge("deprecated")}

\emph{Deprecated}, please use \code{derive_var_agegr_fda()} and \code{derive_var_agegr_ema()} instead.
}
\author{
Ondrej Slama
}
