% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_aage.R
\name{derive_agegr_fda}
\alias{derive_agegr_fda}
\alias{derive_agegr_ema}
\title{Derive age groups}
\usage{
derive_agegr_fda(dataset, age_var, new_var)

derive_agegr_ema(dataset, age_var, new_var)
}
\arguments{
\item{dataset}{Input dataset.}

\item{age_var}{AGE variable.}

\item{new_var}{New variable to be created.}
}
\value{
\code{dataset} with new column \code{new_var} of class factor.
}
\description{
Functions for deriving standardized age groups
}
\details{
\code{derive_agegr_fda} Derive age groups according to FDA ().

\code{derive_agegr_ema} Derive age groups according to EMA ().
}
\examples{
derive_agegr_fda(data.frame(age = 1:100), age_var = age, new_var = agegr1)
derive_agegr_ema(data.frame(age = 1:100), age_var = age, new_var = agegr1)
}
\author{
Ondrej Slama
}
