% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_base.R
\name{derive_baseline}
\alias{derive_baseline}
\title{Derive Basline}
\usage{
derive_baseline(bds_dataset, by, source, target)
}
\arguments{
\item{bds_dataset}{`data.frame`}

\item{by}{`character` vector of columns uniqely identifying a set
of records for which to calculate `BASE`}

\item{source}{The column from which to extract the baseline value,
e.g. `AVAL`}

\item{target}{The name of the newly created baseline column, e.g.
`BASE`}
}
\value{
A new `data.frame` containing all records and variables of the input
dataset plus the `target` variable.
}
\description{
Derive a baseline variable, e.g. `BASE`, in a BDS dataset
}
\author{
Thomas Neitmann
}
