% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_base.R
\name{derive_baseline}
\alias{derive_baseline}
\title{Derive Baseline}
\usage{
derive_baseline(dataset, by_vars, source_var, new_var)
}
\arguments{
\item{dataset}{The input dataset}

\item{by_vars}{Grouping variables uniquely identifying a set
of records for which to calculate \code{new_var}}

\item{source_var}{The column from which to extract the baseline value, e.g. \code{AVAL}}

\item{new_var}{The name of the newly created baseline column, e.g. \code{BASE}}
}
\description{
\verb{r lifecycle::badge("deprecated")}

This function is \emph{deprecated}. Please use \code{\link[=derive_var_base]{derive_var_base()}} instead.
}
\seealso{
\code{\link[=derive_var_base]{derive_var_base()}}
}
\author{
Thomas Neitmann
}
\keyword{bds}
\keyword{derivation}
