% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_baseline.R
\name{derive_baseline}
\alias{derive_baseline}
\title{Derive Basline}
\usage{
derive_baseline(bds_dataset, by, source, target)
}
\arguments{
\item{bds_dataset}{\code{data.frame}}

\item{by}{\code{character} vector of columns uniqely identifying a set
of records for which to calculate \code{BASE}}

\item{source}{The column from which to extract the baseline value,
e.g. \code{AVAL}}

\item{target}{The name of the newly created baseline column, e.g.
\code{BASE}}
}
\value{
A new \code{data.frame} containing all records and variables of the input
dataset plus the \code{target} variable.
}
\description{
Derive a baseline variable, e.g. \code{BASE}, in a BDS dataset
}
\examples{
bds_dataset <- tibble::tribble(
  ~STUDYID, ~USUBJID, ~PARAMCD,  ~AVALC,   ~ABLFL, ~BASETYPE,
  "TEST01", "PAT01",  "PARAM01", "LOW",    "Y",    "LAST",
  "TEST01", "PAT01",  "PARAM01", "LOW",    "N",    "LAST",
  "TEST01", "PAT01",  "PARAM01", "MEDIUM", "N",    "LAST",
  "TEST01", "PAT01",  "PARAM02", "HIGH",   "Y",    "LAST",
  "TEST01", "PAT01",  "PARAM02", "HIGH",   "N",    "LAST",
  "TEST01", "PAT01",  "PARAM02", "MEDIUM", "N",    "LAST"
)
derive_baseline(
  bds_dataset,
  by = c("USUBJID", "PARAMCD", "BASETYPE"),
  source = AVALC,
  target = BASEC
)

}
\author{
Thomas Neitmann
}
