% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_baseline.R
\name{derive_baseline}
\alias{derive_baseline}
\title{Derive Basline}
\usage{
derive_baseline(dataset, by_vars, source_var, new_var)
}
\arguments{
\item{dataset}{`data.frame`}

\item{by_vars}{`character` vector of columns uniqely identifying a set
of records for which to calculate `BASE`}

\item{source_var}{The column from which to extract the baseline value,
e.g. `AVAL`}

\item{new_var}{The name of the newly created baseline column, e.g.
`BASE`}
}
\value{
A new `data.frame` containing all records and variables of the input
dataset plus the `new_var` variable.
}
\description{
Derive a baseline variable, e.g. `BASE`, in a BDS dataset
}
\examples{
dataset <- tibble::tribble(
  ~STUDYID, ~USUBJID, ~PARAMCD,  ~AVALC,   ~ABLFL, ~BASETYPE,
  "TEST01", "PAT01",  "PARAM01", "LOW",    "Y",    "LAST",
  "TEST01", "PAT01",  "PARAM01", "LOW",    "N",    "LAST",
  "TEST01", "PAT01",  "PARAM01", "MEDIUM", "N",    "LAST",
  "TEST01", "PAT01",  "PARAM02", "HIGH",   "Y",    "LAST",
  "TEST01", "PAT01",  "PARAM02", "HIGH",   "N",    "LAST",
  "TEST01", "PAT01",  "PARAM02", "MEDIUM", "N",    "LAST"
)
derive_baseline(
  dataset,
  by_vars = c("USUBJID", "PARAMCD", "BASETYPE"),
  source_var = AVALC,
  new_var = BASEC
)

}
\author{
Thomas Neitmann
}
