% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_base.R
\name{derive_baseline}
\alias{derive_baseline}
\title{Derive BASE or BASEC}
\usage{
derive_baseline(bds_dataset, by, source, target)
}
\arguments{
\item{bds_dataset}{`data.frame`}

\item{by}{`character` vector of columns uniqely identifying a set
of records for which to calculate `BASE`}

\item{source}{Either `AVAL` or `AVALC`}

\item{target}{Either `BASE` or `BASEC`}
}
\value{
A new `data.frame` containing all records and variables of the input
dataset plus the `BASE` or `BASEC` variable
}
\description{
Derive the `BASE` or `BASEC` variable in a BDS dataset
}
\author{
Thomas Neitmann
}
