% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_dthcaus.R
\name{derive_dthcaus}
\alias{derive_dthcaus}
\title{Derive Death Cause and Source Domain}
\usage{
derive_dthcaus(dataset, sources)
}
\arguments{
\item{dataset}{Input dataset. \code{USUBJID} is an expected column.}

\item{sources}{A list of sources.

Each element is a list containing at least
the following fields:
\itemize{
\item dataset: the source dataset.
\item dthdom: name of the source domain as a string, e.g. "AE".
\item dthcaus: a symbol or a string --- if a symbol, e.g., \code{expr(AEDECOD)},
it is the variable in the source dataset to be used to assign values to
\code{DTHCAUS}; if a string, it is the fixed value to be assigned to \code{DTHCAUS}.
}

Additional fields can also be supplied:
\itemize{
\item order: variables to be used to sort the source dataset,
e.g., \code{exprs(AEDTHDTC)}.
\item mode: filtering "first" or "last" when multiple records exist.
}

Note: If a subject has death info from multiple sources, the one from the
first source will be kept.}
}
\value{
The input dataset with \code{DTHCAUS} and \code{DTHDOM} added.
}
\description{
Derives death cause (\code{DTHCAUS}) and source domain (\code{DTHDOM}).
}
\examples{
adsl <- tibble::tribble(
 ~STUDYID, ~USUBJID,
 "STUDY01", "PAT01",
 "STUDY01", "PAT02"
)
ae <- tibble::tribble(
  ~STUDYID, ~USUBJID, ~AEDECOD, ~AEOUT, ~AEDTHDTC,
  "STUDY01", "PAT01", "SUDDEN DEATH" , "FATAL", "2021-04-04"
)
ds <- tibble::tribble(
  ~STUDYID, ~USUBJID, ~DSDECOD, ~DSTERM, ~DSSTDTC,
  "STUDY01", "PAT02", "INFORMED CONSENT OBTAINED", "INFORMED CONSENT OBTAINED", "2021-04-03",
  "STUDY01", "PAT02", "RANDOMIZATION", "RANDOMIZATION", "2021-04-11",
  "STUDY01", "PAT02", "DEATH", "DEATH DUE TO PROGRESSION OF DISEASE", "2022-02-01"
)

src_ae <- list(
  dataset = ae,
  filter = exprs(AEOUT == "FATAL"),
  order = exprs(AEDTHDTC),
  mode = "first",
  dthdom = "AE",
  dthcaus = expr(AEDECOD)
)
src_ds <- list(
  dataset = ds,
  filter = exprs(DSDECOD == "DEATH", grepl("DEATH DUE TO", DSTERM)),
  order = exprs(DSSTDTC),
  mode = "first",
  dthdom = "DS",
  dthcaus = expr(DSTERM)
)

derive_dthcaus(adsl, list(src_ae, src_ds))
}
\author{
Shimeng Huang
}
\keyword{adsl}
