% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_dthcaus.R
\name{derive_dthcaus}
\alias{derive_dthcaus}
\title{Derive Death Cause and Source Domain}
\usage{
derive_dthcaus(dataset, ...)
}
\arguments{
\item{dataset}{Input dataset. \code{USUBJID} is an expected column.}

\item{...}{Objects of class "dthcaus_source" created by \code{dthcaus_source}.
See \code{\link[=dthcaus_source]{dthcaus_source()}}.}
}
\value{
The input dataset with \code{DTHCAUS} and \code{DTHDOM} added.
}
\description{
Derives death cause (\code{DTHCAUS}) and source domain (\code{DTHDOM}).
}
\details{
This function derives \code{DTHCAUS} along with the source domain
\code{DTHDOM}. If a subject has death info from multiple sources, the one from
the first source will be kept, so the user should provide the inputs in
the preferred order.
}
\examples{
adsl <- tibble::tribble(
 ~STUDYID, ~USUBJID,
 "STUDY01", "PAT01",
 "STUDY01", "PAT02"
)
ae <- tibble::tribble(
  ~STUDYID, ~USUBJID, ~AEDECOD, ~AEOUT, ~AEDTHDTC,
  "STUDY01", "PAT01", "SUDDEN DEATH" , "FATAL", "2021-04-04"
)
ds <- tibble::tribble(
  ~STUDYID, ~USUBJID, ~DSDECOD, ~DSTERM, ~DSSTDTC,
  "STUDY01", "PAT02", "INFORMED CONSENT OBTAINED", "INFORMED CONSENT OBTAINED", "2021-04-03",
  "STUDY01", "PAT02", "RANDOMIZATION", "RANDOMIZATION", "2021-04-11",
  "STUDY01", "PAT02", "DEATH", "DEATH DUE TO PROGRESSION OF DISEASE", "2022-02-01"
)

src_ae <- dthcaus_source(
  dataset = ae,
  filter = expr(AEOUT == "FATAL"),
  order = exprs(AEDTHDTC),
  mode = "first",
  dthdom = "AE",
  dthcaus = expr(AEDECOD)
)

src_ds <- dthcaus_source(
  dataset = ds,
  filter = expr(DSDECOD == "DEATH" & grepl("DEATH DUE TO", DSTERM)),
  order = exprs(DSSTDTC),
  mode = "first",
  dthdom = "DS",
  dthcaus = expr(DSTERM)
)

derive_dthcaus(adsl, src_ae, src_ds)
}
\seealso{
\code{\link[=dthcaus_source]{dthcaus_source()}}
}
\author{
Shimeng Huang
}
\keyword{adsl}
