% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_exposure_params.R
\name{derive_exposure_params}
\alias{derive_exposure_params}
\title{Adds a Record Computed from the Aggregated Analysis Value of another Parameter}
\usage{
derive_exposure_params(
  dataset,
  by_vars,
  input_param,
  fns,
  filter = NULL,
  set_values_to = NULL
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{by_vars} parameter, \code{PARAMCD}, \code{AVAL}, \code{AVALC}, \code{ASTDTM}
and \code{AENDTM} are expected.

The variable specified by \code{by_vars} and \code{PARAMCD} must be a unique key of
the input dataset after restricting it by the filter condition (\code{filter}
parameter) and to the parameters specified by \code{parameters}.}

\item{by_vars}{Grouping variables

For each group defined by \code{by_vars} an observation is added to the output
dataset.

\emph{Permitted Values:} list of variables}

\item{input_param}{Required parameter code

The observations where \code{PARAMCD} equals the specified value are considered to compute the
summary record.

\emph{Permitted Values:} A character of \code{PARAMCD} value}

\item{fns}{List of formulas specifying variable to use for aggregations.

This can include base functions like \code{mean()}, \code{min()}, \code{max()}, \code{median()},
\code{sd()}, or \code{sum()} or any other user-defined aggregation function.
For example, \code{fns = list(AVAL~ mean)}.

In general,
\itemize{
\item LHS refer to the one or more variable to use for summarizing.
\item RHS refer to a \strong{single} summary function.
}

In the formula representation e.g., \code{CHG ~ sum(., na.rm = TRUE)}, a \code{.}
serves as the data to be summarized which refers to the variable \code{CHG}.}

\item{filter}{Filter condition

The specified condition is applied to the input dataset before deriving the
new parameter, i.e., only observations fulfilling the condition are taken
into account.

\emph{Permitted Values:} a condition}

\item{set_values_to}{Variables to be set

The specified variables are set to the specified values for the new
observations (e.g. \code{vars(PARAMCD = "TDOSE", PARCAT1 = "OVERALL")}).

\emph{Permitted Values:} List of variable-value pairs}
}
\value{
The input dataset with a new record added
}
\description{
Adds a record computed from the aggregated analysis value of another parameter.
}
\examples{
library(dplyr)
library(lubridate)
library(stringr)
adex <- tibble::tribble(
  ~USUBJID, ~PARAMCD, ~AVAL, ~AVALC, ~VISIT, ~ASTDT, ~AENDT,
  "01-701-1015", "DOSE", 80, NA_character_, "BASELINE", ymd("2014-01-02"), ymd("2014-01-16"),
  "01-701-1015", "DOSE", 85, NA_character_, "WEEK 2", ymd("2014-01-17"), ymd("2014-06-18"),
  "01-701-1015", "DOSE", 82, NA_character_, "WEEK 24", ymd("2014-06-19"), ymd("2014-07-02"),
  "01-701-1015", "ADJ", NA, NA_character_, "BASELINE", ymd("2014-01-02"), ymd("2014-01-16"),
  "01-701-1015", "ADJ", NA, NA_character_, "WEEK 2", ymd("2014-01-17"), ymd("2014-06-18"),
  "01-701-1015", "ADJ", NA, NA_character_, "WEEK 24", ymd("2014-06-19"), ymd("2014-07-02"),
  "01-701-1017", "DOSE", 80, NA_character_, "BASELINE", ymd("2014-01-05"), ymd("2014-01-19"),
  "01-701-1017", "DOSE", 50, NA_character_, "WEEK 2", ymd("2014-01-20"), ymd("2014-05-10"),
  "01-701-1017", "DOSE", 65, NA_character_, "WEEK 24", ymd("2014-05-10"), ymd("2014-07-02"),
  "01-701-1017", "ADJ", NA, NA_character_, "BASELINE", ymd("2014-01-05"), ymd("2014-01-19"),
  "01-701-1017", "ADJ", NA, "ADVERSE EVENT", "WEEK 2", ymd("2014-01-20"), ymd("2014-05-10"),
  "01-701-1017", "ADJ", NA, NA_character_, "WEEK 24", ymd("2014-05-10"), ymd("2014-07-02")
) \%>\%
  mutate(ASTDTM = ymd_hms(paste(ASTDT, "00:00:00")), AENDTM = ymd_hms(paste(AENDT, "00:00:00")))

  # Cumulative dose
  adex \%>\%
  derive_exposure_params(
    by_vars = vars(USUBJID),
    set_values_to = vars(PARAMCD = "TDOSE", PARCAT1 = "OVERALL"),
    input_param = "DOSE",
    fns = AVAL ~ sum(., na.rm = TRUE)
  ) \%>\%
  # average dose in w2-24
  derive_exposure_params(
    by_vars = vars(USUBJID),
    filter = VISIT \%in\% c("WEEK2", "WEEK 24"),
    set_values_to = vars(PARAMCD = "AVDW224", PARCAT1 = "WEEK2-24"),
    input_param = "DOSE",
    fns = AVAL ~ mean(., na.rm = TRUE)
  ) \%>\%
  # Any dose adjustement?
  derive_exposure_params(
    by_vars = vars(USUBJID),
    set_values_to = vars(PARAMCD = "TADJ", PARCAT1 = "OVERALL"),
    input_param = "ADJ",
    fns = AVALC ~ if_else(sum(!is.na(.)) > 0, "Y", NA_character_)
  )
}
\author{
Samia Kabi
}
\keyword{bds}
\keyword{derivation}
