% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_extreme_flag.R
\name{derive_extreme_flag}
\alias{derive_extreme_flag}
\title{Adds a variable flagging the first or last observation within each by group}
\usage{
derive_extreme_flag(dataset, new_var, by_vars, order, mode = "last",
  flag_filter, check_type = "warning")
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{by_vars} parameter are expected.}

\item{new_var}{Variable to add

The specified variable is added to the output dataset. It is set to \code{"Y"}
for the first or last observation (depending on the mode) of each by group.

Permitted Values: list of name-value pairs}

\item{by_vars}{Grouping variables

Permitted Values: list of variables}

\item{order}{Sort order

The first or last observation is determined with respect to the specified
order.

Permitted Values: list of variables or functions of variables}

\item{mode}{Flag mode

Determines of the first or last observation is flagged.

Default: \code{"last"}

Permitted Values: \code{"first"}, \code{"last"}}

\item{flag_filter}{Filter for flag data

Only observations fulfilling the specified condition are taken into account
for flagging. If the parameter is not specified, all observations are
considered.

Permitted Values: a condition}

\item{check_type}{Check uniqueness?

If \code{"warning"} or \code{"error"} is specified, the specified message is issued
if the observations of the input dataset are not unique with respect to the
by variables and the order.

Default: \code{"warning"}

Permitted Values: \code{"none"}, \code{"warning"}, \code{"error"}}
}
\value{
The input dataset with the new flag variable added
}
\description{
Adds a variable flagging the first or last observation within each by group
}
\details{
For each group (with respect to the variables specified for the
\code{by_vars} parameter) the first or last observation (with respect to the
order specified for the \code{order} parameter and the flag mode) is included in
the output dataset.
}
\examples{
library(dplyr)
library(magrittr)

data("vs")

# flag last value for each patient, test, and visit, baseline observations are ignored
derive_extreme_flag(vs,
                    new_var = LASTFL,
                    by_vars = rlang::exprs(USUBJID, VSTESTCD, VISIT),
                    order = rlang::exprs(VSTPTNUM),
                    flag_filter = rlang::expr(VISIT != "BASELINE")) \%>\%
  arrange(USUBJID, VSTESTCD, VISITNUM, VSTPTNUM) \%>\%
  select(USUBJID, VSTESTCD, VISIT, VSTPTNUM, VSSTRESN, LASTFL)

}
\author{
Stefan Bundfuss
}
\keyword{adam}
\keyword{derivation}
