% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_extreme_flag.R
\name{derive_extreme_flag}
\alias{derive_extreme_flag}
\title{Add a Variable Flagging the First or Last Observation Within Each By Group}
\usage{
derive_extreme_flag(
  dataset,
  by_vars,
  order,
  new_var,
  mode,
  filter = deprecated(),
  check_type = "warning"
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{by_vars} parameter are expected.}

\item{by_vars}{Grouping variables

Permitted Values: list of variables}

\item{order}{Sort order

The first or last observation is determined with respect to the specified
order.

Permitted Values: list of variables or \verb{desc(<variable>)} function calls}

\item{new_var}{Variable to add

The specified variable is added to the output dataset. It is set to \code{"Y"}
for the first or last observation (depending on the mode) of each by group.

Permitted Values: variable name}

\item{mode}{Flag mode

Determines if the first or last observation is flagged.

Permitted Values: \code{"first"}, \code{"last"}}

\item{filter}{Deprecated, please use \code{restrict_derivation()} instead (see
examples for \code{derive_var_extreme_flag()}).}

\item{check_type}{Check uniqueness?

If \code{"warning"} or \code{"error"} is specified, the specified message is issued
if the observations of the input dataset are not unique with respect to the
by variables and the order.

Default: \code{"warning"}

Permitted Values: \code{"none"}, \code{"warning"}, \code{"error"}}
}
\value{
The input dataset with the new flag variable added
}
\description{
\verb{r lifecycle::badge("deprecated")}

\emph{Deprecated}, please use \code{derive_var_extreme_flag()} instead.

Add a variable flagging the first or last observation within each by group
}
\details{
For each group (with respect to the variables specified for the
\code{by_vars} parameter), \code{new_var} is set to "Y" for the first or last observation
(with respect to the order specified for the \code{order} parameter and the flag mode
specified for the \code{mode} parameter). Only observations included by the \code{filter} parameter
are considered for flagging.
Otherwise, \code{new_var} is set to \code{NA}.
}
\author{
Stefan Bundfuss
}
\keyword{adam}
\keyword{derivation}
