% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_extreme_records.R
\name{derive_extreme_records}
\alias{derive_extreme_records}
\title{Add the First or Last Observation for Each By Group as New Records}
\usage{
derive_extreme_records(
  dataset,
  by_vars = NULL,
  order,
  mode,
  check_type = "warning",
  filter = NULL,
  set_values_to
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{order} and the \code{by_vars} parameter are
expected.}

\item{by_vars}{Grouping variables

\emph{Permitted Values:} list of variables created by \code{vars()}}

\item{order}{Sort order

Within each by group the observations are ordered by the specified order.

\emph{Permitted Values:} list of variables or \verb{desc(<variable>)} function calls
created by \code{vars()}, e.g., \code{vars(ADT, desc(AVAL))}}

\item{mode}{Selection mode (first or last)

If \code{"first"} is specified, the first observation of each by group is added
to the input dataset. If \code{"last"} is specified, the last observation of
each by group is added to the input dataset.

\emph{Permitted Values:} \code{"first"}, \code{"last"}}

\item{check_type}{Check uniqueness?

If \code{"warning"} or \code{"error"} is specified, the specified message is issued
if the observations of the filtered input dataset are not unique with
respect to the by variables and the order.

\emph{Default:} \code{"warning"}

\emph{Permitted Values:} \code{"none"}, \code{"warning"}, \code{"error"}}

\item{filter}{Filter for observations to consider

Only observations fulfilling the specified condition are taken into account
for selecting the first or last observation. If the parameter is not
specified, all observations are considered.

\emph{Default}: \code{NULL}

\emph{Permitted Values}: a condition}

\item{set_values_to}{Variables to be set

The specified variables are set to the specified values for the new
observations. For example \code{vars(DTYPE = "MINIMUM")} defines the derivation
type for the new observations.

\emph{Permitted Values:} List of variable-value pairs}
}
\value{
The input dataset with the first or last observation of each by group
added as new observations.
}
\description{
Adds the first or last observation for each by group as new observations. It
can be used for example for adding the maximum or minimum value as a separate
visit.
}
\details{
\enumerate{
\item The input dataset is restricted as specified by the \code{filter} parameter.
\item For each group (with respect to the variables specified for the
\code{by_vars} parameter) the first or last observation (with respect to the
order specified for the \code{order} parameter and the mode specified for the
\code{mode} parameter) is selected.
\item The variables specified by the \code{set_values_to} parameter are added to
the selected observations.
\item The observations are added to input dataset.
}
}
\examples{
adlb <- tibble::tribble(
  ~USUBJID, ~AVISITN, ~AVAL, ~LBSEQ,
  "1",      1,        113,   1,
  "1",      2,        113,   2,
  "1",      3,        117,   3,
  "2",      1,        101,   1,
  "2",      2,        101,   2,
  "2",      3,         95,   3
)

# Add a new visit for the minium value (first observation if not unique)
derive_extreme_records(
  adlb,
  by_vars = vars(USUBJID),
  order = vars(AVAL, AVISITN),
  mode = "first",
  set_values_to = vars(
    AVISITN = 97,
    DTYPE = "MINIMUM"
  )
)

# Add a new visit for the maxium value (first observation used if not unique)
derive_extreme_records(
  adlb,
  by_vars = vars(USUBJID),
  order = vars(desc(AVAL), AVISITN),
  mode = "first",
  set_values_to = vars(
    AVISITN = 98,
    DTYPE = "MAXIMUM"
  )
)

# Add a new visit for the last value
derive_extreme_records(
  adlb,
  by_vars = vars(USUBJID),
  order = vars(AVISITN),
  mode = "last",
  set_values_to = vars(
    AVISITN = 99,
    DTYPE = "LOV"
  )
)
}
\author{
Stefan Bundfuss
}
\keyword{bds}
\keyword{derivation}
