% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_last_dose.R
\name{derive_last_dose}
\alias{derive_last_dose}
\title{Derive last dose date(-time)}
\usage{
derive_last_dose(dataset, dataset_ex, filter_ex = NULL,
  by_vars = vars(STUDYID, USUBJID), dose_start, dose_end, analysis_date,
  dataset_seq_var, output_var, output_datetime = TRUE,
  check_dates_only = FALSE, traceability_vars = vars())
}
\arguments{
\item{dataset}{Input AE dataset.}

\item{dataset_ex}{Input EX dataset.}

\item{filter_ex}{Filtering condition applied to EX dataset.
For example, it can be used to filter for valid dose.
Defaults to NULL.}

\item{by_vars}{Variables to join by (type \code{vars}).}

\item{dose_start}{The dose start date variable.}

\item{dose_end}{The dose end date variable.}

\item{analysis_date}{The analysis date variable.}

\item{dataset_seq_var}{The sequence variable
(this together with \code{by_vars} creates the keys of \code{dataset}).}

\item{output_var}{The output variable.}

\item{output_datetime}{Logical. Should only date or date-time variable be returned?
Defaults to \code{TRUE} (i.e. date-time variable).}

\item{check_dates_only}{Logical.
An assumption that start and end dates of treatment match is checked.
By default (\code{FALSE}), the date as well as the time component is checked.
If set to \code{TRUE}, then only the date component of those variables is checked.}

\item{traceability_vars}{A named list returned by \code{vars} listing the traceability variables,
e.g. \code{vars(DTHDOM = "DS", DTHSEQ = DSSEQ)}}
}
\value{
AE dataset with additional column \code{output_var}.
}
\description{
Derive last dose date(-time)
}
\details{
All date (date-time) variables can be characters in standard ISO format or
of date / date-time class.
For ISO format, see \code{\link{impute_dtc}}, parameter \code{dtc} for further details.

The last dose date is derived as follows:
Firstly, the \code{dataset_ex} is filtered using \code{filter_ex}, if provided.
This is useful for, for example, filtering for valid dose only.
Secondly, the datasets \code{dataset} and \code{dataset_ex} are joined using \code{by_vars}
and \code{dataset_seq_var} variable.
Thirdly, the last dose date is derived:
the last dose date is the maximum date where \code{dose_end} is lower to or equal to
\code{analysis_date}, subject to both date values are non-NA.
Lastly, the last dose date is appended to the \code{dataset} and returned to the user.
}
\examples{
data(ae); data(ex_single)
derive_last_dose(
  head(ae, 100),
  head(ex_single, 100),
  filter_ex = (EXDOSE > 0 | (EXDOSE == 0 & stringr::str_detect(EXTRT, "PLACEBO"))) &
    nchar(as.character(EXENDTC)) >= 10,
  dose_start = EXSTDTC,
  dose_end = EXENDTC,
  analysis_date = AESTDTC,
  dataset_seq_var = AESEQ,
  output_var = LDOSEDTM,
  output_datetime = TRUE,
  check_dates_only = FALSE
)

# or with traceability variables
derive_last_dose(
  head(ae, 100),
  head(ex_single, 100),
  filter_ex = (EXDOSE > 0 | (EXDOSE == 0 & stringr::str_detect(EXTRT, "PLACEBO"))) &
    nchar(as.character(EXENDTC)) >= 10,
  dose_start = EXSTDTC,
  dose_end = EXENDTC,
  analysis_date = AESTDTC,
  dataset_seq_var = AESEQ,
  output_var = LDOSEDTM,
  output_datetime = TRUE,
  check_dates_only = FALSE,
  traceability_vars = dplyr::vars(LDOSEDOM = "EX", LDOSESEQ = EXSEQ, LDOSEVAR = "EXSTDTC")
)

}
\author{
Ondrej Slama
}
