% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_last_dose.R
\name{derive_last_dose}
\alias{derive_last_dose}
\title{Derive last dose date(-time)}
\usage{
derive_last_dose(dataset, dataset_ex, filter_ex = exprs((EXDOSE > 0 |
  (EXDOSE == 0 & str_detect(EXTRT, "PLACEBO"))) & nchar(EXENDTC) >= 10),
  by_vars = exprs(STUDYID, USUBJID))
}
\arguments{
\item{dataset}{Input AE dataset.}

\item{dataset_ex}{Input EX dataset.}

\item{filter_ex}{Filtering condition applied to EX dataset.
For example, it can be used to filter for valid dose.}

\item{by_vars}{Variables to join by.}
}
\value{
AE dataset with additional columns \code{LDOSEDTM} and \code{LDOSEDT}.
}
\description{
Derive last dose date(-time)
}
\examples{
data(ae); data(ex)
derive_last_dose(ae, ex)
}
\author{
Ondrej Slama
}
