% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_last_dose_date.R
\name{derive_last_dose_date}
\alias{derive_last_dose_date}
\title{Derive Last Dose Date-Time}
\usage{
derive_last_dose_date(
  dataset,
  dataset_ex,
  filter_ex,
  by_vars = vars(STUDYID, USUBJID),
  dose_start,
  dose_end,
  new_var,
  analysis_date,
  dataset_seq_var,
  output_datetime = TRUE,
  check_dates_only,
  traceability_vars
)
}
\arguments{
\item{dataset}{Input dataset.}

\item{dataset_ex}{Input EX dataset.}

\item{filter_ex}{Filtering condition applied to EX dataset.
For example, it can be used to filter for valid dose.
Defaults to NULL.}

\item{by_vars}{Variables to join by (created by \code{dplyr::vars}).}

\item{dose_start}{The dose start date variable.}

\item{dose_end}{The dose end date variable.}

\item{new_var}{The output variable.}

\item{analysis_date}{The analysis date variable.}

\item{dataset_seq_var}{The sequence variable
(this together with \code{by_vars} creates the keys of \code{dataset}).}

\item{output_datetime}{Display \code{new_var} as datetime or as date}

\item{check_dates_only}{Logical.
An assumption that start and end dates of treatment match is checked.
By default (\code{FALSE}), the date as well as the time component is checked.
If set to \code{TRUE}, then only the date component of those variables is checked.}

\item{traceability_vars}{A named list returned by \code{\link[=vars]{vars()}} listing the traceability variables,
e.g. \code{vars(LDOSEDOM = "EX", LDOSESEQ = EXSEQ)}.
The left-hand side (names of the list elements) gives the names of the traceability variables
in the returned dataset.
The right-hand side (values of the list elements) gives the values of the traceability variables
in the returned dataset.
These can be either strings or symbols referring to existing variables.}
}
\value{
Input dataset with additional column \code{new_var}.
}
\description{
Displays the start date or start datetime from the exposure dataset relative to an adverse event
}
\details{
The last dose date
}
\examples{
library(dplyr, warn.conflicts = FALSE)
data(ae)
data(ex_single)

ae \%>\%
  head(100) \%>\%
  derive_last_dose_date(
    head(ex_single, 100),
    filter_ex = (EXDOSE > 0 | (EXDOSE == 0 & grepl("PLACEBO", EXTRT))) &
      nchar(EXENDTC) >= 10,
    dose_start = EXSTDTC,
    dose_end = EXENDTC,
    analysis_date = AESTDTC,
    dataset_seq_var = AESEQ,
    new_var = LDOSEDTM,
    output_datetime = TRUE,
    check_dates_only = FALSE,
    traceability_vars = dplyr::vars(LDOSEDOM = "EX", LDOSESEQ = EXSEQ, LDOSEVAR = "EXDOSE")
  ) \%>\%
  select(STUDYID, USUBJID, AESEQ, AESTDTC, LDOSEDTM)
}
\seealso{
\code{\link[=get_last_dose]{get_last_dose()}}
}
\author{
Ben Straub
}
\keyword{adae}
\keyword{derivation}
