% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_locf_records.R
\name{derive_locf_records}
\alias{derive_locf_records}
\title{Derive LOCF(Last observation carried forward) records}
\usage{
derive_locf_records(dataset, dataset_expected_obs, by_vars, order)
}
\arguments{
\item{dataset}{Input dataset

The columns specified by the \code{by_vars} and the \code{order}
parameter are expected.}

\item{dataset_expected_obs}{Expected observations dataset

Data frame with all the combinations of \code{PARAMCD}, \code{PARAM}, \code{AVISIT},
\code{AVISITN}, ... which are expected in the dataset is expected.}

\item{by_vars}{Grouping variables

For each group defined by \code{by_vars} an observation is added to the output
dataset. Only variables specified in \code{by_vars} will be populated
in the newly created records.}

\item{order}{List of variables for sorting a dataset

The dataset is sorted by \code{order} before carrying the last
observation (eg. \code{AVAL}) within each \code{by_vars} forward.}
}
\value{
The input dataset with the new \code{LOCF} observations added for each
\code{by_vars}. Note, a variable will only be populated in the new parameter rows
if it is specified in \code{by_vars}.
}
\description{
Adds LOCF records as new observations for each 'by group' when the data set
does not contain observations for missed visits/time points.
}
\details{
For each group (with respect to the variables specified for the
\code{by_vars} parameter) an observation is added to the output dataset from
\code{dataset_expected_obs}, if the input dataset does not contain observations
for missed visits/time points or if AVAL is missing.

For the new observations, \code{AVAL} is set to the non-missing \code{AVAL} of the
previous observation in the input dataset (when sorted by \code{order}) and
\code{DTYPE} is set to \code{LOCF}.
}
\examples{

advs <- tibble::tribble(
  ~STUDYID,  ~USUBJID,      ~PARAMCD, ~PARAM,                          ~AVAL, ~AVISITN, ~AVISIT,
  "CDISC01", "01-701-1015", "DIABP", "Diastolic Blood Pressure (mmHg)", 51,    0,       "BASELINE",
  "CDISC01", "01-701-1015", "DIABP", "Diastolic Blood Pressure (mmHg)", 50,    2,       "WEEK 2",
  "CDISC01", "01-701-1015", "DIABP", "Diastolic Blood Pressure (mmHg)", 51,    4,       "WEEK 4",
  "CDISC01", "01-701-1015", "DIABP", "Diastolic Blood Pressure (mmHg)", 50,    6,       "WEEK 6",
  "CDISC01", "01-701-1015", "SYSBP", "Systolic Blood Pressure (mmHg)", 121,    0,       "BASELINE",
  "CDISC01", "01-701-1015", "SYSBP", "Systolic Blood Pressure (mmHg)", 121,    2,       "WEEK 2",
  "CDISC01", "01-701-1015", "SYSBP", "Systolic Blood Pressure (mmHg)", 121,    4,       "WEEK 4",
  "CDISC01", "01-701-1015", "SYSBP", "Systolic Blood Pressure (mmHg)", 121,    6,       "WEEK 6",
  "CDISC01", "01-701-1028", "DIABP", "Diastolic Blood Pressure (mmHg)", 79,    0,       "BASELINE",
  "CDISC01", "01-701-1028", "DIABP", "Diastolic Blood Pressure (mmHg)", 80,    2,       "WEEK 2",
  "CDISC01", "01-701-1028", "DIABP", "Diastolic Blood Pressure (mmHg)", NA,    4,       "WEEK 4",
  "CDISC01", "01-701-1028", "DIABP", "Diastolic Blood Pressure (mmHg)", NA,    6,       "WEEK 6",
  "CDISC01", "01-701-1028", "SYSBP", "Systolic Blood Pressure (mmHg)", 130,    0,       "BASELINE",
  "CDISC01", "01-701-1028", "SYSBP", "Systolic Blood Pressure (mmHg)", 132,    2,       "WEEK 2"
)


# A dataset with all the combinations of PARAMCD, PARAM, AVISIT, AVISITN, ... which are expected.
advs_expected_obsv <- tibble::tribble(
 ~PARAMCD, ~PARAM, ~AVISITN, ~AVISIT,
  "DIABP", "Diastolic Blood Pressure (mmHg)", 0, "BASELINE",
  "DIABP", "Diastolic Blood Pressure (mmHg)", 2, "WEEK 2",
  "DIABP", "Diastolic Blood Pressure (mmHg)", 4, "WEEK 4",
  "DIABP", "Diastolic Blood Pressure (mmHg)", 6, "WEEK 6",
  "SYSBP", "Systolic Blood Pressure (mmHg)",  0, "BASELINE",
  "SYSBP", "Systolic Blood Pressure (mmHg)",  2, "WEEK 2",
  "SYSBP", "Systolic Blood Pressure (mmHg)",  4, "WEEK 4",
  "SYSBP", "Systolic Blood Pressure (mmHg)",  6, "WEEK 6"
)

# Add a new record for each 'by_vars' when the input dataset does not contain
# observations for missed visits/timepoints or when AVAL is missing.
# Set DTYPE = 'LOCF' for the newly added records.
# Carry forward the AVAL of the previous observation after sorting the dataset
# by 'order' parameter.
Final_advs <- derive_locf_records(data=advs,
                                  dataset_expected_obs=advs_expected_obsv,
                                  by_vars=vars(STUDYID, USUBJID, PARAM, PARAMCD),
                                  order=vars(STUDYID, USUBJID, PARAM, PARAMCD, AVISITN, AVISIT))
}
\seealso{
BDS-Findings Functions that returns variable appended to dataset: 
\code{\link{derive_var_analysis_ratio}()},
\code{\link{derive_var_anrind}()},
\code{\link{derive_var_atoxgr_dir}()},
\code{\link{derive_var_atoxgr}()},
\code{\link{derive_var_basetype}()},
\code{\link{derive_var_base}()},
\code{\link{derive_var_chg}()},
\code{\link{derive_var_ontrtfl}()},
\code{\link{derive_var_pchg}()},
\code{\link{derive_var_shift}()}
}
\author{
G Gayatri
}
\concept{der_bds_findings}
\keyword{der_bds_findings}
