% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param.R
\name{derive_param}
\alias{derive_param}
\title{Derive PARAM and PARAMCD}
\usage{
derive_param(dataset, mapping, by_vars = setdiff(colnames(mapping),
  c("PARAM", "PARAMCD")))
}
\arguments{
\item{dataset}{The input dataset}

\item{mapping}{A dataset containing the mapping of one or more SDTM variables
to \code{PARAM} and \code{PARAMCD}}

\item{by_vars}{The variables used to join \code{dataset} and \code{mapping}. By default,
these are all variables in \code{mapping} but \code{PARAM} and \code{PARAMCD}.}
}
\value{
The input dataset with additional columns \code{PARAM} and \code{PARAMCD}
}
\description{
Derive PARAM and PARAMCD
}
\examples{
data(vs)
data(advs_param)

vs \%>\%
  derive_param(advs_param, by_vars = c("VSTESTCD", "VSPOS")) \%>\%
  select(USUBJID, VSTESTCD, VSPOS, PARAM, PARAMCD, VSSTRESN)

## Disregard the `VSPOS` variable for mapping
vs \%>\%
  derive_param(advs_param, by_vars = c("VSTESTCD")) \%>\%
  select(USUBJID, VSTESTCD, PARAM, PARAMCD, VSSTRESN)
}
\author{
Thomas Neitmann
}
