% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_advs_params.R
\name{derive_param_bmi}
\alias{derive_param_bmi}
\title{Adds a Parameter for BMI}
\usage{
derive_param_bmi(
  dataset,
  by_vars,
  set_values_to = vars(PARAMCD = "BMI"),
  weight_code = "WEIGHT",
  height_code = "HEIGHT",
  unit_var = NULL,
  filter = NULL
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{by_vars} and the \code{unit_var} parameter,
\code{PARAMCD}, and \code{AVAL} are expected.

The variable specified by \code{by_vars} and \code{PARAMCD} must be a unique key of
the input dataset after restricting it by the filter condition (\code{filter}
parameter) and to the parameters specified by \code{weight_code} and \code{height_code}.}

\item{by_vars}{Grouping variables

Permitted Values: list of variables}

\item{set_values_to}{Variables to be set

The specified variables are set to the specified values for the new
observations. For example \code{vars(PARAMCD = "MAP")} defines the parameter
code for the new parameter.

\emph{Permitted Values:} List of variable-value pairs}

\item{weight_code}{WEIGHT parameter code

The observations where \code{PARAMCD} equals the specified value are considered
as the WEIGHT. It is expected that WEIGHT is measured in kg

Permitted Values: character value}

\item{height_code}{HEIGHT parameter code

The observations where \code{PARAMCD} equals the specified value are considered
as the HEIGHT. It is expected that HEIGHT is measured in cm

Permitted Values: character value}

\item{unit_var}{Variable providing the unit of the parameter

The variable is used to check the units of the input parameters and it is
set to \code{"kg/m^2"} for the new parameter.

Permitted Values: A variable of the input dataset}

\item{filter}{Filter condition

The specified condition is applied to the input dataset before deriving the
new parameter, i.e., only observations fulfilling the condition are taken
into account.

\emph{Permitted Values:} a condition}
}
\value{
The input dataset with the new parameter added
}
\description{
Adds a record for BMI/Body Mass Index using Weight and Height each by group
(e.g., subject and visit) where the source parameters are available.
}
\details{
The analysis value of the new parameter is derived as
\deqn{BMI = \frac{WEIGHT}{HEIGHT^2}}
}
\examples{
advs <- tibble::tribble(
  ~USUBJID,      ~PARAMCD, ~PARAM,        ~AVAL, ~AVALU, ~AVISIT,
  "01-701-1015", "HEIGHT", "Height (cm)", 147,   "cm",   "SCREENING",
  "01-701-1015", "WEIGHT", "Weight (kg)", 54.0,  "kg",   "SCREENING",
  "01-701-1015", "WEIGHT", "Weight (kg)", 54.4,  "kg",   "BASELINE",
  "01-701-1015", "WEIGHT", "Weight (kg)", 53.1,  "kg",   "WEEK 2",
  "01-701-1028", "HEIGHT", "Height (cm)", 163,   "cm",   "SCREENING",
  "01-701-1028", "WEIGHT", "Weight (kg)", 78.5,  "kg",   "SCREENING",
  "01-701-1028", "WEIGHT", "Weight (kg)", 80.3,  "kg",   "BASELINE",
  "01-701-1028", "WEIGHT", "Weight (kg)", 80.7,  "kg",   "WEEK 2"
)

derive_param_bmi (
  advs,
  by_vars = vars(USUBJID, AVISIT),
  weight_code = "WEIGHT",
  height_code = "HEIGHT",
  set_values_to = vars(
    PARAMCD = "BMI",
    PARAM = "Body Mass Index (kg/m^2)"
  )
 )
}
\author{
Pavan Kumar
}
\keyword{advs}
\keyword{derivation}
