% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_doseint.R
\name{derive_param_doseint}
\alias{derive_param_doseint}
\title{Adds a parameter for dose intensity}
\usage{
derive_param_doseint(
  dataset,
  filter = NULL,
  new_param = "TNDOSINT",
  tadm_code = "TNDOSE",
  tpadm_code = "TSNDOSE",
  by_vars
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{by_vars} parameter are expected.}

\item{filter}{Filter condition

The specified condition is applied to the input dataset before deriving the
new parameter, i.e., only observations fulfilling the condition are taken
into account.

\emph{Permitted Values:} a condition}

\item{new_param}{Parameter code to add

For the new observations \code{PARAMCD} is set to the specified value.

Permitted Values: character value}

\item{tadm_code}{Total Doses Administered parameter code

The observations where \code{PARAMCD} equals the specified value are considered
as the total dose administered.  This will be the numerator of the dose
intensity calculation.

Permitted Values: character value}

\item{tpadm_code}{Total Doses Planned parameter code

The observations where \code{PARAMCD} equals the specified value are considered
as the total planned dose.  This will be the denominator of the dose
intensity calculation.

Permitted Values: character value}

\item{by_vars}{Grouping variables

Permitted Values: list of variables}
}
\value{
The input dataset with the new flag variable added
}
\description{
Adds a parameter for dose intensity
}
\examples{
adex <- tibble::tribble(
}
\author{
Alice Ehmann
}
\keyword{adec}
\keyword{adex}
\keyword{derivation}
