% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_doseint.R
\name{derive_param_doseint}
\alias{derive_param_doseint}
\title{Adds a Parameter for Dose Intensity}
\usage{
derive_param_doseint(
  dataset,
  by_vars,
  set_values_to = vars(PARAMCD = "TNDOSINT", PARAM = "Dose Intensity (\%)"),
  tadm_code = "TNDOSE",
  tpadm_code = "TSNDOSE",
  zero_doses = NULL,
  filter = NULL
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{by_vars} and the \code{unit_var} parameter,
\code{PARAMCD}, and \code{AVAL} are expected.

The variable specified by \code{by_vars} and \code{PARAMCD} must be a unique key of
the input dataset after restricting it by the filter condition (\code{filter}
parameter) and to the parameters specified by \code{tadm_code} and \code{padm_code}.}

\item{by_vars}{Grouping variables

Permitted Values: list of variables}

\item{set_values_to}{Variables to be set

The specified variables are set to the specified values for the new
observations. For example \code{vars(PARAMCD = "MAP")} defines the parameter
code for the new parameter.

\emph{Permitted Values:} List of variable-value pairs}

\item{tadm_code}{Total Doses Administered parameter code

The observations where \code{PARAMCD} equals the specified value are considered
as the total dose administered.  The \code{AVAL} associated with this \code{PARAMCD}
will be the numerator of the dose intensity calculation.

Permitted Values: character value}

\item{tpadm_code}{Total Doses Planned parameter code

The observations where \code{PARAMCD} equals the specified value are considered
as the total planned dose.  The \code{AVAL} associated with this \code{PARAMCD}
will be the denominator of the dose intensity calculation.

Permitted Values: character value}

\item{zero_doses}{Flag indicating logic for handling 0 planned or
administered doses for a \code{by_vars} group

If some instances, the planned dose (\code{tpadm_code}) or administered dose
(\code{adm_code}) will be \code{NA} or 0.  The following options may be applied for cases
where the planned dose (\code{tpadm_code}) and/or (\code{tadm_code}) are 0.

Permitted Values:
NULL: Default, No record is returned if either the planned (\code{tpadm_code}) or
administered (\code{tadm_code}) \code{AVAL} are \code{NA}.  If the planned dose
(\code{tpadm_code}) is 0, \code{Inf} is returned. If the administered dose
(\code{tadm_code}) is 0 and the planned dose (\code{tpadm_code}) is > 0, 0 is
returned. A record must exist for both \code{tadm_code} and \code{tpadm_code} for the
dose intensity calculation to be done.

\code{Y}: Returns 100 when the planned dose (\code{tpadm_code}) is 0 and the
administered dose (\code{tadm_code}) is > 0. Returns 0 when the planned dose
(\code{tpadm_code}) is 0 and the administered dose (\code{tadm_code}) is 0.}

\item{filter}{Filter condition

The specified condition is applied to the input dataset before deriving the
new parameter, i.e., only observations fulfilling the condition are taken
into account.

\emph{Permitted Values:} a condition}
}
\value{
The input dataset with the new parameter rows added
}
\description{
Adds a record for the dose intensity for each by group
(e.g., subject and visit) where the source parameters are available.
}
\details{
The analysis value of the new parameter is derived as
Total Dose / Planned Dose * 100
}
\examples{
library(dplyr, warn.conflicts = FALSE)

adex <- tibble::tribble(
  derive_param_doseint(adex,
                       by_vars=vars(USUBJID, VISIT),
                       set_values_to = vars(PARAMCD = "TNDOSINT",
                                            PARAM = "Dose Intensity (\%)"),

  derive_param_doseint(adex,
                       by_vars=vars(USUBJID, VISIT),
                       set_values_to = vars(PARAMCD = "TDOSINT2",
                                            PARAM = "Dose Intensity with Zero Flag(\%)"),
                       zero_doses = "Y")
}
\author{
Alice Ehmann
}
\keyword{adec}
\keyword{adex}
\keyword{derivation}
