% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_extreme_record.R
\name{derive_param_extreme_record}
\alias{derive_param_extreme_record}
\title{Adds a Parameter Based on First or Last Record from Multiple Sources}
\usage{
derive_param_extreme_record(
  dataset = NULL,
  sources,
  source_datasets,
  by_vars = NULL,
  order,
  mode,
  set_values_to
)
}
\arguments{
\item{dataset}{Input dataset

\describe{
\item{Default value}{\code{NULL}}
}}

\item{sources}{Sources

A list of \code{records_source()} objects is expected.

\describe{
\item{Default value}{none}
}}

\item{source_datasets}{Source datasets

A named list of datasets is expected. The \code{dataset_name} field of
\code{records_source()} refers to the dataset provided in the list. The variables
specified by the \code{order} and the \code{by_vars} arguments are expected after applying \code{new_vars}.

\describe{
\item{Default value}{none}
}}

\item{by_vars}{Grouping variables

If the argument is specified, for each by group the observations are
selected separately.

\describe{
\item{Default value}{\code{NULL}}
}}

\item{order}{Sort order

If the argument is set to a non-null value, for each by group the first or
last observation from the source datasets is selected with respect to
the specified order. Variables created via \code{new_vars} e.g., imputed date variables,
can be specified as well (see examples below).

Please note that \code{NA} is considered as the last value. I.e., if a order
variable is \code{NA} and \code{mode = "last"}, this observation is chosen while for
\code{mode = "first"} the observation is chosen only if there are no
observations where the variable is not \code{NA}.

\describe{
\item{Permitted values}{list of expressions created by \code{exprs()}, e.g.,
\code{exprs(ADT, desc(AVAL))}}
\item{Default value}{none}
}}

\item{mode}{Selection mode (first or last)

If \code{"first"} is specified, for each by group the first observation with
respect to \code{order} is included in the output dataset. If \code{"last"} is
specified, the last observation is included in the output dataset.

\describe{
\item{Permitted values}{\code{"first"}, \code{"last"}}
\item{Default value}{none}
}}

\item{set_values_to}{Variables to be set

The specified variables are set to the specified values for the new
observations.

A list of variable name-value pairs is expected.
\itemize{
\item LHS refers to a variable.
\item RHS refers to the values to set to the variable. This can be a string, a
symbol, a numeric value or \code{NA}, e.g., \code{exprs(PARAMCD = "PD", PARAM =   "First Progressive Disease")}.
}

\describe{
\item{Default value}{none}
}}
}
\value{
The input dataset with the first or last observation of each by group
added as new observations.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} The \code{derive_param_extreme_record()}
function has been deprecated in favor of \code{derive_extreme_event()}.

Generates parameter based on the first or last observation from multiple
source datasets, based on user-defined filter, order and by group criteria.
All variables of the selected observation are kept.
}
\details{
The following steps are performed to create the output dataset:

\enumerate{
\item For each source dataset the observations as specified by
the \code{filter} element are selected.

\item Variables specified by \code{new_vars} are created for each source dataset.

\item The first or last observation (with respect to the
\code{order} variable) for each by group (specified by \code{by_vars}) from multiple sources
is selected and added to the input dataset. }
}
\seealso{
Other deprecated:
\code{\link[=date_source]{date_source()}},
\code{\link[=derive_var_dthcaus]{derive_var_dthcaus()}},
\code{\link[=derive_var_extreme_dt]{derive_var_extreme_dt()}},
\code{\link[=derive_var_extreme_dtm]{derive_var_extreme_dtm()}},
\code{\link[=derive_var_merged_summary]{derive_var_merged_summary()}},
\code{\link[=dthcaus_source]{dthcaus_source()}},
\code{\link[=get_summary_records]{get_summary_records()}}
}
\concept{deprecated}
\keyword{deprecated}
