% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_adeg_params.R
\name{derive_param_rr}
\alias{derive_param_rr}
\title{Adds a parameter for derived RR}
\usage{
derive_param_rr(dataset, new_param = "RRR", hr_code = "HR", by_vars)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{by_vars} parameter are expected.}

\item{new_param}{Parameter code to add

For the new observations \code{PARAMCD} is set to the specified value.

Permitted Values: character value}

\item{hr_code}{HR parameter code

The observations where \code{PARAMCD} equals the specified value are considered
as the heart rate assessments.

Permitted Values: character value}

\item{by_vars}{Grouping variables

Permitted Values: list of variables}
}
\value{
The input dataset with the new flag variable added
}
\description{
Adds a parameter for derived RR
}
\author{
Stefan Bundfuss
}
\keyword{adeg}
\keyword{derivation}
