% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_tte.R
\name{derive_param_tte}
\alias{derive_param_tte}
\title{Derive a Time-to-event Parameter}
\usage{
derive_param_tte(
  dataset = NULL,
  dataset_adsl,
  start_date = TRTSDT,
  start_imputation_flag = NULL,
  event_conditions,
  censor_conditions,
  set_values_to,
  subject_keys = vars(STUDYID, USUBJID)
)
}
\arguments{
\item{dataset}{Input dataset

The \code{USUBJID} variable is expected.}

\item{dataset_adsl}{ADSL input dataset
The variables specified for \code{start_date} and \code{start_imputation_flag} are
expected.}

\item{start_date}{Time to event origin date
The variable \code{STARTDT} is set to the specified date. The value is taken
from the ADSL dataset.

If the event or censoring date is before the origin date, \code{ADT} is set to
the origin date.

If the specified variable is imputed, the corresponding date imputation
flag must specified for \code{start_imputation_flag}.}

\item{start_imputation_flag}{If the start date is imputed, the corresponding date imputation flag must
be specified. The variable \code{STARTDTF} is set to the specified variable.}

\item{event_conditions}{Sources and conditions defining events. A list of
\code{tte_source()} objects is expected.}

\item{censor_conditions}{Sources and conditions defining censorings. A list
of \code{tte_source()} objects is expected.}

\item{set_values_to}{A named list returned by \code{vars()} defining the variables
to be set for the new parameter, e.g. \code{vars(PARAMCD = "OS", PARAM = "Overall Survival")}. The values must be a symbol, a character string, a
numeric value, or \code{NA}.}

\item{subject_keys}{Variables to uniquely identify a subject

A list of quosures where the expressions are symbols as returned by
\code{vars()} is expected.}
}
\value{
The input dataset with the new parameter added
}
\description{
Add a time-to-event parameter to the input dataset.
}
\details{
The following steps are performed to create the observations of the
new parameter:

\strong{Deriving the events:}

\enumerate{ \item For each event source dataset the observations as
specified by the \code{filter} element are selected. Then for each patient the
first observation (with respect to \code{date}) is selected.

\item The \code{ADT} variable is set to the variable specified by the
\code{date} element. If the date variable is a datetime variable, only
the datepart is copied. If the source variable is a character variable, it
is converted to a date. If the date is incomplete, it is imputed as the
first possible date.

\item The \code{CNSR} variable is added and set to the \code{censor} element.

\item The variables specified by the \code{set_values_to} element are
added.

\item The selected observations of all event source datasets are combined into a
single dataset.

\item For each patient the first observation (with respect to the \code{ADT}
variable) from the single dataset is selected. }

\strong{Deriving the censoring observations:}

\enumerate{ \item For each censoring source dataset the observations as
specified by the \code{filter} element are selected. Then for each patient the
last observation (with respect to \code{date}) is selected.

\item The \code{ADT} variable is set to the variable specified by the
\code{date} element. If the date variable is a datetime variable, only
the datepart is copied. If the source variable is a character variable, it
is converted to a date. If the date is incomplete, it is imputed as the
first possible date.

\item The \code{CNSR} variable is added and set to the \code{censor} element.

\item The variables specified by the \code{set_values_to} element are
added.

\item The selected observations of all censoring source datasets are
combined into a single dataset.

\item For each patient the last observation (with respect to the \code{ADT}
variable) from the single dataset is selected. }

For each subject (as defined by the \code{subject_keys} parameter) an
observation is added to the output dataset. If an event is available the
event observations is added. Otherwise the censoring observation is added.

Finally the variables as defined by the \code{set_values_to} parameter are added
for the new observations.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
data("adsl")

death <- tte_source(
  dataset = adsl,
  filter = DTHFL == "Y",
  date = DTHDT,
  set_values_to =vars(
    EVENTDESC = "DEATH",
    SRCDOM = "ADSL",
    SRCVAR = "DTHDT"))

lstalv <- tte_source(
  dataset = adsl,
  date = LSTALVDT,
  censor = 1,
  set_values_to = vars(
    EVENTDESC = "LAST KNOWN ALIVE DATE",
    SRCDOM = "ADSL",
    SRCVAR = "LSTALVDT"))

derive_param_tte(
  event_conditions = list(death),
  censor_conditions = list(lstalv),
  set_values_to = vars(PARAMCD = "OS",
                       PARAM = "Overall Survival"))
}
\author{
Stefan Bundfuss
}
\keyword{adtte}
\keyword{derivation}
