% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_wbc_abs.R
\name{derive_param_wbc_abs}
\alias{derive_param_wbc_abs}
\title{Add a parameter for lab differentials converted to absolute values}
\usage{
derive_param_wbc_abs(
  dataset,
  by_vars,
  set_values_to,
  filter_diff = LBSTRESU == "fraction of 1",
  wbc_code = "WBC",
  diff_code
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{by_vars} argument are expected to be present.

The variable specified by \code{by_vars} and \code{LBTESTCD} must be a unique key of
the input dataset after restricting it by the \code{filter_diff} condition,
and to the parameters specified by \code{wbc_code} and \code{diff_code}.}

\item{by_vars}{Grouping variables

Permitted Values: list of variables}

\item{set_values_to}{Variables to set

A named list returned by \code{vars()} defining the variables to be set for the
new parameter, e.g. \code{vars(PARAMCD = "LYMPHSI", PARAM = "Lymphocytes Abs (10^9/L) (SI Result)")} is
expected.}

\item{filter_diff}{Filter condition to identify differential records

The specified condition is applied to the input dataset before deriving the
new parameter, only the differential records specified by \code{diff_code} and satisfy the
\code{filter_diff} condition are kept.

Default: \code{LBSTRESU == "fraction of 1"}}

\item{wbc_code}{White Blood Cell (WBC) parameter

The observations where \code{LBTESTCD} equals the specified value are considered
as the WBC absolute results to use for converting the differentials.

Default: \code{"WBC"}

Permitted Values: character value}

\item{diff_code}{white blood differential parameter

The observations where \code{LBTESTCD} equals the specified value are considered
as the white blood differential lab results in percentage value (i.e. \code{LBSTRESU}
is "fraction of 1") to be converted into absolute value.}
}
\value{
The input dataset with the new parameter added
}
\description{
Add a parameter by converting lab differentials from percentage to absolute values
}
\details{
The analysis value of the new parameter is derived as
\deqn{\frac{White Blood Cell Count  * Percentage Value}{100}}.
New records are created for each group of records (grouped by \code{by_vars}) if 1) the white blood
cell component in absolute value is not already available from the input dataset, and 2) the
white blood cell absolute value (identified by \code{wbc_code}) and the white blood cell differential
(identified by \code{diff_code}) are both present.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
test_lb <- tibble::tribble(
~USUBJID, ~LBTESTCD, ~LBSTRESN, ~LBSTRESU, ~VISIT,
"P01",    "WBC",     33,   "10^9/L",       "CYCLE 1 DAY 1",
"P01",    "WBC",     38,   "10^9/L",       "CYCLE 2 DAY 1",
"P01",    "LYMLE",   90,   "fraction of 1", "CYCLE 1 DAY 1",
"P01",    "LYMLE",   70,   "fraction of 1", "CYCLE 2 DAY 1",
"P01",    "ALB",     36,   "g/L",           "CYCLE 2 DAY 1",
"P02",    "WBC",     33,   "10^9/L",        "CYCLE 1 DAY 1",
"P02",    "LYMPHSI", 29,   "10^9/L",        "CYCLE 1 DAY 1",
"P02",    "LYMLE",   87,   "fraction of 1", "CYCLE 1 DAY 1",
"P03",    "LYMLE",   89,   "fraction of 1", "CYCLE 1 DAY 1"
)

derive_param_wbc_abs (dataset = test_lb,
                      by_vars = vars(USUBJID, VISIT),
                      set_values_to = vars(PARAMCD = "LYMPHSI",
                                           PARAM = "Lymphocytes Abs (10^9/L) (SI Result)",
                                           DTYPE = "CALCULATION",
                                           AVALU = "10^9/L"),
                      filter_diff = LBSTRESU == "fraction of 1",
                      wbc_code = "WBC",
                      diff_code = "LYMLE")
}
\author{
Annie Yang
}
\keyword{adlb}
\keyword{bds}
\keyword{derivation}
