% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_query_vars.R
\name{derive_query_vars}
\alias{derive_query_vars}
\title{Derive query variables.}
\usage{
derive_query_vars(dataset, queries, dataset_keys)
}
\arguments{
\item{dataset}{Input dataset.

The columns specified by \code{dataset_keys} are expected.}

\item{queries}{A data.frame containing the following required columns:
\itemize{
\item VAR_PREFIX, e.g., SMQ01, CQ12
\item QUERY_NAME, non NULL
\item QUERY_ID, could be NULL
\item QUERY_SCOPE, ‘BROAD’, ‘NARROW’, or NULL
\item TERM_LEVEL, e.g., AEDECOD, AELLT, ...
\item TERM_NAME, non NULL
This will be check by \code{\link[=validate_queries]{validate_queries()}}.
}}

\item{dataset_keys}{A vector of column names that can be used to identify
a unique record in the \code{dataset}. e.g. for ADAE, this would be
\code{c("USUBJID", "ASTDTM", "AETERM", "AESEQ")}.}
}
\description{
Derive query variables.
}
\details{
If \code{VAR_PREFIX} starts with "SMQ" or "SDG", then the corresponding
"NAM", "CD", and "SC" variables are derived; if starts with "CQ", only
"NAM" variable is derived.
}
\author{
Shimeng Huang
}
\keyword{adae}
\keyword{adcm}
