% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_query_vars.R
\name{derive_query_vars}
\alias{derive_query_vars}
\title{Derive query variables.}
\usage{
derive_query_vars(dataset, queries)
}
\arguments{
\item{dataset}{Input dataset.}

\item{queries}{A data.frame containing required columns \code{VAR_PREFIX},
\code{QUERY_NAME}, \code{TERM_LEVEL}, \code{TERM_NAME}, and optional columns
\code{QUERY_ID}, \code{QUERY_SCOPE}, \code{QUERY_SCOPE_NUM}.

The content of the dataset will be verified by \code{\link[=assert_valid_queries]{assert_valid_queries()}}.}
}
\value{
The input dataset with query variables derived.
}
\description{
Derive query variables.
}
\details{
For each unique element in \code{VAR_PREFIX}, the corresponding "NAM"
variable will be created. For each unique \code{VAR_PREFIX}, if \code{QUERY_ID} is
not "" or NA, then the corresponding "CD" variable is created; similarly,
if \code{QUERY_SCOPE} is not "" or NA, then the corresponding "SC" variable will
be created; if \code{QUERY_SCOPE_NUM} is not "" or NA, then the corresponding
"SCN" variable will be created.

For each record in \code{dataset}, the "NAM" variable takes the value of
\code{QUERY_NAME} if the value of \code{TERM_NAME} in \code{queries} matches the value
in the \code{TERM_LEVEL} column of \code{dataset}. The "CD", "SC", and "SCN"
variables are derived accordingly based on \code{QUERY_ID}, \code{QUERY_SCOPE}, and
\code{QUERY_SCOPE_NUM} respectively, whenever not missing.
}
\examples{
data("queries")
adae <- tibble::tribble(
~USUBJID, ~ASTDTM, ~AETERM, ~AESEQ, ~AEDECOD, ~AELLT,
"01", "2020-06-02 23:59:59", "ALANINE AMINOTRANSFERASE ABNORMAL",
3, "Alanine aminotransferase abnormal", NA_character_,
"02", "2020-06-05 23:59:59", "BASEDOW'S DISEASE",
5, "Basedow's disease", NA_character_,
"03", "2020-06-07 23:59:59", "SOME TERM",
2, "Some query", "Some term",
"05", "2020-06-09 23:59:59", "ALVEOLAR PROTEINOSIS",
7, "Alveolar proteinosis", NA_character_)
derive_query_vars(adae, queries)
}
\seealso{
\code{\link[=assert_valid_queries]{assert_valid_queries()}}
}
\author{
Ondrej Slama, Shimeng Huang
}
\keyword{adae}
\keyword{adcm}
\keyword{derivations}
