% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_query_vars.R
\name{derive_query_vars}
\alias{derive_query_vars}
\title{Derive query variables.}
\usage{
derive_query_vars(dataset, queries)
}
\arguments{
\item{dataset}{Input dataset.

The columns specified by \code{dataset_keys} are expected.}

\item{queries}{A data.frame containing the following required columns:
\itemize{
\item VAR_PREFIX, e.g., SMQ01, CQ12
\item QUERY_NAME, non NULL
\item QUERY_ID, could be NULL
\item QUERY_SCOPE, ‘BROAD’, ‘NARROW’, or NULL
\item TERM_LEVEL, e.g., AEDECOD, AELLT, ...
\item TERM_NAME, non NULL
This will be check by \code{\link[=assert_valid_queries]{assert_valid_queries()}}.
}}
}
\description{
Derive query variables.
}
\details{
For each unique element in \code{VAR_PREFIX}, the corresponding "NAM"
variable is created. For "SMQ" or "SGD", if \code{QUERY_ID} is not "" or NA,
then the corresponding "CD" variable is created ; similarly, if \code{QUERY_SCOPE}
is not "" or NA, then the corresponding "SC" variable is created.
}
\examples{
data("queries")
adae <- tibble::tribble(
  ~USUBJID, ~ASTDTM, ~AETERM, ~AESEQ, ~AEDECOD, ~AELLT,
  "01", "2020-06-02 23:59:59", "ALANINE AMINOTRANSFERASE ABNORMAL",
  3, "Alanine aminotransferase abnormal", NA_character_,
  "02", "2020-06-05 23:59:59", "BASEDOW'S DISEASE",
  5, "Basedow's disease", NA_character_,
  "02", "2020-06-05 23:59:59", "ALVEOLAR PROTEINOSIS",
  1, "Alveolar proteinosis", NA_character_,
  "03", "2020-06-07 23:59:59", "SOME TERM", 2, "Some query", "Some term"
)
derive_query_vars(adae, queries)
}
\seealso{
\code{\link[=assert_valid_queries]{assert_valid_queries()}}
}
\author{
Ondrej Slama, Shimeng Huang
}
\keyword{adae}
\keyword{adcm}
