% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_suppqual_vars.R
\name{derive_suppqual_vars}
\alias{derive_suppqual_vars}
\title{Join Supplementary Qualifier Variables into the Parent SDTM Domain}
\usage{
derive_suppqual_vars(dataset, dataset_suppqual, domain = NULL)
}
\description{
\verb{r lifecycle::badge("deprecated")}

\code{derive_suppqual_vars()} was renamed to \code{derive_vars_suppqual()} to create a
more consistent API.
}
\keyword{internal}
