% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_ady.R
\name{derive_var_ady}
\alias{derive_var_ady}
\title{Derive Analysis Study Day}
\usage{
derive_var_ady(dataset, startdate = TRTSDT, enddate = ADT)
}
\arguments{
\item{dataset}{Input dataset

  The columns specified by the `startdate` and the `enddate` parameter are
  expected.}

\item{startdate}{The start date column, e.g., date of first treatment

  A date or date-time object column is expected.

  The default is ``TRTSDT``.}

\item{enddate}{The end date column for which the study day should be derived

  A date or date-time object column is expected.

  The default is ``ADT``}
}
\value{
The input dataset with ``ADY`` column added
}
\description{
Adds the analysis study day (``ADY``) to the dataset, i.e., study day of analysis date.
}
\details{
The study day is derived as number of days from the start date
  to the end date. If it is nonnegative, one is added. I.e., the study day of the
  start date is 1.
}
\examples{
data <- tibble::tribble(
  ~TRTSDT, ~ADT,
  ymd('2020-01-01'), ymd('2020-02-24'))

derive_var_ady(data)

}
\author{
Stefan Bundfuss
}
