% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_ady.R
\name{derive_var_ady}
\alias{derive_var_ady}
\title{Derive Analysis Study Day}
\usage{
derive_var_ady(dataset, reference_date = TRTSDT, date = ADT)
}
\arguments{
\item{dataset}{Input dataset

The columns specified by the \code{reference_date} and the \code{date} parameter are
expected.}

\item{reference_date}{The start date column, e.g., date of first treatment

A date or date-time object column is expected.

The default is \code{TRTSDT}.}

\item{date}{The end date column for which the study day should be derived

A date or date-time object column is expected.

The default is \code{ADT}}
}
\value{
The input dataset with \code{ADY} column added
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#questioning'><img src='figures/lifecycle-questioning.svg' alt='Questioning lifecycle'></a>}}{\strong{Questioning}}

Adds the analysis study day (\code{ADY}) to the dataset, i.e., study day of analysis date.
}
\details{
The study day is derived as number of days from the start date
to the end date. If it is non-negative, one is added. I.e., the study day of the
start date is 1.
}
\examples{
data <- tibble::tribble(
  ~TRTSDT, ~ADT,
  lubridate::ymd("2020-01-01"), lubridate::ymd("2020-02-24")
)

derive_var_ady(data)
}
\author{
Stefan Bundfuss
}
\keyword{bds}
\keyword{derivation}
\keyword{timing}
