% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_astdy.R
\name{derive_var_astdy}
\alias{derive_var_astdy}
\title{Derive Analysis Start Relative Day}
\usage{
derive_var_astdy(dataset, reference_date = TRTSDT, date = ASTDT)
}
\arguments{
\item{dataset}{Input dataset

The columns specified by the \code{reference_date} and the \code{date} parameter are
expected.}

\item{reference_date}{The start date column, e.g., date of first treatment

A date or date-time object column is expected.

Refer to \code{derive_vars_dt()} to impute and derive a date from a date character vector to a date object.

The default is \code{TRTSDT}.}

\item{date}{The end date column for which the study day should be derived

A date or date-time object column is expected.

Refer to \code{derive_vars_dt()} to impute and derive a date from a date character vector to a date object.

The default is \code{ASTDT}}
}
\value{
The input dataset with \code{ASTDY} column added
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is \emph{deprecated}, please use \code{derive_vars_dy()} instead.

Adds the analysis start relative day (\code{ASTDY}) to the dataset, i.e., study
day of analysis start date.
}
\details{
The study day is derived as number of days from the start date
to the end date. If it is nonnegative, one is added. I.e., the study day of the
start date is 1.
}
\author{
Stefan Bundfuss
}
\keyword{bds}
\keyword{derivation}
\keyword{occds}
\keyword{timing}
