% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_atoxgr.R
\name{derive_var_atoxgr_dir}
\alias{derive_var_atoxgr_dir}
\title{Derive Lab toxicity grade 0 - 4}
\usage{
derive_var_atoxgr_dir(
  dataset,
  new_var,
  tox_description_var,
  meta_criteria,
  criteria_direction
)
}
\arguments{
\item{dataset}{Input data set

The columns specified by \code{tox_description_var} parameter is expected.}

\item{new_var}{Name of the character grade variable to create.}

\item{tox_description_var}{Variable containing the description of the grading
criteria eg. Anemia.}

\item{meta_criteria}{metadata data set holding the criteria (normally a case statement)

The metadata should have the following variables:
\itemize{
\item \code{TERM}: variable to hold the term describing the criteria applied to a particular lab test,
eg. 'Anemia' or 'INR Increased'. Note: the variable is case insensitive.
\item \code{DIRECTION}: variable to hold the direction of the abnormality of a particular lab test
value. 'L' is for LOW values, 'H' is for HIGH values. Note: the variable is case insensitive.
\item \code{SI_UNIT_CHECK}: variable to hold unit of particular lab test. Used to check against input data
if criteria is based on absolute values.
\item \code{IMPLEMENTATION}: variable to hold case statement to create grade based on defined criteria.
}}

\item{criteria_direction}{Direction (L= Low, H = High) of toxicity grade.}
}
\value{
The input dataset with the character grade added
}
\description{
Derives a character variable containing lab grade based on severity/toxicity criteria.
}
\details{
\code{new_var} is derived with values NA, '0', '1', '2', '3', '4', where '4' is the most
severe grade
\itemize{
\item '4' is where the lab value satisfies the criteria for grade 4.
\item '3' is where the lab value satisfies the criteria for grade 3.
\item '2' is where the lab value satisfies the criteria for grade 2.
\item '1' is where the lab value satisfies the criteria for grade 1.
\item '0' is where a grade can be derived and is not grade '1', '2', '3' or '4'.
\item NA is where a grade cannot be derived.
}
}
\examples{
library(dplyr, warn.conflicts = FALSE)

data <- tibble::tribble(
  ~ATOXDSCL,                     ~AVAL,  ~ANRLO,   ~ANRHI,
  "Hypoglycemia",                119,    4,        7,
  "Hypoglycemia",                120,    4,        7,
  "Anemia",                      129,    120,      180,
  "White blood cell decreased",  10,     5,        20,
  "White blood cell decreased",   15,     5,        20,
  "Anemia",                      140,    120,      180,
)

derive_var_atoxgr_dir(adlb,
  new_var = ATOXGRL,
  tox_description_var = ATOXDSCL,
  meta_criteria = atoxgr_criteria_ctcv4,
  criteria_direction = "L"
)

data <- tibble::tribble(
  ~ATOXDSCH,                     ~AVAL,  ~ANRLO,   ~ANRHI,
  "Hyperglycemia",               119,    4,        7,
  "Hyperglycemia",               120,    4,        7,
  "Hemoglobin increased",        129,    120,      180,
  "Lymphocyte count increased",  4,      1,        4,
  "Lymphocyte count increased",  2,      1,        4,
  "Hemoglobin increased",        140,    120,      180,
)

derive_var_atoxgr_dir(adlb,
  new_var = ATOXGRH,
  tox_description_var = ATOXDSCH,
  meta_criteria = atoxgr_criteria_ctcv4,
  criteria_direction = "H"
)
}
\author{
Gordon Miller
}
\keyword{adam}
\keyword{adlb}
\keyword{bds}
\keyword{derivation}
