% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_baseline.R
\name{derive_var_base}
\alias{derive_var_base}
\title{Derive BASE}
\usage{
derive_var_base(bds_dataset, by = c("USUBJID", "PARAMCD", "BASETYPE"))
}
\arguments{
\item{bds_dataset}{\code{data.frame}}

\item{by}{\code{character} vector of columns uniqely identifying a set
of records for which to calculate \code{BASE}}
}
\value{
A new \code{data.frame} containing all records and variables of the input
dataset plus the \code{BASE} variable
}
\description{
Derive the \code{BASE} variable in a BDS dataset
}
\examples{
bds_dataset <- tibble::tribble(
  ~STUDYID, ~USUBJID, ~PARAMCD,  ~AVAL, ~ABLFL, ~BASETYPE,
  "TEST01", "PAT01",  "PARAM01", 10.12, "Y",    "LAST",
  "TEST01", "PAT01",  "PARAM01",  9.7,  "N",    "LAST",
  "TEST01", "PAT01",  "PARAM01", 15.01, "N",    "LAST",
  "TEST01", "PAT01",  "PARAM02",  8.35, "Y",    "LAST",
  "TEST01", "PAT01",  "PARAM02", NA,    "N",    "LAST",
  "TEST01", "PAT01",  "PARAM02",  8.35, "N",    "LAST"
)
derive_var_base(bds_dataset)

}
\author{
Thomas Neitmann
}
