% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_baseline.R
\name{derive_var_base}
\alias{derive_var_base}
\title{Derive BASE Variable}
\usage{
derive_var_base(dataset, by_vars)
}
\arguments{
\item{dataset}{The input dataset}

\item{by_vars}{Grouping variables uniquely identifying a set
of records for which to calculate \code{BASE}}
}
\value{
A new \code{data.frame} containing all records and variables of the input
dataset plus the \code{BASE} variable
}
\description{
Derive the \code{BASE} variable in a BDS dataset
}
\examples{
dataset <- tibble::tribble(
  ~STUDYID, ~USUBJID, ~PARAMCD,  ~AVAL, ~AVISIT,    ~ABLFL,
  "TEST01", "PAT01",  "PARAM01", 10.12, "Baseline", "Y",
  "TEST01", "PAT01",  "PARAM01",  9.7,  "Day 7",    "N",
  "TEST01", "PAT01",  "PARAM01", 15.01, "Day 14",   "N",
  "TEST01", "PAT01",  "PARAM02",  8.35, "Baseline", "Y",
  "TEST01", "PAT01",  "PARAM02", NA,    "Day 7",    "N",
  "TEST01", "PAT01",  "PARAM02",  8.35, "Day 14",   "N"
)
derive_var_base(dataset, by_vars = vars(USUBJID, PARAMCD))
}
\author{
Thomas Neitmann
}
\keyword{bds}
\keyword{derivation}
