% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_baseline.R
\name{derive_var_base}
\alias{derive_var_base}
\title{Derive BASE Variable}
\usage{
derive_var_base(dataset, by_vars)
}
\arguments{
\item{dataset}{The input dataset}

\item{by_vars}{Grouping variables uniquely identifying a set
of records for which to calculate \code{BASE}}
}
\value{
A new \code{data.frame} containing all records and variables of the input
dataset plus the \code{BASE} variable
}
\description{
Derive the \code{BASE} variable in a BDS dataset
}
\details{
For each \code{by_vars} group the baseline record is identified by filtering using the
condition \code{ABLFL == "Y"}. Subsequently, every value of the \code{BASE} variable for the \code{by_vars}
group is set to the value of the \code{AVAL} variable of the baseline record. In case there
are multiple baseline records within \code{by_vars} an error is issued.
}
\examples{
dataset <- tibble::tribble(
  ~STUDYID, ~USUBJID, ~PARAMCD,  ~AVAL, ~AVISIT,    ~ABLFL,
  "TEST01", "PAT01",  "PARAM01", 10.12, "Baseline", "Y",
  "TEST01", "PAT01",  "PARAM01",  9.7,  "Day 7",    "N",
  "TEST01", "PAT01",  "PARAM01", 15.01, "Day 14",   "N",
  "TEST01", "PAT01",  "PARAM02",  8.35, "Baseline", "Y",
  "TEST01", "PAT01",  "PARAM02", NA,    "Day 7",    "N",
  "TEST01", "PAT01",  "PARAM02",  8.35, "Day 14",   "N"
)
derive_var_base(dataset, by_vars = vars(USUBJID, PARAMCD))
}
\seealso{
\code{\link[=derive_var_basec]{derive_var_basec()}}, \code{\link[=derive_baseline]{derive_baseline()}}
}
\author{
Thomas Neitmann
}
\keyword{bds}
\keyword{derivation}
