% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_baseline.R
\name{derive_var_basec}
\alias{derive_var_basec}
\title{Derive BASEC}
\usage{
derive_var_basec(bds_dataset, by = c("USUBJID", "PARAMCD", "BASETYPE"))
}
\arguments{
\item{bds_dataset}{\code{data.frame}}

\item{by}{\code{character} vector of columns uniqely identifying a set
of records for which to calculate \code{BASE}}
}
\value{
A new \code{data.frame} containing all records and variables of the input
dataset plus the \code{BASEC} variable
}
\description{
Derive the \code{BASEC} variable in a BDS dataset
}
\examples{
bds_dataset <- tibble::tribble(
  ~STUDYID, ~USUBJID, ~PARAMCD,  ~AVALC,   ~ABLFL, ~BASETYPE,
  "TEST01", "PAT01",  "PARAM01", "LOW",    "Y",    "LAST",
  "TEST01", "PAT01",  "PARAM01", "LOW",    "N",    "LAST",
  "TEST01", "PAT01",  "PARAM01", "MEDIUM", "N",    "LAST",
  "TEST01", "PAT01",  "PARAM02", "HIGH",   "Y",    "LAST",
  "TEST01", "PAT01",  "PARAM02", "HIGH",   "N",    "LAST",
  "TEST01", "PAT01",  "PARAM02", "MEDIUM", "N",    "LAST"
)
derive_var_basec(bds_dataset)

}
\author{
Thomas Neitmann
}
