% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_baseline.R
\name{derive_var_basec}
\alias{derive_var_basec}
\title{Derive BASEC Variable}
\usage{
derive_var_basec(dataset, by_vars)
}
\arguments{
\item{dataset}{The input dataset}

\item{by_vars}{Grouping variables uniquely identifying a set
of records for which to calculate \code{BASE}}
}
\value{
A new \code{data.frame} containing all records and variables of the input
dataset plus the \code{BASEC} variable
}
\description{
Derive the \code{BASEC} variable in a BDS dataset
}
\details{
For each \code{by_vars} group the baseline record is identified by filtering using the
condition \code{ABLFL == "Y"}. Subsequently, every value of the \code{BASEC} variable for the \code{by_vars}
group is set to the value of the \code{AVALC} variable of the baseline record. In case there
are multiple baseline records within \code{by_vars} an error is issued.
}
\examples{
dataset <- tibble::tribble(
  ~STUDYID, ~USUBJID, ~PARAMCD,  ~AVALC,   ~AVISIT,    ~ABLFL,
  "TEST01", "PAT01",  "PARAM01", "LOW",    "Baseline", "Y",
  "TEST01", "PAT01",  "PARAM01", "LOW",    "Day 7",    "N",
  "TEST01", "PAT01",  "PARAM01", "MEDIUM", "Day 14",   "N",
  "TEST01", "PAT01",  "PARAM02", "HIGH",   "Baseline", "Y",
  "TEST01", "PAT01",  "PARAM02", "HIGH",   "Day 7",    "N",
  "TEST01", "PAT01",  "PARAM02", "MEDIUM", "Day 14",   "N"
)
derive_var_basec(dataset, by_vars = vars(USUBJID, PARAMCD))
}
\seealso{
\code{\link[=derive_var_base]{derive_var_base()}}, \code{\link[=derive_baseline]{derive_baseline()}}
}
\author{
Thomas Neitmann
}
\keyword{bds}
\keyword{derivation}
