% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_base.R
\name{derive_var_basec}
\alias{derive_var_basec}
\title{Derive BASEC}
\usage{
derive_var_basec(bds_dataset)
}
\arguments{
\item{bds_dataset}{`data.frame`}
}
\value{
A new `data.frame` containing all records and variables of the input
dataset plus the `BASEC` variable
}
\description{
Derive the `BASEC` variable in a BDS dataset
}
\examples{
bds_dataset <- tibble::tribble(
  ~STUDYID, ~USUBJID, ~PARAMCD,  ~AVALC,   ~ABLFL,
  "TEST01", "PAT01",  "PARAM01", "LOW,     "Y",
  "TEST01", "PAT01",  "PARAM01", "LOW,     "N",
  "TEST01", "PAT01",  "PARAM01", "MEDIUM", "N",
  "TEST01", "PAT01",  "PARAM02", "HIGH",   "Y",
  "TEST01", "PAT01",  "PARAM02", "HIGH",   "N",
  "TEST01", "PAT01",  "PARAM02", "MEDIUM", "N",
)
derive_var_basec(bds_dataset)

}
\author{
Thomas Neitmann
}
