% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_base.R
\name{derive_var_basec}
\alias{derive_var_basec}
\title{Derive BASEC Variable}
\usage{
derive_var_basec(dataset, by_vars)
}
\arguments{
\item{dataset}{The input dataset}

\item{by_vars}{Grouping variables uniquely identifying a set
of records for which to calculate \code{new_var}}
}
\description{
\verb{r lifecycle::badge("deprecated")}

This function is \emph{deprecated}. Please use \code{\link[=derive_var_base]{derive_var_base()}} instead.
}
\seealso{
\code{\link[=derive_var_base]{derive_var_base()}}
}
\author{
Thomas Neitmann
}
\keyword{bds}
\keyword{derivation}
