% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_dthcaus.R
\name{derive_var_dthcaus}
\alias{derive_var_dthcaus}
\title{Derive Death Cause}
\usage{
derive_var_dthcaus(
  dataset,
  ...,
  source_datasets,
  subject_keys = get_admiral_option("subject_keys")
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{subject_keys} argument are expected to be in the dataset.

\describe{
\item{Default value}{none}
}}

\item{...}{Objects of class "dthcaus_source" created by \code{\link[=dthcaus_source]{dthcaus_source()}}.

\describe{
\item{Default value}{none}
}}

\item{source_datasets}{A named \code{list} containing datasets in which to search for the
death cause

\describe{
\item{Default value}{none}
}}

\item{subject_keys}{Variables to uniquely identify a subject

A list of expressions where the expressions are symbols as returned by
\code{exprs()} is expected.

\describe{
\item{Default value}{\code{get_admiral_option("subject_keys")}}
}}
}
\value{
The input dataset with \code{DTHCAUS} variable added.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} The \code{derive_var_dthcaus()}
function has been deprecated in favor of \code{derive_vars_extreme_event()}.

Derive death cause (\code{DTHCAUS}) and add traceability variables if required.
}
\details{
This function derives \code{DTHCAUS} along with the user-defined traceability
variables, if required. If a subject has death info from multiple sources,
the one from the source with the earliest death date will be used. If dates are
equivalent, the first source will be kept, so the user should provide the inputs in
the preferred order.
}
\seealso{
\code{\link[=dthcaus_source]{dthcaus_source()}}

Other deprecated:
\code{\link[=date_source]{date_source()}},
\code{\link[=derive_param_extreme_record]{derive_param_extreme_record()}},
\code{\link[=derive_var_extreme_dt]{derive_var_extreme_dt()}},
\code{\link[=derive_var_extreme_dtm]{derive_var_extreme_dtm()}},
\code{\link[=derive_var_merged_summary]{derive_var_merged_summary()}},
\code{\link[=dthcaus_source]{dthcaus_source()}},
\code{\link[=get_summary_records]{get_summary_records()}}
}
\concept{deprecated}
\keyword{deprecated}
