% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_dthcaus.R
\name{derive_var_dthcaus}
\alias{derive_var_dthcaus}
\title{Derive Death Cause}
\usage{
derive_var_dthcaus(dataset, ...)
}
\arguments{
\item{dataset}{Input dataset. \code{USUBJID} is an expected column.}

\item{...}{Objects of class "dthcaus_source" created by \code{\link[=dthcaus_source]{dthcaus_source()}}.}
}
\value{
The input dataset with \code{DTHCAUS} variable added.
}
\description{
Derive death cause (\code{DTHCAUS}) and add traceability variables if required.
}
\details{
This function derives \code{DTHCAUS} along with the user-defined traceability
variables, if required. If a subject has death info from multiple sources,
the one from the source with the earliest death date will be used. If dates are
equivalent, the first source will be kept, so the user should provide the inputs in
the preferred order.
}
\examples{
adsl <- tibble::tribble(
  ~STUDYID, ~USUBJID,
  "STUDY01", "PAT01",
  "STUDY01", "PAT02"
)
ae <- tibble::tribble(
  ~STUDYID, ~USUBJID, ~AESEQ, ~AEDECOD, ~AEOUT, ~AEDTHDTC,
  "STUDY01", "PAT01", 12, "SUDDEN DEATH", "FATAL", "2021-04-04"
)
ds <- tibble::tribble(
  ~STUDYID, ~USUBJID, ~DSSEQ, ~DSDECOD, ~DSTERM, ~DSSTDTC,
  "STUDY01", "PAT02", 1, "INFORMED CONSENT OBTAINED", "INFORMED CONSENT OBTAINED", "2021-04-03",
  "STUDY01", "PAT02", 2, "RANDOMIZATION", "RANDOMIZATION", "2021-04-11",
  "STUDY01", "PAT02", 3, "DEATH", "DEATH DUE TO PROGRESSION OF DISEASE", "2022-02-01"
)

# Derive `DTHCAUS` only
src_ae <- dthcaus_source(
  dataset = ae,
  filter = AEOUT == "FATAL",
  date_var = AEDTHDTC,
  mode = "first",
  dthcaus = AEDECOD
)

src_ds <- dthcaus_source(
  dataset = ds,
  filter = DSDECOD == "DEATH" & grepl("DEATH DUE TO", DSTERM),
  date = DSSTDTC,
  mode = "first",
  dthcaus = DSTERM
)

derive_var_dthcaus(adsl, src_ae, src_ds)

# Derive `DTHCAUS` and add traceability variables
src_ae <- dthcaus_source(
  dataset = ae,
  filter = AEOUT == "FATAL",
  date_var = AEDTHDTC,
  mode = "first",
  dthcaus = AEDECOD,
  traceabilty_vars = vars(DTHDOM = "AE", DTHSEQ = AESEQ)
)

src_ds <- dthcaus_source(
  dataset = ds,
  filter = DSDECOD == "DEATH" & grepl("DEATH DUE TO", DSTERM),
  date_var = DSSTDTC,
  mode = "first",
  dthcaus = DSTERM,
  traceabilty_vars = vars(DTHDOM = "DS", DTHSEQ = DSSEQ)
)

derive_var_dthcaus(adsl, src_ae, src_ds)
}
\seealso{
\code{\link[=dthcaus_source]{dthcaus_source()}}
}
\author{
Shimeng Huang
Samia Kabi
}
\keyword{adsl}
