% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_extreme_flag.R
\name{derive_var_extreme_flag}
\alias{derive_var_extreme_flag}
\title{Add a Variable Flagging the First or Last Observation Within Each By Group}
\usage{
derive_var_extreme_flag(
  dataset,
  by_vars,
  order,
  new_var,
  mode,
  true_value = "Y",
  false_value = NA_character_,
  flag_all = FALSE,
  check_type = "warning"
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{by_vars} argument are expected to be in the dataset.}

\item{by_vars}{Grouping variables}

\item{order}{Sort order

The first or last observation is determined with respect to the specified
order.

For handling of \code{NA}s in sorting variables see the "Sort Order" section in \code{vignette("generic")}.}

\item{new_var}{Variable to add

The specified variable is added to the output dataset. It is set to the value
set in \code{true_value} for the first or last observation (depending on the mode) of each by group.}

\item{mode}{Flag mode

Determines of the first or last observation is flagged.}

\item{true_value}{True value

The value for the specified variable \code{new_var}, applicable to
the first or last observation (depending on the mode) of each by group.}

\item{false_value}{False value

The value for the specified variable \code{new_var}, NOT applicable to
the first or last observation (depending on the mode) of each by group.}

\item{flag_all}{Flag setting

A logical value where if set to \code{TRUE}, all records are flagged
and no error or warning is issued if the first or last record is not unique.}

\item{check_type}{Check uniqueness?

If \code{"warning"} or \code{"error"} is specified, the specified message is issued
if the observations of the input dataset are not unique with respect to the
by variables and the order.}
}
\value{
The input dataset with the new flag variable added
}
\description{
Add a variable flagging the first or last observation within each by group
}
\details{
For each group (with respect to the variables specified for the
\code{by_vars} parameter), \code{new_var} is set to \code{"Y"} for the first or last observation
(with respect to the order specified for the \code{order} parameter and the flag mode
specified for the \code{mode} parameter). In the case where the user wants to flag multiple records
of a grouping, for example records that all happen on the same visit and time, the argument
\code{flag_all} can be set to \code{TRUE}.
Otherwise, \code{new_var} is set to \code{NA}. Thus, the direction of "worst" is considered fixed for
all parameters in the dataset depending on the \code{order} and the \code{mode}, i.e. for every
parameter the first or last record will be flagged across the whole dataset.
}
\seealso{
General Derivation Functions for all ADaMs that returns variable appended to dataset:
\code{\link{derive_var_joined_exist_flag}()},
\code{\link{derive_var_merged_ef_msrc}()},
\code{\link{derive_var_merged_exist_flag}()},
\code{\link{derive_var_obs_number}()},
\code{\link{derive_var_relative_flag}()},
\code{\link{derive_vars_cat}()},
\code{\link{derive_vars_computed}()},
\code{\link{derive_vars_joined}()},
\code{\link{derive_vars_joined_summary}()},
\code{\link{derive_vars_merged}()},
\code{\link{derive_vars_merged_lookup}()},
\code{\link{derive_vars_merged_summary}()},
\code{\link{derive_vars_transposed}()}
}
\concept{der_gen}
\keyword{der_gen}
