% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_extreme_date.R
\name{derive_var_lstalvdt}
\alias{derive_var_lstalvdt}
\title{Derive Last Known Alive Date}
\usage{
derive_var_lstalvdt(
  dataset,
  ...,
  source_datasets,
  subject_keys = vars(STUDYID, USUBJID)
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by \code{subject_keys} are required.}

\item{...}{Source(s) of known alive dates. One or more \code{lstalvdt_source()} objects are
expected.}

\item{source_datasets}{A named \code{list} containing datasets in which to search for the
last known alive date}

\item{subject_keys}{Variables to uniquely identify a subject

A list of quosures where the expressions are symbols as returned by
\code{vars()} is expected.}
}
\value{
The input dataset with the \code{LSTALVDT} variable added.
}
\description{
\ifelse{html}{\figure{lifecycle-deprecated.svg}{options: alt='Deprecated lifecycle'}}{\strong{Deprecated}}

\emph{Deprecated}, please use \code{derive_var_extreme_dt()} instead. Add the last
known alive date (\code{LSTALVDT}) to the dataset.
}
\author{
Stefan Bundfuss, Thomas Neitmann
}
\keyword{adsl}
\keyword{derivation}
