% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_merged.R
\name{derive_var_merged_cat}
\alias{derive_var_merged_cat}
\title{Merge a Categorized Variable}
\usage{
derive_var_merged_cat(
  dataset,
  dataset_add,
  by_vars,
  order = NULL,
  new_var,
  source_var,
  cat_fun,
  filter_add = NULL,
  mode = NULL
)
}
\description{
Merge a Categorized Variable
}
\examples{
library(admiral.test)
data("dm")
data("vs")
wgt_cat <- function(wgt) {
  case_when(wgt < 50 ~ "low",
            wgt > 90 ~ "high",
            TRUE ~ "normal")
}

derive_var_merged_cat(
  dm,
  dataset_add = vs,
  by_vars = vars(STUDYID, USUBJID),
  order = vars(VSDTC, VSSEQ),
  filter_add = VSTESTCD == "WEIGHT" & substr(VISIT, 1, 9) == "SCREENING" ,
  new_var = WGTBLCAT,
  source_var = VSSTRESN,
  cat_fun = wgt_cat,
  mode = "last"
)
}
