% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_merged.R
\name{derive_var_merged_character}
\alias{derive_var_merged_character}
\title{Merge a Character Variable}
\usage{
derive_var_merged_character(
  dataset,
  dataset_add,
  by_vars,
  order = NULL,
  new_var,
  source_var,
  case = NULL,
  filter_add = NULL,
  mode = NULL
)
}
\description{
Merge a Character Variable
}
\examples{
library(admiral.test)
data("dm")
data("ds")
derive_var_merged_character(
  dm,
  dataset_add = ds,
  by_vars = vars(STUDYID, USUBJID),
  new_var = DISPSTAT,
  filter_add = DSCAT == "DISPOSITION EVENT",
  source_var = DSDECOD,
  case = "title"
)
}
