% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_merged.R
\name{derive_var_merged_exist_flag}
\alias{derive_var_merged_exist_flag}
\title{Merge a Existence Flag}
\usage{
derive_var_merged_exist_flag(
  dataset,
  dataset_add,
  by_vars,
  new_var,
  condition,
  true_value = "Y",
  false_value = NA_character_,
  filter_add
)
}
\description{
Merge a Existence Flag
}
\examples{

library(admiral.test)
data("dm")
data("ae")
derive_var_merged_exist_flag(
  dm,
  dataset_add = ae,
  by_vars = vars(STUDYID, USUBJID),
  new_var = AESERFL,
  condition = AESER == "Y"
)
}
