% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_chg.R
\name{derive_var_pchg}
\alias{derive_var_pchg}
\title{Derive Percent Change from Baseline}
\usage{
derive_var_pchg(bds_dataset)
}
\arguments{
\item{bds_dataset}{`data.frame`. Required variables are `BASE` and
`CHG`.}
}
\value{
The input dataset with an additional column named `PCHG`
}
\description{
Derive percent change from baseline (`PCHG`) in a BDS dataset
}
\details{
Percent change from baseline is calculated by dividing change from
baseline by the baseline value and multiplying the result by `100`.
}
\examples{
advs <- tibble::tribble(
  ~USUBJID, ~PARAMCD, ~AVAL, ~ABLFL, ~BASE, ~CHG,
  "P01",    "WEIGHT", 80,    "Y",    80,    0,
  "P01",    "WEIGHT", 80.8,  "",     80,    0.8,
  "P01",    "WEIGHT", 81.4,  "",     80,    1.4,
  "P02",    "WEIGHT", 75.3,  "Y",    75.3,  0,
  "P02",    "WEIGHT", 76,    "",     75.3,  0.7
)
derive_var_pchg(advs)

}
\author{
Thomas Neitmann
}
