% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_aage.R
\name{derive_vars_aage}
\alias{derive_vars_aage}
\title{Derive Analysis Age}
\usage{
derive_vars_aage(
  dataset,
  start_date = BRTHDT,
  end_date = RANDDT,
  unit = "years"
)
}
\arguments{
\item{dataset}{Input dataset

The columns specified by the \code{start_date} and the \code{end_date} parameter are
expected.}

\item{start_date}{The start date

A date or date-time object is expected.

Refer to \code{derive_vars_dt()} to impute and derive a date from a date character
vector to a date object.

Default: \code{BRTHDT}}

\item{end_date}{The end date

A date or date-time object is expected.

Refer to \code{derive_vars_dt()} to impute and derive a date from a date character
vector to a date object.

Default: \code{RANDDT}}

\item{unit}{Unit

The age is derived in the specified unit

Default: 'years'

Permitted Values: 'years', 'months', 'weeks', 'days', 'hours', 'minutes', 'seconds'}
}
\value{
The input dataset with \code{AAGE} and \code{AAGEU} added
}
\description{
Derives analysis age (\code{AAGE}) and analysis age unit (\code{AAGEU})
}
\details{
The age is derived as the integer part of the duration from start to
end date in the specified unit. When 'years' or 'months' are specified in the \code{out_unit}
parameter, because of the underlying \code{lubridate::time_length()} function that is used
here, results are calculated based on the actual calendar length of months or years
rather than assuming equal days every month (30.4375 days) or every year (365.25 days).
}
\examples{
library(tibble)
library(lubridate)

data <- tribble(
  ~BRTHDT, ~RANDDT,
  ymd("1984-09-06"), ymd("2020-02-24")
)

derive_vars_aage(data)
}
\seealso{
\code{\link[=derive_vars_duration]{derive_vars_duration()}}

ADSL Functions that returns variable appended to dataset: 
\code{\link{derive_var_age_years}()},
\code{\link{derive_var_disposition_status}()},
\code{\link{derive_var_dthcaus}()},
\code{\link{derive_var_extreme_dtm}()},
\code{\link{derive_var_extreme_dt}()},
\code{\link{derive_vars_disposition_reason}()},
\code{\link{derive_vars_period}()}
}
\author{
Stefan Bundfuss
}
\concept{der_adsl}
\keyword{der_adsl}
