% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_cat.R
\name{derive_vars_cat}
\alias{derive_vars_cat}
\title{Derive Categorization Variables Like \code{AVALCATy} and \code{AVALCAyN}}
\usage{
derive_vars_cat(dataset, definition, by_vars = NULL)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{by_vars} and \code{definition} arguments are expected to be in the dataset.}

\item{definition}{List of expressions created by \code{exprs()}.
Must be in rectangular format and specified using the same syntax as when creating
a \code{tibble} using the \code{tribble()} function.
The \code{definition} object will be converted to a \code{tibble} using \code{tribble()} inside this function.

Must contain:
\itemize{
\item the column \code{condition} which will be converted to a logical expression and
will be used on the \code{dataset} input.
\item at least one additional column with the new column name and
the category value(s) used by the logical expression.
\item the column specified in \code{by_vars} (if \code{by_vars} is specified)
}

e.g. if \code{by_vars} is not specified:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{exprs(~condition,   ~AVALCAT1, ~AVALCA1N,
      AVAL >= 140, ">=140 cm",         1,
      AVAL < 140,   "<140 cm",         2)
}\if{html}{\out{</div>}}

e.g. if \code{by_vars} is specified as \code{exprs(VSTEST)}:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{exprs(~VSTEST,   ~condition,  ~AVALCAT1, ~AVALCA1N,
      "Height", AVAL >= 140, ">=140 cm",         1,
      "Height",  AVAL < 140,  "<140 cm",         2)
}\if{html}{\out{</div>}}}

\item{by_vars}{list of expressions with one element. \code{NULL} by default.
Allows for specifying by groups, e.g. \code{exprs(PARAMCD)}.
Variable must be present in both \code{dataset} and \code{definition}.
The conditions in \code{definition} are applied only to those records that match \code{by_vars}.
The categorization variables are set to \code{NA} for records
not matching any of the by groups in \code{definition}.}
}
\value{
The input dataset with the new variables defined in \code{definition} added
}
\description{
Derive Categorization Variables Like \code{AVALCATy} and \code{AVALCAyN}
}
\details{
If conditions are overlapping, the row order of \code{definitions} must be carefully considered.
The \strong{first} match will determine the category.
i.e. if

\code{AVAL = 155}

and the \code{definition} is:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{definition <- exprs(
  ~VSTEST,   ~condition,  ~AVALCAT1, ~AVALCA1N,
  "Height",  AVAL > 170,  ">170 cm",         1,
  "Height", AVAL <= 170, "<=170 cm",         2,
  "Height", AVAL <= 160, "<=160 cm",         3
)
}\if{html}{\out{</div>}}

then \code{AVALCAT1} will be \code{"<=170 cm"}, as this is the first match for \code{AVAL}.
If you specify:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{definition <- exprs(
  ~VSTEST,   ~condition,  ~AVALCAT1, ~AVALCA1N,
  "Height", AVAL <= 160, "<=160 cm",         3,
  "Height", AVAL <= 170, "<=170 cm",         2,
  "Height",  AVAL > 170,  ">170 cm",         1
)
}\if{html}{\out{</div>}}

Then \code{AVAL <= 160} will lead to \code{AVALCAT1 == "<=160 cm"},
\code{AVAL} in-between \code{160} and \code{170} will lead to \code{AVALCAT1 == "<=170 cm"},
and \code{AVAL <= 170} will lead to \code{AVALCAT1 == ">170 cm"}.

However, we suggest to be more explicit when defining the \code{condition}, to avoid overlap.
In this case, the middle condition should be:
\code{AVAL <= 170 & AVAL > 160}
}
\seealso{
General Derivation Functions for all ADaMs that returns variable appended to dataset:
\code{\link{derive_var_extreme_flag}()},
\code{\link{derive_var_joined_exist_flag}()},
\code{\link{derive_var_merged_ef_msrc}()},
\code{\link{derive_var_merged_exist_flag}()},
\code{\link{derive_var_obs_number}()},
\code{\link{derive_var_relative_flag}()},
\code{\link{derive_vars_computed}()},
\code{\link{derive_vars_joined}()},
\code{\link{derive_vars_joined_summary}()},
\code{\link{derive_vars_merged}()},
\code{\link{derive_vars_merged_lookup}()},
\code{\link{derive_vars_merged_summary}()},
\code{\link{derive_vars_transposed}()}
}
\concept{der_gen}
\keyword{der_gen}
