% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_cat.R
\name{derive_vars_cat}
\alias{derive_vars_cat}
\title{Derive pair of variables}
\usage{
derive_vars_cat(dataset, definition)
}
\arguments{
\item{dataset}{data frame containing the variables specified in the conditions}

\item{definition}{of condition and values defined as \code{exprs()} object.
e.g. exprs(~condition, ~AVALCAT1, ~AVALCA1N
AVAL >= 140, ">=140 cm", 1,
AVAL < 140, "<140 cm", 2)}
}
\value{
data frame
}
\description{
Derive pair of variables
}
\examples{

advs <- tibble::tribble(
  ~USUBJID, ~PARAMCD, ~AVAL,
  "01",     "HEIGHT", 150,
  "02",     "HEIGHT", 135,
  "03",     "HEIGHT", 145
)
advs <- pharmaverseadam::advs

definition <- exprs(
  ~condition,  ~AVALCAT1,  ~AVALCA1N, ~NEWCOL,
  VSTESTCD == "HEIGHT" & AVAL > 140,  ">140 cm",          1,  "extra1",
  VSTESTCD == "HEIGHT" & AVAL <= 140, "<=140 cm",         2,  "extra2"
)

derive_vars_cat(dataset = advs \%>\% filter(VSTESTCD == "HEIGHT"),
                definition = definition) \%>\%
  select(USUBJID, VSTESTCD, AVAL, AVALCA1N, AVALCAT1)
}
