% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dt.R
\name{derive_vars_dt}
\alias{derive_vars_dt}
\title{Derive/Impute a Date from a Character Date}
\usage{
derive_vars_dt(
  dataset,
  new_vars_prefix,
  dtc,
  highest_imputation = "n",
  date_imputation = "first",
  flag_imputation = "auto",
  min_dates = NULL,
  min_dates_strict = NULL,
  max_dates = NULL,
  max_dates_strict = NULL,
  preserve = FALSE
)
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{dtc} argument are expected to be in the dataset.

\describe{
\item{Permitted values}{a dataset, i.e., a \code{data.frame} or tibble}
\item{Default value}{none}
}}

\item{new_vars_prefix}{Prefix used for the output variable(s).

A character scalar is expected. For the date variable (\verb{*DT}) is appended to
the specified prefix and for the date imputation flag (\verb{*DTF}), i.e., for
\code{new_vars_prefix = "AST"} the variables \code{ASTDT} and \code{ASTDTF} are created.

\describe{
\item{Permitted values}{a character scalar, i.e., a character vector of length one}
\item{Default value}{none}
}}

\item{dtc}{The \code{--DTC} date to impute

A character date is expected in a format like \code{yyyy-mm-dd} or
\code{yyyy-mm-ddThh:mm:ss}. Trailing components can be omitted and \code{-} is a
valid "missing" value for any component.

\describe{
\item{Permitted values}{a character date variable}
\item{Default value}{none}
}}

\item{highest_imputation}{Highest imputation level

The \code{highest_imputation} argument controls which components of the \code{--DTC}
value are imputed if they are missing. All components up to the specified
level are imputed.

If a component at a higher level than the highest imputation level is
missing, \code{NA} is returned. For example, for \code{highest_imputation = "D"}
\code{"2020"} results in \code{NA} because the month is missing.

If \code{"n"} (none, lowest level) is specified no imputation is performed, i.e.,
if any component is missing, \code{NA} is returned.

If \code{"Y"} (year, highest level) is specified, \code{date_imputation} must be \code{"first"} or \code{"last"}
and \code{min_dates} or \code{max_dates} must be specified respectively. Otherwise,
an error is thrown.

\describe{
\item{Permitted values}{\code{"Y"} (year, highest level), \code{"M"} (month), \code{"D"}
(day), \code{"n"} (none, lowest level)}
\item{Default value}{\code{"n"}}
}}

\item{date_imputation}{The value to impute the day/month when a datepart is
missing.

A character value is expected.
\itemize{
\item The\code{"first"} and \code{"last"} keywords allow imputation to the first/last
day/month. They can also be used to impute the year if used in conjunction
with the \code{min_dates} or \code{max_dates} arguments. Some examples of this are available
\href{https://pharmaverse.github.io/admiral/cran-release/articles/imputation.html#minimummaximum-dates}{here}.
\item When \code{highest_imputation} is \code{"M"} or \code{"D"}, the \code{"mid"} keyword can also be
specified to impute missing components to the middle of the possible range:
\itemize{
\item If both month and day are missing, they are imputed as \code{"06-30"}
(middle of the year).
\item If only day is missing, it is imputed as \code{"15"} (middle of the month).
}
\item \code{"<dd>"} can be specified only if \code{highest_imputation = "D"}. Missing days are
imputed by the specified day, e.g. \code{"10"} for the 10th day of the month.
The specified day should be valid for all months as otherwise an error might be
issued. For example, \code{date_imputation = "30"} results in an invalid date of
"2024-02-30" for the partial date "2024-02".
\item \code{"<mm>-<dd>"} can be specified only if \code{highest_imputation} is \code{"M"},  e.g. \code{"06-15"}
for the 15th of June.
}

\describe{
\item{Permitted values}{a key-word, i.e. \code{"first"}, \code{"mid"}, \code{"last"}, or \code{"<mm>-<dd>"} or \code{"<dd>"}}
\item{Default value}{\code{"first"}}
}}

\item{flag_imputation}{Whether the date imputation flag must also be derived.

If \code{"auto"} is specified and \code{highest_imputation} argument is not \code{"n"},
then date imputation flag is derived.

If \code{"date"} is specified, then date imputation flag is derived.

If \code{"none"} is specified, then no date imputation flag is derived.

Please note that CDISC requirements dictate the need for a date imputation
flag if any imputation is performed, so \code{flag_imputation = "none"} should
only be used if the imputed variable is not part of the final ADaM dataset.

\describe{
\item{Permitted values}{\code{"auto"}, \code{"date"} or \code{"none"}}
\item{Default value}{\code{"auto"}}
}}

\item{min_dates}{Minimum dates

A list of dates is expected. It is ensured that the imputed date is not
before any of the specified dates, e.g., that the imputed adverse event start
date is not before the first treatment date. Only dates which are in the
range of possible dates of the \code{dtc} value are considered. The possible dates
are defined by the missing parts of the \code{dtc} date (see example below). This
ensures that the non-missing parts of the \code{dtc} date are not changed.
A date or date-time object is expected.
For example

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(lubridate, warn.conflicts = FALSE)
impute_dtc_dt(
  "2020-11",
  min_dates = list(
   ymd("2020-12-06"),
   ymd("2020-11-11")
  ),
  highest_imputation = "M"
)
#> [1] "2020-11-11"
}\if{html}{\out{</div>}}

returns \code{"2020-11-11"} because the possible dates for \code{"2020-11"} range from
\code{"2020-11-01"} to \code{"2020-11-30"}. Therefore \code{"2020-12-06"} is ignored.
Returning \code{"2020-12-06"} would have changed the month although it is not
missing (in the \code{dtc} date).

\describe{
\item{Permitted values}{a list of dates, e.g. \code{list(ymd_hms("2021-07-01T04:03:01"), ymd_hms("2022-05-12T13:57:23"))}}
\item{Default value}{\code{NULL}}
}}

\item{min_dates_strict}{Minimum dates (strict)

The argument works like the \code{min_dates} argument but it affects the behavior
of the \code{min_dates} and \code{max_dates} arguments. The range that is used to
determine which of the \code{min_dates} and \code{max_dates} are considered is
restricted by the dates specified for \code{min_dates_strict} (see example below
and the example in the "Avoid Imputed Dates Before a Particular Date" section
in \code{vignette("imputation")}).

This argument is useful if the \code{max_dates} argument is used and there are
strict restrictions on the imputed date, like the event start date when the
event end date is imputed.

For example

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(lubridate)
impute_dtc_dt(
  c("2020-11", "2020-11"),
  min_dates_strict = list(
   c(ymd("2020-11-24"), ymd("2020-11-11"))
  ),
  max_dates = list(
    c(ymd("2020-11-22"), ymd("2020-11-22"))
  ),
  highest_imputation = "M",
  date_imputation = "last"
)
#> [1] "2020-11-30" "2020-11-22"
}\if{html}{\out{</div>}}

returns \code{"2020-11-30"}. The possible dates for \code{"2020-11"} range from
\code{"2020-11-01"} to \code{"2020-11-30"}.

For the first element, the \code{min_dates_strict} argument restricts this range
to \code{"2020-11-24"} to \code{"2020-11-30"}. Therefore \code{"2020-11-22"} (from the
\code{max_dates} argument) is ignored as it is not within the restricted range.

For the second element, the \code{min_dates_strict} argument restricts this range
to \code{"2020-11-11"} to \code{"2020-11-30"}. In this case \code{"2020-11-22"} (from the
\code{max_dates} argument) is within the restricted range and is considered.

\describe{
\item{Permitted values}{a list of dates, e.g. \code{list(ymd_hms("2021-07-01T04:03:01"), ymd_hms("2022-05-12T13:57:23"))}}
\item{Default value}{\code{NULL}}
}}

\item{max_dates}{Maximum dates

A list of dates is expected. It is ensured that the imputed date is not after
any of the specified dates, e.g., that the imputed date is not after the data
cut off date. Only dates which are in the range of possible dates are
considered. A date or date-time object is expected.

\describe{
\item{Permitted values}{a list of dates, e.g. \code{list(ymd_hms("2021-07-01T04:03:01"), ymd_hms("2022-05-12T13:57:23"))}}
\item{Default value}{\code{NULL}}
}}

\item{max_dates_strict}{Maximum dates (strict)

The argument works like the \code{max_dates} argument but it affects the behavior
of the \code{min_dates} and \code{max_dates} arguments. The range that is used to
determine which of the \code{min_dates} and \code{max_dates} are considered is
restricted by the dates specified for \code{max_dates_strict} (see example below).

This argument is useful if the \code{min_dates} argument is used and there are
strict restrictions on the imputed date, like date of death or the event end
date when the event start date is imputed.

For example

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(lubridate)
impute_dtc_dt(
  c("2020-11", "2020-11"),
  max_dates_strict = list(
   c(ymd("2020-11-24"), ymd("2020-11-11"))
  ),
  min_dates = list(
    c(ymd("2020-11-22"), ymd("2020-11-22"))
  ),
  highest_imputation = "M",
  date_imputation = "first"
)
#> [1] "2020-11-22" "2020-11-01"
}\if{html}{\out{</div>}}

returns (\code{"2020-11-22"}, \code{"2020-11-01"}). The possible dates for \code{"2020-11"}
range from \code{"2020-11-01"} to \code{"2020-11-30"}.

For the first element, the \code{max_dates_strict} argument restricts this range
to \code{"2020-11-01"} to \code{"2020-11-24"}. The date \code{"2020-11-22"} (from the
\code{min_dates} argument) is within the restricted range and is therefore
considered.

For the second element, the \code{max_dates_strict} argument restricts this range
to \code{"2020-11-01"} to \code{"2020-11-11"}. In this case \code{"2020-11-22"} (from the
\code{max_dates} argument) is outside the restricted range and thus it is ignored.

\describe{
\item{Permitted values}{a list of dates, e.g. \code{list(ymd_hms("2021-07-01T04:03:01"), ymd_hms("2022-05-12T13:57:23"))}}
\item{Default value}{\code{NULL}}
}}

\item{preserve}{Preserve day if month is missing and day is present

For example \code{"2019---07"} would return \verb{"2019-06-07} if \code{preserve = TRUE}
(and \code{date_imputation = "MID"}).

\describe{
\item{Permitted values}{\code{TRUE}, \code{FALSE}}
\item{Default value}{\code{FALSE}}
}}
}
\value{
The input dataset with the date \verb{*DT} (and the date imputation flag \verb{*DTF}
if requested) added.
}
\description{
Derive a date (\verb{*DT}) from a character date (\code{--DTC}).
The date can be imputed (see \code{date_imputation} argument)
and the date imputation flag (\verb{*DTF}) can be added.
}
\details{
In \code{{admiral}} we don't allow users to pick any single part of the date/time to
impute, we only enable to impute up to a highest level, i.e. you couldn't
choose to say impute months, but not days.

The presence of a \verb{*DTF} variable is checked and if it already exists in the input dataset,
a warning is issued and \verb{*DTF} will be overwritten.

Additionally, the function will throw an error if imputation rules cause an
invalid date (e.g. "2020-02-31") to be generated. In this case,  the user should adjust
the imputation rules.
}
\seealso{
\code{vignette("imputation")}

Date/Time Derivation Functions that returns variable appended to dataset:
\code{\link[=derive_var_trtdurd]{derive_var_trtdurd()}},
\code{\link[=derive_vars_dtm]{derive_vars_dtm()}},
\code{\link[=derive_vars_dtm_to_dt]{derive_vars_dtm_to_dt()}},
\code{\link[=derive_vars_dtm_to_tm]{derive_vars_dtm_to_tm()}},
\code{\link[=derive_vars_duration]{derive_vars_duration()}},
\code{\link[=derive_vars_dy]{derive_vars_dy()}}
}
\concept{der_date_time}
\keyword{der_date_time}
\keyword{der_gen}
\section{Examples}{
\subsection{Derive a date variable without imputation}{

In this example, we derive \code{ASTDT} from \code{MHSTDTC} with no imputation
done for partial dates.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(tibble)
library(lubridate)

mhdt <- tribble(
  ~MHSTDTC,
  "2019-07-18T15:25:40",
  "2019-07-18T15:25",
  "2019-07-18",
  "2019-02",
  "2019",
  "2019---07",
  ""
)

derive_vars_dt(
  mhdt,
  new_vars_prefix = "AST",
  dtc = MHSTDTC
)
#> # A tibble: 7 × 2
#>   MHSTDTC               ASTDT     
#>   <chr>                 <date>    
#> 1 "2019-07-18T15:25:40" 2019-07-18
#> 2 "2019-07-18T15:25"    2019-07-18
#> 3 "2019-07-18"          2019-07-18
#> 4 "2019-02"             NA        
#> 5 "2019"                NA        
#> 6 "2019---07"           NA        
#> 7 ""                    NA        }\if{html}{\out{</div>}}}
\subsection{Impute partial dates (\code{highest_imputation})}{

Imputation is requested by the \code{highest_imputation} argument. Here
\code{highest_imputation = "M"} for month imputation is used, i.e. the highest
imputation done on a partial date is up to the month. By default, missing date
components are imputed to the first day/month/year. A date imputation flag variable, \code{ASTDTF},
is automatically created. The flag variable indicates if imputation was done
on the date.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{derive_vars_dt(
  mhdt,
  new_vars_prefix = "AST",
  dtc = MHSTDTC,
  highest_imputation = "M",
  date_imputation = "first"
)
#> # A tibble: 7 × 3
#>   MHSTDTC               ASTDT      ASTDTF
#>   <chr>                 <date>     <chr> 
#> 1 "2019-07-18T15:25:40" 2019-07-18 <NA>  
#> 2 "2019-07-18T15:25"    2019-07-18 <NA>  
#> 3 "2019-07-18"          2019-07-18 <NA>  
#> 4 "2019-02"             2019-02-01 D     
#> 5 "2019"                2019-01-01 M     
#> 6 "2019---07"           2019-01-01 M     
#> 7 ""                    NA         <NA>  }\if{html}{\out{</div>}}

It is also possible to just impute the day, with \code{highest_imputation = "D"}. Here
dates with just a missing day have it imputed to the 10th of the month. Note that in this
case care needs to be taken to ensure invalid dates are not created, e.g.
\code{date_imputation = "30"} would create an invalid date of \code{"2020-02-30"} when trying to impute
the day for \code{"2020-02"}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{derive_vars_dt(
  mhdt,
  new_vars_prefix = "AST",
  dtc = MHSTDTC,
  highest_imputation = "D",
  date_imputation = "10"
)
#> # A tibble: 7 × 3
#>   MHSTDTC               ASTDT      ASTDTF
#>   <chr>                 <date>     <chr> 
#> 1 "2019-07-18T15:25:40" 2019-07-18 <NA>  
#> 2 "2019-07-18T15:25"    2019-07-18 <NA>  
#> 3 "2019-07-18"          2019-07-18 <NA>  
#> 4 "2019-02"             2019-02-10 D     
#> 5 "2019"                NA         <NA>  
#> 6 "2019---07"           NA         <NA>  
#> 7 ""                    NA         <NA>  }\if{html}{\out{</div>}}}
\subsection{Impute to the last day/month (\code{date_imputation = "last"})}{

In this example, we derive \code{ADT} impute partial dates to last day/month, i.e.
\code{date_imputation = "last"}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{qsdt <- tribble(
  ~QSDTC,
  "2019-07-18T15:25:40",
  "2019-07-18T15:25",
  "2019-07-18",
  "2019-02",
  "2019",
  "2019---07",
  ""
)

derive_vars_dt(
  qsdt,
  new_vars_prefix = "A",
  dtc = QSDTC,
  highest_imputation = "M",
  date_imputation = "last"
)
#> # A tibble: 7 × 3
#>   QSDTC                 ADT        ADTF 
#>   <chr>                 <date>     <chr>
#> 1 "2019-07-18T15:25:40" 2019-07-18 <NA> 
#> 2 "2019-07-18T15:25"    2019-07-18 <NA> 
#> 3 "2019-07-18"          2019-07-18 <NA> 
#> 4 "2019-02"             2019-02-28 D    
#> 5 "2019"                2019-12-31 M    
#> 6 "2019---07"           2019-12-31 M    
#> 7 ""                    NA         <NA> }\if{html}{\out{</div>}}}
\subsection{Impute to the middle (\code{date_imputation = "mid"}) and suppress
imputation flag (\code{flag_imputation = "none"})}{

In this example, we will derive \code{TRTSDT} with date imputation flag
(\verb{*DTF}) suppressed. Since \code{date_imputation = "mid"}, partial date imputation
will be set to June 30th for missing month and 15th for missing day only.
The \code{flag_imputation = "none"} call ensures no date imputation flag is
created. In practice, as per CDISC requirements this option can only be
selected if the imputed variable is not part of the final ADaM dataset.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{
exdt <- tribble(
  ~EXSTDTC,
  "2019-07-18T15:25:40",
  "2019-07-18T15:25",
  "2019-07-18",
  "2019-02",
  "2019",
  "2019---07",
  ""
)
derive_vars_dt(
  exdt,
  new_vars_prefix = "TRTS",
  dtc = EXSTDTC,
  highest_imputation = "M",
  date_imputation = "mid",
  flag_imputation = "none"
)
#> # A tibble: 7 × 2
#>   EXSTDTC               TRTSDT    
#>   <chr>                 <date>    
#> 1 "2019-07-18T15:25:40" 2019-07-18
#> 2 "2019-07-18T15:25"    2019-07-18
#> 3 "2019-07-18"          2019-07-18
#> 4 "2019-02"             2019-02-15
#> 5 "2019"                2019-06-30
#> 6 "2019---07"           2019-06-30
#> 7 ""                    NA        }\if{html}{\out{</div>}}}
\subsection{Impute to a specific date (\code{date_imputation = "04-06"})}{

In this example, we derive \code{ASTDT} with specific date imputation, i.e.
\code{date_imputation = "04-06"}. Note that day portion, \code{"-06"}, is used in the
imputation of the record with \code{"2019-02"}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{derive_vars_dt(
  mhdt,
  new_vars_prefix = "AST",
  dtc = MHSTDTC,
  highest_imputation = "M",
  date_imputation = "04-06"
)
#> # A tibble: 7 × 3
#>   MHSTDTC               ASTDT      ASTDTF
#>   <chr>                 <date>     <chr> 
#> 1 "2019-07-18T15:25:40" 2019-07-18 <NA>  
#> 2 "2019-07-18T15:25"    2019-07-18 <NA>  
#> 3 "2019-07-18"          2019-07-18 <NA>  
#> 4 "2019-02"             2019-02-06 D     
#> 5 "2019"                2019-04-06 M     
#> 6 "2019---07"           2019-04-06 M     
#> 7 ""                    NA         <NA>  }\if{html}{\out{</div>}}}
\subsection{Applying a lower boundary to date imputation with (\code{min_dates})}{

In this example, we derive \code{ASTDT} where \code{AESTDTC} is all partial dates in
need of imputation. Using \code{min_dates = exprs(TRTSDTM)}, we are telling the function
to apply the treatment start date \code{(TRTSDTM)} as a lower boundary for imputation
via the \code{min_dates} argument.
This means:
\itemize{
\item For partial dates that could potentially include \code{TRTSDTM} (case 1 & 2),
the imputed date is adjusted to \code{TRTSDTM}
\item For partial dates that are entirely before \code{TRTSDTM} (case 3 & 4),
standard imputation rules apply without adjustment
\item For partial dates that are entirely after \code{TRTSDTM} (case 5),
standard imputation rules apply
}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{adae <- tribble(
  ~case, ~AESTDTC,  ~TRTSDTM,
  1,     "2020-12", ymd_hms("2020-12-06T12:12:12"),
  2,     "2020",    ymd_hms("2020-12-06T12:12:12"),
  3,     "2020-11", ymd_hms("2020-12-06T12:12:12"),
  4,     "2020-01", ymd_hms("2020-12-06T12:12:12"),
  5,     "2021-01", ymd_hms("2020-12-06T12:12:12")
)

derive_vars_dt(
  adae,
  dtc = AESTDTC,
  new_vars_prefix = "AST",
  highest_imputation = "M",
  date_imputation = "first",
  min_dates = exprs(TRTSDTM)
)
#> # A tibble: 5 × 5
#>    case AESTDTC TRTSDTM             ASTDT      ASTDTF
#>   <dbl> <chr>   <dttm>              <date>     <chr> 
#> 1     1 2020-12 2020-12-06 12:12:12 2020-12-06 D     
#> 2     2 2020    2020-12-06 12:12:12 2020-12-06 M     
#> 3     3 2020-11 2020-12-06 12:12:12 2020-11-01 D     
#> 4     4 2020-01 2020-12-06 12:12:12 2020-01-01 D     
#> 5     5 2021-01 2020-12-06 12:12:12 2021-01-01 D     }\if{html}{\out{</div>}}}
\subsection{Restricting the range for \code{min_dates} (\code{max_dates_strict})}{

As in the previous example, we derive \code{ASTDT} from \code{AESTDTC}. Here we
additionally consider \code{AENDT} to determine if the potential start dates
include \code{TRTSDTM}.
\itemize{
\item For partial dates that could potentially include \code{TRTSDTM} and where
\code{AENDT} is at or after \code{TRTSDTM} (case 1), the imputed date is adjusted to
\code{TRTSDTM}.
\item For partial dates that could potentially include \code{TRTSDTM} but \code{AENDT} is
before \code{TRTSDTM} (case 2), standard imputation rules apply without
adjustment.
}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{adae <- tribble(
  ~case, ~AESTDTC,  ~AENDT,            ~TRTSDTM,
  1,     "2020-12", ymd("2020-12-12"), ymd_hms("2020-12-06T12:12:12"),
  2,     "2020",    ymd("2020-11-11"), ymd_hms("2020-12-06T12:12:12"),
  3,     "2020-11", ymd("2020-11-03"), ymd_hms("2020-12-06T12:12:12"),
  4,     "2020-01", ymd("2020-01-04"), ymd_hms("2020-12-06T12:12:12"),
  5,     "2021-01", ymd("2021-01-17"), ymd_hms("2020-12-06T12:12:12")
)

derive_vars_dt(
  adae,
  dtc = AESTDTC,
  new_vars_prefix = "AST",
  highest_imputation = "M",
  date_imputation = "first",
  min_dates = exprs(TRTSDTM),
  max_dates_strict = exprs(AENDT)
)
#> # A tibble: 5 × 6
#>    case AESTDTC AENDT      TRTSDTM             ASTDT      ASTDTF
#>   <dbl> <chr>   <date>     <dttm>              <date>     <chr> 
#> 1     1 2020-12 2020-12-12 2020-12-06 12:12:12 2020-12-06 D     
#> 2     2 2020    2020-11-11 2020-12-06 12:12:12 2020-01-01 M     
#> 3     3 2020-11 2020-11-03 2020-12-06 12:12:12 2020-11-01 D     
#> 4     4 2020-01 2020-01-04 2020-12-06 12:12:12 2020-01-01 D     
#> 5     5 2021-01 2021-01-17 2020-12-06 12:12:12 2021-01-01 D     }\if{html}{\out{</div>}}}
\subsection{Applying an upper boundary to date imputation with (\code{max_dates})}{

In this example, we derive \code{ASTDT} where \code{AESTDTC} is all partial dates in
need of imputation. Using \code{max_dates = exprs(TRTEDTM)}, we are telling the function
to apply the treatment end date \code{(TRTEDTM)} as an upper boundary for imputation
via the \code{max_dates} argument.
This means:
\itemize{
\item For partial dates that could potentially include \code{TRTEDTM} (case 1 & 2),
the imputed date is adjusted to \code{TRTEDTM}
\item For partial dates that are entirely before \code{TRTEDTM} (case 3 & 4),
standard imputation rules apply without adjustment
\item For partial dates that are entirely after \code{TRTEDTM} (case 5),
standard imputation rules apply
}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{adae <- tribble(
  ~case, ~AESTDTC,  ~TRTEDTM,
  1,     "2020-12", ymd_hms("2020-12-20T23:59:59"),
  2,     "2020",    ymd_hms("2020-12-20T23:59:59"),
  3,     "2020-11", ymd_hms("2020-12-20T23:59:59"),
  4,     "2020-01", ymd_hms("2020-12-20T23:59:59"),
  5,     "2021-01", ymd_hms("2020-12-20T23:59:59")
)

derive_vars_dt(
  adae,
  dtc = AESTDTC,
  new_vars_prefix = "AST",
  highest_imputation = "M",
  date_imputation = "last",
  max_dates = exprs(TRTEDTM)
)
#> # A tibble: 5 × 5
#>    case AESTDTC TRTEDTM             ASTDT      ASTDTF
#>   <dbl> <chr>   <dttm>              <date>     <chr> 
#> 1     1 2020-12 2020-12-20 23:59:59 2020-12-20 D     
#> 2     2 2020    2020-12-20 23:59:59 2020-12-20 M     
#> 3     3 2020-11 2020-12-20 23:59:59 2020-11-30 D     
#> 4     4 2020-01 2020-12-20 23:59:59 2020-01-31 D     
#> 5     5 2021-01 2020-12-20 23:59:59 2021-01-31 D     }\if{html}{\out{</div>}}}
\subsection{Preserve lower components if higher ones were imputed (\code{preserve})}{

The \code{preserve} argument can be used to "preserve" information from the partial dates.
For example, \code{"2019---07"}, will be displayed as \code{"2019-06-07"} rather than
\code{"2019-06-30"} with \code{preserve = TRUE} and \code{date_imputation = "mid"} .

\if{html}{\out{<div class="sourceCode r">}}\preformatted{derive_vars_dt(
  mhdt,
  new_vars_prefix = "AST",
  dtc = MHSTDTC,
  highest_imputation = "M",
  date_imputation = "mid",
  preserve = TRUE
)
#> # A tibble: 7 × 3
#>   MHSTDTC               ASTDT      ASTDTF
#>   <chr>                 <date>     <chr> 
#> 1 "2019-07-18T15:25:40" 2019-07-18 <NA>  
#> 2 "2019-07-18T15:25"    2019-07-18 <NA>  
#> 3 "2019-07-18"          2019-07-18 <NA>  
#> 4 "2019-02"             2019-02-15 D     
#> 5 "2019"                2019-06-30 M     
#> 6 "2019---07"           2019-06-07 M     
#> 7 ""                    NA         <NA>  }\if{html}{\out{</div>}}}
\subsection{Further examples}{

Further example usages of this function can be found in the
\code{vignette("imputation")}.}}

