% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_date_vars.R
\name{derive_vars_dt}
\alias{derive_vars_dt}
\title{Derive/Impute a date from a date character vector}
\usage{
derive_vars_dt(dataset, new_vars_prefix, dtc, date_imputation = NULL,
  flag_imputation = TRUE)
}
\arguments{
\item{dataset}{Input dataset.

The date character vector (dtc) must be present.}

\item{new_vars_prefix}{Prefix used for the output variable(s).

a character is expected: e.g new_vars_prefix="AST".}

\item{dtc}{The date character vector used to derive/impute the date.

A character date is expected in a format like yyyy-mm-dd or yyyy-mm-ddThh:mm:ss.
if the year part is not recorded (missing date), no imputation is done.}

\item{date_imputation}{The value to impute the day/month when a datepart is missing.

If NULL: no date imputation is performed and partial dates are returned as missing.

Otherwise, a character value is expected, either as a
\itemize{
\item format with day and month specified as 'dd-mm': e.g. '15-06' for the 15th of June,
\item or as a keyword: 'FIRST', 'MID', 'LAST' to impute to the first/mid/last day/month.
}

Default is NULL.}

\item{flag_imputation}{Whether the date imputation flag must also be derived.

A logical value

Default: TRUE}
}
\value{
the input dataset with the date '--DT' (and the date imputation flag '--DTF') added.
}
\description{
Derive a date ('--DT') from a date character vector ('---DTC').
The date can be imputed (see date_imputation parameter)
and the date imputation flag ('--DTF') can be added.
}
\examples{
mhdt <- tibble::tribble(
  ~MHSTDTC,
  "2019-07-18T15:25:40",
  "2019-07-18T15:25",
  "2019-07-18",
  "2019-02",
  "2019",
  "2019---07",
  ""
)

# Create ASTDT and ASTDTF
# no imputation for partial date
derive_vars_dt(
  mhdt,
  new_vars_prefix = "AST",
  dtc = MHSTDTC
)

# Create ASTDT and ASTDTF
# Impute partial dates to first day/month
derive_vars_dt(
  mhdt,
  new_vars_prefix = "AST",
  dtc = MHSTDTC,
  date_imputation = "FIRST"
)

# Impute partial dates to 4th of june
derive_vars_dt(
  mhdt,
  new_vars_prefix = "AST",
  dtc = MHSTDTC,
  date_imputation = "04-06"
)

# Create AENDT and AENDTF
# Impute partial dates to last day/month
derive_vars_dt(
  mhdt,
  new_vars_prefix = "AEN",
  dtc = MHSTDTC,
  date_imputation = "LAST"
)

# Create BIRTHDT
# Impute partial dates to 15th of June. No DTF
derive_vars_dt(
  mhdt,
  new_vars_prefix = "BIRTH",
  dtc = MHSTDTC,
  date_imputation = "MID",
  flag_imputation = FALSE
)
}
\author{
Samia Kabi
}
\keyword{adam}
\keyword{derivation}
\keyword{timing}
