% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_date_vars.R
\name{derive_vars_dtm}
\alias{derive_vars_dtm}
\title{Derive --DTM (and --DTF/--TMF)}
\usage{
derive_vars_dtm(dataset, new_vars_prefix, dtc, date_imputation = NULL,
  time_imputation = "00:00:00", flag_imputation = TRUE)
}
\arguments{
\item{dataset}{Input dataset

The --DTC variable must be present.}

\item{new_vars_prefix}{Prefix used for the output variable(s)

a character is expected, e.g new_vars_prefix="AST"}

\item{dtc}{The --DTC date used to derive/impute --DT

A character date is expected in a format like yyyy-mm-dd or yyyy-mm-ddThh:mm:ss.
If the year part is not recorded (missing date), no imputation is performed.}

\item{date_imputation}{The value to impute the day/month when a datepart is missing.

If NULL: no date imputation is performed and partial dates are returned as missing.

Otherwise, a character value is expected, either as a
\itemize{
\item format with day and month specified as 'dd-mm': e.g. '15-06' for the 15th of June
\item or as a keyword: 'FIRST', 'MID', 'LAST' to impute to the first/mid/last day/month
}

Default is NULL}

\item{time_imputation}{The value to impute the time when a timepart is missing

A character value is expected, either as a
\itemize{
\item format with hour, min and sec specified as 'hh:mm:ss': e.g. '00:00:00' for
the start of the day
\item or as a keyword: 'FIRST','LAST' to impute to the start/end of a day
}

Default is '00:00:00'}

\item{flag_imputation}{Whether the --DTF and or --TMF variable must also be derived

A logical value

Default: TRUE}
}
\description{
Derive --DTM based on --DTC. --DTM is imputed based on user input
Derive --DTF/--TMF if needed based on --DTC and --DT
}
\details{
The presence of a --DTF variable is checked and the variable is not derived
if it already exists in the input dataset. However, if --TMF already exists
in the input dataset, a warning is issued and --TMF will be overwritten
}
\examples{
mhdt <- tibble::tribble(
  ~MHSTDTC,
  "2019-07-18T15:25:40",
  "2019-07-18T15:25",
  "2019-07-18",
  "2019-02",
  "2019",
  "2019---07",
  ""
)

derive_vars_dtm(
  mhdt,
  new_vars_prefix = "AST",
  dtc = MHSTDTC,
  date_imputation = "FIRST",
  time_imputation = "FIRST"
)
}
\author{
Samia Kabi
}
