% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dtm_to_dt.R
\name{derive_vars_dtm_to_dt}
\alias{derive_vars_dtm_to_dt}
\title{Derive Date Variables from Datetime Variables}
\usage{
derive_vars_dtm_to_dt(dataset, source_vars)
}
\arguments{
\item{dataset}{Input dataset}

\item{source_vars}{A list of datetime variables from which dates are to be extracted}
}
\value{
A data frame containing the input dataset with the corresponding date (--DT)
variable(s) of all datetime variables (--DTM) specified in source_vars.
}
\description{
This function creates date(s) as output from datetime variable(S)
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(lubridate, warn.conflicts = FALSE)

adcm <- tibble::tribble(
  ~USUBJID, ~TRTSDTM,              ~ASTDTM,               ~AENDTM,
  "PAT01",  "2012-02-25 23:00:00", "2012-02-28 19:00:00", "2012-02-25 23:00:00",
  "PAT01",  NA,                    "2012-02-28 19:00:00", NA,
  "PAT01",  "2017-02-25 23:00:00", "2013-02-25 19:00:00", "2014-02-25 19:00:00",
  "PAT01",  "2017-02-25 16:00:00", "2017-02-25 14:00:00", "2017-03-25 23:00:00",
  "PAT01",  "2017-02-25 16:00:00", "2017-02-25 14:00:00", "2017-04-29 14:00:00",
) \%>\%
  mutate(
    TRTSDTM = as_datetime(TRTSDTM),
    ASTDTM = as_datetime(ASTDTM),
    AENDTM = as_datetime(AENDTM)
  )

derive_vars_dtm_to_dt(adcm, vars(TRTSDTM, ASTDTM, AENDTM))
}
\author{
Teckla Akinyi
}
\keyword{ADaM}
\keyword{Date}
\keyword{Timing}
