% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_merged.R
\name{derive_vars_merged}
\alias{derive_vars_merged}
\title{Merge Variables}
\usage{
derive_vars_merged(
  dataset,
  dataset_add,
  by_vars,
  order = NULL,
  new_vars = NULL,
  mode = NULL,
  filter_add = NULL,
  check_type = "warning"
)
}
\description{
Merge Variables
}
\examples{
library(admiral.test)
data("vs")
data("dm")

derive_vars_merged(vs,
                   dataset_add = select(dm, -DOMAIN),
                   by_vars = vars(STUDYID, USUBJID))
}
