% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_merged.R
\name{derive_vars_merged_dt}
\alias{derive_vars_merged_dt}
\title{Merge a (imputed) Date Variable}
\usage{
derive_vars_merged_dt(
  dataset,
  dataset_add,
  by_vars,
  order,
  new_vars_prefix,
  filter_add,
  mode,
  dtc,
  date_imputation = NULL,
  flag_imputation = TRUE,
  min_dates = NULL,
  max_dates = NULL
)
}
\description{
Merge a (imputed) Date Variable
}
\examples{
library(admiral.test)
data("dm")
data("ex")
derive_vars_merged_dt(
  select(dm, STUDYID, USUBJID),
  dataset_add = ex,
  by_vars = vars(STUDYID, USUBJID),
  new_vars_prefix = "TRTS",
  dtc = EXSTDTC,
  date_imputation = "first",
  order = vars(TRTSDT),
  mode = "first"
)

derive_vars_merged_dt(
  select(dm, STUDYID, USUBJID),
  dataset_add = ex,
  by_vars = vars(STUDYID, USUBJID),
  new_vars_prefix = "TRTE",
  dtc = EXENDTC,
  date_imputation = "last",
  order = vars(TRTEDT),
  mode = "last"
)
}
