% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{arg_name}
\alias{arg_name}
\title{Extract Argument Name from an Expression}
\usage{
arg_name(expr)
}
\arguments{
\item{expr}{An expression created inside a function using \code{substitute()}}
}
\description{
Extract Argument Name from an Expression
}
\examples{
test_fun <- function(something) {
  admiral:::arg_name(substitute(something))
}

inner_function <- function(x) x
test_fun2 <- function(something) {
  admiral:::arg_name(substitute(inner_function(something)))
}
}
\author{
Thomas Neitmann, Ondrej Slama
}
\keyword{dev_utility}
