% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{backquote}
\alias{backquote}
\title{Wrap a String in Backquotes}
\usage{
backquote(x)
}
\arguments{
\item{x}{A \code{character} vector}
}
\description{
Wrap a String in Backquotes
}
\examples{
admiral:::backquote("USUBJID")
}
\author{
Thomas Neitmann
}
\keyword{dev_utility}
