% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{dquote}
\alias{dquote}
\title{Wrap a String in Double Quotes}
\usage{
dquote(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
If the input is \code{NULL}, the text \code{"NULL"} is returned. Otherwise, the
input in double quotes is returned.
}
\description{
Wrap a string in double quotes, e.g., for displaying character values in
messages.
}
\examples{
dquote("foo")
dquote(NULL)
}
\author{
Stefan Bundfuss
}
\keyword{dev_utility}
