% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_dthcaus.R
\name{dthcaus_source}
\alias{dthcaus_source}
\title{Create a \code{dthcaus_source} Object}
\usage{
dthcaus_source(
  dataset_name,
  filter,
  date,
  order = NULL,
  mode = "first",
  dthcaus,
  traceability_vars = NULL
)
}
\arguments{
\item{dataset_name}{The name of the dataset, i.e. a string, used to search for
the death cause.}

\item{filter}{An expression used for filtering \code{dataset}.}

\item{date}{A date or datetime variable or an expression to be used for
sorting \code{dataset}.}

\item{order}{Sort order

Additional variables/expressions to be used for sorting the \code{dataset}. The
dataset is ordered by \code{date} and \code{order}. Can be used to avoid duplicate
record warning.

\emph{Permitted Values}: list of expressions created by \code{exprs()}, e.g.,
\code{exprs(ADT, desc(AVAL))} or \code{NULL}}

\item{mode}{One of \code{"first"} or \code{"last"}.
Either the \code{"first"} or \code{"last"} observation is preserved from the \code{dataset}
which is ordered by \code{date}.}

\item{dthcaus}{A variable name, an expression, or a string literal

If a variable name is specified, e.g., \code{AEDECOD}, it is the variable in the
source dataset to be used to assign values to \code{DTHCAUS}; if an expression,
e.g., \code{str_to_upper(AEDECOD)}, it is evaluated in the source dataset and
the results is assigned to \code{DTHCAUS}; if a string literal, e.g. \code{"Adverse Event"}, it is the fixed value to be assigned to \code{DTHCAUS}.}

\item{traceability_vars}{A named list returned by \code{\link[=exprs]{exprs()}} listing the
traceability variables, e.g. \code{exprs(DTHDOM = "DS", DTHSEQ = DSSEQ)}. The
left-hand side (names of the list elements) gives the names of the
traceability variables in the returned dataset. The right-hand side (values
of the list elements) gives the values of the traceability variables in the
returned dataset. These can be either strings, numbers, symbols, or
expressions referring to existing variables.}
}
\value{
An object of class "dthcaus_source".
}
\description{
Create a \code{dthcaus_source} Object
}
\examples{
# Deaths sourced from AE
src_ae <- dthcaus_source(
  dataset_name = "ae",
  filter = AEOUT == "FATAL",
  date = AEDTHDT,
  mode = "first",
  dthcaus = AEDECOD
)

# Deaths sourced from DS
src_ds <- dthcaus_source(
  dataset_name = "ds",
  filter = DSDECOD == "DEATH",
  date = convert_dtc_to_dt(DSSTDTC),
  mode = "first",
  dthcaus = DSTERM
)
}
\seealso{
\code{\link[=derive_var_dthcaus]{derive_var_dthcaus()}}

Source Objects: 
\code{\link{basket_select}()},
\code{\link{censor_source}()},
\code{\link{date_source}()},
\code{\link{death_event}},
\code{\link{event_source}()},
\code{\link{event}()},
\code{\link{query}()},
\code{\link{records_source}()},
\code{\link{tte_source}()}
}
\concept{source_specifications}
\keyword{source_specifications}
