% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_dthcaus.R
\name{dthcaus_source}
\alias{dthcaus_source}
\title{Create an \code{dthcaus_source} object}
\usage{
dthcaus_source(dataset, filter, order, mode = "first", dthdom, dthcaus)
}
\arguments{
\item{dataset}{A data.frame containint a source dataset.}

\item{filter}{A symbol returned by \code{expr} to be used for filtering \code{dataset}.}

\item{order}{Alist returned by \code{exprs} to be used for sorting \code{dataset}.}

\item{mode}{One of "first" or "last".}
}
\description{
Create an \code{dthcaus_source} object
}
