% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_dthcaus.R
\name{dthcaus_source}
\alias{dthcaus_source}
\title{Create a \code{dthcaus_source} Object}
\usage{
dthcaus_source(
  dataset_name,
  filter,
  date,
  date_imputation = NULL,
  order = NULL,
  mode = "first",
  dthcaus,
  traceability_vars = NULL
)
}
\arguments{
\item{dataset_name}{The name of the dataset, i.e. a string, used to search for
the death cause.}

\item{filter}{An expression used for filtering \code{dataset}.}

\item{date}{A date or datetime variable to be used for sorting \code{dataset}.}

\item{date_imputation}{The value to impute the day/month when a datepart is missing.

If \code{NULL}: no date imputation is performed and partial dates are returned as missing.

Otherwise, a character value is expected, either as a
\itemize{
\item format with day and month specified as 'mm-dd': e.g. '06-15' for the 15th
of June
\item or as a keyword: 'FIRST', 'MID', 'LAST' to impute to the first/mid/last day/month.
}

Default is \code{NULL}}

\item{order}{Sort order

Additional variables to be used for sorting the \code{dataset} which is ordered by the
(imputed) \code{date} and \code{order}. Can be used to avoid duplicate record warning.

\emph{Default}: \code{NULL}

\emph{Permitted Values}: list of variables or \verb{desc(<variable>)} function calls
created by \code{vars()}, e.g., \code{vars(ADT, desc(AVAL))} or \code{NULL}}

\item{mode}{One of \code{"first"} or \code{"last"}.
Either the \code{"first"} or \code{"last"} observation is preserved from the \code{dataset}
which is ordered by \code{date}.}

\item{dthcaus}{A variable name or a string literal --- if a variable name, e.g., \code{AEDECOD},
it is the variable in the source dataset to be used to assign values to
\code{DTHCAUS}; if a string literal, e.g. \code{"Adverse Event"}, it is the fixed value
to be assigned to \code{DTHCAUS}.}

\item{traceability_vars}{A named list returned by \code{\link[=vars]{vars()}} listing the traceability variables,
e.g. \code{vars(DTHDOM = "DS", DTHSEQ = DSSEQ)}.
The left-hand side (names of the list elements) gives the names of the traceability variables
in the returned dataset.
The right-hand side (values of the list elements) gives the values of the traceability variables
in the returned dataset.
These can be either strings or symbols referring to existing variables.}
}
\value{
An object of class "dthcaus_source".
}
\description{
Create a \code{dthcaus_source} Object
}
\seealso{
\code{\link[=derive_var_dthcaus]{derive_var_dthcaus()}}

Source Specifications: 
\code{\link{censor_source}()},
\code{\link{date_source}()},
\code{\link{death_event}},
\code{\link{event_source}()},
\code{\link{filter_date_sources}()},
\code{\link{list_tte_source_objects}()},
\code{\link{params}()},
\code{\link{query}()},
\code{\link{sdg_select}()},
\code{\link{smq_select}()},
\code{\link{tte_source}()}
}
\author{
Shimeng Huang
}
\concept{source_specifications}
\keyword{source_specifications}
