% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_date_vars.R
\name{dtm_level}
\alias{dtm_level}
\title{Create a \code{dtm_level} object}
\usage{
dtm_level(level)
}
\arguments{
\item{level}{Datetime level

\emph{Permitted Values}: \code{"Y"} (year, highest level), \code{"M"} (month), \code{"D"}
(day), \code{"h"} (hour), \code{"m"} (minute), \code{"s"} (second, lowest level), \code{"n"}
(none)}
}
\value{
A \code{dtm_level} object
}
\description{
Create a \code{dtm_level} object
}
\details{
A \code{dtm_level} object is an ordered factor, i.e., two objects can be
compared.
}
\seealso{
Other utils_impute: 
\code{\link{dt_level}()},
\code{\link{get_imputation_target_date}()},
\code{\link{get_imputation_target_time}()},
\code{\link{get_partialdatetime}()},
\code{\link{restrict_imputed_dtc_dtm}()},
\code{\link{restrict_imputed_dtc_dt}()}
}
\author{
Stefan Bundfuss
}
\concept{utils_impute}
\keyword{utils_impute}
