% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enumerate}
\alias{enumerate}
\title{Enumerate Multiple Strings}
\usage{
enumerate(x, quote_fun = backquote, conjunction = "and")
}
\arguments{
\item{x}{A \code{character} vector}

\item{quote_fun}{Quoting function, defaults to \code{backquote}.}

\item{conjunction}{Character to be used in the message, defaults to "and".}
}
\description{
Enumerate Multiple Strings
}
\examples{
admiral:::enumerate(c("STUDYID", "USUBJID", "PARAMCD"))
admiral:::enumerate(letters[1:6], quote_fun = admiral:::squote)
admiral:::enumerate(
  c("date", "time", "both"),
  quote_fun = admiral:::squote,
  conjunction = "or"
)
}
\author{
Thomas Neitmann
}
\keyword{dev_utility}
