% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_extreme_event.R
\name{event}
\alias{event}
\title{Create a \code{event} Object}
\usage{
event(condition, set_values_to = NULL)
}
\arguments{
\item{condition}{An unquoted condition for selecting the observations, which
will contribute to the extreme event.}

\item{set_values_to}{A named list returned by \code{exprs()} defining the variables
to be set for the extreme answer, e.g. \verb{exprs(PARAMCD = "WSP", PARAM  = "Worst Sleeping Problems"}. The values must be a symbol, a
character string, a numeric value, or \code{NA}.}
}
\value{
An object of class \code{event}
}
\description{
The \code{event} object is used to define events as input for the
\code{derive_extreme_event()} function.
}
\seealso{
\code{\link[=derive_extreme_event]{derive_extreme_event()}}

Source Specifications: 
\code{\link{assert_db_requirements}()},
\code{\link{assert_terms}()},
\code{\link{assert_valid_queries}()},
\code{\link{basket_select}()},
\code{\link{censor_source}()},
\code{\link{date_source}()},
\code{\link{death_event}},
\code{\link{dthcaus_source}()},
\code{\link{event_source}()},
\code{\link{extend_source_datasets}()},
\code{\link{filter_date_sources}()},
\code{\link{format.basket_select}()},
\code{\link{list_tte_source_objects}()},
\code{\link{params}()},
\code{\link{query}()},
\code{\link{tte_source}()},
\code{\link{validate_basket_select}()},
\code{\link{validate_query}()}
}
\concept{source_specifications}
\keyword{source_specifications}
