% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_tte.R
\name{event_source}
\alias{event_source}
\title{Create an \code{event_source} Object}
\usage{
event_source(dataset_name, filter = NULL, date, set_values_to = NULL)
}
\arguments{
\item{dataset_name}{The name of the source dataset

The name refers to the dataset provided by the \code{source_datasets} parameter
of \code{derive_param_tte()}.}

\item{filter}{An unquoted condition for selecting the observations from
\code{dataset} which are events or possible censoring time points.}

\item{date}{A variable providing the date of the event or censoring. A date,
a datetime, or a character variable containing ISO 8601 dates can be
specified. An unquoted symbol is expected.}

\item{set_values_to}{A named list returned by \code{vars()} defining the variables
to be set for the event or censoring, e.g. \code{vars(EVENTDESC = "DEATH", SRCDOM = "ADSL", SRCVAR = "DTHDT")}. The values must be a symbol, a
character string, a numeric value, or \code{NA}.}
}
\value{
An object of class "event_source".
}
\description{
The \code{event_source} object is used to define events as input for the
\code{derive_param_tte()} function.
}
\author{
Stefan Bundfuss
}
\keyword{source_specifications}
