% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_single}
\alias{ex_single}
\title{Single Dose Exposure Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 22439 rows and 17 columns.
}
\source{
Derived from the \link{ex} dataset using \code{{admiral}} and \code{{dplyr}} (\url{https://github.com/pharmaverse/admiral/blob/main/inst/example_scripts/derive_single_dose.R})
}
\usage{
ex_single
}
\description{
A derived dataset with single dose per date.
}
\keyword{datasets}
