% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{extract_duplicate_records}
\alias{extract_duplicate_records}
\title{Extract Duplicate Records}
\usage{
extract_duplicate_records(dataset, by_vars)
}
\arguments{
\item{dataset}{A data frame}

\item{by_vars}{A list of variables created using \code{vars()} identifying groups of
records in which to look for duplicates}
}
\value{
A \code{data.frame} of duplicate records within \code{dataset}
}
\description{
Extract Duplicate Records
}
\examples{
data(admiral_adsl)

# Duplicate the first record
adsl <- rbind(adsl[1L, ], adsl)

extract_duplicate_records(adsl, vars(USUBJID))
}
\author{
Thomas Neitmann
}
\keyword{dev_utility}
