% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_utils.R
\name{extract_unit}
\alias{extract_unit}
\title{Extract Unit From Parameter Description}
\usage{
extract_unit(x)
}
\arguments{
\item{x}{A parameter description

\describe{
\item{Default value}{none}
}}
}
\value{
A string
}
\description{
Extract the unit of a parameter from a description like "Param (unit)".
}
\examples{
extract_unit("Height (cm)")

extract_unit("Diastolic Blood Pressure (mmHg)")
}
\seealso{
Utilities used within Derivation functions: 
\code{\link{get_flagged_records}()},
\code{\link{get_not_mapped}()},
\code{\link{get_vars_query}()}
}
\concept{utils_help}
\keyword{utils_help}
