% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R, R/user_utils.R
\name{extract_unit}
\alias{extract_unit}
\title{Extract Unit From Parameter Description}
\usage{
extract_unit(x)

extract_unit(x)
}
\arguments{
\item{x}{A parameter description}
}
\value{
A string
}
\description{
Extract the unit of a parameter from a description like "Param (unit)".

Extract the unit of a parameter from a description like "Param (unit)".
}
\examples{
extract_unit("Height (cm)")

extract_unit("Diastolic Blood Pressure (mmHg)")
extract_unit("Height (cm)")

extract_unit("Diastolic Blood Pressure (mmHg)")
}
\seealso{
Utilities used within Derivation functions: 
\code{\link{call_user_fun}()},
\code{\link{get_not_mapped}()},
\code{\link{negate_vars}()},
\code{\link{signal_duplicate_records}()},
\code{\link{vars2chr}()}

Utilities used within Derivation functions: 
\code{\link{call_user_fun}()},
\code{\link{get_not_mapped}()},
\code{\link{negate_vars}()},
\code{\link{signal_duplicate_records}()},
\code{\link{vars2chr}()}
}
\concept{utils_help}
\keyword{utils_help}
