% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_unit}
\alias{extract_unit}
\title{Extract Unit From Parameter Description}
\usage{
extract_unit(x)
}
\arguments{
\item{x}{A parameter description}
}
\description{
Extract the unit of a parameter from a description like "Param (unit)".
}
\examples{
extract_unit("Height (cm)")

extract_unit("Diastolic Blood Pressure (mmHg)")
}
\keyword{user_utility}
