% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{extract_vars}
\alias{extract_vars}
\title{Extract All Symbols from a List of Quosures}
\usage{
extract_vars(x, side = "lhs")
}
\arguments{
\item{x}{An \code{R} object}

\item{side}{One of \code{"lhs"} (the default) or \code{"rhs"}}
}
\value{
A list of \code{quosures}
}
\description{
Extract All Symbols from a List of Quosures
}
\examples{
extract_vars(vars(STUDYID, USUBJID, desc(ADTM)))
}
\seealso{
Functions moved to admiraldev package: 
\code{\link{arg_name}()},
\code{\link{assert_db_requirements}()},
\code{\link{assert_terms}()},
\code{\link{assert_valid_queries}()},
\code{\link{backquote}()},
\code{\link{contains_vars}()},
\code{\link{dquote}()},
\code{\link{enumerate}()},
\code{\link{extend_source_datasets}()},
\code{\link{format.sdg_select}()},
\code{\link{format.smq_select}()},
\code{\link{get_constant_vars}()},
\code{\link{get_duplicates}()},
\code{\link{get_source_vars}()},
\code{\link{is_auto}()},
\code{\link{quo_c}()},
\code{\link{replace_values_by_names}()},
\code{\link{squote}()},
\code{\link{valid_time_units}()},
\code{\link{validate_query}()},
\code{\link{validate_sdg_select}()},
\code{\link{validate_smq_select}()},
\code{\link{what_is_it}()}
}
\author{
Thomas Neitmann
}
\concept{move_adm_dev}
\keyword{move_adm_dev}
