% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{extract_vars}
\alias{extract_vars}
\title{Extract All Symbols from a List of Quosures}
\usage{
extract_vars(x, side = "lhs")
}
\arguments{
\item{x}{An \code{R} object}

\item{side}{One of \code{"lhs"} (the default) or \code{"rhs"}}
}
\value{
A list of \code{quosures}
}
\description{
Extract All Symbols from a List of Quosures
}
\examples{
extract_vars(vars(STUDYID, USUBJID, desc(ADTM)))
}
\author{
Thomas Neitmann
}
\keyword{dev_utility}
