% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{extract_vars}
\alias{extract_vars}
\title{Extract All Symbols from a List of Expressions}
\usage{
extract_vars(x, side = "lhs")
}
\arguments{
\item{x}{An \code{R} object}

\item{side}{One of \code{"lhs"} (the default) or \code{"rhs"} for formulas}
}
\value{
A list of expressions
}
\description{
Extract All Symbols from a List of Expressions
}
\examples{
library(rlang)
extract_vars(exprs(PARAMCD, (BASE - AVAL) / BASE + 100))
extract_vars(AVAL ~ ARMCD + AGEGR1)
extract_vars(AVAL ~ ARMCD + AGEGR1, side = "rhs")
}
\seealso{
Developer Utility Functions: 
\code{\link{\%notin\%}()},
\code{\link{\%or\%}()},
\code{\link{arg_name}()},
\code{\link{contains_vars}()},
\code{\link{convert_dtm_to_dtc}()},
\code{\link{filter_if}()},
\code{\link{friendly_type_of}()},
\code{\link{valid_time_units}()},
\code{\link{vars2chr}()}
}
\concept{dev_utility}
\keyword{dev_utility}
