% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_extreme.R
\name{filter_extreme}
\alias{filter_extreme}
\title{Filter out the first or last observation for each group}
\usage{
filter_extreme(dataset, order, by_vars, mode = "last")
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{order} and the \code{by_vars} parameter are
expected.}

\item{order}{Sort order

Within each by group the observations are ordered by the specified order.

Permitted Values: list of variables or functions of variables}

\item{by_vars}{Grouping variables

Permitted Values: list of variables}

\item{mode}{Selection mode (first or last)

If \code{"first"} is specified, the first observation of each by group is
included in the output dataset. If \code{"last"} is specified, the last
observation of each by group is included in the output dataset.

Permitted Values:  \code{"first"}, \code{"last"}}
}
\value{
A dataset containing the first or last observation of each by group
}
\description{
Filters out the first or last observation for each group.
}
\details{
For each group (with respect to the variables specified for the
\code{by_vars} parameter) the first or last observation (with respect to the
order specified for the \code{order} parameter and the mode specified for the
\code{mode} parameter) is included in the output dataset.
}
\examples{
data("ex")

# selecting first dose for each patient
filter_extreme(ex,
               order = rlang::exprs(EXSEQ),
               by_vars = rlang::exprs(USUBJID),
               mode = 'first')

# selecting highest dose for each patient
filter_extreme(ex,
               order = rlang::exprs(EXDOSE),
               by_vars = rlang::exprs(USUBJID))

}
\author{
Stefan Bundfuss
}
\keyword{adam}
