% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_extreme.R
\name{filter_extreme}
\alias{filter_extreme}
\title{Filter out the first or last observation for each group}
\usage{
filter_extreme(dataset, order, by_vars, mode, check_type = "warning")
}
\arguments{
\item{dataset}{Input dataset

The variables specified by the \code{order} and the \code{by_vars} parameter are
expected.}

\item{order}{Sort order

Within each by group the observations are ordered by the specified order.

Permitted Values: list of variables or functions of variables}

\item{by_vars}{Grouping variables

Permitted Values: list of variables}

\item{mode}{Selection mode (first or last)

If \code{"first"} is specified, the first observation of each by group is
included in the output dataset. If \code{"last"} is specified, the last
observation of each by group is included in the output dataset.

Permitted Values:  \code{"first"}, \code{"last"}}

\item{check_type}{Check uniqueness?

If \code{"warning"} or \code{"error"} is specified, the specified message is issued
if the observations of the input dataset are not unique with respect to the
by variables and the order.

Default: \code{"none"}

Permitted Values: \code{"none"}, \code{"warning"}, \code{"error"}}
}
\value{
A dataset containing the first or last observation of each by group
}
\description{
Filters out the first or last observation for each group.
}
\details{
For each group (with respect to the variables specified for the
\code{by_vars} parameter) the first or last observation (with respect to the
order specified for the \code{order} parameter and the mode specified for the
\code{mode} parameter) is included in the output dataset.
}
\examples{
library(dplyr)
library(magrittr)

data("ex")

# selecting first dose for each patient
filter_extreme(ex,
               order = exprs(EXSEQ),
               by_vars = exprs(USUBJID),
               mode = "first") \%>\%
  select(USUBJID, EXSEQ)

# selecting highest dose for each patient
filter_extreme(ex,
               order = exprs(EXDOSE),
               by_vars = exprs(USUBJID),
               mode = "last",
               check_type = "none") \%>\%
  select(USUBJID, EXDOSE)
}
\author{
Stefan Bundfuss
}
\keyword{adam}
