% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_if}
\alias{filter_if}
\title{Optional Filter}
\usage{
filter_if(dataset, filter)
}
\arguments{
\item{dataset}{Input dataset}

\item{filter}{A filter condition. Must be a quosure.}
}
\description{
Filters the input dataset if the provided expression is not \code{NULL}
}
\examples{
data(vs)
admiral:::filter_if(vs, rlang::quo(NULL))
admiral:::filter_if(vs, rlang::quo(VSTESTCD == "Weight"))
}
\author{
Thomas Neitmann
}
\keyword{dev_utility}
