% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query_data.R
\name{format.basket_select}
\alias{format.basket_select}
\title{Returns a Character Representation of a \code{basket_select()} Object}
\usage{
\method{format}{basket_select}(x, ...)
}
\arguments{
\item{x}{A \code{basket_select()} object

\describe{
\item{Default value}{none}
}}

\item{...}{Not used

\describe{
\item{Default value}{none}
}}
}
\value{
A character representation of the \code{basket_select()} object
}
\description{
The function returns a character representation of a \code{basket_select()} object.
It can be used for error messages for example.
}
\examples{

format(basket_select(id = 42, scope = "NARROW", type = "smq"))
}
\seealso{
\code{\link[=basket_select]{basket_select()}}

Other internal:
\code{\link{admiral-package}},
\code{\link[=as_admiral_df]{as_admiral_df()}},
\code{\link[=extract_duplicate_records]{extract_duplicate_records()}},
\code{\link[=signal_duplicate_records]{signal_duplicate_records()}}
}
\concept{internal}
\keyword{internal}
