% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query_data.R
\name{format.basket_select}
\alias{format.basket_select}
\title{Returns a Character Representation of a \code{basket_select()} Object}
\usage{
\method{format}{basket_select}(x, ...)
}
\arguments{
\item{x}{A \code{basket_select()} object

\describe{
\item{Default value}{none}
}}

\item{...}{Not used

\describe{
\item{Default value}{none}
}}
}
\value{
A character representation of the \code{basket_select()} object
}
\description{
The function returns a character representation of a \code{basket_select()} object.
It can be used for error messages for example.
}
\examples{

format(basket_select(id = 42, scope = "NARROW", type = "smq"))
}
\seealso{
\code{\link[=basket_select]{basket_select()}}

Other internal:
\code{\link{admiral-package}},
\code{\link[=as_admiral_df]{as_admiral_df()}},
\code{\link[=extract_duplicate_records]{extract_duplicate_records()}},
\code{\link[=get_admiral_df_type]{get_admiral_df_type()}},
\code{\link[=infer_admiral_keys]{infer_admiral_keys()}},
\code{\link[=is_adsl_structure]{is_adsl_structure()}},
\code{\link[=minimal_unique_key]{minimal_unique_key()}},
\code{\link[=print_adsl_summary]{print_adsl_summary()}},
\code{\link[=print_bds_summary]{print_bds_summary()}},
\code{\link[=print_occds_summary]{print_occds_summary()}},
\code{\link[=print_summary_checks]{print_summary_checks()}},
\code{\link[=print_vs_adsl_summary]{print_vs_adsl_summary()}},
\code{\link[=render_summary_table]{render_summary_table()}},
\code{\link[=signal_duplicate_records]{signal_duplicate_records()}},
\code{\link[=summarize_adsl]{summarize_adsl()}},
\code{\link[=summarize_bds]{summarize_bds()}},
\code{\link[=summarize_occds]{summarize_occds()}},
\code{\link[=summarize_vs_adsl]{summarize_vs_adsl()}}
}
\concept{internal}
\keyword{internal}
