% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query_data.R
\name{format.basket_select}
\alias{format.basket_select}
\title{Returns a Character Representation of a \code{basket_select()} Object}
\usage{
\method{format}{basket_select}(x, ...)
}
\arguments{
\item{x}{A \code{basket_select()} object}

\item{...}{Not used}
}
\value{
A character representation of the \code{basket_select()} object
}
\description{
The function returns a character representation of a \code{basket_select()} object.
It can be used for error messages for example.
}
\examples{

format(basket_select(id = 42, scope = "NARROW", type = "smq"))
}
\seealso{
\code{\link[=basket_select]{basket_select()}}

Other Advanced Functions: 
\code{\link{assert_db_requirements}()},
\code{\link{assert_terms}()},
\code{\link{assert_valid_queries}()},
\code{\link{extend_source_datasets}()},
\code{\link{filter_date_sources}()},
\code{\link{list_tte_source_objects}()},
\code{\link{params}()},
\code{\link{validate_basket_select}()},
\code{\link{validate_query}()}
}
\concept{other_advanced}
\keyword{other_advanced}
