% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_disposition_status.R
\name{format_eoxxstt_default}
\alias{format_eoxxstt_default}
\title{Default Format for Disposition Status}
\usage{
format_eoxxstt_default(status)
}
\arguments{
\item{status}{the disposition variable used for the mapping (e.g. \code{DSDECOD}).}
}
\value{
A \code{character} vector derived based on the values given in \code{status}:
"NOT STARTED" if \code{status} is "SCREEN FAILURE" or "SCREENING NOT COMPLETED",
"COMPLETED" if \code{status} is "COMPLETED",
"DISCONTINUED" if \code{status} is not in ("COMPLETED","SCREEN FAILURE",
"SCREENING NOT COMPLETED") nor NA,
"ONGOING" otherwise.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is \emph{deprecated}. This function is a default for \code{derive_var_disposition_status()}
for the \code{format_new_var} argument. Please define your own function and use that as input for the
\code{cat_fun} argument in \code{derive_var_merged_cat()} instead.

Define a function to map the disposition status. To be used as an input for
\code{derive_var_disposition_status()}.
}
\details{
Usually this function can not be used with \verb{\%>\%}.
}
\seealso{
Other deprecated: 
\code{\link{derive_derived_param}()},
\code{\link{derive_param_first_event}()},
\code{\link{derive_var_agegr_fda}()},
\code{\link{derive_var_confirmation_flag}()},
\code{\link{derive_var_disposition_status}()},
\code{\link{derive_var_worst_flag}()},
\code{\link{derive_vars_disposition_reason}()},
\code{\link{derive_vars_merged_dtm}()},
\code{\link{derive_vars_merged_dt}()},
\code{\link{filter_confirmation}()},
\code{\link{format_reason_default}()},
\code{\link{vars}()}
}
\concept{deprecated}
\keyword{deprecated}
