% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_disposition_status.R
\name{format_eoxxstt_default}
\alias{format_eoxxstt_default}
\title{Default format for the disposition status}
\usage{
format_eoxxstt_default(x)
}
\arguments{
\item{x}{the disposition variable used for the mapping (e.g. \code{DSDECOD}).}
}
\value{
A \code{character} vector derived based on the values given in \code{x}:
"COMPLETED" if \code{x} == "COMPLETED",
"DISCONTINUED" if \code{x} is not "COMPLETED" or NA,
"ONGOING" otherwise.
}
\description{
Define a function to map the disposition status.
}
\author{
Samia Kabi
}
\keyword{adsl}
\keyword{computation}
\keyword{user_utility}
