% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_var_disposition_status.R
\name{format_eoxxstt_default}
\alias{format_eoxxstt_default}
\title{Default Format for Disposition Status}
\usage{
format_eoxxstt_default(x)
}
\arguments{
\item{x}{the disposition variable used for the mapping (e.g. \code{DSDECOD}).}
}
\value{
A \code{character} vector derived based on the values given in \code{x}:
"COMPLETED" if \code{x} is "COMPLETED",
"DISCONTINUED" if \code{x} is not "COMPLETED" nor NA,
"ONGOING" otherwise.
}
\description{
Define a function to map the disposition status. To be used as an input for
\code{derive_var_disposition_status()}.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
library(admiraltest)
data("admiral_dm")
data("admiral_ds")

admiral_dm \%>\%
  derive_var_disposition_status(
    dataset_ds = admiral_ds,
    new_var = EOSSTT,
    status_var = DSDECOD,
    format_new_var = format_eoxxstt_default,
    filter_ds = DSCAT == "DISPOSITION EVENT"
  ) \%>\%
  select(STUDYID, USUBJID, EOSSTT)
}
\seealso{
\code{\link[=derive_var_disposition_status]{derive_var_disposition_status()}}
}
\author{
Samia Kabi
}
\keyword{adsl}
\keyword{computation}
\keyword{user_utility}
