% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_disposition_reason.R
\name{format_reason_default}
\alias{format_reason_default}
\title{Default format for the disposition reason}
\usage{
format_reason_default(x, y = NULL)
}
\arguments{
\item{x}{the disposition variable used for the mapping (e.g. \code{DSDECOD}).}

\item{y}{the disposition variable used for the mapping of the details if required (e.g. \code{DSTERM}).}
}
\description{
Define a function to map the disposition reason
}
