% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_disposition_reason.R
\name{format_reason_default}
\alias{format_reason_default}
\title{Default Format for the Disposition Reason}
\usage{
format_reason_default(reason, reason_spe = NULL)
}
\arguments{
\item{reason}{the disposition variable used for the mapping (e.g. \code{DSDECOD}).}

\item{reason_spe}{the disposition variable used for the mapping of the details
if required (e.g. \code{DSTERM}).}
}
\value{
A \code{character} vector
}
\description{
Define a function to map the disposition reason
}
\details{
\code{format_reason_default(DSDECOD)} returns \code{DSDECOD} when \code{DSDECOD} is not
'COMPLETED' nor \code{NA}. \code{format_reason_default(DSDECOD, DSTERM)} returns
\code{DSTERM} when \code{DSDECOD} is not 'COMPLETED' nor \code{NA}.

For example:\preformatted{DCSREAS = format_reason_default(DSDECOD)
DCSREASP = format_reason_default(DSDECOD, DSTERM)
}
}
\author{
Samia Kabi
}
\keyword{adsl}
\keyword{computation}
\keyword{user_utility}
