% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiral_df.R
\name{get_admiral_df_type}
\alias{get_admiral_df_type}
\title{Determine the ADaM Structure of a Dataset}
\usage{
get_admiral_df_type(dataset)
}
\arguments{
\item{dataset}{A data frame

\describe{
\item{Default value}{none}
}}
}
\value{
A character scalar: one of \code{"ADSL"}, \code{"BDS"}, \code{"OCCDS"}, \code{"TTE"}, or
\code{"other"}.
}
\description{
Classifies a data frame as one of the common ADaM structures based on the
variables it contains. This is used by \code{\link[=summary.admiral_df]{summary.admiral_df()}} to tailor the
diagnostic output.
}
\details{
The following precedence is used (the order matters as, e.g., both \code{"BDS"}
and \code{"TTE"} datasets contain \code{PARAMCD}):
\enumerate{
\item \code{"TTE"} -- contains \code{PARAMCD}, \code{CNSR}, and \code{STARTDT}.
\item \code{"OCCDS"} -- does not contain \code{PARAMCD} but contains a \code{--DECOD}/\code{--TERM}
variable and/or \code{TRTEMFL}/\code{AOCCxxFL}.
\item \code{"BDS"} -- contains \code{PARAMCD} and \code{AVAL} or \code{AVALC}.
\item \code{"ADSL"} -- does not contain \code{PARAMCD} and has one record per subject
(with respect to \code{get_admiral_option("subject_keys")}) or contains
typical subject-level treatment variables (\code{TRT01P}/\code{TRTSDT}).
\item \code{"other"} -- none of the above.
}
}
\seealso{
Other internal:
\code{\link{admiral-package}},
\code{\link[=as_admiral_df]{as_admiral_df()}},
\code{\link[=extract_duplicate_records]{extract_duplicate_records()}},
\code{\link[=format.basket_select]{format.basket_select()}},
\code{\link[=is_adsl_structure]{is_adsl_structure()}},
\code{\link[=signal_duplicate_records]{signal_duplicate_records()}}
}
\concept{internal}
\keyword{internal}
