% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_constant_vars}
\alias{get_constant_vars}
\title{Get constant variables}
\usage{
get_constant_vars(dataset, by_vars)
}
\arguments{
\item{dataset}{A data frame.}

\item{by_vars}{By variables
The groups defined by the by variables are considered separately. I.e., if
a variable is constant within each by group, it is returned.}
}
\value{
Variable vector.
}
\description{
Get constant variables
}
\keyword{dev_utility}
