% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_constant_vars}
\alias{get_constant_vars}
\title{Get Constant Variables}
\usage{
get_constant_vars(dataset, by_vars, ignore_vars = NULL)
}
\arguments{
\item{dataset}{A data frame.}

\item{by_vars}{By variables
The groups defined by the by variables are considered separately. I.e., if
a variable is constant within each by group, it is returned.}

\item{ignore_vars}{Variables to ignore
The specified variables are not considered, i.e., they are not returned
even if they are constant (unless they are included in the by variables).

\emph{Permitted Values:} A list of variable names or selector function calls
like \code{starts_with("EX")}}
}
\value{
Variable vector.
}
\description{
Get Constant Variables
}
\seealso{
Brings something to you!?!: 
\code{\link{get_duplicates}()},
\code{\link{get_source_vars}()}
}
\concept{get}
\keyword{get}
