% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_dataset}
\alias{get_dataset}
\title{Retrieve a Dataset from the \code{.datasets} environment}
\usage{
get_dataset(name)
}
\arguments{
\item{name}{The name of the dataset to retrieve}
}
\value{
A \code{data.frame}
}
\description{
Retrieve a Dataset from the \code{.datasets} environment
}
\seealso{
Other datasets: 
\code{\link{set_dataset}()}
}
\author{
Thomas Neitmann
}
\concept{datasets}
\keyword{datasets}
