% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{get_duplicates}
\alias{get_duplicates}
\title{Returns Duplicated Values of a Vector}
\usage{
get_duplicates(x)
}
\arguments{
\item{x}{Input vector}
}
\value{
A vector of the values which occur more than once in the input vector
}
\description{
Returns all values which occur more than once in the vector
}
\examples{

get_duplicates(c(1, 2, 1, 3, 1, 2, 4))
}
\keyword{dev_utility}
