% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{get_duplicates}
\alias{get_duplicates}
\title{Returns Duplicated Values of a Vector}
\usage{
get_duplicates(x)
}
\arguments{
\item{x}{Input vector}
}
\value{
A vector of the values which occur more than once in the input vector
}
\description{
Returns all values which occur more than once in the vector
}
\examples{

get_duplicates(c(1, 2, 1, 3, 1, 2, 4))
}
\seealso{
Functions moved to admiraldev package: 
\code{\link{arg_name}()},
\code{\link{assert_db_requirements}()},
\code{\link{assert_terms}()},
\code{\link{assert_valid_queries}()},
\code{\link{backquote}()},
\code{\link{contains_vars}()},
\code{\link{dquote}()},
\code{\link{enumerate}()},
\code{\link{extend_source_datasets}()},
\code{\link{extract_vars}()},
\code{\link{format.sdg_select}()},
\code{\link{format.smq_select}()},
\code{\link{get_constant_vars}()},
\code{\link{get_source_vars}()},
\code{\link{is_auto}()},
\code{\link{quo_c}()},
\code{\link{replace_values_by_names}()},
\code{\link{squote}()},
\code{\link{valid_time_units}()},
\code{\link{validate_query}()},
\code{\link{validate_sdg_select}()},
\code{\link{validate_smq_select}()},
\code{\link{what_is_it}()}
}
\concept{move_adm_dev}
\keyword{move_adm_dev}
