% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dt_dtm_utils.R
\name{get_imputation_targets}
\alias{get_imputation_targets}
\title{Get Imputation Targets}
\usage{
get_imputation_targets(partial, date_imputation, time_imputation, is_datetime)
}
\arguments{
\item{partial}{A list of partial date/time components.

\describe{
\item{Default value}{none}
}}

\item{date_imputation}{A character specifying the date imputation method.

\describe{
\item{Default value}{none}
}}

\item{time_imputation}{A character specifying the time imputation method.

\describe{
\item{Default value}{none}
}}

\item{is_datetime}{A logical indicating whether it's a datetime imputation.

\describe{
\item{Default value}{none}
}}
}
\value{
A list of imputation targets for date and (if applicable) time components.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Determines the imputation targets for date and time components.
}
\keyword{internal}
