% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_merged.R
\name{get_not_mapped}
\alias{get_not_mapped}
\title{Get list of records not mapped from the lookup table.}
\usage{
get_not_mapped()
}
\value{
A \code{data.frame} or \code{NULL}
}
\description{
Get list of records not mapped from the lookup table.
}
\seealso{
Utilities used within Derivation functions:
\code{\link[=extract_unit]{extract_unit()}},
\code{\link[=get_admiral_keys]{get_admiral_keys()}},
\code{\link[=get_flagged_records]{get_flagged_records()}},
\code{\link[=get_vars_query]{get_vars_query()}},
\code{\link[=set_admiral_keys]{set_admiral_keys()}}
}
\concept{utils_help}
\keyword{utils_help}
