% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_date_vars.R
\name{get_partialdatetime}
\alias{get_partialdatetime}
\title{Parse DTC variable and Determine Components}
\usage{
get_partialdatetime(dtc)
}
\arguments{
\item{dtc}{The \code{'--DTC'} date to parse

A character date is expected in a format like \code{yyyy-mm-dd} or
\code{yyyy-mm-ddThh:mm:ss}. Trailing components can be omitted and \code{-} is a
valid value for any component.}
}
\value{
A list of character vectors. The elements of the list are named
"year", "month", "day", "hour", "minute", and "second". Missing components
are set to \code{NA_character_}.
}
\description{
Parse DTC variable and Determine Components
}
\details{
The function can be replaced by the parttime parser once it is
available.
}
\seealso{
\code{\link[=impute_dtc_dtm]{impute_dtc_dtm()}}, \code{\link[=impute_dtc_dt]{impute_dtc_dt()}}

Other utils_impute: 
\code{\link{dt_level}()},
\code{\link{dtm_level}()},
\code{\link{get_imputation_target_date}()},
\code{\link{get_imputation_target_time}()},
\code{\link{restrict_imputed_dtc_dtm}()},
\code{\link{restrict_imputed_dtc_dt}()}
}
\author{
Stefan Bundfuss
}
\concept{utils_impute}
\keyword{utils_impute}
