% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_source_vars}
\alias{get_source_vars}
\title{Get Source Variables from a List of Quosures}
\usage{
get_source_vars(quosures)
}
\arguments{
\item{quosures}{A list of quosures}
}
\value{
A list of quosures
}
\description{
Get Source Variables from a List of Quosures
}
\seealso{
Brings something to you!?!: 
\code{\link{get_constant_vars}()},
\code{\link{get_duplicates}()}
}
\author{
Stefan Bundfuss
}
\concept{get}
\keyword{get}
