% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{\%or\%}
\alias{\%or\%}
\title{Or}
\usage{
lhs \%or\% rhs
}
\arguments{
\item{lhs}{Any valid R expression}

\item{rhs}{Any valid R expression}
}
\value{
Either the result of evaluating \code{lhs}, \code{rhs} or an error
}
\description{
Or
}
\details{
The function evaluates the expression \code{lhs} and if this expression results
in an error, it catches that error and proceeds with evaluating the expression
\code{rhs} and returns that result.
}
\seealso{
Developer Utility Functions: 
\code{\link{\%notin\%}()},
\code{\link{arg_name}()},
\code{\link{contains_vars}()},
\code{\link{convert_dtm_to_dtc}()},
\code{\link{extract_vars}()},
\code{\link{filter_if}()},
\code{\link{valid_time_units}()},
\code{\link{vars2chr}()}
}
\concept{dev_utility}
\keyword{dev_utility}
