% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{hello_admiral}
\alias{hello_admiral}
\title{Derive Extension Example}
\usage{
hello_admiral(hw = TRUE)
}
\arguments{
\item{hw}{TRUE or FALSE}
}
\value{
Happy Message
}
\description{
Says hello admiral
}
\details{
In the roxygen documentation you will find tags for family and keywords.
This is to create organized sections for the Reference tab on the pkgdown website.
You can modify the \verb{_pkgdown.yml} as necessary to create appropriate sections as necessary.
Under \code{./man/roxygen/meta.R}, you will find where to store these family/keywords.
}
\examples{
hello_admiral(hw = FALSE)
}
\seealso{
Other internal: 
\code{\link{admiral-package}},
\code{\link{extract_duplicate_records}()},
\code{\link{format.basket_select}()},
\code{\link{signal_duplicate_records}()}
}
\concept{internal}
\keyword{internal}
