% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fnc.R
\name{hello_admiral}
\alias{hello_admiral}
\title{Greetings}
\usage{
hello_admiral(hw = TRUE)
}
\arguments{
\item{hw}{logical}
}
\value{
Happy message
}
\description{
Greetings
}
\examples{
hello_admiral(hw = FALSE)
}
\seealso{
Other internal: 
\code{\link{admiral-package}},
\code{\link{extract_duplicate_records}()},
\code{\link{format.basket_select}()},
\code{\link{signal_duplicate_records}()}
}
\concept{internal}
\keyword{internal}
