% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_function.R
\name{hello_admiral}
\alias{hello_admiral}
\title{hello_admiral}
\usage{
hello_admiral(hw = TRUE)
}
\arguments{
\item{hw}{logical (TRUE or FALSE)}
}
\value{
string
}
\description{
hello_admiral
}
\details{
Use roxygen2
Under ./man/roxygen/meta.R
}
\examples{
hello_admiral(hw = TRUE)
}
\concept{der_adx}
\keyword{der_adxx}
