% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{hello_admiral}
\alias{hello_admiral}
\title{My first function, happy greeting}
\usage{
hello_admiral(hw = TRUE)
}
\arguments{
\item{hw}{TRUE or FALSE

\describe{
\item{Default value}{\code{TRUE}}
}}
}
\value{
Happy Message
}
\description{
My first function, happy greeting
}
\details{
In the roxygen documentation you will find tags for family and keywords.
This is to create organized sections for the Reference tab on the pkgdown website.
You can modify the \verb{_pkgdown.yml} as necessary to create appropriate sections as necessary.
Under \code{./man/roxygen/meta.R}, you will find where to store these family/keywords.
}
\examples{
hello_admiral(hw = FALSE)
}
\concept{example}
\keyword{example}
