% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{hello_admiral}
\alias{hello_admiral}
\title{My first function, happy greeting}
\usage{
hello_admiral(smile = TRUE)
}
\arguments{
\item{hw}{TRUE or FALSE

\describe{
\item{Default value}{\code{NULL}}
}}
}
\value{
Happy Message
}
\description{
My first function, happy greeting
}
\details{
This is a message of happiness.
}
\examples{
hello_admiral(smile = FALSE)
}
\concept{example}
\keyword{example}
