% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{hello_admiral}
\alias{hello_admiral}
\title{A dummy first commit}
\usage{
hello_admiral(hw = FALSE)
}
\arguments{
\item{hw}{TRUE or FALSE}
}
\value{
Message
}
\description{
A dummy first commit for issue #1839
}
\details{
Jordan Tuia's first commit/test to pharmaverse
}
\examples{
library(tibble)
hello_admiral()
}
\concept{der_addx}
\keyword{hello_admiral}
