% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{hello_admiral}
\alias{hello_admiral}
\title{Display a Welcome Message to the Admiral Family}
\usage{
hello_admiral(hw = TRUE)
}
\arguments{
\item{hw}{\code{TRUE} to include \code{hw} in the welcome message, \code{FALSE} otherwise.
Defaults to \code{TRUE}.

\describe{
\item{Default value}{\code{TRUE}}
}}
}
\value{
Happy message
}
\description{
This function displays a welcome message to users of the admiral package.
It is for dummy testing.
}
\examples{
# Display welcome message with hw
hello_admiral()

# Display welcome message without hw
hello_admiral(hw = FALSE)

}
\concept{utils_misc}
\keyword{utils_misc}
