% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{hello_admiral}
\alias{hello_admiral}
\title{Welcome message function}
\usage{
hello_admiral(exclamation = TRUE)
}
\arguments{
\item{exclamation}{\code{TRUE} or \code{FALSE}}
}
\value{
A string that welcomes the user to the admiral family, with or without an exclamation point.
}
\description{
This function displays a welcome message to users and is used for on-boarding purposes.
}
\details{
\code{exclamation} argument refers to the presence/absence of an exclamation point in the output.
}
\examples{
# Display welcome message without an exclamation point
hello_admiral(exclamation = FALSE)

# Display welcome message with an exclamation point
hello_admiral()

}
\seealso{
Utilities used for examples and template scripts: 
\code{\link{list_all_templates}()},
\code{\link{use_ad_template}()}
}
\concept{utils_examples}
\keyword{utils_examples}
