% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{hello_admiral}
\alias{hello_admiral}
\title{Derive Extension Example}
\usage{
hello_admiral(hw = TRUE)
}
\arguments{
\item{hw}{TRUE or FALSE}
}
\value{
Happy Message
}
\description{
Says hello admiral
}
\details{
Something Something
}
\examples{
hello_admiral(hw = FALSE)
}
\author{
Open Source
}
\keyword{ADaM}
\keyword{derivation}
\keyword{hello}
\keyword{timing}
