% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dt_dtm_utils.R
\name{impute_values}
\alias{impute_values}
\title{Impute Missing Values}
\usage{
impute_values(partial, target, components)
}
\arguments{
\item{partial}{A list of partial date/time components.

\describe{
\item{Default value}{none}
}}

\item{target}{A list of target values for imputation.

\describe{
\item{Default value}{none}
}}

\item{components}{A character vector of component names.

\describe{
\item{Default value}{none}
}}
}
\value{
A list of imputed date/time components.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This is a helping function for \code{get_dt_dtm_range()}
Imputes missing values in partial date/time components using target values.
}
\keyword{internal}
