% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiral_df.R
\name{is_adsl_structure}
\alias{is_adsl_structure}
\title{Check Whether a Dataset Has an ADSL (Subject-Level) Structure}
\usage{
is_adsl_structure(dataset, cols)
}
\arguments{
\item{dataset}{A data frame

\describe{
\item{Default value}{none}
}}

\item{cols}{The column names of \code{dataset}

\describe{
\item{Default value}{none}
}}
}
\value{
\code{TRUE} if the dataset has one record per subject (with respect to
\code{get_admiral_option("subject_keys")}) or contains typical subject-level
treatment variables (\code{TRT01P}/\code{TRTSDT}), \code{FALSE} otherwise.
}
\description{
Check Whether a Dataset Has an ADSL (Subject-Level) Structure
}
\seealso{
Other internal:
\code{\link{admiral-package}},
\code{\link[=as_admiral_df]{as_admiral_df()}},
\code{\link[=extract_duplicate_records]{extract_duplicate_records()}},
\code{\link[=format.basket_select]{format.basket_select()}},
\code{\link[=get_admiral_df_type]{get_admiral_df_type()}},
\code{\link[=signal_duplicate_records]{signal_duplicate_records()}}
}
\concept{internal}
\keyword{internal}
