% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{is_auto}
\alias{is_auto}
\title{Checks if the argument equals the auto keyword}
\usage{
is_auto(arg)
}
\arguments{
\item{arg}{argument to check}
}
\value{
\code{TRUE} if the argument equals the auto keyword, i.e., it is a quosure
of a symbol named auto.
}
\description{
Checks if the argument equals the auto keyword
}
\examples{

example_fun <- function(arg) {
  arg <- rlang::enquo(arg)
  if (admiral:::is_auto(arg)) {
    "auto keyword was specified"
  } else {
    arg
  }
}

example_fun("Hello World!")

example_fun(auto)
}
\author{
Stefan Bundfuss
}
\keyword{check}
