% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_auto}
\alias{is_auto}
\title{Checks if the argument equals the auto keyword}
\usage{
is_auto(arg)
}
\arguments{
\item{arg}{argument to check}
}
\value{
\code{TRUE} if the argument equals the auto keyword, i.e., it is an
expression of a symbol named auto.
}
\description{
Checks if the argument equals the auto keyword
}
\seealso{
Identifies type of Object with return of TRUE/FALSE: 
\code{\link{is_named}()},
\code{\link{is_order_vars}()},
\code{\link{is_valid_dtc}()}
}
\concept{is}
\keyword{is}
