% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_date}
\alias{is_date}
\title{Is Date/Date-time?}
\usage{
is_date(arg)
}
\arguments{
\item{arg}{The argument to check}
}
\value{
\code{TRUE} if the argument is a date or date-time, \code{FALSE} otherwise
}
\description{
Checks if a date or date-time vector was specified
}
\examples{
refdate <- lubridate::ymd("2020-01-02")
date <- lubridate::ymd("2020-02-03")
assertthat::assert_that(admiral:::is_date(refdate), admiral:::is_date(date))
}
\author{
Stefan Bundfuss
}
\keyword{check}
