% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_order_vars}
\alias{is_order_vars}
\title{Is order vars?}
\usage{
is_order_vars(arg)
}
\arguments{
\item{arg}{A<n R object}
}
\value{
A TRUE if input was created with \code{vars()} or calls involving \code{desc()}
otherwise throws an error
}
\description{
Check if inputs are created using \code{vars()} or calls involving \code{desc()}
}
\seealso{
Identifies type of Object with return of TRUE/FALSE: 
\code{\link{is_auto}()},
\code{\link{is_named}()},
\code{\link{is_valid_dtc}()}
}
\concept{is}
\keyword{is}
