% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_order_vars}
\alias{is_order_vars}
\title{Is order vars?}
\usage{
is_order_vars(arg)
}
\arguments{
\item{arg}{A<n R object}
}
\value{
A \code{logical} vector
}
\description{
Is order vars?
}
\seealso{
Identifies type of Object with return of TRUE/FALSE: 
\code{\link{is_auto}()},
\code{\link{is_date}()},
\code{\link{is_named}()},
\code{\link{is_timeunit}()},
\code{\link{is_valid_date_entry}()},
\code{\link{is_valid_day}()},
\code{\link{is_valid_dtc}()},
\code{\link{is_valid_hour}()},
\code{\link{is_valid_month}()},
\code{\link{is_valid_sec_min}()},
\code{\link{is_valid_time_entry}()}
}
\concept{is}
\keyword{is}
