% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_timeunit}
\alias{is_timeunit}
\title{Is Time Unit?}
\usage{
is_timeunit(arg)
}
\arguments{
\item{arg}{The argument to check}
}
\value{
\code{TRUE} if the argument is a time unit, \code{FALSE} otherwise
}
\description{
Checks if a string is a time unit, i.e., 'years', 'months', 'days', 'hours',
'minutes', or 'seconds'.
}
\seealso{
Identifies type of Object with return of TRUE/FALSE: 
\code{\link{is_auto}()},
\code{\link{is_date}()},
\code{\link{is_named}()},
\code{\link{is_order_vars}()},
\code{\link{is_valid_date_entry}()},
\code{\link{is_valid_day}()},
\code{\link{is_valid_dtc}()},
\code{\link{is_valid_hour}()},
\code{\link{is_valid_month}()},
\code{\link{is_valid_sec_min}()},
\code{\link{is_valid_time_entry}()}
}
\author{
Stefan Bundfuss
}
\concept{is}
\keyword{is}
