% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_valid_dtc}
\alias{is_valid_dtc}
\title{Is this string a valid DTC}
\usage{
is_valid_dtc(arg)
}
\arguments{
\item{arg}{A \code{character} vector}
}
\value{
\code{TRUE} if the argument is a valid \code{--DTC} string, \code{FALSE} otherwise
}
\description{
Is this string a valid DTC
}
\seealso{
Identifies type of Object with return of TRUE/FALSE: 
\code{\link{is_auto}()},
\code{\link{is_order_vars}()}
}
\concept{is}
\keyword{is}
