% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_valid_month}
\alias{is_valid_month}
\title{Check Validity of the Month Portion in the Date Input}
\usage{
is_valid_month(arg)
}
\arguments{
\item{arg}{The argument to check}
}
\value{
\code{TRUE} if the argument is a month input, \code{FALSE} otherwise
}
\description{
Days are expected to range from 1 to 12
}
\seealso{
Identifies type of Object with return of TRUE/FALSE: 
\code{\link{is_auto}()},
\code{\link{is_date}()},
\code{\link{is_named}()},
\code{\link{is_order_vars}()},
\code{\link{is_timeunit}()},
\code{\link{is_valid_date_entry}()},
\code{\link{is_valid_day}()},
\code{\link{is_valid_dtc}()},
\code{\link{is_valid_hour}()},
\code{\link{is_valid_sec_min}()},
\code{\link{is_valid_time_entry}()}
}
\author{
Samia Kabi
}
\concept{is}
\keyword{is}
