% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_valid_time_entry}
\alias{is_valid_time_entry}
\title{Check Validity of the Time Imputation Input}
\usage{
is_valid_time_entry(arg)
}
\arguments{
\item{arg}{The argument to check}
}
\value{
\code{TRUE} if the argument is a valid time_imputation input, \code{FALSE} otherwise
}
\description{
Time_imputation format should be specified as "hh:mm:ss" (e.g. "00:00:00")
or as a keyword: "FIRST", "LAST"
}
\seealso{
Identifies type of Object with return of TRUE/FALSE: 
\code{\link{is_auto}()},
\code{\link{is_date}()},
\code{\link{is_named}()},
\code{\link{is_order_vars}()},
\code{\link{is_timeunit}()},
\code{\link{is_valid_date_entry}()},
\code{\link{is_valid_day}()},
\code{\link{is_valid_dtc}()},
\code{\link{is_valid_hour}()},
\code{\link{is_valid_month}()},
\code{\link{is_valid_sec_min}()}
}
\author{
Samia Kabi
}
\concept{is}
\keyword{is}
