% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{anti_join}
\alias{anti_join}
\alias{inner_join}
\alias{left_join}
\title{Join Functions}
\usage{
anti_join(x, y, by = NULL, copy = FALSE, ...)

inner_join(x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...)

left_join(x, y, by = NULL, copy = FALSE, suffix = c(".x", ".y"), ...)
}
\arguments{
\item{x}{\code{data.frame}}

\item{y}{\code{data.frame}}

\item{by}{\code{character} vector}

\item{copy}{\code{logical}}

\item{...}{Additional arguments}

\item{suffix}{\code{character} vector}
}
\value{
\code{data.frame}
}
\description{
The \verb{*_join()} functions from {dplyr} without a warning on different attributes
in datasets.
}
\keyword{joins}
