% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_helpers.R
\name{list_all_templates}
\alias{list_all_templates}
\title{List All Available ADaM Templates}
\usage{
list_all_templates()
}
\description{
List All Available ADaM Templates
}
\examples{
list_all_templates()
}
\author{
Shimeng Huang, Thomas Neitmann
}
\keyword{user_utility}
