% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_helpers.R
\name{list_all_templates}
\alias{list_all_templates}
\title{List all templates provided by {admiral}}
\usage{
list_all_templates()
}
\description{
List all templates provided by {admiral}
}
\examples{
list_all_templates()
}
\author{
Shimeng Huang
}
\keyword{user_utility}
