% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_helpers.R
\name{list_all_templates}
\alias{list_all_templates}
\title{List All Available ADaM Templates}
\usage{
list_all_templates(package = "admiral")
}
\arguments{
\item{package}{The R package in which to look for templates. By default \code{"admiral"}.}
}
\value{
A \code{character} vector of all available templates
}
\description{
List All Available ADaM Templates
}
\examples{
list_all_templates()
}
\seealso{
Helper Utilities: 
\code{\link{call_user_fun}()},
\code{\link{convert_blanks_to_na}()},
\code{\link{extract_unit}()},
\code{\link{negate_vars}()},
\code{\link{signal_duplicate_records}()},
\code{\link{use_ad_template}()},
\code{\link{vars2chr}()}
}
\author{
Shimeng Huang, Thomas Neitmann
}
\concept{utils_help}
\keyword{utils_help}
