% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_tte.R
\name{list_tte_source_objects}
\alias{list_tte_source_objects}
\title{List all \code{tte_source} Objects Available in a Package}
\usage{
list_tte_source_objects(package = "admiral")
}
\arguments{
\item{package}{The name of the package in which to search for \code{tte_source} objects}
}
\value{
A \code{data.frame} where each row corresponds to one \code{tte_source} object or \code{NULL}
if \code{package} does not contain any \code{tte_source} objects
}
\description{
List all \code{tte_source} Objects Available in a Package
}
\examples{
list_tte_source_objects()
}
