% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_param_tte.R
\name{list_tte_source_objects}
\alias{list_tte_source_objects}
\title{List all \code{tte_source} Objects Available in a Package}
\usage{
list_tte_source_objects(package = "admiral")
}
\arguments{
\item{package}{The name of the package in which to search for \code{tte_source} objects}
}
\value{
A \code{data.frame} where each row corresponds to one \code{tte_source} object or \code{NULL}
if \code{package} does not contain any \code{tte_source} objects
}
\description{
List all \code{tte_source} Objects Available in a Package
}
\examples{
list_tte_source_objects()
}
\seealso{
Source Specifications: 
\code{\link{assert_db_requirements}()},
\code{\link{assert_terms}()},
\code{\link{assert_valid_queries}()},
\code{\link{basket_select}()},
\code{\link{censor_source}()},
\code{\link{date_source}()},
\code{\link{death_event}},
\code{\link{dthcaus_source}()},
\code{\link{event_source}()},
\code{\link{extend_source_datasets}()},
\code{\link{filter_date_sources}()},
\code{\link{format.basket_select}()},
\code{\link{params}()},
\code{\link{query}()},
\code{\link{sdg_select}()},
\code{\link{smq_select}()},
\code{\link{tte_source}()},
\code{\link{validate_basket_select}()},
\code{\link{validate_query}()}
}
\author{
Thomas Neitmann
}
\concept{source_specifications}
\keyword{source_specifications}
