% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{my_first_fcn.R}
\alias{my_first_fcn.R}
\alias{hello_admiral}
\title{Hello Admiral Function}
\usage{
hello_admiral(hw = TRUE)
}
\arguments{
\item{hw}{TRUE or FALSE}
}
\value{
Happy Message
}
\description{
Hello Admiral Function
}
\details{
First function to generate welcome message
}
\examples{
hello_admiral(hw = FALSE)
}
\concept{der_adxx}
\keyword{der_adxx}
