% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{my_first_fcn}
\alias{my_first_fcn}
\title{Welcome Message Function}
\usage{
my_first_fcn(hw = TRUE)
}
\arguments{
\item{hw}{Logical. Determines whether to display a slightly more enthusiastic welcome.
Defaults to TRUE.}
}
\value{
None. This function's primary purpose is to print a message to the console.
}
\description{
Displays a friendly welcome message, specifically designed for the "admiral family".
}
\details{
This is my first trial to write a function
}
\examples{
my_first_fcn()
my_first_fcn(hw = FALSE)
}
