% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{my_first_fcn}
\alias{my_first_fcn}
\title{Dummy function for on boarding}
\usage{
my_first_fcn(input)
}
\arguments{
\item{input}{\itemize{
\item required input information in any format
}}
}
\value{
a welcome message
}
\description{
Dummy function for on boarding
}
\details{
create a function to output a text message Welcome to the admiral family!
}
\examples{
my_first_fcn("Y")
my_first_fcn(NULL)
}
\author{
Lina Patil
}
