% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{my_first_fcn}
\alias{my_first_fcn}
\title{Derive on boarding dummy function}
\usage{
my_first_fcn(first_time = TRUE)
}
\arguments{
\item{first_time}{TRUE or FALSE}
}
\value{
a Welcome message
}
\description{
Derive on boarding dummy function
}
\details{
outputs "Welcome to the admiral family!".
}
\examples{
my_first_fcn(first_time = FALSE)
}
\seealso{
Utilities for printing: 
\code{\link{print.adam_templates}()},
\code{\link{print.duplicates}()},
\code{\link{print.source}()},
\code{\link{print_named_list}()}
}
\concept{utils_print}
\keyword{utils_print}
