% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_first_Fcn_AZ.R
\name{my_first_fcn}
\alias{my_first_fcn}
\title{Prints a welcome message to the Admiral family.}
\usage{
my_first_fcn(hw = TRUE)
}
\arguments{
\item{hw}{true or false}
}
\value{
character vector containing the welcome message.
}
\description{
Prints a welcome message to the Admiral family.
}
\examples{
my_first_fcn(hw = TRUE)
my_first_fcn()
}
