% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{my_first_fcn}
\alias{my_first_fcn}
\title{Send welcoming message to the user}
\usage{
my_first_fcn()
}
\description{
Send welcoming message to the user
}
\details{
This is a function that prints a welcoming message to the terminal that
greets the user.
}
