% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn_ld.R
\name{my_first_fcn}
\alias{my_first_fcn}
\title{My First Function}
\usage{
my_first_fcn()
}
\value{
No return value, called for side effects.
}
\description{
Prints a welcome message to the admiral family.
}
\examples{
my_first_fcn()
}
