% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{my_first_fcn}
\alias{my_first_fcn}
\title{Print Welcome to the admiral family!}
\usage{
my_first_fcn()
}
\description{
Print Welcome to the admiral family! quit
}
\examples{
my_first_fcn()

}
\concept{der_adxx}
\keyword{der_adxx}
