% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{my_first_fcn}
\alias{my_first_fcn}
\title{Onboard dummy function}
\usage{
my_first_fcn()
}
\value{
A character string with a welcome message
}
\description{
This function displays a welcome message and return a string literal to the caller
}
\examples{
my_first_fcn()

}
\concept{utils_misc}
\keyword{utils_misc}
