% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{my_first_fcn}
\alias{my_first_fcn}
\title{Welcome message}
\usage{
my_first_fcn(hw = TRUE)
}
\arguments{
\item{hw}{TRUE or FALSE}
}
\value{
A message to the user
}
\description{
Welcome message
}
\examples{
my_first_fcn()
}
