% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{my_first_fcn}
\alias{my_first_fcn}
\title{My First Function}
\usage{
my_first_fcn()
}
\description{
Prints a welcome message.
}
\details{
Prints "Welcome to the admiral family!" to the console.
}
\concept{admiral_onboarding}
\keyword{admiral_onboarding}
