% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{my_first_fcn}
\alias{my_first_fcn}
\title{My First Function}
\usage{
my_first_fcn()
}
\value{
A string with a welcome message "Welcome to admiral!"
}
\description{
This function outputs a welcome message for onboarding purposes.
}
