% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{my_first_fcn}
\alias{my_first_fcn}
\title{Print a Greeting Message}
\usage{
my_first_fcn()
}
\value{
No return value (\code{NULL}).
}
\description{
Prints a fixed greeting \code{message}.
}
\examples{
my_first_fcn()
}
\seealso{
Other Advanced Functions: 
\code{\link{list_tte_source_objects}()},
\code{\link{params}()}
}
\concept{other_advanced}
\keyword{other_advanced}
