% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_first_fcn.R
\name{my_first_fcn}
\alias{my_first_fcn}
\title{my_first_fcn}
\usage{
my_first_fcn()
}
\value{
This function returns the Character "Welcome to the admiral family!".
}
\description{
my first function
}
\details{
This is a function created as dummy.
}
\examples{
my_first_fcn() ## returns "Welcome to the admiral family!"
}
\seealso{
Other Advanced Functions: 
\code{\link{list_tte_source_objects}()},
\code{\link{params}()}
}
\concept{other_advanced}
\keyword{other_advanced}
