% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_derivation.R
\name{params}
\alias{params}
\title{Create a Set of Parameters}
\usage{
params(...)
}
\arguments{
\item{...}{One or more named arguments}
}
\value{
An object of class \code{params}
}
\description{
Create a set of variable parameters to be used in an iteration of \code{\link[=call_derivation]{call_derivation()}}
}
\examples{
params(
  fns = list(VSSTRESN ~ mean(., na.rm = TRUE)),
  set_values_to = vars(DTYPE = "AVERAGE")
)
}
\author{
Thomas Neitmann
}
