% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dt_dtm_utils.R
\name{parse_partial_date_time}
\alias{parse_partial_date_time}
\title{Parse Partial Date or Datetime}
\usage{
parse_partial_date_time(dtc, is_datetime)
}
\arguments{
\item{dtc}{A character vector of date or datetime strings to be parsed.

\describe{
\item{Default value}{none}
}}

\item{is_datetime}{A logical value indicating whether the input strings include
time information.

\describe{
\item{Default value}{none}
}}
}
\value{
A list of character vectors, each representing a component of the date
or datetime.
For dates, the components are "year", "month", and "day".
For datetimes, the components also include "hour", "minute", and "second".
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This is a helping function for \code{get_dt_dtm_range()}
This function parses a vector of date or datetime strings into their component
parts.
}
\details{
The function uses different parsing methods depending on whether the input is
a date or a datetime:
\itemize{
\item For dates, it calls \code{get_partialdate()}.
\item For datetimes, it calls \code{get_partialdatetime()}.
}
}
\keyword{internal}
