% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_helpers.R
\name{print.adam_templates}
\alias{print.adam_templates}
\title{Print \code{adam_templates} Objects}
\usage{
\method{print}{adam_templates}(x, ...)
}
\arguments{
\item{x}{A \code{adam_templates} object}

\item{...}{Not used}
}
\value{
No return value, called for side effects
}
\description{
Print \code{adam_templates} Objects
}
\examples{
templates <- list_all_templates()
print(templates)
}
\seealso{
\code{\link[=list_all_templates]{list_all_templates()}}

Utilities for printing: 
\code{\link{print.source}()},
\code{\link{print_named_list}()}
}
\author{
Thomas Neitmann
}
\concept{utils_print}
\keyword{utils_print}
