% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{print.duplicates}
\alias{print.duplicates}
\title{Print \code{duplicates} Objects}
\usage{
\method{print}{duplicates}(x, ...)
}
\arguments{
\item{x}{A \code{duplicates} object}

\item{...}{Not used}
}
\value{
No return value, called for side effects
}
\description{
Print \code{duplicates} Objects
}
\seealso{
Utilities for printing: 
\code{\link{print.adam_templates}()},
\code{\link{print.source}()},
\code{\link{print_named_list}()}
}
\concept{utils_print}
\keyword{utils_print}
