% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_utils.R
\name{print.source}
\alias{print.source}
\title{Print \code{source} Objects}
\usage{
\method{print}{source}(x, ...)
}
\arguments{
\item{x}{An \code{source} object}

\item{...}{Not used}
}
\value{
No return value, called for side effects
}
\description{
Print \code{source} Objects
}
\examples{
print(death_event)
}
\seealso{
\code{\link[=tte_source]{tte_source()}}, \code{\link[=censor_source]{censor_source()}}, \code{\link[=event_source]{event_source()}}

Other internal: 
\code{\link{admiral}},
\code{\link{print.adam_templates}()}
}
\author{
Stefan Bundfuss
}
\concept{internal}
\keyword{internal}
