% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_utils.R
\name{print.source}
\alias{print.source}
\title{Print \code{source} Objects}
\usage{
\method{print}{source}(x, ...)
}
\arguments{
\item{x}{An \code{source} object}

\item{...}{If \verb{indent = <numeric value>} is specified the output is indented
by the specified number of characters.}
}
\value{
No return value, called for side effects
}
\description{
Print \code{source} Objects
}
\examples{
print(death_event)
}
\seealso{
Utilities for printing: 
\code{\link{print.adam_templates}()},
\code{\link{print_named_list}()}
}
\author{
Stefan Bundfuss
}
\concept{utils_print}
\keyword{utils_print}
