% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiral_df.R
\name{print.summary_admiral_df}
\alias{print.summary_admiral_df}
\title{Print a Summary of an \code{admiral_df} Dataset}
\usage{
\method{print}{summary_admiral_df}(x, ...)
}
\arguments{
\item{x}{A \code{summary_admiral_df} object created by \code{\link[=summary.admiral_df]{summary.admiral_df()}}

\describe{
\item{Default value}{none}
}}

\item{...}{Not used

\describe{
\item{Default value}{none}
}}
}
\value{
No return value, called for side effects. The \code{summary_admiral_df}
object is returned invisibly.
}
\description{
Print a Summary of an \code{admiral_df} Dataset
}
\seealso{
Utilities for printing:
\code{\link[=print.adam_templates]{print.adam_templates()}},
\code{\link[=print.duplicates]{print.duplicates()}},
\code{\link[=print.source]{print.source()}},
\code{\link[=print_named_list]{print_named_list()}},
\code{\link[=summary.admiral_df]{summary.admiral_df()}}
}
\concept{utils_print}
\keyword{utils_print}
\section{Examples}{
\subsection{Printing a summary of an \code{admiral_df} dataset}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(tibble)

advs <- tribble(
  ~STUDYID,  ~USUBJID,      ~PARAMCD, ~AVAL, ~AVISIT,
  "PILOT01", "01-701-1015", "DIABP",     51, "BASELINE",
  "PILOT01", "01-701-1015", "SYSBP",    121, "BASELINE"
) \%>\%
  as_admiral_df()

print(summary(advs))}\if{html}{\out{</div>}}}}

