% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_utils.R
\name{print_named_list}
\alias{print_named_list}
\title{Print Named List}
\usage{
print_named_list(list, indent = 0)
}
\arguments{
\item{list}{A named list}

\item{...}{If \verb{indent = <numeric value>} is specified the output is indented
by the specified number of characters.}
}
\value{
No return value, called for side effects
}
\description{
Print Named List
}
\examples{
print_named_list(death_event)
}
\seealso{
Other internal: 
\code{\link{admiral}},
\code{\link{print.adam_templates}()},
\code{\link{print.source}()}
}
\author{
Stefan Bundfuss
}
\concept{internal}
\keyword{internal}
