% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dt_dtm_utils.R
\name{propagate_na_values}
\alias{propagate_na_values}
\title{Propagate NA Values}
\usage{
propagate_na_values(partial)
}
\arguments{
\item{partial}{A list of partial date/time components.}
}
\value{
A list of date/time components with propagated NA values.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This is a helping function for \code{get_dt_dtm_range()}.
Propagates NA values through date/time components.
}
\details{
This function ensures that if a higher-order component (e.g., month) is NA,
all lower-order components (e.g., day, hour, etc.) are also set to NA.
}
\keyword{internal}
