% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{queries}
\alias{queries}
\title{Queries Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15 rows and 8 columns.
}
\usage{
queries
}
\description{
Queries Dataset
}
\keyword{ADAE}
\keyword{ADCM}
\keyword{An}
\keyword{and}
\keyword{be}
\keyword{dataset}
\keyword{datasets}
\keyword{deriving}
\keyword{example}
\keyword{in}
\keyword{of}
\keyword{query}
\keyword{standard}
\keyword{to}
\keyword{used}
\keyword{variables}
