% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_query_data.R
\name{query}
\alias{query}
\title{Create an \code{query} object}
\usage{
query(
  prefix,
  name = auto,
  id = NULL,
  scope = NULL,
  scope_num = NULL,
  definition = NULL
)
}
\arguments{
\item{prefix}{Prefix of the new variables, e.g., \code{"SMQ03"}}

\item{name}{The variable \verb{<prefix>NAM} is set to the specified value for
observations matching the query definition. If the \code{auto} keyword is
specified, the variable is set to the name of the query in the SMQ/SDG
database.

\emph{Permitted Values}: A character scalar or the \code{auto} keyword. The \code{auto}
keyword is permitted only for queries which are defined by a \code{smq_select()}
or \code{sdg_select()} object.}

\item{id}{The variable \verb{<prefix>CD} is set to the specified value for
observations matching the query definition if the parameter is specified.
Otherwise the variable is not created. If the \code{auto} keyword is specified,
the variable is set to the id of the query in the SMQ/SDG database.

\emph{Permitted Values}: A integer scalar or the \code{auto} keyword. The \code{auto}
keyword is permitted only for queries which are defined by a \code{smq_select()}
or \code{sdg_select()} object.}

\item{scope}{The variable \verb{<prefix>SC} is set to the specified value for
observations matching the query definition if the parameter is specified.
Otherwise the variable is not created.

\emph{Permitted Values}: \code{"BROAD"}, \code{"NARROW"}, \code{NULL}}

\item{scope_num}{The variable \verb{<prefix>SCN} is set to the specified value for
observations matching the query definition if the parameter is specified.
Otherwise the variable is not created.

\emph{Permitted Values}: \code{1}, \code{2}, \code{NULL}}

\item{definition}{Definition of terms belonging to the query

There are three different ways to define the terms:
\itemize{
\item A \code{smq_select()} object is specified to select a query from the SMQ
database. The database contains Standard MedDRA Queries (SMQs) and AE
Grouping Terms (AEGTs).
\item A \code{sdg_select()} object is specified to select a query from the SDG
database.
\item A data frame with columns \code{TERM_LEVEL} and \code{TERM_NAME} or \code{TERM_ID} can
be specified to define the terms of the query. The \code{TERM_LEVEL} should be
set to the name of the variable which should be used to select the terms,
e.g., \code{"AEDECOD"} or \code{"AELLTCD"}. \code{TERM_LEVEL} does not need to be
constant within a query. For example a query can be based on \code{AEDECOD}
and \code{AELLT}.

If \code{TERM_LEVEL} refers to a character variable, \code{TERM_NAME} should be set
to the value the variable. If it refers to a numeric variable, \code{TERM_ID}
should be set to the value of the variable. If only character variables
or only numeric variables are used, \code{TERM_ID} or \code{TERM_NAME} respectively
can be omitted.
}}
}
\value{
An object of class "query".
}
\description{
An \code{query} object defines a query, e.g., a Standard MedDRA Query (SMQ), a
Standardised Drug Grouping (SDG), or a customized query (CQ).
}
\details{
The definition includes
\itemize{
\item the variables which should be created for the query, e.g., \code{SMQ02NAM} and
\code{SMQ02SC},
\item the values which should be assigned to the variables, and
\item which terms belong to the query.
}
}
\author{
Stefan Bundfuss
}
\keyword{source_specifications}
