% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quo.R
\name{quo_c}
\alias{quo_c}
\title{Concatenate One or More Quosure(s)}
\usage{
quo_c(...)
}
\arguments{
\item{...}{One or more objects of class \code{quosure} or \code{quosures}}
}
\value{
An object of class \code{quosures}
}
\description{
Concatenate One or More Quosure(s)
}
\seealso{
Helpers for working with Quosures: 
\code{\link{add_suffix_to_vars}()},
\code{\link{quo_not_missing}()},
\code{\link{replace_symbol_in_quo}()},
\code{\link{replace_values_by_names}()}
}
\author{
Thomas Neitmann
}
\concept{quo}
\keyword{quo}
