% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quo.R
\name{quo_c}
\alias{quo_c}
\title{Concatenate One or More Quosure(s)}
\usage{
quo_c(...)
}
\arguments{
\item{...}{One or more objects of class \code{quosure} or \code{quosures}}
}
\value{
An object of class \code{quosures}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
This function is \emph{deprecated}, please use \code{expr_c()} instead.
}
\seealso{
Helpers for working with Quosures: 
\code{\link{add_suffix_to_vars}()},
\code{\link{expr_c}()},
\code{\link{quo_not_missing}()},
\code{\link{replace_symbol_in_expr}()},
\code{\link{replace_symbol_in_quo}()},
\code{\link{replace_values_by_names}()}
}
\concept{quo}
\keyword{quo}
