% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{quo_c}
\alias{quo_c}
\title{Concatenate One or More Quosure(s)}
\usage{
quo_c(...)
}
\arguments{
\item{...}{One or more objects of class \code{quosure} or \code{quosures}}
}
\value{
An object of class \code{quosures}
}
\description{
Concatenate One or More Quosure(s)
}
\examples{
quo_c(rlang::quo(USUBJID))
quo_c(rlang::quo(STUDYID), rlang::quo(USUBJID))
quo_c(vars(USUBJID, ADTM))
quo_c(rlang::quo(BASETYPE), vars(USUBJID, PARAM), rlang::quo(ADTM))
}
\author{
Thomas Neitmann
}
\keyword{dev_utility}
