% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{quo_not_missing}
\alias{quo_not_missing}
\title{Check whether an argument is a quosure with a symbol or not.}
\usage{
quo_not_missing(x)
}
\arguments{
\item{x}{Test object}
}
\value{
TRUE or error.
}
\description{
Check whether an argument is a quosure with a symbol or not.
}
\examples{
test_fun <- function(x) {x <- rlang::enquo(x); assert_that(quo_not_missing(x))}
test_fun(my_variable) # works, returns TRUE
\dontrun{
test_fun() # throws error because argument x is missing and has no default
}
}
\author{
Thomas Neitmann, Ondrej Slama
}
