% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{quo_not_missing}
\alias{quo_not_missing}
\title{Check whether an argument is not a quosure of a missing argument}
\usage{
quo_not_missing(x)
}
\arguments{
\item{x}{Test object}
}
\value{
TRUE or error.
}
\description{
Check whether an argument is not a quosure of a missing argument
}
\examples{
test_fun <- function(x) {x <- rlang::enquo(x); assertthat::assert_that(quo_not_missing(x))}
test_fun(my_variable) # no missing argument -> returns TRUE
\dontrun{
test_fun() # missing argument -> throws error
}
}
\author{
Thomas Neitmann, Ondrej Slama
}
