% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quo.R
\name{quo_not_missing}
\alias{quo_not_missing}
\title{Check Whether an Argument Is Not a Quosure of a Missing Argument}
\usage{
quo_not_missing(x)
}
\arguments{
\item{x}{Test object}
}
\value{
TRUE or error.
}
\description{
Check Whether an Argument Is Not a Quosure of a Missing Argument
}
\seealso{
Helpers for working with Quosures: 
\code{\link{quo_c}()}
}
\author{
Thomas Neitmann, Ondrej Slama
}
\concept{quo}
\keyword{quo}
