% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{filter_if}
\alias{negate_vars}
\alias{vars2chr}
\alias{vars}
\alias{desc}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{admiraldev}{\code{\link[admiraldev]{filter_if}}, \code{\link[admiraldev]{negate_vars}}, \code{\link[admiraldev]{vars2chr}}}

  \item{dplyr}{\code{\link[dplyr]{desc}}, \code{\link[dplyr]{vars}}}
}}

