% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{expr}
\alias{exprs}
\alias{vars}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{vars}}}

  \item{rlang}{\code{\link[rlang]{expr}}, \code{\link[rlang]{exprs}}}
}}

