% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quo.R
\name{replace_symbol_in_quo}
\alias{replace_symbol_in_quo}
\title{Replace Symbols in a Quosure}
\usage{
replace_symbol_in_quo(quosure, target, replace)
}
\arguments{
\item{quosure}{Quosure}

\item{target}{Target symbol}

\item{replace}{Replacing symbol}
}
\value{
The quosure where every occurrence of the symbol \code{target} is replaced
by \code{replace}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
This function is \emph{deprecated}, please use \code{replace_symbol_in_expr()} instead.
}
\seealso{
Other deprecated: 
\code{\link{assert_order_vars}()},
\code{\link{quo_c}()},
\code{\link{quo_not_missing}()}
}
\concept{deprecated}
\keyword{deprecated}
