% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quo.R
\name{replace_values_by_names}
\alias{replace_values_by_names}
\title{Replace Expression Value with Name}
\usage{
replace_values_by_names(expressions, quosures)
}
\arguments{
\item{expressions}{A list of expressions}

\item{quosures}{\emph{Deprecated}, please use \code{expressions} instead.}
}
\value{
A list of expressions
}
\description{
Replace Expression Value with Name
}
\examples{
library(rlang)
replace_values_by_names(exprs(AVAL, ADT = convert_dtc_to_dt(EXSTDTC)))
}
\seealso{
Helpers for working with Quosures: 
\code{\link{add_suffix_to_vars}()},
\code{\link{expr_c}()},
\code{\link{replace_symbol_in_expr}()}
}
\concept{quo}
\keyword{quo}
