% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quo.R
\name{replace_values_by_names}
\alias{replace_values_by_names}
\title{Replace Quosure Value with Name}
\usage{
replace_values_by_names(quosures)
}
\arguments{
\item{quosures}{A list of quosures}
}
\value{
A list of quosures
}
\description{
Replace Quosure Value with Name
}
\seealso{
Helpers for working with Quosures: 
\code{\link{add_suffix_to_vars}()},
\code{\link{quo_c}()},
\code{\link{quo_not_missing}()},
\code{\link{replace_symbol_in_quo}()}
}
\author{
Thomas Neitmann
}
\concept{quo}
\keyword{quo}
