% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{replace_values_by_names}
\alias{replace_values_by_names}
\title{Replace Quosure Value with Name}
\usage{
replace_values_by_names(quosures)
}
\arguments{
\item{quosures}{A list of quosures}
}
\value{
A list of quosures
}
\description{
Replace Quosure Value with Name
}
\seealso{
Developer Utility Functions: 
\code{\link{\%notin\%}()},
\code{\link{\%or\%}()},
\code{\link{arg_name}()},
\code{\link{as_name}()},
\code{\link{convert_dtm_to_dtc}()},
\code{\link{extract_vars}()},
\code{\link{filter_if}()},
\code{\link{negate_vars}()},
\code{\link{valid_time_units}()},
\code{\link{vars2chr}()}
}
\author{
Thomas Neitmann
}
\concept{dev_utility}
\keyword{dev_utility}
