% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrict_derivation.R
\name{restrict_derivation}
\alias{restrict_derivation}
\title{Execute a Derivation on a Subset of the Input Dataset}
\usage{
restrict_derivation(dataset, derivation, args, filter)
}
\arguments{
\item{dataset}{Input dataset}

\item{derivation}{Derivation}

\item{args}{Arguments of the derivation

A \code{param()} object is expected.}

\item{filter}{Filter condition}
}
\description{
Execute a derivation on a subset of the input dataset.
}
\seealso{
\code{\link[=param]{param()}}
}
\author{
Stefan Bundfuss
}
