% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dt.R
\name{restrict_imputed_dtc_dt}
\alias{restrict_imputed_dtc_dt}
\title{Restrict Imputed \code{--DTC} date to Minimum/Maximum Dates}
\usage{
restrict_imputed_dtc_dt(
  dtc,
  imputed_dtc,
  min_dates,
  min_dates_strict,
  max_dates,
  max_dates_strict
)
}
\arguments{
\item{dtc}{The \code{--DTC} date to impute

A character date is expected in a format like \code{yyyy-mm-dd} or
\code{yyyy-mm-ddThh:mm:ss}. Trailing components can be omitted and \code{-} is a
valid "missing" value for any component.

\describe{
\item{Permitted values}{a character date variable}
\item{Default value}{none}
}}

\item{imputed_dtc}{The imputed \code{--DTC} date

\describe{
\item{Default value}{none}
}}

\item{min_dates}{Minimum dates

A list of dates is expected. It is ensured that the imputed date is not
before any of the specified dates, e.g., that the imputed adverse event start
date is not before the first treatment date. Only dates which are in the
range of possible dates of the \code{dtc} value are considered. The possible dates
are defined by the missing parts of the \code{dtc} date (see example below). This
ensures that the non-missing parts of the \code{dtc} date are not changed.
A date or date-time object is expected.
For example

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(lubridate)
impute_dtc_dt(
  "2020-11",
  min_dates = list(
   ymd("2020-12-06"),
   ymd("2020-11-11")
  ),
  highest_imputation = "M"
)
#> [1] "2020-11-11"
}\if{html}{\out{</div>}}

returns \code{"2020-11-11"} because the possible dates for \code{"2020-11"} range from
\code{"2020-11-01"} to \code{"2020-11-30"}. Therefore \code{"2020-12-06"} is ignored.
Returning \code{"2020-12-06"} would have changed the month although it is not
missing (in the \code{dtc} date).

\describe{
\item{Permitted values}{a list of dates, e.g. \code{list(ymd_hms("2021-07-01T04:03:01"), ymd_hms("2022-05-12T13:57:23"))}}
\item{Default value}{\code{NULL}}
}}

\item{min_dates_strict}{Minimum dates (strict)

The argument works like the \code{min_dates} argument but it affects the behavior
of the \code{min_dates} and \code{max_dates} arguments. The range which is used to
determine which of the \code{min_dates} and \code{max_dates} are considered is
restricted by the dates specified for \code{min_dates_strict} (see example below).

This argument is useful if the \code{max_dates} argument is used and there are
strict restrictions on the imputed date like the event start date if event
end date is imputed.

For example

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(lubridate)
impute_dtc_dt(
  c("2020-11", "2020-11"),
  min_dates_strict = list(
   c(ymd("2020-11-24"), ymd("2020-11-11"))
  ),
  max_dates = list(
    c(ymd("2020-11-22"), ymd("2020-11-22"))
  ),
  highest_imputation = "M",
  date_imputation = "last"
)
#> [1] "2020-11-30" "2020-11-22"
}\if{html}{\out{</div>}}

returns \code{"2020-11-30"}. The possible dates for \code{"2020-11"} range from
\code{"2020-11-01"} to \code{"2020-11-30"}.

For the first element, the \code{min_dates_strict} argument restricts this range
to \code{"2020-11-24"} to \code{"2020-11-30"}. Therefore \code{"2020-11-22"} (from the
\code{max_dates} argument) is ignored as it is not within the restricted range.

For the second element, the \code{min_dates_strict} argument restricts this range
to \code{"2020-11-11"} to \code{"2020-11-30"}. In this case \code{"2020-11-22"} (from the
\code{max_dates} argument) is within the restricted range and is considered.

\describe{
\item{Permitted values}{a list of dates, e.g. \code{list(ymd_hms("2021-07-01T04:03:01"), ymd_hms("2022-05-12T13:57:23"))}}
\item{Default value}{\code{NULL}}
}}

\item{max_dates}{Maximum dates

A list of dates is expected. It is ensured that the imputed date is not after
any of the specified dates, e.g., that the imputed date is not after the data
cut off date. Only dates which are in the range of possible dates are
considered. A date or date-time object is expected.

\describe{
\item{Permitted values}{a list of dates, e.g. \code{list(ymd_hms("2021-07-01T04:03:01"), ymd_hms("2022-05-12T13:57:23"))}}
\item{Default value}{\code{NULL}}
}}

\item{max_dates_strict}{Maximum dates (strict)

The argument works like the \code{max_dates} argument but it affects the behavior
of the \code{min_dates} and \code{max_dates} arguments. The range which is used to
determine which of the \code{min_dates} and \code{max_dates} are considered is
restricted by the dates specified for \code{max_dates_strict} (see example below).

This argument is useful if the \code{min_dates} argument is used and there are
strict restrictions on the imputed date like date of death or the event end
date if event start date is imputed.

For example

\if{html}{\out{<div class="sourceCode r">}}\preformatted{library(lubridate)
impute_dtc_dt(
  c("2020-11", "2020-11"),
  max_dates_strict = list(
   c(ymd("2020-11-24"), ymd("2020-11-11"))
  ),
  min_dates = list(
    c(ymd("2020-11-22"), ymd("2020-11-22"))
  ),
  highest_imputation = "M",
  date_imputation = "first"
)
#> [1] "2020-11-22" "2020-11-01"
}\if{html}{\out{</div>}}

returns (\code{"2020-11-22"}, \code{"2020-11-01"}). The possible dates for \code{"2020-11"}
range from \code{"2020-11-01"} to \code{"2020-11-30"}.

For the first element, the \code{max_dates_strict} argument restricts this range
to \code{"2020-11-01"} to \code{"2020-11-24"}. The date \code{"2020-11-22"} (from the
\code{min_dates} argument) is within the restricted range and is therefore it is
considered.

For the second element, the \code{max_dates_strict} argument restricts this range
to \code{"2020-11-01"} to \code{"2020-11-11"}. In this case \code{"2020-11-22"} (from the
\code{max_dates} argument) is outside the restricted range and thus it is ignored.

\describe{
\item{Permitted values}{a list of dates, e.g. \code{list(ymd_hms("2021-07-01T04:03:01"), ymd_hms("2022-05-12T13:57:23"))}}
\item{Default value}{\code{NULL}}
}}
}
\value{
\itemize{
\item The last of the minimum dates (\code{min_dates} and \code{min_dates_strict}) which
are in the range of the partial \code{--DTC} date (\code{dtc})
\item The first of the maximum dates (\code{max_dates} and \code{max_dates_strict}) which
are in the range of the partial \code{--DTC} date (\code{dtc})
\item \code{imputed_dtc} if the partial \code{--DTC} date (\code{dtc}) is not in range of any of
the minimum or maximum dates.
}
}
\description{
Restrict Imputed \code{--DTC} date to Minimum/Maximum Dates
}
\seealso{
\code{\link[=impute_dtc_dtm]{impute_dtc_dtm()}}, \code{\link[=impute_dtc_dt]{impute_dtc_dt()}}

Utilities used for date imputation:
\code{\link[=dt_level]{dt_level()}},
\code{\link[=dtm_level]{dtm_level()}},
\code{\link[=get_imputation_target_date]{get_imputation_target_date()}},
\code{\link[=get_imputation_target_time]{get_imputation_target_time()}},
\code{\link[=get_partialdatetime]{get_partialdatetime()}},
\code{\link[=restrict_imputed_dtc_dtm]{restrict_imputed_dtc_dtm()}}
}
\concept{utils_impute}
\keyword{internal}
