% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_vars_dtm.R
\name{restrict_imputed_dtc_dtm}
\alias{restrict_imputed_dtc_dtm}
\title{Restrict Imputed \code{--DTC} date to Minimum/Maximum Dates}
\usage{
restrict_imputed_dtc_dtm(
  dtc,
  imputed_dtc,
  min_dates,
  min_dates_strict,
  max_dates,
  max_dates_strict
)
}
\arguments{
\item{dtc}{The \code{--DTC} date to impute

A character date is expected in a format like \code{yyyy-mm-dd} or
\code{yyyy-mm-ddThh:mm:ss}. Trailing components can be omitted and \code{-} is a
valid "missing" value for any component.

\describe{
\item{Permitted values}{a character date variable}
\item{Default value}{none}
}}

\item{imputed_dtc}{The imputed \code{--DTC} date

\describe{
\item{Default value}{none}
}}

\item{min_dates}{Minimum dates

A list of dates is expected. It is ensured that the imputed date is not
before any of the specified dates, e.g., that the imputed adverse event start
date is not before the first treatment date. Only dates which are in the
range of possible dates of the \code{dtc} value are considered. The possible dates
are defined by the missing parts of the \code{dtc} date (see example below). This
ensures that the non-missing parts of the \code{dtc} date are not changed.
A date or date-time object is expected.
For example

\if{html}{\out{<div class="sourceCode r">}}\preformatted{impute_dtc_dtm(
  "2020-11",
  min_dates = list(
   ymd_hm("2020-12-06T12:12"),
   ymd_hm("2020-11-11T11:11")
  ),
  highest_imputation = "M"
)
}\if{html}{\out{</div>}}

returns \code{"2020-11-11T11:11:11"} because the possible dates for \code{"2020-11"}
range from \code{"2020-11-01T00:00:00"} to \code{"2020-11-30T23:59:59"}. Therefore
\code{"2020-12-06T12:12:12"} is ignored. Returning \code{"2020-12-06T12:12:12"} would
have changed the month although it is not missing (in the \code{dtc} date).

For date variables (not datetime) in the list the time is imputed to
\code{"00:00:00"}. Specifying date variables makes sense only if the date is
imputed. If only time is imputed, date variables do not affect the result.

\describe{
\item{Permitted values}{a list of dates, e.g. \code{list(ymd_hms("2021-07-01T04:03:01"), ymd_hms("2022-05-12T13:57:23"))}}
\item{Default value}{\code{NULL}}
}}

\item{min_dates_strict}{Minimum dates (strict)

The argument works like the \code{min_dates} argument but it affects the behavior
of the \code{min_dates} and \code{max_dates} arguments. The range which is used to
determine which of the \code{min_dates} and \code{max_dates} are considered is
restricted by the dates specified for \code{min_dates_strict}.

For example

\if{html}{\out{<div class="sourceCode r">}}\preformatted{impute_dtc_dtm(
  "2020-11",
  min_dates_strict = list(
   ymd_hm("2020-11-24T00:00")
  ),
  max_dates = list(
    ymd_hm("2020-11-22T12:12")
  ),
  highest_imputation = "M",
  date_imputation = "last",
  time_imputation = "last"
)
}\if{html}{\out{</div>}}

returns \code{"2020-11-30T23:59:59"}. The possible dates for \code{"2020-11"} range
from \code{"2020-11-01T00:00:00"} to \code{"2020-11-30T23:59:59"}. The
\code{min_dates_strict} argument restricts this range to \code{"2020-11-24T00:00:00"}
to \code{"2020-11-30T23:59:59"}. Therefore \code{"2020-11-22T12:12:12"} (from the
\code{max_dates} argument) is ignored as it is not within the restricted range.

\describe{
\item{Permitted values}{a list of dates, e.g. \code{list(ymd_hms("2021-07-01T04:03:01"), ymd_hms("2022-05-12T13:57:23"))}}
\item{Default value}{\code{NULL}}
}}

\item{max_dates}{Maximum dates

A list of dates is expected. It is ensured that the imputed date is not after
any of the specified dates, e.g., that the imputed date is not after the data
cut off date. Only dates which are in the range of possible dates are
considered. A date or date-time object is expected.

For date variables (not datetime) in the list the time is imputed to
\code{"23:59:59"}. Specifying date variables makes sense only if the date is
imputed. If only time is imputed, date variables do not affect the result.

\describe{
\item{Permitted values}{a list of dates, e.g. \code{list(ymd_hms("2021-07-01T04:03:01"), ymd_hms("2022-05-12T13:57:23"))}}
\item{Default value}{\code{NULL}}
}}

\item{max_dates_strict}{Maximum dates (strict)

The argument works like the \code{max_dates} argument but it affects the behavior
of the \code{min_dates} and \code{max_dates} arguments. The range which is used to
determine which of the \code{min_dates} and \code{max_dates} are considered is
restricted by the dates specified for \code{max_dates_strict}.

\describe{
\item{Permitted values}{a list of dates, e.g. \code{list(ymd_hms("2021-07-01T04:03:01"), ymd_hms("2022-05-12T13:57:23"))}}
\item{Default value}{\code{NULL}}
}}
}
\value{
\itemize{
\item The last of the minimum dates (\code{min_dates}) which are in the range of the
partial \code{--DTC} date (\code{dtc})
\item The first of the maximum dates (\code{max_dates}) which are in the range of the
partial \code{--DTC} date (\code{dtc})
\item \code{imputed_dtc} if the partial \code{--DTC} date (\code{dtc}) is not in range of any of
the minimum or maximum dates.
}
}
\description{
Restrict Imputed \code{--DTC} date to Minimum/Maximum Dates
}
\details{
The function will throw an error if imputation rules cause an
invalid datetime (e.g. "2020-02-01T25:00:00") to be generated. In this case,
the user should adjust the imputation rules.
}
\seealso{
\code{\link[=impute_dtc_dtm]{impute_dtc_dtm()}}, \code{\link[=impute_dtc_dt]{impute_dtc_dt()}}

Utilities used for date imputation:
\code{\link[=dt_level]{dt_level()}},
\code{\link[=dtm_level]{dtm_level()}},
\code{\link[=get_imputation_target_date]{get_imputation_target_date()}},
\code{\link[=get_imputation_target_time]{get_imputation_target_time()}},
\code{\link[=get_partialdatetime]{get_partialdatetime()}},
\code{\link[=restrict_imputed_dtc_dt]{restrict_imputed_dtc_dt()}}
}
\concept{utils_impute}
\keyword{internal}
