% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query_data.R
\name{sdg_select}
\alias{sdg_select}
\title{Create an \code{sdg_select} object}
\usage{
sdg_select(name = NULL, id = NULL)
}
\arguments{
\item{name}{Name of the query used to select the definition of the query
from the company database.}

\item{id}{Identifier of the query used to select the definition of the query
from the company database.}
}
\value{
An object of class \code{sdg_select}.
}
\description{
Create an \code{sdg_select} object
}
\details{
Exactly one \code{name} or \code{id} must be specified.
}
\seealso{
\code{\link[=create_query_data]{create_query_data()}}, \code{\link[=query]{query()}}

Source Specifications: 
\code{\link{assert_db_requirements}()},
\code{\link{assert_terms}()},
\code{\link{assert_valid_queries}()},
\code{\link{censor_source}()},
\code{\link{date_source}()},
\code{\link{death_event}},
\code{\link{derive_var_dthcaus}()},
\code{\link{event_source}()},
\code{\link{extend_source_datasets}()},
\code{\link{filter_date_sources}()},
\code{\link{format.sdg_select}()},
\code{\link{format.smq_select}()},
\code{\link{list_tte_source_objects}()},
\code{\link{params}()},
\code{\link{query}()},
\code{\link{smq_select}()},
\code{\link{tte_source}()},
\code{\link{validate_query}()},
\code{\link{validate_sdg_select}()},
\code{\link{validate_smq_select}()}
}
\author{
Stefan Bundfuss
}
\concept{source_specifications}
\keyword{source_specifications}
