% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query_data.R
\name{sdg_select}
\alias{sdg_select}
\title{Create an \code{sdg_select} object}
\usage{
sdg_select(name = NULL, id = NULL)
}
\arguments{
\item{name}{Name of the query used to select the definition of the query
from the company database.}

\item{id}{Identifier of the query used to select the definition of the query
from the company database.}
}
\value{
An object of class \code{sdg_select}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
This function is \emph{deprecated}, please use \code{basket_select()} instead.

Exactly one \code{name} or \code{id} must be specified.
}
\seealso{
\code{\link[=create_query_data]{create_query_data()}}, \code{\link[=query]{query()}}

Source Specifications: 
\code{\link{assert_db_requirements}()},
\code{\link{assert_terms}()},
\code{\link{assert_valid_queries}()},
\code{\link{basket_select}()},
\code{\link{censor_source}()},
\code{\link{date_source}()},
\code{\link{death_event}},
\code{\link{dthcaus_source}()},
\code{\link{event_source}()},
\code{\link{extend_source_datasets}()},
\code{\link{filter_date_sources}()},
\code{\link{format.basket_select}()},
\code{\link{list_tte_source_objects}()},
\code{\link{params}()},
\code{\link{query}()},
\code{\link{smq_select}()},
\code{\link{tte_source}()},
\code{\link{validate_basket_select}()},
\code{\link{validate_query}()}
}
\author{
Stefan Bundfuss
}
\concept{source_specifications}
\keyword{source_specifications}
