% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_query_data.R
\name{smq_select}
\alias{smq_select}
\title{Create an \code{smq_select} object}
\usage{
smq_select(name = NULL, id = NULL)
}
\arguments{
\item{name}{Name of the query used to select the definition of the query
from the central database.

The query with \code{BASKET_NAME == "<specified value>"} is selected.}

\item{id}{Identifier of the query used to select the definition of the query
from the central database.

The query with \verb{REFERENCE_NO == <specified value>} is selected.}
}
\value{
An object of class "smq_select".
}
\description{
Create an \code{smq_select} object
}
\details{
Exactly one \code{name} or \code{id} must be specified.
}
\author{
Stefan Bundfuss
}
\keyword{source_specifications}
