% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{squote}
\alias{squote}
\title{Wrap a String in Single Quotes}
\usage{
squote(x)
}
\arguments{
\item{x}{A \code{character} vector}
}
\description{
Wrap a String in Single Quotes
}
\examples{
admiral:::squote("foo")
}
\author{
Thomas Neitmann
}
\keyword{dev_utility}
