% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{suppress_warning}
\alias{suppress_warning}
\title{Suppress Specific Warnings}
\usage{
suppress_warning(expr, regexpr)
}
\arguments{
\item{expr}{Expression to be executed}

\item{regexpr}{Regular expression matching warnings to suppress}
}
\value{
Return value of the expression
}
\description{
Suppress certain warnings issued by an expression.
}
\details{
All warnings which are issued by the expression and match the regular expression
are suppressed.
}
\seealso{
Function that provide users with custom warnings
\code{\link{warn_if_incomplete_dtc}()},
\code{\link{warn_if_inconsistent_list}()},
\code{\link{warn_if_invalid_dtc}()},
\code{\link{warn_if_vars_exist}()}
}
\concept{warnings}
\keyword{warnings}
