% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tte_sources.R
\docType{data}
\name{death_event}
\alias{death_event}
\alias{lastalive_censor}
\alias{ae_event}
\alias{ae_ser_event}
\alias{ae_gr1_event}
\alias{ae_gr2_event}
\alias{ae_gr3_event}
\alias{ae_gr4_event}
\alias{ae_gr5_event}
\alias{ae_gr35_event}
\alias{ae_sev_event}
\alias{ae_wd_event}
\title{Pre-Defined Time-to-Event Source Objects}
\format{An object of class \code{event_source} (inherits from \code{tte_source}, \code{source}, \code{list}) of length 5.}
\usage{
death_event

lastalive_censor

ae_event

ae_ser_event

ae_gr1_event

ae_gr2_event

ae_gr3_event

ae_gr4_event

ae_gr5_event

ae_gr35_event

ae_sev_event

ae_wd_event
}
\description{
These pre-defined \code{tte_source} objects can be used as input to \code{derive_param_tte()}.
}
\details{
To see the definition of the various objects simply print the object in the
R console, e.g. \code{print(death_event)}.
}
\seealso{
\code{\link[=derive_param_tte]{derive_param_tte()}}, \code{\link[=tte_source]{tte_source()}}, \code{\link[=event_source]{event_source()}}, \code{\link[=censor_source]{censor_source()}}
}
\keyword{tte_source}
