% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_helpers.R
\name{use_ad_template}
\alias{use_ad_template}
\title{A helper function to create a template script.}
\usage{
use_ad_template(save_path, dom_name = "adsl", open = interactive())
}
\arguments{
\item{save_path}{Path to save the script.}

\item{dom_name}{An ADaM domain name.}

\item{open}{Whether to open the script right away.}
}
\description{
A helper function to create a template script.
}
\examples{
use_ad_template("./ad_adsl.R", "adsl")
}
\author{
Shimeng Huang
}
