% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_helpers.R
\name{use_ad_template}
\alias{use_ad_template}
\title{Open an ADaM Template Script}
\usage{
use_ad_template(
  adam_name = "adsl",
  save_path = paste0("./", adam_name, ".R"),
  overwrite = FALSE,
  open = interactive()
)
}
\arguments{
\item{adam_name}{An ADaM dataset name.}

\item{save_path}{Path to save the script.}

\item{overwrite}{Whether to overwrite an existing file named \code{save_path}.}

\item{open}{Whether to open the script right away.}
}
\value{
No return values, called for side effects
}
\description{
Open an ADaM Template Script
}
\details{
Use \code{list_all_templates()} to discover which templates are available
}
\examples{
if (interactive()) {
  use_ad_template("adsl")
}
}
\author{
Shimeng Huang, Thomas Neitmann
}
\keyword{user_utility}
