% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{valid_time_units}
\alias{valid_time_units}
\title{Valid Time Units}
\usage{
valid_time_units()
}
\value{
A \code{character} vector of valid time units
}
\description{
Contains the acceptable character vector of valid time units
}
\seealso{
Developer Utility Functions: 
\code{\link{\%notin\%}()},
\code{\link{\%or\%}()},
\code{\link{arg_name}()},
\code{\link{contains_vars}()},
\code{\link{convert_dtm_to_dtc}()},
\code{\link{extract_vars}()},
\code{\link{filter_if}()},
\code{\link{vars2chr}()}
}
\concept{dev_utility}
\keyword{dev_utility}
