% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query_data.R
\name{validate_basket_select}
\alias{validate_basket_select}
\title{Validate an object is indeed a \code{basket_select} object}
\usage{
validate_basket_select(obj)
}
\arguments{
\item{obj}{An object to be validated.}
}
\value{
The original object.
}
\description{
Validate an object is indeed a \code{basket_select} object
}
\seealso{
\code{\link[=basket_select]{basket_select()}}

Other Advanced Functions: 
\code{\link{assert_db_requirements}()},
\code{\link{assert_terms}()},
\code{\link{assert_valid_queries}()},
\code{\link{extend_source_datasets}()},
\code{\link{filter_date_sources}()},
\code{\link{format.basket_select}()},
\code{\link{list_tte_source_objects}()},
\code{\link{params}()},
\code{\link{validate_query}()}
}
\concept{other_advanced}
\keyword{other_advanced}
