% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query_data.R
\name{validate_query}
\alias{validate_query}
\title{Validate an object is indeed a \code{query} object}
\usage{
validate_query(obj)
}
\arguments{
\item{obj}{An object to be validated.}
}
\value{
The original object.
}
\description{
Validate an object is indeed a \code{query} object
}
\seealso{
\code{\link[=query]{query()}}

Functions moved to admiraldev package: 
\code{\link{assert_db_requirements}()},
\code{\link{assert_terms}()},
\code{\link{assert_valid_queries}()},
\code{\link{extend_source_datasets}()},
\code{\link{format.sdg_select}()},
\code{\link{format.smq_select}()},
\code{\link{validate_sdg_select}()},
\code{\link{validate_smq_select}()}
}
\author{
Stefan Bundfuss
}
\concept{move_adm_dev}
\keyword{move_adm_dev}
