% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query_data.R
\name{validate_query}
\alias{validate_query}
\title{Validate an object is indeed a \code{query} object}
\usage{
validate_query(obj)
}
\arguments{
\item{obj}{An object to be validated.}
}
\value{
The original object.
}
\description{
Validate an object is indeed a \code{query} object
}
\seealso{
\code{\link[=query]{query()}}
}
\author{
Stefan Bundfuss
}
\keyword{bye_fcn_amd_dev}
