% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query_data.R
\name{validate_smq_select}
\alias{validate_smq_select}
\title{Validate an object is indeed a \code{smq_select} object}
\usage{
validate_smq_select(obj)
}
\arguments{
\item{obj}{An object to be validated.}
}
\value{
The original object.
}
\description{
Validate an object is indeed a \code{smq_select} object
}
\seealso{
\code{\link[=smq_select]{smq_select()}}

Source Specifications: 
\code{\link{assert_db_requirements}()},
\code{\link{assert_terms}()},
\code{\link{assert_valid_queries}()},
\code{\link{censor_source}()},
\code{\link{date_source}()},
\code{\link{death_event}},
\code{\link{derive_var_dthcaus}()},
\code{\link{event_source}()},
\code{\link{extend_source_datasets}()},
\code{\link{filter_date_sources}()},
\code{\link{format.sdg_select}()},
\code{\link{format.smq_select}()},
\code{\link{list_tte_source_objects}()},
\code{\link{params}()},
\code{\link{query}()},
\code{\link{sdg_select}()},
\code{\link{smq_select}()},
\code{\link{tte_source}()},
\code{\link{validate_query}()},
\code{\link{validate_sdg_select}()}
}
\author{
Stefan Bundfuss
}
\concept{source_specifications}
\keyword{source_specifications}
