% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_utils.R
\name{vars2chr}
\alias{vars2chr}
\title{Turn a List of Quosures into a Character Vector}
\usage{
vars2chr(quosures)
}
\arguments{
\item{quosures}{A \code{list} of \code{quosures} created using \code{\link[=vars]{vars()}}}
}
\value{
A character vector
}
\description{
Turn a List of Quosures into a Character Vector
}
\examples{
vars2chr(vars(USUBJID, AVAL))
}
\seealso{
Helper Utilities: 
\code{\link{extract_unit}()},
\code{\link{list_all_templates}()},
\code{\link{negate_vars}()},
\code{\link{use_ad_template}()}
}
\author{
Thomas Neitmann
}
\concept{utils_help}
\keyword{utils_help}
