% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vars2chr}
\alias{vars2chr}
\title{Turn a List of Quosures into a Character Vector}
\usage{
vars2chr(quosures)
}
\arguments{
\item{quosures}{A \code{list} of \code{quosures} created using \code{\link[=vars]{vars()}}}
}
\description{
Turn a List of Quosures into a Character Vector
}
\examples{
vars2chr(vars(USUBJID, AVAL))
}
\author{
Thomas Neitmann
}
\keyword{user_utility}
