% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utilities.R
\name{vars2chr}
\alias{vars2chr}
\title{Turn a List of Quosures into a Character Vector}
\usage{
vars2chr(quosures)
}
\arguments{
\item{quosures}{A \code{list} of \code{quosures} created using \code{\link[=vars]{vars()}}}
}
\value{
A character vector
}
\description{
Turn a List of Quosures into a Character Vector
}
\examples{
library(dplyr, warn.conflicts = FALSE)

vars2chr(vars(USUBJID, AVAL))
}
\seealso{
Developer Utility Functions: 
\code{\link{\%notin\%}()},
\code{\link{\%or\%}()},
\code{\link{arg_name}()},
\code{\link{as_name}()},
\code{\link{contains_vars}()},
\code{\link{convert_dtm_to_dtc}()},
\code{\link{extract_vars}()},
\code{\link{filter_if}()},
\code{\link{valid_time_units}()}
}
\author{
Thomas Neitmann
}
\concept{dev_utility}
\keyword{dev_utility}
