% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{warn_if_param_missing}
\alias{warn_if_param_missing}
\title{Warn If Mapping a \code{PARAM} Failed}
\usage{
warn_if_param_missing(dataset, by_vars)
}
\arguments{
\item{dataset}{A \code{data.frame}}

\item{vars}{\code{character} vector of columns to check for in \code{dataset}}
}
\description{
Warn if a variable already exists inside a dataset
}
\author{
Thomas Neitmann
}
