% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev_utils.R
\name{what_is_it}
\alias{what_is_it}
\title{What Kind of Object is This?}
\usage{
what_is_it(x)
}
\arguments{
\item{x}{Any R object}
}
\description{
Returns a string describing what kind of object the input is.
}
\examples{
what_is_it(mtcars)
what_is_it(NA)
what_is_it(TRUE)
what_is_it(lm(hp ~ mpg, data = mtcars))
what_is_it(letters)
}
\seealso{
Functions moved to admiraldev package: 
\code{\link{arg_name}()},
\code{\link{assert_db_requirements}()},
\code{\link{assert_terms}()},
\code{\link{assert_valid_queries}()},
\code{\link{backquote}()},
\code{\link{contains_vars}()},
\code{\link{dquote}()},
\code{\link{enumerate}()},
\code{\link{extend_source_datasets}()},
\code{\link{extract_vars}()},
\code{\link{format.sdg_select}()},
\code{\link{format.smq_select}()},
\code{\link{get_constant_vars}()},
\code{\link{get_duplicates}()},
\code{\link{get_source_vars}()},
\code{\link{is_auto}()},
\code{\link{quo_c}()},
\code{\link{replace_values_by_names}()},
\code{\link{squote}()},
\code{\link{valid_time_units}()},
\code{\link{validate_query}()},
\code{\link{validate_sdg_select}()},
\code{\link{validate_smq_select}()}
}
\author{
Thomas Neitmann
}
\concept{move_adm_dev}
\keyword{move_adm_dev}
