% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{what_is_it}
\alias{what_is_it}
\title{What Kind of Object is This?}
\usage{
what_is_it(x)
}
\arguments{
\item{x}{Any R object}
}
\description{
Returns a string describing what kind of object the input is.
}
\examples{
admiral:::what_is_it(mtcars)
admiral:::what_is_it(NA)
admiral:::what_is_it(TRUE)
admiral:::what_is_it(lm(hp ~ mpg, data = mtcars))
admiral:::what_is_it(letters)
}
\author{
Thomas Neitmann
}
\keyword{dev_utility}
