% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what.R
\name{what_is_it}
\alias{what_is_it}
\title{What Kind of Object is This?}
\usage{
what_is_it(x)
}
\arguments{
\item{x}{Any R object}
}
\value{
A \code{character} description of the type of \code{x}
}
\description{
Returns a string describing what kind of object the input is.
}
\examples{
what_is_it("abc")
what_is_it(1L)
what_is_it(1:10)
what_is_it(mtcars)
}
\concept{what}
\keyword{what}
