% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_utils.R
\name{yn_to_numeric}
\alias{yn_to_numeric}
\title{Map \code{"Y"} and \code{"N"} to Numeric Values}
\usage{
yn_to_numeric(arg)
}
\arguments{
\item{arg}{Character vector}
}
\value{
\code{1} if \code{arg} equals \code{"Y"}, \code{0} if \code{arg} equals \code{"N"}, \code{NA_real_} otherwise
}
\description{
Map \code{"Y"} and \code{"N"} to numeric values.
}
\examples{

yn_to_numeric(c("Y", "N", NA_character_))
}
\seealso{
Utilities for Formatting Observations: 
\code{\link{convert_blanks_to_na}()},
\code{\link{convert_na_to_blanks}()},
\code{\link{format_eoxxstt_default}()},
\code{\link{format_reason_default}()}
}
\author{
Stefan Bundfuss
}
\concept{utils_fmt}
\keyword{utils_fmt}
